% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondCashFlows.R
\docType{methods}
\name{Accrued,BondCashFlows-method}
\alias{Accrued,BondCashFlows-method}
\title{Method to get Accrued from S4 class BondCashFlows}
\usage{
\S4method{Accrued}{BondCashFlows}(object)
}
\arguments{
\item{object}{the name of the S4 object}
}
\description{
Method to get Accrued from S4 class BondCashFlows
}
