% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageCashFlow.R
\docType{methods}
\name{Accrued,MortgageCashFlow-method}
\alias{Accrued,MortgageCashFlow-method}
\title{Method to extract Accrued from S4 class}
\usage{
\S4method{Accrued}{MortgageCashFlow}(object)
}
\arguments{
\item{object}{the nameof the S4 object}
}
\description{
Method to extract Accrued from S4 class
}

