% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondBasis.R
\name{ActFactor}
\alias{ActFactor}
\title{Act/Act day count}
\usage{
ActFactor(settlement.date, lastpmt.date, nextpmt.date, ..., start.date = NULL)
}
\arguments{
\item{settlement.date}{the settlement date}

\item{lastpmt.date}{the last payment bond coupon and or principal payment date}

\item{nextpmt.date}{the next bond coupon and or principal payment date}

\item{...}{option value used to compute time between payments}

\item{start.date}{the dated date of the bond or alternatively the settlement date.
the default value is null.}
}
\description{
Calculate Act/Act factor IMCA. used for UST corporates, etc
}
\seealso{
Other BondBasis: \code{\link{ActualFactor}},
  \code{\link{BondBasisConversion}},
  \code{\link{is.leapyear}}
}
