% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrincipalPmtDateCalcs.R
\name{AddMonths}
\alias{AddMonths}
\title{A function to add months to a date}
\usage{
AddMonths(Date = "character", Months = numeric())
}
\arguments{
\item{Date}{A character string the date to which the number of months are added}

\item{Months}{A numeric value the number of months to add to Date}
}
\description{
A function to add months to a date
}
\examples{
AddMonths("03-01-2015", 360)
}

