% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBSDetails.R
\name{AmortizationType}
\alias{AmortizationType}
\title{A standard generic function to access the slot AmortizationType}
\usage{
AmortizationType(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A standard generic function to access the slot AmortizationType
}

