% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelFunctions.R
\name{ArcTanIncentive}
\alias{ArcTanIncentive}
\title{A standard generic function to access the slot ArcTanIncentive from the 
class PrepaymentModelFunctions}
\usage{
ArcTanIncentive(object)
}
\arguments{
\item{object}{an S4 object}
}
\description{
A standard generic function to access the slot ArcTanIncentive from the 
class PrepaymentModelFunctions
}
