% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtomsData.R
\docType{class}
\name{AtomsAnalytics-class}
\alias{AtomsAnalytics-class}
\title{An S4 class representing ATOMs analytics}
\description{
The class is a superclass containing AtomsData class
and the AtomsScenario class.  The AtomsScenario class
is a superclass which contains the class Scenario defined
in the file MtgScenario
}

