% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtomsData.R
\docType{class}
\name{AtomsData-class}
\alias{AtomsData-class}
\title{An S4 classs representing the results of the ATOMs analytics}
\description{
An S4 classs representing the results of the ATOMs analytics
}
\section{Slots}{

\describe{
\item{\code{ISpread}}{a numeric value the interpolated spread over the U.S. Treasury coupon curve using consensus PSA}

\item{\code{NSpread}}{a numeric value the interpolated spread over the Swap coupon curve using consensus PSA}

\item{\code{ZSpread}}{a numeric value the spread over the U.S. Treasury spot rate curve}

\item{\code{ModDuration}}{a numeric value the modified duration based on the prepayment model vector}

\item{\code{Convexity}}{a numeric value the convexity based on the prepayment model vector}

\item{\code{EffDuration}}{a numeric value the sum of the key rate durations}

\item{\code{EffConvexity}}{a numeric value the}

\item{\code{KeyRateTenor}}{a length 11 numeric vector the tenors of the key rates}

\item{\code{KeyRateDuration}}{a length 11 numeric vector the key rate durations}

\item{\code{KeyRateConvexity}}{a length 11 numeric vector the key rate convexities}

\item{\code{OAD}}{a numeric value the option adjusted duration}

\item{\code{OAC}}{a numeric value the option adjusted convexity}

\item{\code{USTOAS}}{a numeric value the OAS to the U.S. Treasury curve}

\item{\code{USTZVSpread}}{a numeric value the zero volatltity OAS to the U.S.Treasury curve}

\item{\code{LIBOROAS}}{a numeric value the OAS to the swap curve}

\item{\code{LIBORZVSpread}}{a numeric value the zero volatility OAS to the swap curve}
}}

