% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AtomsData.R
\name{AtomsData}
\alias{AtomsData}
\title{A Function to calculate ATOMs Index Analytics}
\usage{
AtomsData(Bond.id = "character", original.bal = numeric(),
  price = numeric(), trade.date = "character",
  settlement.date = "character", swapspread = list(),
  method = "character")
}
\arguments{
\item{Bond.id}{A character vector the id or cusip of the bond}

\item{original.bal}{A numeric value the original balance}

\item{price}{A numeric value the price of the pass-through security}

\item{trade.date}{A numeric value the trade date}

\item{settlement.date}{A numeric value the settlement date}

\item{method}{A character vector the method used to fit the term structure}
}
\description{
AtomsAnalytics is the analytic function for the ATOMs Index
}

