% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtomsData.R
\docType{class}
\name{AtomsScenario-class}
\alias{AtomsScenario-class}
\title{An S4 class representing ATOMs Scenario results}
\description{
contains the class Scenario which is defined in the file
MtgScenario.R
}
\section{Slots}{

\describe{
\item{\code{HorizonReturn}}{a numeric value the horizon rate of return}

\item{\code{contains}}{Scenario class this class is set in MtgScenario.R}
}}

