% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PassThroughConstructor.R
\docType{methods}
\name{BalloonPmt,MBSDetails-method}
\alias{BalloonPmt,MBSDetails-method}
\title{Method to extract the slot BalloonPmt from the class MBSDetails}
\usage{
\S4method{BalloonPmt}{MBSDetails}(object)
}
\arguments{
\item{object}{the name of the object of the type MBSDetails}
}
\description{
Method to extract the slot BalloonPmt from the class MBSDetails
}

