% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StructuringFunctions.R
\name{BeginBal}
\alias{BeginBal}
\title{Function to assign the beginning principal balance to REMIC tranche}
\usage{
BeginBal(node, period = numeric())
}
\arguments{
\item{node}{}
}
\description{
This function takes the ending balance of the previous period and assigns
the value to the beginning balance in the current period
}

