% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelTune.R
\docType{methods}
\name{BeginCDR,PrepaymentModelTune-method}
\alias{BeginCDR,PrepaymentModelTune-method}
\title{A method to extract BeginCDR from class PrepaymentModelTune}
\usage{
\S4method{BeginCDR}{PrepaymentModelTune}(object)
}
\arguments{
\item{object}{The name of the object of class PrepaymentModelTune}
}
\description{
A method to extract BeginCDR from class PrepaymentModelTune
}

