% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageCashFlow.R
\docType{methods}
\name{BeginningBal,MortgageCashFlow-method}
\alias{BeginningBal,MortgageCashFlow-method}
\title{Method to extract the PrepaidPrin from class MortgageCashFlow}
\usage{
\S4method{BeginningBal}{MortgageCashFlow}(object)
}
\arguments{
\item{object}{the name of the object of type MortgageCashFlow}
}
\description{
Method to extract the PrepaidPrin from class MortgageCashFlow
}
