% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\docType{methods}
\name{BeginningBal,Mtg.Scenario-method}
\alias{BeginningBal,Mtg.Scenario-method}
\title{A method to extract BeginningBal from S4 class Mtg.Scenario}
\usage{
\S4method{BeginningBal}{Mtg.Scenario}(object)
}
\arguments{
\item{object}{The name of an S4 class of type Mtg.Scenario}
}
\description{
A method to extract BeginningBal from S4 class Mtg.Scenario
}

