% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurveSpreads.R
\docType{methods}
\name{BenchMark,CurveSpreads-method}
\alias{BenchMark,CurveSpreads-method}
\title{A method to extract the BenchMark from object CurveSpreads}
\usage{
\S4method{BenchMark}{CurveSpreads}(object)
}
\arguments{
\item{object}{An object of the type CurveSpreads}
}
\description{
A method to extract the BenchMark from object CurveSpreads
}

