% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurveSpreads.R
\name{BenchMark}
\alias{BenchMark}
\title{A standard generic function to access BenchMark}
\usage{
BenchMark(object)
}
\arguments{
\item{object}{An object of the type CurveSpreads}
}
\description{
A standard generic function to access BenchMark
}
