% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BillCashFlows.R
\docType{class}
\name{BillCashFlows-class}
\alias{BillCashFlows-class}
\title{An S4 class representing standared bill cash flows}
\description{
An S4 class representing standared bill cash flows
}
\section{Slots}{

\describe{
\item{\code{Price}}{a character the price of the bill}

\item{\code{Discount}}{a numeric value the discount rate}

\item{\code{YieldToMaturity}}{a numeric value the bill yield expressed as semi-annual
bond equivalent}

\item{\code{WAL}}{a numeric value the weighted average life of the bond}

\item{\code{ModDuration}}{a numeric value the bill modified duration}

\item{\code{Convexity}}{a numeric value the bill convexity}

\item{\code{Period}}{a numeric value the index of the payment to the investor}

\item{\code{PmtDate}}{a character string the payment date to the investor the format
is mm-dd-YYYY.}

\item{\code{TimePeriod}}{a numeric value the time period between payment dates made 
to the investor.}

\item{\code{PrincipalOutstanding}}{a numeric value the outstanding principal balance}

\item{\code{CouponPmt}}{a numeric value the outstanding principal balance}

\item{\code{TotalCashFlow}}{a numeric value the sum of the principal and interest 
payment made in each period.}
}}

