% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BillCashFlows.R
\name{BillCashFlows}
\alias{BillCashFlows}
\title{Bill Cash Flow engine for a standard discount bill}
\usage{
BillCashFlows(bill.id, principal, settlement.date, price)
}
\arguments{
\item{bill.id}{A character the referencing an object of type BillDetails}

\item{principal}{A numeric value the principal or face amount of the bill}

\item{settlement.date}{A character the settlment date mm-dd-YYYY}

\item{price}{A character the price of the bill}
}
\description{
Bill Cash Flow engine for a standard discount bill
}

