% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BillDetails.R
\docType{class}
\name{BillDetails-class}
\alias{BillDetails-class}
\title{An S4 class representing the cusip detail of a discount bill}
\description{
An S4 class representing the cusip detail of a discount bill
}
\section{Slots}{

\describe{
\item{\code{Cusip}}{a character string of length 9 the bond's cusip}

\item{\code{ID}}{a character string the ID of the bond}

\item{\code{BillType}}{a character string the type of the bond UST, Corp}

\item{\code{Sector}}{a character string the bond sector 
(Govt, Financial, Utility, etc.)}

\item{\code{Issuer}}{a character string the issuer of the bill, note, or bond.}

\item{\code{Underwriter}}{a character string the deal underwriter}

\item{\code{OfferAmount}}{a numeric value the OfferAmount Issued.}

\item{\code{IssueDiscountRate}}{a numeric value the bond's coupon}

\item{\code{IssueDate}}{A character the issue date mm-dd-YYYY}

\item{\code{DatedDate}}{A character the dated date mm-dd-YYYY}

\item{\code{Maturity}}{a character string the maturity date in mm/dd/YYYY format}

\item{\code{Moody}}{A character string the Moody credit rating}

\item{\code{SP}}{A character string the S&P credit rating}

\item{\code{BondLab}}{A character string the BondLab or investor 
assigned credit rating}

\item{\code{BondBasis}}{a character string the bond's interest calculation basis}
}}

