% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BillDetails.R
\name{BillDetails}
\alias{BillDetails}
\title{BillDetails}
\usage{
BillDetails(Cusip, ID, BillType, Sector, Issuer, Underwriter, OfferAmount,
  IssueDiscountRate, IssueDate, DatedDate, MaturityDate, Moody, SP, BondLab,
  BondBasis)
}
\arguments{
\item{Cusip}{A character the bill cusip}

\item{ID}{A character the bill id}

\item{BillType}{A character the type of bill}

\item{Sector}{A character the sector}

\item{Issuer}{A character the Issuer}

\item{Underwriter}{A character the Underwriter}

\item{OfferAmount}{A numeric value the offer amount}

\item{IssueDiscountRate}{A numeric value the discount rate}

\item{IssueDate}{A character mm-dd-YYYY}

\item{DatedDate}{A character mm-dd-YYYY}

\item{Moody}{A character the Moody Rating}

\item{SP}{A character the SP Rating}

\item{BondLab}{A character the BondLab (Investor) assigned rating}

\item{BondBasis}{A character the interest calculation basis}

\item{Maturity}{A character mm-dd-YYYY}
}
\description{
BillDetails creates BillDetails cusip object in the 
local environment
}
