% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscountToPrice.R
\name{BillPriceToYield}
\alias{BillPriceToYield}
\title{Price to Yield bill}
\usage{
BillPriceToYield(bill.id, price, day.count, settlement.date)
}
\arguments{
\item{bill.id}{A character string the cusip number or bill.id}

\item{price}{A character string the bill price}

\item{day.count}{A numeric value the day count used for pricing}

\item{settlement.date}{A character the settlement date mm-dd-YYYY}
}
\description{
Returns the bond equivalent yield of bill given price
}
\seealso{
Other Pricing: \code{\link{BenchMark,CurveSpreads-method}},
  \code{\link{BillYieldToPrice}},
  \code{\link{CurveSpreads-class}},
  \code{\link{CurveSpreads}},
  \code{\link{SpreadToBenchmark,CurveSpreads-method}},
  \code{\link{SpreadToCurve,CurveSpreads-method}},
  \code{\link{SpreadToPriceBond}},
  \code{\link{SpreadToPriceMBS}},
  \code{\link{ZVSpreadToPriceBond}},
  \code{\link{ZVSpreadToPriceMBS}},
  \code{\link{ZeroVolSpread,CurveSpreads-method}}
}
