% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BillDetails.R
\docType{methods}
\name{BillType,BillDetails-method}
\alias{BillType,BillDetails-method}
\title{Method to extract BillType from an S4 class of type BillDetails}
\usage{
\S4method{BillType}{BillDetails}(object)
}
\arguments{
\item{object}{The name of the S4 object of type BillDetails}
}
\description{
Method to extract BillType from an S4 class of type BillDetails
}

