% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BillDetails.R
\name{BillType}
\alias{BillType}
\title{A standard generic function to get the BillType}
\usage{
BillType(object)
}
\arguments{
\item{object}{An S4 class object}
}
\description{
A standard generic function to get the BillType
}

