% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ReadConnectionFunctions.R
\name{Bond}
\alias{Bond}
\title{Bond A connection function to the BondData folder}
\usage{
Bond(Bond.id = "character")
}
\arguments{
\item{Bond.id}{A character string the bond's cusip number or id}
}
\description{
Opens a connection to bond data folder
}

