% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondAnalytics.R
\name{BondAnalytics}
\alias{BondAnalytics}
\title{Bond Analytics function}
\usage{
BondAnalytics(bond.id, principal, price, trade.date, settlement.date,
  method = "dl")
}
\arguments{
\item{bond.id}{A character string the cusip number or id of the bond}

\item{principal}{A numeric value the principal amount of the bond}

\item{price}{A numeric value the price of the bond}

\item{trade.date}{A character string the trade date}

\item{settlement.date}{A character string the settlement date}

\item{method}{A character string the method used to fit the term 
structure passed to the package termstruc}
}
\description{
The constructor function for the BondAnalytics class.  Returns
the BondAnalytics object methods may be dispatched against slot generics.
}
\examples{
\dontrun{BondAnalytics(bond.id = "bondlab10", principal = 1000, price = 100, 
trade.date = "01-10-2013", settlement.date = "01-16-2013")}
}
\seealso{
Other Parametric: \code{\link{BondAnalytics-class}}
}

