% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BillDetails.R
\name{BondBasis,BillDetails-method}
\alias{BondBasis,BillDetails-method}
\title{Method to extract BondBasis of the S4 class of type BillDetails}
\usage{
\S4method{BondBasis}{BillDetails}(object)
}
\arguments{
\item{object}{The name of the S4 class of type BillDetails}
}
\description{
Method to extract BondBasis of the S4 class of type BillDetails
}
