% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondBasis.R
\name{BondBasisConversion}
\alias{BondBasisConversion}
\title{Bond Basis (day count) conventions}
\usage{
BondBasisConversion(issue.date, start.date, end.date, settlement.date,
  lastpmt.date, nextpmt.date, type)
}
\arguments{
\item{issue.date}{A character value the issue date of the security}

\item{start.date}{A character value the start date for interest payment
(dated date)}

\item{end.date}{A character value the final payment date}

\item{settlement.date}{A character value the settlement date}

\item{lastpmt.date}{A character value the last payment date to the investor}

\item{nextpmt.date}{A character value the next payment date to the investor}

\item{type}{a character vector the interest basis day count type}
}
\description{
applies the correct bond basis (day count) convention to interest
calculations
}
\seealso{
Other BondBasis: \code{\link{ActualFactor}},
  \code{\link{is.leapyear}}
}

