% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondCashFlow.R
\docType{class}
\name{BondCashFlows-class}
\alias{BondCashFlows-class}
\title{An S4 class representing standard bond cash flows}
\description{
An S4 class representing standard bond cash flows
}
\section{Slots}{

\describe{
\item{\code{Price}}{a numeric value the price of the bond}

\item{\code{Accrued}}{a numeric value the accrued interest}

\item{\code{YieldToMaturity}}{a numeric value the bond yield to maturity}

\item{\code{WAL}}{a numeric value the weighted average life of the bond}

\item{\code{ModDuration}}{a numeric value the bond duration}

\item{\code{Convexity}}{a numeric value the bond convexity}

\item{\code{Period}}{a numeric value an index of the payment to the investor}

\item{\code{PmtDate}}{a character string the payment date to the investor format 
is mm/dd/YYYY}

\item{\code{TimePeriod}}{a numeric value the time period between payments made
to the investor}

\item{\code{PrincipalOutstanding}}{a numeric value the outstanding principal balance}

\item{\code{CouponPmt}}{a numeric value the coupon payment amount}

\item{\code{TotalCashFlow}}{a numeric value the sum of the principal and interest payment
made in each period}
}}

