% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondConstructor.R
\docType{class}
\name{BondDetails-class}
\alias{BondDetails-class}
\title{An S4 class representing the cusip detail of the a standard bond}
\description{
An S4 class representing the cusip detail of the a standard bond
}
\section{Slots}{

\describe{
\item{\code{Cusip}}{a character string of length 9 the bond's cusip}

\item{\code{ID}}{a character string the ID of the bond}

\item{\code{BondType}}{a character string the type of the bond UST, Corp}

\item{\code{Sector}}{a character string the bond sector 
(Govt, Financial, Utility, etc.)}

\item{\code{Coupon}}{a numeric value the bond's coupon}

\item{\code{IssueDate}}{a character string the issue date in mm/dd/YYYY format}

\item{\code{DatedDate}}{a character string the dated date in mm/dd/YYYY format}

\item{\code{Maturity}}{a character string the maturity date in mm/dd/YYYY format}

\item{\code{LastPmtDate}}{a character string the payment date the date of the last
payment received by the investor in mm/dd/YYYY format}

\item{\code{NextPmtDate}}{a character string the payment date of the next scheduled
payment to be received by the investor}

\item{\code{Moody}}{A character string the Moody credit rating}

\item{\code{SP}}{A character string the S&P credit rating}

\item{\code{BondLab}}{A character string the BondLab or investor 
assigned credit rating}

\item{\code{Frequency}}{A numeric value the payment frequency of the bond 
(the number of payments made in one-year)}

\item{\code{BondBasis}}{a character string the bond's interest calculation basis}

\item{\code{Callable}}{a logical indicating the bond is callable}

\item{\code{Putable}}{a logical indicating the bond is putable}

\item{\code{SinkingFund}}{a logical indicating the bond has a sinking fund schedule}
}}

