% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondDetails.R
\name{BondDetails}
\alias{BondDetails}
\title{BondDetails}
\usage{
BondDetails(Cusip, ID, BondType, Sector, Issuer, Underwriter, OfferAmount,
  Coupon, IssueDate, DatedDate, Maturity, LastPmtDate, NextPmtDate, Moody, SP,
  BondLab, Frequency, BondBasis, Callable, Putable, SinkingFund)
}
\arguments{
\item{Cusip}{A character the cusip number}

\item{ID}{a character the bond ticker}

\item{BondType}{A character string the type of bond Bond, MBS, etc}

\item{Sector}{A character string the description of the sector}

\item{Issuer}{A character string the name of the issuer}

\item{Underwriter}{A character string the name of the underwriter}

\item{OfferAmount}{A numeric value the original offering amount}

\item{Coupon}{A numeric value the coupon}

\item{IssueDate}{A character string the issue date mm-dd-YYYY}

\item{DatedDate}{A character string the dated date mm-dd-YYYY}

\item{Maturity}{A character string the maturity date mm-dd-YYYY}

\item{LastPmtDate}{A character string the last pmt date mm-dd-YYYY}

\item{NextPmtDate}{A character string the next pmt date mm-dd-YYYY}

\item{Moody}{A character string the Moody rating}

\item{SP}{A character string the SP rating}

\item{BondLab}{A character string the BondLab rating}

\item{Frequency}{A numeric value the frequency of payments made to the
investor}

\item{BondBasis}{A character string the basis on which interest 
is calcualted}

\item{Callable}{A logical indicating if the bond is callable}

\item{Putable}{A logical indicating if the bond is putable}

\item{SinkingFund}{A logical indicating if the bond has a sinking fund
provision}
}
\description{
BondDetails creates Bond cusip object in the local environment
useful for connection to an extenal database and parallel computing
}
\seealso{
Other BondDetails: \code{\link{MakeBondDetails}}
}

