% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondFunctions.R
\name{BondEffectiveConvexity}
\alias{BondEffectiveConvexity}
\title{A function to compute effective convexity}
\usage{
BondEffectiveConvexity(
  Rate.Delta = numeric(),
  cashflow = vector(),
  discount.rates = vector(),
  time.period = vector()
)
}
\arguments{
\item{Rate.Delta}{A numeric value the interest rate shift in basis points}

\item{cashflow}{A numeric vector of cashflow}

\item{discount.rates}{A numeric vector of the up discount rates}

\item{time.period}{A numeric vector of the down discount rates}
}
\description{
Calculates effective convexity based on a discount vector (zero coupon)
cashflow vector, and rate delta
}
