\name{BondLab}
\alias{BondLab - package}
\docType{package}
\title{
A package for the analysis of Mortgage and Asset Backed Securities
}
\description{
Bond Lab offers functions for the analysis and valuation of Mortgage and Asset Backed Securities. The implementation of Bond Lab focuses the valuation, structuring, and risk management of structured product investments.
}


\details{
\tabular{ll}{
Package: \tab BondLab\cr
Type: \tab Package\cr
Version: \tab 0.0.0\cr
Date: \tab 2013-12-08\cr
License: \tab GPL >= 3.0 restrictions: non-commerical use
}
The package provides a suite of software for the analysis of strucutured products
}
\author{
Glenn M. Schultz, CFA

Maintainer: Glenn Schultz<glennmschultz@me.com>
}
\references{
Glenn M. Schultz, CFA (2015): Investing in Mortgage and Asset Backed Securities with R and Open Source Analytics
}

\keyword{ Fixed Income }

