% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BillDetails.R
\name{BondLabRating,BillDetails-method}
\alias{BondLabRating,BillDetails-method}
\title{Method to extract BondLab rating from S4 class of type BillDetails}
\usage{
\S4method{BondLabRating}{BillDetails}(object)
}
\arguments{
\item{object}{The name of the S4 class of type BillDetails}
}
\description{
Method to extract BondLab rating from S4 class of type BillDetails
}
