% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondDetails.R
\docType{methods}
\name{BondLabRating,BondDetails-method}
\alias{BondLabRating,BondDetails-method}
\title{Method to extract BondLab rating from S4 class of type BondDetails}
\usage{
\S4method{BondLabRating}{BondDetails}(object)
}
\arguments{
\item{object}{the name of the object of type BondDetails}
}
\description{
Method to extract BondLab rating from S4 class of type BondDetails
}

