% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBSDetails.R
\name{BondLabRating,MBSDetails-method}
\alias{BondLabRating,MBSDetails-method}
\title{Method to extract the slot BondLab from the class MBSDetails}
\usage{
\S4method{BondLabRating}{MBSDetails}(object)
}
\arguments{
\item{object}{the name of the object of the type MBSDetails}
}
\description{
Method to extract the slot BondLab from the class MBSDetails
}
