% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Functions.R
\name{bondprice}
\alias{bondprice}
\title{Determine the price a bond on the interest payment date}
\usage{
bondprice(yield.to.maturity = numeric(), coupon = numeric(),
  coupon.frequency = numeric(), years.mat = numeric(),
  face.value = numeric())
}
\arguments{
\item{yield.to.maturity}{A numeric value expressing the yield to
maturity (discount rate) as an annual percentage.}

\item{coupon}{A numeric value the coupon paid by the bond as a percentage of
the bond's principal amount}

\item{coupon.frequency}{A numeirc value expressing the frequency of payments
over one year}

\item{years.mat}{A numeric value expressing the years to maturity}

\item{face.value}{A numeric value expressing the face value (principal amount) of the bond}
}
\description{
This is generic function to determine the price of a bond
given the yield to maturity (YTM).  It is a nominal example of
pricing a bond given its yield to maturity.  The equation assumes pricing
from one payment date to the next.  It does not account for acrrued interest.
}
\examples{
bondprice(yield.to.maturity = .05, coupon = .05, coupon.frequency = 2,
years.mat = 10, face.value = 100)
}

