% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondScenario.R
\docType{class}
\name{BondReturn-class}
\alias{BondReturn-class}
\title{Bond Total Return Metrics}
\description{
The class BondReturn holds the results of bond total return analysis.
Bond total return is the sum of coupon income, principal received,
reinvestment income, and price appreciation or depreciation. In addition,
most investors also track horizon current balance.  The BondReturn class
differs from the MortgageReturn class in that the MortgageReturn class
includes scheduled principal and prepaid principal.
}
\section{Slots}{

\describe{
\item{\code{CouponIncome}}{A numeric value the coupon income received over the 
investment horizon}

\item{\code{PrincipalReceived}}{A numeric value the scheduled principal 
received over the investment horizon}

\item{\code{ReinvestmentIncome}}{A numeric value the reivestment income received 
over the investment horizon}

\item{\code{HorizonCurrBal}}{A numeric value the current balance at the end of the
investment horizon}

\item{\code{HorizonPrice}}{A numeric the price at the end of the horizon}

\item{\code{HorizonReturn}}{A numeric value the horizon total return}

\item{\code{HorizonMos}}{A numeric value the number of months to 
the scenario horizon date}
}}
\seealso{
Other Scenario.Analysis: \code{\link{BondScenario-class}},
  \code{\link{BondScenario}},
  \code{\link{CouponIncome,BondReturn-method}},
  \code{\link{HorizonCurrBal,BondReturn-method}},
  \code{\link{HorizonMos,BondReturn-method}},
  \code{\link{HorizonPrice,BondReturn-method}},
  \code{\link{HorizonReturn,BondReturn-method}},
  \code{\link{Name,Scenario-method}},
  \code{\link{PrincipalReceived,BondReturn-method}},
  \code{\link{PrincipalReceived}},
  \code{\link{ReinvestmentIncome,BondReturn-method}},
  \code{\link{Scenario-class}}
}

