% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondScenario.R
\docType{class}
\name{BondScenario-class}
\alias{BondScenario-class}
\title{BondScenario class}
\description{
The class \strong{BondScenario} contains the following super classes: 
TermStructure, BondCashFlow, BondTermStructure, BondReturn, CurveSpreads, 
and Scenario.
}
\seealso{
Other Bond Scenario Analysis: \code{\link{BondReturn-class}},
  \code{\link{BondScenario}},
  \code{\link{CouponIncome,BondReturn-method}},
  \code{\link{HorizonCurrBal,BondReturn-method}},
  \code{\link{HorizonMos,BondReturn-method}},
  \code{\link{HorizonPrice,BondReturn-method}},
  \code{\link{HorizonReturn,BondReturn-method}},
  \code{\link{PrincipalReceived,BondReturn-method}},
  \code{\link{ReinvestmentIncome,BondReturn-method}},
  \code{\link{ScenarioHorizonMos,ScenarioCurve-method}}
}
\concept{Bond Scenario Analysis}
