% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondScenario.R
\name{BondScenario}
\alias{BondScenario}
\title{Bond Scenario Analysis}
\usage{
BondScenario(bond.id = "character", settlement.date = "character",
  rates.data = "character", price = numeric(), par.amount = numeric(),
  scenario = "character", horizon.months = numeric(), method = "ns", ...,
  horizon.spot.spread = NULL, horizon.nominal.spread = NULL,
  horizon.OAS = NULL, horizon.price = NULL)
}
\arguments{
\item{bond.id}{A character string referencing an object of the type BondDetails}

\item{settlement.date}{A character string the settlement data "mm-dd-YYYY".}

\item{rates.data}{A character string an object yield curve}

\item{price}{A character string in decimal equivalent (.) or 32nds (-)}

\item{par.amount}{A numeric value the par amount.}

\item{scenario}{A character string the scenario}

\item{horizon.months}{A numeric value the time horizon}

\item{method}{A character string the method used to fit the term structure}

\item{...}{Optional values to select term structure and horizon price method}

\item{horizon.spot.spread}{A numeric value the horizon zero volatility 
spread}

\item{horizon.nominal.spread}{A numeric value the horizon nominal spread
or spread to  the curve}

\item{horizon.OAS}{A numeric value the horizon option adjusted spread
(not currently implemented)}

\item{horizon.price}{A numeric value the horizon price}
}
\description{
A function to compute the total return of a Bond
}

