% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondScenario.R
\name{BondScenarioSet}
\alias{BondScenarioSet}
\title{BondScenarioSet is a constructor function for the class BondScenarioSet}
\usage{
BondScenarioSet(ScenarioSet = "list")
}
\arguments{
\item{ScenarioSet}{a list of the scenario classes used for analysis}
}
\description{
BondScenarioSet is a constructor function for the class BondScenarioSet
}

