% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondKeyRate.R
\docType{class}
\name{BondTermStructure-class}
\alias{BondTermStructure-class}
\title{An S4 class representing the bond term structure exposure}
\description{
An S4 class representing the bond term structure exposure
}
\section{Slots}{

\describe{
\item{\code{SpotSpread}}{a numeric value the spread to the spot curve}

\item{\code{EffDuration}}{a numeric value the effective duration}

\item{\code{EffConvexity}}{a numeric value the effective convexity}

\item{\code{KeyRateTenor}}{a vector of values the key rate tenors}

\item{\code{KeyRateDuration}}{a vector of values the key rate durations}

\item{\code{KeyRateConvexity}}{a vector of values the key rate convexities}
}}

