% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondKeyRate.R
\name{BondTermStructure}
\alias{BondTermStructure}
\title{A function to calculate a bond key rate duration}
\usage{
BondTermStructure(bond.id = "character", Rate.Delta = numeric(),
  TermStructure = "character", principal = numeric(), price = "character",
  cashflow = "character")
}
\arguments{
\item{bond.id}{A character string referencing an object of 
type bond details.}

\item{Rate.Delta}{A numeric value the rate delta used to calcualte the 
bond KRDs.}

\item{TermStructure}{A character string referencing an object of typre
TermStructure.}

\item{principal}{A numeric value the principal or face amount of the bond.}

\item{price}{A character the price of the bond}

\item{cashflow}{A character string referencing an object of the 
BondCashFlow.}
}
\description{
This is a generic function used to contruct the object BondTermStructure
}

