% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBSDetails.R
\name{BondType,MBSDetails-method}
\alias{BondType,MBSDetails-method}
\title{Method to extract the slot BondType from the class MBSDetails}
\usage{
\S4method{BondType}{MBSDetails}(object)
}
\arguments{
\item{object}{the name of the object of type MBSDetails}
}
\description{
Method to extract the slot BondType from the class MBSDetails
}
