% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{BondValue,RAID-method}
\alias{BondValue,RAID-method}
\title{A method to access the slot BondValue from class RAID}
\usage{
\S4method{BondValue}{RAID}(object)
}
\arguments{
\item{object}{an S4 class of type RAID}
}
\description{
A method to access the slot BondValue from class RAID
}

