% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{BondValueCap,RAID-method}
\alias{BondValueCap,RAID-method}
\title{A method to access the slot BondValueCap from class RAID}
\usage{
\S4method{BondValueCap}{RAID}(object)
}
\arguments{
\item{object}{an S4 class of type RAID}
}
\description{
A method to access the slot BondValueCap from class RAID
}

