% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{BondValueDiscountRate,RAID-method}
\alias{BondValueDiscountRate,RAID-method}
\title{A method to access the slot BondValueDiscountRate from class RAID}
\usage{
\S4method{BondValueDiscountRate}{RAID}(object)
}
\arguments{
\item{object}{an S4 class of the type BondValueDiscountRate}
}
\description{
A method to access the slot BondValueDiscountRate from class RAID
}

