% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{BondValueMethod,RAID-method}
\alias{BondValueMethod,RAID-method}
\title{A method to access the slot BondValueMethod from class RAID}
\usage{
\S4method{BondValueMethod}{RAID}(object)
}
\arguments{
\item{object}{an S4 class of type RAID}
}
\description{
A method to access the slot BondValueMethod from class RAID
}

