% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\name{BondValueMethod}
\alias{BondValueMethod}
\title{A generic function to access the slot BondValueMethod}
\usage{
BondValueMethod(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A generic function to access the slot BondValueMethod
}

