% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{BondValueReinvestmentRate,RAID-method}
\alias{BondValueReinvestmentRate,RAID-method}
\title{A method to access the slot BondValueReinvestmentRate from class RAID}
\usage{
\S4method{BondValueReinvestmentRate}{RAID}(object)
}
\arguments{
\item{object}{an S4 class of the type RAID}
}
\description{
A method to access the slot BondValueReinvestmentRate from class RAID
}

