% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\name{BondValueReinvestmentRate}
\alias{BondValueReinvestmentRate}
\title{A generic function to access the slot BondValueReinvestmentRate}
\usage{
BondValueReinvestmentRate(object)
}
\arguments{
\item{object}{an S4 clss object}
}
\description{
A generic function to access the slot BondValueReinvestmentRate
}

