% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelFunctions.R
\name{BorrowerBurnout}
\alias{BorrowerBurnout}
\title{A standard generic function to access the slot BorrowerBurnout from the 
class PrepaymentModelFunctions}
\usage{
BorrowerBurnout(object)
}
\arguments{
\item{object}{an S4 object}
}
\description{
A standard generic function to access the slot BorrowerBurnout from the 
class PrepaymentModelFunctions
}
