% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentFunctions.R
\name{BorrowerIncentive}
\alias{BorrowerIncentive}
\title{BorrowerIncentive}
\usage{
BorrowerIncentive(midpoint, max.cpr, steepness, spread)
}
\arguments{
\item{midpoint}{The function mid point}

\item{max.cpr}{The function max CPR}

\item{steepness}{the function steepness}

\item{spread}{The spread between the note rate and the prevailing mortgage rate}
}
\description{
The function is the logit used to model incentive.  The 
function may be used to fit incentive curve based on either wac spread to 
the mortgage rate or annuity ratio.
}
\seealso{
Other Prepayment and Default: \code{\link{BurnoutFunction}},
  \code{\link{SDARamp}}, \code{\link{Seasonals}},
  \code{\link{Turnover}}
}

