% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HullWhite.R
\name{Bterm}
\alias{Bterm}
\title{Hull White B term}
\usage{
Bterm(mean.reversion, delta.time)
}
\arguments{
\item{mean.reversion}{The mean reversion rate}

\item{delta.time}{The time step (maturity of zero coupon bond)}
}
\description{
A function to compute B term of the Hull White Bond Price
}
\seealso{
Other Interest Rate Models: 
\code{\link{CreateCashFlowMatrix}()},
\code{\link{CreateMaturitiesMatrix}()},
\code{\link{HullWhite}()}
}
\concept{Interest Rate Models}
