% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PassThroughConstructor.R
\docType{methods}
\name{BurnOut,MBSDetails-method}
\alias{BurnOut,MBSDetails-method}
\title{Mehthod to extract the slot Burnout from the class MBSDetails}
\usage{
\S4method{BurnOut}{MBSDetails}(object)
}
\arguments{
\item{object}{the name of the object of the type MBSDetails}
}
\description{
Mehthod to extract the slot Burnout from the class MBSDetails
}

