% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PassThroughConstructor.R
\name{BurnOut}
\alias{BurnOut}
\title{A standard generic function to access the slot Burnout}
\usage{
BurnOut(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A standard generic function to access the slot Burnout
}

