% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelTune.R
\docType{methods}
\name{BurnoutBetaOne,PrepaymentModelTune-method}
\alias{BurnoutBetaOne,PrepaymentModelTune-method}
\title{A method to extract Burnout.beta.1 from class PrepaymentModelTune}
\usage{
\S4method{BurnoutBetaOne}{PrepaymentModelTune}(object)
}
\arguments{
\item{object}{The name of the object of class PrepaymentModelTune}
}
\description{
A method to extract Burnout.beta.1 from class PrepaymentModelTune
}
