% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelTune.R
\name{BurnoutBetaOne}
\alias{BurnoutBetaOne}
\title{A standard generic functin to access the slot Burnout.beta.1}
\usage{
BurnoutBetaOne(object)
}
\arguments{
\item{object}{an S4 object}
}
\description{
A standard generic functin to access the slot Burnout.beta.1
}
