% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelTune.R
\docType{methods}
\name{BurnoutBetaTwo,PrepaymentModelTune-method}
\alias{BurnoutBetaTwo,PrepaymentModelTune-method}
\title{A method to extract Burnout.beta.2 from class PrepaymentModelTune}
\usage{
\S4method{BurnoutBetaTwo}{PrepaymentModelTune}(object)
}
\arguments{
\item{object}{The name of the object of class PrepaymentModelTune}
}
\description{
A method to extract Burnout.beta.2 from class PrepaymentModelTune
}
