% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelTune.R
\name{BurnoutBetaTwo}
\alias{BurnoutBetaTwo}
\title{A standard generic function to access the slot Burnout.beta.e}
\usage{
BurnoutBetaTwo(object)
}
\arguments{
\item{object}{an S4 object}
}
\description{
A standard generic function to access the slot Burnout.beta.e
}
