% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentFunctions.R
\name{BurnoutFunction}
\alias{BurnoutFunction}
\title{BurnoutFunction}
\usage{
BurnoutFunction(fast.fast = 0.97, slow.fast = 0.01, fast.pop = 0.95,
  slow.pop = 0.05, age)
}
\arguments{
\item{fast.fast}{The transition probablity of a fast payer to a fast payer in
any given period.  The higher the fast.fast probablity the slower the rate of 
burnout.  Defaults to 0.97}

\item{slow.fast}{The transition probability of slow payer to a fast paying in
any given period.  The higher the slow.fast transition the faster the burnout.
Defaults to 0.01}

\item{fast.pop}{The initial estimate of the fast payers in the pool defaults to 0.95}

\item{slow.pop}{The initial estimate of the slow payers in the pool defaults to 0.05}

\item{age}{The loan or pool age}
}
\description{
The function is a Markov process of population evolution of fast
and slow payers in a mortgage pool.
}
\seealso{
Other Prepayment and Default: \code{\link{BorrowerIncentive}},
  \code{\link{SDARamp}}, \code{\link{Seasonals}},
  \code{\link{Turnover}}
}

