% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CalibrateCIR.R
\name{CalibrateCIR}
\alias{CalibrateCIR}
\title{Calibrate CIR}
\usage{
CalibrateCIR(trade.date = character, sigma = numeric())
}
\arguments{
\item{trade.date}{A numeric value the trade date}

\item{sigma}{the volatility in absolute percent the annualized value is sigma times sqrt(240)}
}
\description{
This function calibrates the CIR Model to the yield curve
}

