% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageOAS.R
\name{CalibrateRates}
\alias{CalibrateRates}
\title{CalibrateRates}
\usage{
CalibrateRates(rates.data, num.paths = 5000)
}
\arguments{
\item{rates.data}{A character referencing a rates.data object}

\item{num.paths}{the number of simulations to fit to the market 
defaults to 5,000}
}
\description{
optimization to fit 3-mo, 2-year, and 10-year rates to the
forward rate curve
}
\seealso{
Other mortgage pass through option adjusted spread: \code{\link{FitMarket}},
  \code{\link{MonteCarloMeanReversion}},
  \code{\link{MortgageOAS-class}},
  \code{\link{OASRatesArray}}, \code{\link{SimCurve}},
  \code{\link{SimRates}}
}

