% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondDetails.R
\docType{methods}
\name{Callable,BondDetails-method}
\alias{Callable,BondDetails-method}
\title{Method to extract the slot Callable from S4 class of type BondDetails}
\usage{
\S4method{Callable}{BondDetails}(object)
}
\arguments{
\item{object}{the name of the object of type BondDetails}
}
\description{
Method to extract the slot Callable from S4 class of type BondDetails
}
