% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CashFlowEngine.R
\name{CashFlowBill}
\alias{CashFlowBill}
\title{Bond Lab function to cash flow a discount bill.}
\usage{
CashFlowBill(bill.id, principal, settlement.date)
}
\arguments{
\item{bill.id}{a character or connection referring to an object of type BillDetails.}

\item{principal}{A numeric value the principal.}

\item{settlement.date}{A character the settlement date mm-dd-YYYY}
}
\description{
Calculates the bill cash flow table used in the BillCashFlow function.
}
