% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CashFlowEngine.R
\name{CashFlowBond}
\alias{CashFlowBond}
\title{Bond Lab function to cash flow a bond}
\usage{
CashFlowBond(bond.id, principal, settlement.date)
}
\arguments{
\item{bond.id}{a character or connection referring to an object of type BondDetails.}

\item{principal}{the investor principal amount.}

\item{settlement.date}{a character the settlement data mm-dd-YYYY.}
}
\description{
calculates the bond cashflow table used in function BondCashflows.
}

