% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CashFlowEngine.R
\name{CashFlowEngine}
\alias{CashFlowEngine}
\title{Bond Lab function the residential mortgage cash flow engine}
\usage{
CashFlowEngine(bond.id = "character", settlement.date = "character",
  principal = numeric(), PrepaymentAssumption = "character")
}
\arguments{
\item{bond.id}{A character string referring to an object of type MBSDetails}

\item{settlement.date}{A character string the settlement date}

\item{principal}{A numeric value the principal balance}

\item{PrepaymentAssumption}{A character string referencing 
an object type PrepaymentModel}
}
\description{
Bond Lab function the residential mortgage cash flow engine
}
