% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertToCashFlowTable.R
\name{CashFlowTable}
\alias{CashFlowTable}
\title{A function to convert an S4 MortgaageCashFlow class to an array}
\usage{
CashFlowTable(CashFlow = "character")
}
\arguments{
\item{CashFlow}{A character string referring to an object of type MortgageCashFlow}
}
\description{
The function is used as part of the structuring waterfall script
to convert the MortgageCashFlow class to array to facilitate the division
of principal and interest
}

