% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageFunctions.R
\name{CDR.To.MDR}
\alias{CDR.To.MDR}
\title{A function to convert the CDR to a single monthly mortality rate MDR}
\usage{
CDR.To.MDR(CDR = numeric())
}
\arguments{
\item{CDR}{A numeric value}
}
\description{
Standard Generic Function used to convert CPR to SMM
}
\examples{
CDR.To.MDR(CDR = .06)
}
