% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CIRBondPrice.R
\name{CIRBondPrice}
\alias{CIRBondPrice}
\title{Calculate the price of a zero coupon bond using the closed form equation given by Cox, Ingersoll, and Ross}
\usage{
CIRBondPrice(shortrate = vector(), T = numeric(), step = numeric(),
  kappa = numeric(), lambda = numeric(), sigma = numeric(),
  theta = numeric(), result = character)
}
\arguments{
\item{shortrate}{A vector of simulated short-term rates}

\item{T}{A numeric value the time over which the short-term rate is simulated}

\item{step}{A numeric value the time step on the simulation}

\item{kappa}{A numeric value the speed of mean reversion of the simulation}

\item{lambda}{A numeric value the risk premium}

\item{sigma}{A numeric value the volatility entered as a standard deviation}

\item{theta}{A numeric vlaue the long term forward rate}

\item{result}{A character string the desired output either "p" - Price, "y" - Yield, or "l" - Limit}
}
\description{
This function implements the CIR closed form solution to a bond price given the short-term rate
}

