% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CIRSimulation.R
\name{CIRSim}
\alias{CIRSim}
\title{CIR Sim}
\usage{
CIRSim(shortrate = numeric(), kappa = numeric(), theta = numeric(),
  T = numeric(), step = numeric(), sigma = numeric(), N = numeric())
}
\arguments{
\item{shortrate}{A numeric value the shortrate}

\item{kappa}{A numeric value the rate of mean reversion}

\item{theta}{A numeric value the long term forward rate}

\item{T}{A numeric value time in years}

\item{step}{A numeric value the time step}

\item{sigma}{A numeric value the volatility not annualized}

\item{N}{the number of paths}
}
\description{
The short rate simulation engine
}

