% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{class}
\name{Collateral-class}
\alias{Collateral-class}
\title{A S4 class of the REMIC collateral}
\description{
Collateral class holds collateral details for the collateral group
}
\section{Slots}{

\describe{
\item{\code{Group}}{a numeric value the collateral group number}

\item{\code{Cusip}}{a list of the pool cusip(s) in the collateral group}

\item{\code{OrigBal}}{a list of the Original Balance of each cusip in the 
collateral group}
}}

