% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{CollateralAmount,RAID-method}
\alias{CollateralAmount,RAID-method}
\title{A method to access the slot CollateralAmount from class RAID}
\usage{
\S4method{CollateralAmount}{RAID}(object)
}
\arguments{
\item{object}{an S4 class of type RAID}
}
\description{
A method to access the slot CollateralAmount from class RAID
}

