% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{CollateralAvgLife,RAID-method}
\alias{CollateralAvgLife,RAID-method}
\title{A method to access the slot CollateralAvgLife from class RAID}
\usage{
\S4method{CollateralAvgLife}{RAID}(object)
}
\arguments{
\item{object}{an S4 class of type RAID}
}
\description{
A method to access the slot CollateralAvgLife from class RAID
}

