% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICStructuringFunctions.R
\name{CollateralCashFlow}
\alias{CollateralCashFlow}
\title{A function to calcualte the cashflow of the REMIC collateral}
\usage{
CollateralCashFlow(bond.id = "character", settlement.date = "character",
  term.structure = "character", principal = numeric(), PSA = numeric())
}
\arguments{
\item{bond.id}{a character the bond cusip or id}

\item{settlement.date}{a character the settlement date}

\item{term.structure}{a character refering to object of type TermStructure}

\item{principal}{a character the current balance contributed to the REMIC}

\item{PSA}{the PSA assumtion used to price the deal}
}
\description{
Calculate the REMIC collateral cashflow
}

