% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{class}
\name{CollateralGroup-class}
\alias{CollateralGroup-class}
\title{An S4 class the groups underlying the REMIC}
\description{
CollateralGroup is an aggregator of the collatreal class
}
\section{Slots}{

\describe{
\item{\code{Group}}{a list of the collateral groups}
}}

