% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\name{CollateralGroup}
\alias{CollateralGroup}
\title{A function to aggregate the collateral group information}
\usage{
CollateralGroup(NumberofGroups = numeric(), DealName = "character")
}
\arguments{
\item{NumberofGroups}{A numeric value the number collateral groups}

\item{DealName}{A character string the Deal Name}
}
\description{
The function is used by the REMIC Constructor
}

