% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\name{CollateralGroups}
\alias{CollateralGroups}
\title{A standard generic function to access the CollateralGroup list}
\usage{
CollateralGroups(object)
}
\arguments{
\item{object}{an list of S4 objects}
}
\description{
A standard generic function to access the CollateralGroup list
}

