% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BillCashFlows.R
\docType{methods}
\name{Convexity,BillCashFlows-method}
\alias{Convexity,BillCashFlows-method}
\title{Method to get Convexity from S4 class BillCashFlows}
\usage{
\S4method{Convexity}{BillCashFlows}(object)
}
\arguments{
\item{object}{the name of the S4 object BillCashFlows}
}
\description{
Method to get Convexity from S4 class BillCashFlows
}

