% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondCashFlow.R
\docType{methods}
\name{Convexity,BondCashFlows-method}
\alias{Convexity,BondCashFlows-method}
\title{Method to extract Convexity from S4 class BondCashFlows}
\usage{
\S4method{Convexity}{BondCashFlows}(object)
}
\arguments{
\item{object}{the name of the S4 object}
}
\description{
Method to extract Convexity from S4 class BondCashFlows
}

