% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageCashFlow.R
\name{Convexity}
\alias{Convexity}
\title{A standard generic function to access the slot Convexity}
\usage{
Convexity(object)
}
\arguments{
\item{object}{an S4 class object of type MortgageCashFlow}
}
\description{
A standard generic function to access the slot Convexity
}

