% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondConstructor.R
\docType{methods}
\name{Coupon,BondDetails-method}
\alias{Coupon,BondDetails-method}
\title{Method to extract Coupon from an S4 class of type BondDetails}
\usage{
\S4method{Coupon}{BondDetails}(object)
}
\arguments{
\item{object}{the name of the S4 object of type BondDetails}
}
\description{
Method to extract Coupon from an S4 class of type BondDetails
}

