% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondTermStructure.R
\name{Coupon,BondTermStructure-method}
\alias{Coupon,BondTermStructure-method}
\title{Coupon}
\usage{
\S4method{Coupon}{BondTermStructure}(object)
}
\arguments{
\item{object}{object of type BondTermStructure}
}
\description{
Get Coupon from object BondTermStructure
}
