% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBSDetails.R
\docType{methods}
\name{Coupon,MBSDetails-method}
\alias{Coupon,MBSDetails-method}
\title{Method to extract the slot Coupon from the class MBSDetails}
\usage{
\S4method{Coupon}{MBSDetails}(object)
}
\arguments{
\item{object}{the name of the object of type MBSDetails}
}
\description{
Method to extract the slot Coupon from the class MBSDetails
}
