% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageOAS.R
\docType{methods}
\name{Coupon,MortgageOAS-method}
\alias{Coupon,MortgageOAS-method}
\title{A Method to access the slot Coupon}
\usage{
\S4method{Coupon}{MortgageOAS}(object)
}
\arguments{
\item{object}{MortgageOAS object}
}
\description{
Access the slot Coupon from object MortgageOAS
}

