% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageKeyRate.R
\name{Coupon,MortgageTermStructure-method}
\alias{Coupon,MortgageTermStructure-method}
\title{Coupon}
\usage{
\S4method{Coupon}{MortgageTermStructure}(object)
}
\arguments{
\item{object}{MortgageTermStructure}
}
\description{
Get Coupon from object of type MortgageTermStructure
}
