% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CouponTypes.R
\docType{methods}
\name{CouponBasis,CouponTypes-method}
\alias{CouponBasis,CouponTypes-method}
\title{A method to extract CouponBasis from the slot of CouponTypes}
\usage{
\S4method{CouponBasis}{CouponTypes}(object)
}
\arguments{
\item{object}{an S4 object of the type CouponTypes}
}
\description{
A method to extract CouponBasis from the slot of CouponTypes
}

