% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CouponTypes.R
\docType{methods}
\name{CouponBasis<-,CouponTypes-method}
\alias{CouponBasis<-,CouponTypes-method}
\title{A method to set CouponBasis in the slot of CouponTypes}
\usage{
\S4method{CouponBasis}{CouponTypes}(object) <- value
}
\arguments{
\item{object}{an S4 object of the type CouponTypes}

\item{value}{the replacement value of the slot}
}
\description{
A method to set CouponBasis in the slot of CouponTypes
}
