% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CouponTypes.R
\name{CouponBasis<-}
\alias{CouponBasis<-}
\title{A standard generic function ot replace the slot CouponBasis}
\usage{
CouponBasis(object) <- value
}
\arguments{
\item{object}{an S4 class object}

\item{value}{the replacement value of the slot}
}
\description{
A standard generic function ot replace the slot CouponBasis
}

