% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CouponTypes.R
\name{CouponBasis}
\alias{CouponBasis}
\title{A standard generic function to access the slot CouponBasis}
\usage{
CouponBasis(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A standard generic function to access the slot CouponBasis
}

