% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CouponTypes.R
\docType{methods}
\name{CouponDecimal,CouponTypes-method}
\alias{CouponDecimal,CouponTypes-method}
\title{A method to extract CouponDecimal from object CouponTypes}
\usage{
\S4method{CouponDecimal}{CouponTypes}(object)
}
\arguments{
\item{object}{an S4 object of the type CouponTypes}
}
\description{
A method to extract CouponDecimal from object CouponTypes
}

