% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CouponTypes.R
\name{CouponDecimal<-,CouponTypes-method}
\alias{CouponDecimal<-,CouponTypes-method}
\title{A method to set CouponDecimal in the slot of CouponTypes}
\usage{
\S4method{CouponDecimal}{CouponTypes}(object) <- value
}
\arguments{
\item{object}{an S4 object of the type CouponTypes}

\item{value}{the replacement value of the slot}
}
\description{
A method to set CouponDecimal in the slot of CouponTypes
}
