% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CouponTypes.R
\name{CouponDecimal<-}
\alias{CouponDecimal<-}
\title{A standard generic function to set the slot CouponDecimal}
\usage{
CouponDecimal(object) <- value
}
\arguments{
\item{object}{an S4 class object of the type CouponTypes}

\item{value}{the replacement value of the slot}
}
\description{
A standard generic function to set the slot CouponDecimal
}

