% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CouponTypes.R
\docType{methods}
\name{CouponDecimalString,CouponTypes-method}
\alias{CouponDecimalString,CouponTypes-method}
\title{A method to extract CouponDecimalString from the object CouponTypes}
\usage{
\S4method{CouponDecimalString}{CouponTypes}(object)
}
\arguments{
\item{object}{an S4 object of the type CouponTypes}
}
\description{
A method to extract CouponDecimalString from the object CouponTypes
}
