% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CouponTypes.R
\name{CouponDecimalString<-}
\alias{CouponDecimalString<-}
\title{A standard generic functon to set the slot CouponDecimalString}
\usage{
CouponDecimalString(object) <- value
}
\arguments{
\item{object}{an S4 class object}

\item{value}{the replacement value of the slot}
}
\description{
A standard generic functon to set the slot CouponDecimalString
}

