% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CouponTypes.R
\name{CouponDecimalString}
\alias{CouponDecimalString}
\title{A standard generic function to get the slot CouponDecimalString}
\usage{
CouponDecimalString(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A standard generic function to get the slot CouponDecimalString
}

