% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondScenario.R
\docType{methods}
\name{CouponIncome,BondReturn-method}
\alias{CouponIncome,BondReturn-method}
\title{A method to extract CouponIncome from S4 class BondReturn}
\usage{
\S4method{CouponIncome}{BondReturn}(object)
}
\arguments{
\item{object}{the name of an S4 class of type BondReturn}
}
\description{
A method to extract CouponIncome from S4 class BondReturn
}

