% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\docType{methods}
\name{CouponIncome,MortgageReturn-method}
\alias{CouponIncome,MortgageReturn-method}
\title{CouponIncome method, class MortgageReturn}
\usage{
\S4method{CouponIncome}{MortgageReturn}(object)
}
\arguments{
\item{object}{A class of the type MortgageReturn}
}
\description{
A method to get the  \strong{CouponIncome} paid to the 
to the investor over the scenrio horizon.  CouponIncome is reported as the 
sum of the coupon income received by the investor.
}
\seealso{
Other Mortgage Scenario Analysis: \code{\link{MortgageReturn-class}},
  \code{\link{ScheduledPrinReceived,MortgageReturn-method}}
}

