% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\name{CouponIncome}
\alias{CouponIncome}
\title{A standard generic function to access the slot CouponIncome}
\usage{
CouponIncome(object)
}
\arguments{
\item{object}{An S4 object of the type MtgScenario}
}
\description{
A standard generic function to access the slot CouponIncome
}

