% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondCashFlow.R
\docType{methods}
\name{CouponPmt,BondCashFlows-method}
\alias{CouponPmt,BondCashFlows-method}
\title{Method to extract CouponPmt from S4 class BondCashFlows}
\usage{
\S4method{CouponPmt}{BondCashFlows}(object)
}
\arguments{
\item{object}{the name of the S4 object}
}
\description{
Method to extract CouponPmt from S4 class BondCashFlows
}

