% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondCashFlows.R
\name{CouponPmt}
\alias{CouponPmt}
\title{A standard generic function to access the slot CouponPmt}
\usage{
CouponPmt(object)
}
\arguments{
\item{object}{An S4 class object}
}
\description{
A standard generic function to access the slot CouponPmt
}

