% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CouponTypes.R
\docType{class}
\name{CouponTypes-class}
\alias{CouponTypes-class}
\title{An S4 class representing the bond coupon}
\description{
This class is used to pass coupon types reported to investors and used in
analytics.  For example coupon is often reported as a decimal number
(e.g. 5.0%) but the coupon basis is used to calculate interest due to the 
investor.  In BondLab coupon is a numeric value as opposed to price which is
a character value.
}
\section{Slots}{

\describe{
\item{\code{CouponDecimal}}{A numeric value the coupon expressed in decimal notation
eg 5.50}

\item{\code{CouponBasis}}{A numeric value the coupon expressed in basis notation
eg 0.055}

\item{\code{CouponDecimalString}}{A character value the coupon expressed as a string
using decimal notation eg "5.50".}
}}

