% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CouponTypes.R
\name{CouponTypes}
\alias{CouponTypes}
\title{CouponTypes is a constructor function for the CouponTypes class}
\usage{
CouponTypes(coupon = numeric())
}
\arguments{
\item{coupon}{A numeric value the state coupon of the bond}
}
\description{
The CouponTypes class converts coupon as a numeric value to a basis value
(CouponBasis) used to calculate interest due and character string 
CouponDecimalString
}

