% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/MortgageFunctions.R
=======
% Please edit documentation in R/Functions.R
>>>>>>> master
\name{CPR.To.SMM}
\alias{CPR.To.SMM}
\title{A function to convert the CPR to a single monthly mortality rate SMM}
\usage{
CPR.To.SMM(CPR = numeric())
}
\arguments{
\item{CPR}{A numeric value}
}
\description{
Standard Generic Function used to convert CPR to SMM
}
\examples{
CPR.To.SMM(CPR = .06)
}

