% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelToCPR.R
\name{CPRLife,ModelToCPR-method}
\alias{CPRLife,ModelToCPR-method}
\title{A method to extract Life CPR from S4 class}
\usage{
\S4method{CPRLife}{ModelToCPR}(object)
}
\arguments{
\item{object}{An object of type ModelToCPR}
}
\description{
A method to extract Life CPR from S4 class
}
