% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelToCPR.R
\name{CPRLife}
\alias{CPRLife}
\title{A standard generic to access the CPRLife slot of class ModelToCPR}
\usage{
CPRLife(object)
}
\arguments{
\item{object}{An S4 object class}
}
\description{
A standard generic to access the CPRLife slot of class ModelToCPR
}

