% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CashFlowMatrix.R
\name{CreateCashFlowMatrix}
\alias{CreateCashFlowMatrix}
\title{CreateCashFlowMatrix}
\usage{
CreateCashFlowMatrix(trade.date, bonddata)
}
\arguments{
\item{trade.date}{The trade date}

\item{bonddata}{A dataframe of bond data.  The complete dataframe must
consist of the following names columns: cusip, interestrate, issuedate,
dateddate, maturitydate, type, close.}
}
\description{
An dataframe containing bond cash flow
}
\seealso{
Other Interest Rate Models: \code{\link{CreateMaturitiesMatrix}}
}
