% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBSDetails.R
\name{CurrentBal,MBSDetails-method}
\alias{CurrentBal,MBSDetails-method}
\title{Method to extract the slot CurrentBal from the class MBSDetails}
\usage{
\S4method{CurrentBal}{MBSDetails}(object)
}
\arguments{
\item{object}{the name of the object of the type MBSDetails}
}
\description{
Method to extract the slot CurrentBal from the class MBSDetails
}
