% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBSDetails.R
\docType{methods}
\name{CurrentBal<-,MBSDetails-method}
\alias{CurrentBal<-,MBSDetails-method}
\title{Method to replace the slot CurrentBal from the class MBSDetails}
\usage{
\S4method{CurrentBal}{MBSDetails}(object) <- value
}
\arguments{
\item{object}{the name of the object of type MBSDetails}

\item{value}{the replacement value of the slot}
}
\description{
Method to replace the slot CurrentBal from the class MBSDetails
}
