% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBSDetails.R
\name{CurrentBal<-}
\alias{CurrentBal<-}
\title{A standard generic function to replace the slot CurrentBal}
\usage{
CurrentBal(object) <- value
}
\arguments{
\item{object}{an S4 class object}

\item{value}{the replacement value of the slot}
}
\description{
A standard generic function to replace the slot CurrentBal
}
