% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelFunctions.R
\name{Curtailment}
\alias{Curtailment}
\title{A standard generic function to access the slot Curatialment from the class
PrepaymentModelFunctions}
\usage{
Curtailment(object)
}
\arguments{
\item{object}{an S4 object}
}
\description{
A standard generic function to access the slot Curatialment from the class
PrepaymentModelFunctions
}

