% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurveSpreads.R
\docType{class}
\name{CurveSpreads-class}
\alias{CurveSpreads-class}
\title{An S4 class to hold curve spread values}
\description{
An S4 class to hold curve spread values
}
\section{Slots}{

\describe{
\item{\code{BenchMark}}{A numeric value the closest curve benchmark}

\item{\code{SpreadToBenchmark}}{A numeric value to nominal spread to the nearest
pricing benchmark}

\item{\code{SpreadToCurve}}{A numeric value the nominal spread to 
the interpolated curve}

\item{\code{ZeroVolSpread}}{A numeric value the Zero Volatilty spread}
}}

