% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BillDetails.R
\docType{methods}
\name{Cusip,BillDetails-method}
\alias{Cusip,BillDetails-method}
\title{Method to extract Cusip from an S4 class of type BillDetails}
\usage{
\S4method{Cusip}{BillDetails}(object)
}
\arguments{
\item{object}{The name of the S4 object of type BillDetails}
}
\description{
Method to extract Cusip from an S4 class of type BillDetails
}

