% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondTermStructure.R
\name{Cusip,BondTermStructure-method}
\alias{Cusip,BondTermStructure-method}
\title{Cusip}
\usage{
\S4method{Cusip}{BondTermStructure}(object)
}
\arguments{
\item{object}{object of type BondTermStructure}
}
\description{
Get Cusip from object BondTermStructure
}
