% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PassThroughConstructor.R
\docType{methods}
\name{Cusip,MBSDetails-method}
\alias{Cusip,MBSDetails-method}
\title{Method to extract the slot Cusip from the class MBSDetails}
\usage{
\S4method{Cusip}{MBSDetails}(object)
}
\arguments{
\item{object}{the name of the object of type MBSDetails}
}
\description{
Method to extract the slot Cusip from the class MBSDetails
}

