% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageOAS.R
\docType{methods}
\name{Cusip,MortgageOAS-method}
\alias{Cusip,MortgageOAS-method}
\title{A method to access the slot cusip}
\usage{
\S4method{Cusip}{MortgageOAS}(object)
}
\arguments{
\item{object}{MortgageOAS object}
}
\description{
Access the slot Cusip from object MortgageOAS
}
