% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageKeyRate.R
\docType{methods}
\name{Cusip,MortgageTermStructure-method}
\alias{Cusip,MortgageTermStructure-method}
\title{Cusip}
\usage{
\S4method{Cusip}{MortgageTermStructure}(object)
}
\arguments{
\item{object}{MortgageTermStructure}
}
\description{
Get Cusip from object of type MortgageTermStructure
}

