% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBSDetails.R
\name{Cusip}
\alias{Cusip}
\title{A standard generic function to access the slot Cusip}
\usage{
Cusip(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A standard generic function to access the slot Cusip
}
