% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{CusipList,Collateral-method}
\alias{CusipList,Collateral-method}
\title{A method ot access the cusip list from S4 class Collateral}
\usage{
\S4method{CusipList}{Collateral}(object)
}
\arguments{
\item{object}{An S4 class object of the type Collateral}
}
\description{
A method ot access the cusip list from S4 class Collateral
}

