% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\name{CusipList}
\alias{CusipList}
\title{A generic function to access a cusip list slot}
\usage{
CusipList(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A generic function to access a cusip list slot
}

