% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/REMICCusip.R
\name{CusipRecord}
\alias{CusipRecord}
\title{A function to move REMIC tranche cusip data to the Bond Data Director}
\usage{
CusipRecord(source.path = "character", target.path = "character")
}
\arguments{
\item{source.path}{A charachter string the source path to the Traches folder}

\item{target.path}{A character string the target path to the BondData folder}
}
\description{
This function moves REMIC tranche data from the Tranches data folder which is used
by the REMIC constructor to the BondData folder.  This is the final step after the
REMIC structure has been validated.
}

