% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BillDetails.R
\name{DatedDate,BillDetails-method}
\alias{DatedDate,BillDetails-method}
\title{Method to extract DatedDate from an S4 class of type BillDetails}
\usage{
\S4method{DatedDate}{BillDetails}(object)
}
\arguments{
\item{object}{The name of the S4 class of type BillDetails}
}
\description{
Method to extract DatedDate from an S4 class of type BillDetails
}
