% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondDetails.R
\docType{methods}
\name{DatedDate,BondDetails-method}
\alias{DatedDate,BondDetails-method}
\title{Method to extract DatedDate from an S4 class of type BondDetails}
\usage{
\S4method{DatedDate}{BondDetails}(object)
}
\arguments{
\item{object}{the name of the S4 class of type BondDetails}
}
\description{
Method to extract DatedDate from an S4 class of type BondDetails
}
