% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICStructuringFunctions.R
\name{DealCurrBal}
\alias{DealCurrBal}
\title{A function to calculate REMIC deal original balance on data.tree}
\usage{
DealCurrBal(self)
}
\description{
Function calculates original balance across tree
}

