% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{DealName,RAID-method}
\alias{DealName,RAID-method}
\title{A method to access the slot DealName from class RAID}
\usage{
\S4method{DealName}{RAID}(object)
}
\arguments{
\item{object}{an S4 class of the type RAID}
}
\description{
A method to access the slot DealName from class RAID
}

