% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{DealName,TrancheDetails-method}
\alias{DealName,TrancheDetails-method}
\title{A method to access the slot DealName from class TrancheDetails}
\usage{
\S4method{DealName}{TrancheDetails}(object)
}
\arguments{
\item{object}{an S4 class of type TrancheDetails}
}
\description{
A method to access the slot DealName from class TrancheDetails
}

