% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{DealNumber,RAID-method}
\alias{DealNumber,RAID-method}
\title{A method to access the slot DealNumber from class RAID}
\usage{
\S4method{DealNumber}{RAID}(object)
}
\arguments{
\item{object}{an S4 class of the type RAID}
}
\description{
A method to access the slot DealNumber from class RAID
}

