% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{DealSize,RAID-method}
\alias{DealSize,RAID-method}
\title{A method to access the slot DealSize from class RAID}
\usage{
\S4method{DealSize}{RAID}(object)
}
\arguments{
\item{object}{an S4 class of type RAID}
}
\description{
A method to access the slot DealSize from class RAID
}

