% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HullWhite.R
\name{DecimalSet}
\alias{DecimalSet}
\title{Set Decimal Length}
\usage{
DecimalSet(x, k)
}
\arguments{
\item{x}{the number to be rounded}

\item{k}{the number of decimal places}
}
\description{
A function to set the number of decimal places.  This is used
in the Hull White Model to address the floating point issue that may arise
when simulating interest rates.
}
