% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelFunctions.R
\name{DefaultOrigLTVMult}
\alias{DefaultOrigLTVMult}
\title{A standard generic function to access the slot OrigLTVMult from the 
class PrepaymentModelFunctions}
\usage{
DefaultOrigLTVMult(object)
}
\arguments{
\item{object}{an S4 object}
}
\description{
A standard generic function to access the slot OrigLTVMult from the 
class PrepaymentModelFunctions
}
