% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelFunctions.R
\docType{methods}
\name{DefaultRamp,PrepaymentModelFunctions-method}
\alias{DefaultRamp,PrepaymentModelFunctions-method}
\title{A method to extract the slot DefaultRamp from the object PrepaymentModelFunctions}
\usage{
\S4method{DefaultRamp}{PrepaymentModelFunctions}(object)
}
\arguments{
\item{object}{an S4 object of the class PrepaymentModelFunctions}
}
\description{
A method to extract the slot DefaultRamp from the object PrepaymentModelFunctions
}

