% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelFunctions.R
\name{DefaultRamp}
\alias{DefaultRamp}
\title{A standard generic function to access the slot DefaultRamp from the 
class PrepaymentModelFunctions}
\usage{
DefaultRamp(object)
}
\arguments{
\item{object}{an S4 object}
}
\description{
A standard generic function to access the slot DefaultRamp from the 
class PrepaymentModelFunctions
}
