% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelFunctions.R
\name{DefaultSATOMult}
\alias{DefaultSATOMult}
\title{A standard generic function to access the slot SATOMult from the 
class PrepaymentModelFunctions}
\usage{
DefaultSATOMult(object)
}
\arguments{
\item{object}{an S4 object}
}
\description{
A standard generic function to access the slot SATOMult from the 
class PrepaymentModelFunctions
}

