% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelFunctions.R
\name{DefaultUpdatedLTVMult}
\alias{DefaultUpdatedLTVMult}
\title{A standard generic function to access the slot UpdatedLTVMult from the 
class PrepaymentModelFunctions}
\usage{
DefaultUpdatedLTVMult(object)
}
\arguments{
\item{object}{an S4 object}
}
\description{
A standard generic function to access the slot UpdatedLTVMult from the 
class PrepaymentModelFunctions
}
