% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BillCashFlows.R
\docType{methods}
\name{Discount,BillCashFlows-method}
\alias{Discount,BillCashFlows-method}
\title{Method to get discount from S4 class BillCashFlows}
\usage{
\S4method{Discount}{BillCashFlows}(object)
}
\arguments{
\item{object}{the name of the S4 object BillCashFlows}
}
\description{
Method to get discount from S4 class BillCashFlows
}

