% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BillCashFlows.R
\name{Discount}
\alias{Discount}
\title{A standard generic to get the slot Discount}
\usage{
Discount(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A standard generic to get the slot Discount
}

