% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscountTypes.R
\docType{methods}
\name{DiscountBasis<-,DiscountTypes-method}
\alias{DiscountBasis<-,DiscountTypes-method}
\title{A method to set the slot DiscountBasis in the object DiscountTypes}
\usage{
\S4method{DiscountBasis}{DiscountTypes}(object) <- value
}
\arguments{
\item{object}{an S4 object of the type DiscountTypes}

\item{value}{the replacement value of the slot}
}
\description{
A method to set the slot DiscountBasis in the object DiscountTypes
}
