% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscountTypes.R
\docType{methods}
\name{DiscountDecimal,DiscountTypes-method}
\alias{DiscountDecimal,DiscountTypes-method}
\title{A method to extract the slot DiscountDecimal from the object DiscountTypes}
\usage{
\S4method{DiscountDecimal}{DiscountTypes}(object)
}
\arguments{
\item{object}{an S4 object of the type DiscountTypes}
}
\description{
A method to extract the slot DiscountDecimal from the object DiscountTypes
}

