% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscountTypes.R
\name{DiscountDecimal<-}
\alias{DiscountDecimal<-}
\title{A standard generic to set the slot DiscountDecimal}
\usage{
DiscountDecimal(object) <- value
}
\arguments{
\item{object}{an S4 class of the type DiscountTypes}

\item{value}{the replacement value of the slot}
}
\description{
A standard generic to set the slot DiscountDecimal
}

