% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscountTypes.R
\name{DiscountDecimal}
\alias{DiscountDecimal}
\title{A standard generic to get the slot DiscountDecimal}
\usage{
DiscountDecimal(object)
}
\arguments{
\item{object}{an S4 class object of the type DiscountTypes}
}
\description{
A standard generic to get the slot DiscountDecimal
}

