% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscountTypes.R
\docType{methods}
\name{DiscountDecimalString<-,DiscountTypes-method}
\alias{DiscountDecimalString<-,DiscountTypes-method}
\title{A method to set the slot DiscountDecimalString from the object DiscountTypes}
\usage{
\S4method{DiscountDecimalString}{DiscountTypes}(object) <- value
}
\arguments{
\item{object}{an object of the type DiscountTypes}

\item{value}{the replacement value of the slot}
}
\description{
A method to set the slot DiscountDecimalString from the object DiscountTypes
}
