% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscountTypes.R
\name{DiscountDecimalString<-}
\alias{DiscountDecimalString<-}
\title{A standard  generic to set the slot DiscountDecimalString}
\usage{
DiscountDecimalString(object) <- value
}
\arguments{
\item{object}{an S4 class of the type DiscountTypes}

\item{value}{the replacement value of the slot}
}
\description{
A standard  generic to set the slot DiscountDecimalString
}

