% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscountTypes.R
\name{DiscountDecimalString}
\alias{DiscountDecimalString}
\title{A standard generic to get the slot DiscountDecimalString}
\usage{
DiscountDecimalString(object)
}
\arguments{
\item{object}{an S4 class of the type DiscountTypes}
}
\description{
A standard generic to get the slot DiscountDecimalString
}
