% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscountToPrice.R
\name{DiscountToPrice}
\alias{DiscountToPrice}
\title{Discount To Price bill}
\usage{
DiscountToPrice(bill.id, discount, day.count, settlement.date)
}
\arguments{
\item{bill.id}{A character string the cusip number or bill.id}

\item{discount}{A numeric value the discount rate stated as a percentage}

\item{day.count}{A numeric value the day count used for pricing}

\item{settlement.date}{A character the settlement date mm-dd-YYYY}
}
\description{
Returns the price of a discount bill given a discount rate
and day count
}
\seealso{
Other Pricing: \code{\link{BenchMark,CurveSpreads-method}},
  \code{\link{CurveSpreads-class}},
  \code{\link{CurveSpreads}},
  \code{\link{PriceToBondYield}},
  \code{\link{PriceToDiscountYield}},
  \code{\link{SpreadToBenchmark,CurveSpreads-method}},
  \code{\link{SpreadToCurve,CurveSpreads-method}},
  \code{\link{SpreadToPriceBond}},
  \code{\link{SpreadToPriceMBS}},
  \code{\link{ZeroVolSpread,CurveSpreads-method}}
}
