% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscountTypes.R
\docType{class}
\name{DiscountTypes-class}
\alias{DiscountTypes-class}
\title{An S4 class representing the discount rate applied to bill and strips}
\description{
This class is used to create and pass discount types reported to investors 
used in analytics.  For example, discount is often reported as a decimal number
(e.g. 5.0%) but the discount basis is used to calculate the price of the security
In BondLab discount is a numeric value in all slots as opposed to 
price which is a character value in all slots.
}
\section{Slots}{

\describe{
\item{\code{DiscountDecimal}}{A numeric value the discount expressed in decimal 
notation eg 1.05}

\item{\code{DiscountBasis}}{A numeric value discount expressed in basis notation
eg 0.015}

\item{\code{DiscountDecimalString}}{A character value the discount rate expressed
as a string expressed in decimal notation eg "5.50"}
}}

