% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscountTypes.R
\name{DiscountTypes}
\alias{DiscountTypes}
\title{DiscountTypes is the constructor function for the DiscountTypes class}
\usage{
DiscountTypes(discount.rate = numeric())
}
\arguments{
\item{discount.rate}{A numeric value the stated discount rate of the bill}
}
\description{
The DiscountTypes class converts the discount rate from a numeric value to a
basis value (divided by 100) and a string (basis value)
}
