% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TermStructure.R
\docType{methods}
\name{DiscRate<-,TermStructure-method}
\alias{DiscRate<-,TermStructure-method}
\title{Method to replace DiscRate from the class TermStructure}
\usage{
\S4method{DiscRate}{TermStructure}(object) <- value
}
\arguments{
\item{object}{the name of the object of type TermStructure}

\item{value}{the replacement value}
}
\description{
Method to replace DiscRate from the class TermStructure
}
