% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TermStructure.R
\name{DiscRate}
\alias{DiscRate}
\title{A generic function to access the slot DiscRate}
\usage{
DiscRate(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A generic function to access the slot DiscRate
}
