% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DollarRoll.R
\name{DollarRoll}
\alias{DollarRoll}
\title{Dollar Roll Calculates the one-month Dollar Roll}
\usage{
DollarRoll(bond.id = "character", price = numeric(), drop = numeric(),
  original.bal = numeric(), settlement.date = "character",
  fwd.settlement.date = "character", reinvestment.rate = numeric(),
  finance.rate = numeric(), MortgageCashFlow = "character")
}
\arguments{
\item{bond.id}{A character string the cusip or id of the MBS}

\item{price}{A numeric value the price of the MBS}

\item{drop}{A numeric value the one-month drop.}

\item{original.bal}{A numeric value the original balance of the MBS pool}

\item{settlement.date}{A character string the current settlement date}

\item{fwd.settlement.date}{A character string the forward settlement date}

\item{reinvestment.rate}{A numeric value the investor's reinvestment rate}

\item{finance.rate}{A numeric value the investor's alternative finance or repo rate}

\item{MortgageCashFlow}{A character string the name of a MortgageCashFlow Object}
}
\description{
One-month dollar roll analysis.  This function is called by Dollar Roll Analytics
}

