% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DollarRollAnalytics.R
\name{DollarRollAnalytics}
\alias{DollarRollAnalytics}
\title{A constructor of Dollar Roll Analytics object Analysis of the one-month Dollar Roll}
\usage{
DollarRollAnalytics(bond.id = "character", original.bal = numeric(),
  price = numeric(), drop = numeric(), trade.date = "character",
  settlement.date = "character", fwd.settlement.date = "character",
  reinvestment.rate = numeric(), finance.rate = numeric(), method = "ns",
  PrepaymentAssumption = "character", ..., begin.cpr = numeric(),
  end.cpr = numeric(), seasoning.period = numeric(), CPR = numeric())
}
\arguments{
\item{bond.id}{A character string the MBS cusip or bond id}

\item{original.bal}{A numeric value the MBS original balance}

\item{price}{A numeric value the price of the MBS pass-through}

\item{drop}{A numeric value the quoted drop}

\item{trade.date}{A character string the trade date}

\item{settlement.date}{A character string the settlement date}

\item{fwd.settlement.date}{A character string the forward settlement date of the roll}

\item{reinvestment.rate}{A numeric value the investor's reinvestment rate}

\item{finance.rate}{A numeric value the investor's alternative finance or repo rate}

\item{method}{A character string the method passed to termstrc to fit the term structure}

\item{PrepaymentAssumption}{A character string the model used to value the dollar roll may be CPR, PPC, or MODEL}

\item{...}{Optional values for input when using CPR or PPC}

\item{begin.cpr}{optional, A numeric value the beginning cpr that must be specified when PPC prepayment assumption is used}

\item{end.cpr}{optional, A numeric vlaue the ending cpr that must be specified  when PPC prepayment assumption is used}

\item{seasoning.period}{optional, A numeric value the length of the PPC seasoning ramp that must be specified when PPC prepayment assumption is used}

\item{CPR}{optional, A numeric value the CPR assumption that must be specified when CPR prepayment assumption is used}
}
\description{
Analysis of the one-month dollar roll for a pass through or cohort representation
}

