% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondTermStructure.R
\docType{methods}
\name{EffConvexity,BondTermStructure-method}
\alias{EffConvexity,BondTermStructure-method}
\title{A method to get EffConvexity from the class BondTermStructure}
\usage{
\S4method{EffConvexity}{BondTermStructure}(object)
}
\arguments{
\item{object}{an object of the type BondTermStructure}
}
\description{
A method to get EffConvexity from the class BondTermStructure
}

