% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageEffectiveMeasures.R
\docType{methods}
\name{EffConvexity,MortgageEffectiveMeasures-method}
\alias{EffConvexity,MortgageEffectiveMeasures-method}
\title{A method to extract the EffConvexity from class MortgageEffectiveMeasures}
\usage{
\S4method{EffConvexity}{MortgageEffectiveMeasures}(object)
}
\arguments{
\item{object}{The name of the S4 object of type MortgageEffectiveMeasures}
}
\description{
A method to extract the EffConvexity from class MortgageEffectiveMeasures
}

