% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageKeyRate.R
\docType{methods}
\name{EffConvexity,MortgageTermStructure-method}
\alias{EffConvexity,MortgageTermStructure-method}
\title{Method to extract EffConvexity from S4 class}
\usage{
\S4method{EffConvexity}{MortgageTermStructure}(object)
}
\arguments{
\item{object}{The name of the S4 object of type MortgageTermStructure}
}
\description{
Method to extract EffConvexity from S4 class
}

