% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\docType{methods}
\name{EffConvexity,Mtg.Scenario-method}
\alias{EffConvexity,Mtg.Scenario-method}
\title{A method to extract EffConvexity from S4 class Mtg.Scenario}
\usage{
\S4method{EffConvexity}{Mtg.Scenario}(object)
}
\arguments{
\item{object}{the name of an S4 class of type Mtg.Scenario}
}
\description{
A method to extract EffConvexity from S4 class Mtg.Scenario
}

