% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondTermStructure.R
\docType{methods}
\name{EffDuration,BondTermStructure-method}
\alias{EffDuration,BondTermStructure-method}
\title{A method to get EffDuration from the class BondTermStructure}
\usage{
\S4method{EffDuration}{BondTermStructure}(object)
}
\arguments{
\item{object}{an object of the type BondTermStructure}
}
\description{
A method to get EffDuration from the class BondTermStructure
}
