% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageKeyRate.R
\docType{methods}
\name{EffDuration,MortgageTermStructure-method}
\alias{EffDuration,MortgageTermStructure-method}
\title{Method to extract EffDuration from S4 class}
\usage{
\S4method{EffDuration}{MortgageTermStructure}(object)
}
\arguments{
\item{object}{The name of the S4 object of type MortgageTermStructure}
}
\description{
Method to extract EffDuration from S4 class
}
