% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageKeyRate.R
\name{EffDuration}
\alias{EffDuration}
\title{A standard generic function to access the slot EffDuration}
\usage{
EffDuration(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A standard generic function to access the slot EffDuration
}

