% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EffectiveMeasure.R
\name{Effective.Measure}
\alias{Effective.Measure}
\title{A function to compute effective duration and convexity}
\usage{
Effective.Measure(Rate.Delta = numeric(), cashflow = vector(),
  discount.rates = vector(), time.period = vector(), type = "character")
}
\arguments{
\item{Rate.Delta}{A numeric value the interest rate shift in basis points}

\item{cashflow}{A numeric vector of cashflow}

\item{discount.rates}{A numeric vector of the discount rates}

\item{time.period}{A numeric vector of the time period}

\item{type}{A character vector to specify either duration or convexity}
}
\description{
Calculates the effective duration and based on discount vector (zero coupon)
cashflow vector, and rate delta
}

