% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelTune.R
\docType{methods}
\name{EndCDR,PrepaymentModelTune-method}
\alias{EndCDR,PrepaymentModelTune-method}
\title{A method to extract EndCDR from class PrepaymentModelTune}
\usage{
\S4method{EndCDR}{PrepaymentModelTune}(object)
}
\arguments{
\item{object}{The name of the object of class PrepaymentModelTune}
}
\description{
A method to extract EndCDR from class PrepaymentModelTune
}

