% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StructuringFunctions.R
\name{EndingBal}
\alias{EndingBal}
\title{Function to assign the ending principal balance to REMIC tranche}
\usage{
EndingBal(node, period = numeric())
}
\arguments{
\item{node}{The entry point of the deal's tree structure}

\item{period}{A numeric value the period in which payment is received stated as integer}
}
\description{
This function takes the beginning balance of the current period to calculate the 
ending balance of the current period based on the principal paid in the period.  The
principal paid in the period is passed via the node$Principal argument 
Structuring functions run on a tree structure created by the package data.tree
}

