% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageCashFlow.R
\docType{methods}
\name{EndingBalance,MortgageCashFlow-method}
\alias{EndingBalance,MortgageCashFlow-method}
\title{Method to extract the Ending Balance from class MortgageCashFlow}
\usage{
\S4method{EndingBalance}{MortgageCashFlow}(object)
}
\arguments{
\item{object}{the name of the object of type MortgageCashFlow}
}
\description{
Method to extract the Ending Balance from class MortgageCashFlow
}
