% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelTune.R
\docType{methods}
\name{EndMonth,PrepaymentModelTune-method}
\alias{EndMonth,PrepaymentModelTune-method}
\title{A method to extract EndMonth from class PrepaymentModelTune}
\usage{
\S4method{EndMonth}{PrepaymentModelTune}(object)
}
\arguments{
\item{object}{The name of the object of class PrepaymentModelTune}
}
\description{
A method to extract EndMonth from class PrepaymentModelTune
}
