% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelTune.R
\name{EndMonth}
\alias{EndMonth}
\title{A standard generic function to access the slot EndMonth}
\usage{
EndMonth(object)
}
\arguments{
\item{object}{an S4 object}
}
\description{
A standard generic function to access the slot EndMonth
}

