% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Functions.R
\name{EstimYTM}
\alias{EstimYTM}
\title{A function to estimate the yield to maturity of a standard bond}
\usage{
EstimYTM(coupon = numeric(), coupon.frequency = numeric(),
  years.mat = numeric(), face.value = numeric(), price = numeric())
}
\arguments{
\item{coupon}{A numeric value expressing the bond's coupon as a
percentage of its face amount}

\item{coupon.frequency}{A numeric value expressing the frequency of payments
over one year}

\item{years.mat}{A numeric value expressing the years to maturity}

\item{face.value}{A numeric value expressing the face value (principal amount) of the bond}

\item{price}{A numeric value expressing the price of the bond (not percentage of face value)
for example a price of$102 is entered as 102.00}
}
\description{
Estimate a bond's yield to maturity given a price.  It is a
nominall example and does not account for acrrued interest
}
\examples{
EstimYTM(coupon = .04, coupon.frequency = 2, years.mat = 10, face.value = 1000, price = 100)
}

