% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{ExpenseBasisPointFee,RAID-method}
\alias{ExpenseBasisPointFee,RAID-method}
\title{A method to access the slot ExpenseBasisPointFee from class RAID}
\usage{
\S4method{ExpenseBasisPointFee}{RAID}(object)
}
\arguments{
\item{object}{an S4 class of the type RAID}
}
\description{
A method to access the slot ExpenseBasisPointFee from class RAID
}

