% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModel.R
\name{FinalPmtDate,PrepaymentModel-method}
\alias{FinalPmtDate,PrepaymentModel-method}
\title{A method to extract FinalPmtDate from S4 class PrepaymentModel}
\usage{
\S4method{FinalPmtDate}{PrepaymentModel}(object)
}
\arguments{
\item{object}{the name of the S4 class}
}
\description{
A method to extract FinalPmtDate from S4 class PrepaymentModel
}
