% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PassThroughConstructor.R
\docType{methods}
\name{FirstPmtDate,MBSDetails-method}
\alias{FirstPmtDate,MBSDetails-method}
\title{Method to extract the slot FirstPmtDate from the class MBSDetails}
\usage{
\S4method{FirstPmtDate}{MBSDetails}(object)
}
\arguments{
\item{object}{the name of the object of type MBSDetails}
}
\description{
Method to extract the slot FirstPmtDate from the class MBSDetails
}

