% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModel.R
\docType{methods}
\name{FirstPmtDate,PrepaymentAssumption-method}
\alias{FirstPmtDate,PrepaymentAssumption-method}
\title{A method to extract FirstPmtDate from S4 class PrepaymentAssumption}
\usage{
\S4method{FirstPmtDate}{PrepaymentAssumption}(object)
}
\arguments{
\item{object}{the name of the S4 class}
}
\description{
A method to extract FirstPmtDate from S4 class PrepaymentAssumption
}

