% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModel.R
\name{FirstPmtDate,PrepaymentModel-method}
\alias{FirstPmtDate,PrepaymentModel-method}
\title{A method to extract FirstPmtDate from S4 class PrepaymentModel}
\usage{
\S4method{FirstPmtDate}{PrepaymentModel}(object)
}
\arguments{
\item{object}{the name of the S4 class}
}
\description{
A method to extract FirstPmtDate from S4 class PrepaymentModel
}
