% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageCashFlow.R
\docType{methods}
\name{FirstPrinPaymentDate,MortgageCashFlow-method}
\alias{FirstPrinPaymentDate,MortgageCashFlow-method}
\title{Method to extract the FirstPrinPaymentDate from the class MortgageCashFlow}
\usage{
\S4method{FirstPrinPaymentDate}{MortgageCashFlow}(object)
}
\arguments{
\item{object}{the name of the object of type MortgageCashFlow}
}
\description{
Method to extract the FirstPrinPaymentDate from the class MortgageCashFlow
}

