% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{Fixed,TrancheDetails-method}
\alias{Fixed,TrancheDetails-method}
\title{A method to access the slot Fixed from the class TrancheDetails}
\usage{
\S4method{Fixed}{TrancheDetails}(object)
}
\arguments{
\item{object}{an S4 object of the type TrancheDetails}
}
\description{
A method to access the slot Fixed from the class TrancheDetails
}

