% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICStructuringFunctions.R
\name{FloatCoupon}
\alias{FloatCoupon}
\title{A function to calculate the floater coupon}
\usage{
FloatCoupon(cap, floor, margin, index, multiplier)
}
\arguments{
\item{cap}{a numeric value the floater cap}

\item{floor}{a numeric value the floater floor}

\item{margin}{a numeric value the floater margin}

\item{index}{a numeric value the index value}

\item{multiplier}{a numeric value the index multiplier}
}
\description{
The function calculates the floater coupon given
}

