% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{FloaterFirstResetDate,TrancheDetails-method}
\alias{FloaterFirstResetDate,TrancheDetails-method}
\title{A method to access the slot FloaterFirstResetDate from class TrancheDetails}
\usage{
\S4method{FloaterFirstResetDate}{TrancheDetails}(object)
}
\arguments{
\item{object}{an S4 object o the type TrancheDetails}
}
\description{
A method to access the slot FloaterFirstResetDate from class TrancheDetails
}

