% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{FloaterInitialCoupon,TrancheDetails-method}
\alias{FloaterInitialCoupon,TrancheDetails-method}
\title{A method to access the slot FloaterInitialCoupon from the 
class TrancheDetails}
\usage{
\S4method{FloaterInitialCoupon}{TrancheDetails}(object)
}
\arguments{
\item{object}{an S4 object of the type TrancheDetails}
}
\description{
A method to access the slot FloaterInitialCoupon from the 
class TrancheDetails
}

