% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{FloaterMultiplier,TrancheDetails-method}
\alias{FloaterMultiplier,TrancheDetails-method}
\title{A method to access the slot FloaterMultiplier from class TrancheDetails}
\usage{
\S4method{FloaterMultiplier}{TrancheDetails}(object)
}
\arguments{
\item{object}{an S4 object of the type TrancheDetails}
}
\description{
A method to access the slot FloaterMultiplier from class TrancheDetails
}

