% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\docType{methods}
\name{Formula,Scenario-method}
\alias{Formula,Scenario-method}
\title{A Mthod to extract the shift Formula from an S4 class Scenario}
\usage{
\S4method{Formula}{Scenario}(object)
}
\arguments{
\item{object}{The name of the object of the S4 class of type Scenario}
}
\description{
A Mthod to extract the shift Formula from an S4 class Scenario
}

