% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\name{Formula}
\alias{Formula}
\title{A standard generic to access the slot Formula}
\usage{
Formula(object)
}
\arguments{
\item{object}{An S4 class object of the type Scenario}
}
\description{
A standard generic to access the slot Formula
}

