% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HorizonBondData.R
\name{ForwardBond}
\alias{ForwardBond}
\title{A Function to "Roll Forward a bond cusip object BondDetails}
\usage{
ForwardBond(bond.id = "character", par.amount = numeric(),
  horizon.months = numeric())
}
\arguments{
\item{bond.id}{A character string referring to an object of the type BondDetails}

\item{par.amount}{A numeric value the par amount of the bond}

\item{horizon.months}{A numeric value the investment horizon in months}
}
\description{
A standard generic function to roll forward bond details to the scenario
horizon for pricing using the spot spread (nomimal spread), spread to the 
benchmark, option adjusted spread or price.
}
