% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TermStructure.R
\docType{methods}
\name{ForwardDate<-,TermStructure-method}
\alias{ForwardDate<-,TermStructure-method}
\title{Method to replace ForwardDate from the class TermStructure}
\usage{
\S4method{ForwardDate}{TermStructure}(object) <- value
}
\arguments{
\item{object}{the name of the object of type TermStructure}

\item{value}{the value of the replacement}
}
\description{
Method to replace ForwardDate from the class TermStructure
}

