% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TermStructure.R
\name{ForwardDate,TermStructure-method}
\alias{ForwardDate,TermStructure-method}
\title{Method to extract Date from the class TermStructure}
\usage{
\S4method{ForwardDate}{TermStructure}(object)
}
\arguments{
\item{object}{the name of the object of type TermStructure}
}
\description{
Method to extract Date from the class TermStructure
}
