% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TermStructure.R
\name{ForwardDate}
\alias{ForwardDate}
\title{A generic function to access the slot Date in class TermStructure}
\usage{
ForwardDate(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A generic function to access the slot Date in class TermStructure
}

