% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/HorizonBondData.R
\name{ForwardPassThrough}
\alias{ForwardPassThrough}
\title{A Function to "Roll Forward" the mortgage pass through cusip object (MBS Details)}
\usage{
ForwardPassThrough(bond.id = "character", ProjectedCashFlow = "character",
  HorizonMonths = "numeric")
}
\arguments{
\item{bond.id}{A character string referring to an object type of MBSDetails}

\item{ProjectedCashFlow}{A character string referring to an object type of MortgageCashFlows}
}
\description{
A standard generic function to roll forward MBS cusip detail to the scenario horizon
for pricing using spot spread, nominal spread, or option adjusted spread
}

