% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TermStructure.R
\docType{methods}
\name{ForwardRate<-,TermStructure-method}
\alias{ForwardRate<-,TermStructure-method}
\title{Method to replace ForwardRate in the class TermStructure}
\usage{
\S4method{ForwardRate}{TermStructure}(object) <- value
}
\arguments{
\item{object}{the name of the object of the type TermStructure}

\item{value}{the replacement value}
}
\description{
Method to replace ForwardRate in the class TermStructure
}
