% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TermStructure.R
\name{ForwardRate}
\alias{ForwardRate}
\title{A generic function to access the slot ForwardRate}
\usage{
ForwardRate(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A generic function to access the slot ForwardRate
}

