% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondDetails.R
\docType{methods}
\name{Frequency,BondDetails-method}
\alias{Frequency,BondDetails-method}
\title{Method to extract the slot Frequency from S4 class of type BondDetails}
\usage{
\S4method{Frequency}{BondDetails}(object)
}
\arguments{
\item{object}{the name of the object of type BondDetails}
}
\description{
Method to extract the slot Frequency from S4 class of type BondDetails
}

