% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\name{GetTrancheNumber}
\alias{GetTrancheNumber}
\title{A generic function to access the slot Tranches}
\usage{
GetTrancheNumber(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A generic function to access the slot Tranches
}

