% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetUSTCurve.R
\name{GetUSTCurve}
\alias{GetUSTCurve}
\title{A function to call UST constant maturity curve data from the St, Louis Fed  
FRED website.}
\usage{
GetUSTCurve(CurveDate, FREDAPI, ..., save.curve = FALSE)
}
\arguments{
\item{CurveDate}{a character the UST curve date}

\item{FREDAPI}{a character the user's FRED API Key}

\item{...}{optional value to save curve to rates file}

\item{save.curve}{a logical indicating whether to save the rates data TRUE of FALSE
defaults to FALSE}
}
\description{
A function to call UST constant maturity curve data from the St, Louis Fed  
FRED website.
}
