% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PassThroughConstructor.R
\name{GFee}
\alias{GFee}
\title{A standard generic function to access the slot GFee}
\usage{
GFee(object)
}
\arguments{
\item{object}{S4 class object}
}
\description{
A standard generic function to access the slot GFee
}

