% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFeeTypes.R
\docType{methods}
\name{GFeeBasis,GFeeTypes-method}
\alias{GFeeBasis,GFeeTypes-method}
\title{A method to get the slot GFeeBasis in the object GFeeTypes}
\usage{
\S4method{GFeeBasis}{GFeeTypes}(object)
}
\arguments{
\item{object}{an object of the type GFeeTypes}
}
\description{
A method to get the slot GFeeBasis in the object GFeeTypes
}

