% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFeeTypes.R
\name{GFeeBasis<-,GFeeTypes-method}
\alias{GFeeBasis<-,GFeeTypes-method}
\title{A method to set the value of the slot GFeeBasis in the object GFeeTypes}
\usage{
\S4method{GFeeBasis}{GFeeTypes}(object) <- value
}
\arguments{
\item{object}{an object of the type GFeeTypes}

\item{value}{the replacement value of the slot}
}
\description{
A method to set the value of the slot GFeeBasis in the object GFeeTypes
}
