% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFeeTypes.R
\name{GFeeBasis}
\alias{GFeeBasis}
\title{A standard generic function to get the value of the slot GFeeBasis}
\usage{
GFeeBasis(object)
}
\arguments{
\item{object}{an S4 object of the type GFeeTypes}
}
\description{
A standard generic function to get the value of the slot GFeeBasis
}

