% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFeeTypes.R
\docType{methods}
\name{GFeeDecimal,GFeeTypes-method}
\alias{GFeeDecimal,GFeeTypes-method}
\title{A method to access the slot GFeeDecimal from the object GFeeTypes}
\usage{
\S4method{GFeeDecimal}{GFeeTypes}(object)
}
\arguments{
\item{object}{an S4 object of the type GFeeTypes}
}
\description{
A method to access the slot GFeeDecimal from the object GFeeTypes
}

