% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFeeTypes.R
\name{GFeeDecimal<-,GFeeTypes-method}
\alias{GFeeDecimal<-,GFeeTypes-method}
\title{A method to set the slot GFeeDecimal in the object GFeeTypes}
\usage{
\S4method{GFeeDecimal}{GFeeTypes}(object) <- value
}
\arguments{
\item{object}{an object of the type GFeeTypes}

\item{value}{the replacement value of the slot}
}
\description{
A method to set the slot GFeeDecimal in the object GFeeTypes
}
