% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFeeTypes.R
\name{GFeeDecimal<-}
\alias{GFeeDecimal<-}
\title{A standard generic function to set the value of the slot GFeeDecimal}
\usage{
GFeeDecimal(object) <- value
}
\arguments{
\item{object}{an S4 object of the type GFeeTypes}

\item{value}{the replacement value of the slot}
}
\description{
A standard generic function to set the value of the slot GFeeDecimal
}
