% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFeeTypes.R
\name{GFeeDecimal}
\alias{GFeeDecimal}
\title{A standard generic function to access the slot GFeeDecimal}
\usage{
GFeeDecimal(object)
}
\arguments{
\item{object}{an S4 object of the type GFeeTypes}
}
\description{
A standard generic function to access the slot GFeeDecimal
}

