% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFeeTypes.R
\name{GFeeDecimalString<-,GFeeTypes-method}
\alias{GFeeDecimalString<-,GFeeTypes-method}
\title{A method to set the value of the slot GFeeDecimalString in the 
object GFeeTypes}
\usage{
\S4method{GFeeDecimalString}{GFeeTypes}(object) <- value
}
\arguments{
\item{object}{an S4 object of the type GFeeTypes}

\item{value}{the replacement value of the slot}
}
\description{
A method to set the value of the slot GFeeDecimalString in the 
object GFeeTypes
}
