% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFeeTypes.R
\name{GFeeDecimalString<-}
\alias{GFeeDecimalString<-}
\title{A standard generic function to set the slot GFeeDecimalString}
\usage{
GFeeDecimalString(object) <- value
}
\arguments{
\item{object}{an object of the type GFeeTypes}

\item{value}{the replacement value of the slot}
}
\description{
A standard generic function to set the slot GFeeDecimalString
}
