% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFeeTypes.R
\name{GFeeDecimalString}
\alias{GFeeDecimalString}
\title{A standard generic functon to get the slot GFeeDecimalString}
\usage{
GFeeDecimalString(object)
}
\arguments{
\item{object}{an object of the type GFeeTypes}
}
\description{
A standard generic functon to get the slot GFeeDecimalString
}
