% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageCashFlow.R
\name{GFeePremium,MortgageCashFlow-method}
\alias{GFeePremium,MortgageCashFlow-method}
\title{Method to extract the GFeePremium from the class MortgageCashFlow}
\usage{
\S4method{GFeePremium}{MortgageCashFlow}(object)
}
\arguments{
\item{object}{the name of the object of type MortgageCashFlow}
}
\description{
Method to extract the GFeePremium from the class MortgageCashFlow
}
