% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBSDetails.R
\name{GFeePremium}
\alias{GFeePremium}
\title{A standard generic function to access the slot of GFeePremium}
\usage{
GFeePremium(object)
}
\arguments{
\item{object}{an S4 class object of the type MortgageCashFlow}
}
\description{
A standard generic function to access the slot of GFeePremium
}
