% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFeeTypes.R
\docType{class}
\name{GFeeTypes-class}
\alias{GFeeTypes-class}
\title{An S4 class representing GFee (guarantee fee) GFeeTypes}
\description{
This class is used to calculate and pass along the guarantee fee paid 
by the borrower or turst to guarantee agency pooling the MBS.  For example,
GFee is quoted as numeric decimal or decimal string but basis is used in 
cashflow calcualtions to determine the amount paid to guarantor
}
\section{Slots}{

\describe{
\item{\code{GFeeDecimal}}{a numeric value GFee expressed as numeric decimal}

\item{\code{GFeeBasis}}{a numeric value GFee expressed as basis in numeric decimal 
form}

\item{\code{GFeeDecimalString}}{a character string the GFee expressed as numeric 
decimal}
}}

