% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFeeTypes.R
\name{GFeeTypes}
\alias{GFeeTypes}
\title{GFeeTypes is a constructor function for the class GFeeTypes}
\usage{
GFeeTypes(GFee = numeric())
}
\arguments{
\item{GFee}{a numeric value the GFee reported to the investor and reported
as a numeric decimal}
}
\description{
GFeeTypes is a constructor function for the class GFeeTypes
}
