% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{Group,Collateral-method}
\alias{Group,Collateral-method}
\title{A method to access the GroupNumber from S4 class Collateral}
\usage{
\S4method{Group}{Collateral}(object)
}
\arguments{
\item{object}{An S4 class object of the type Collateral}
}
\description{
A method to access the GroupNumber from S4 class Collateral
}

