% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{Group,TrancheDetails-method}
\alias{Group,TrancheDetails-method}
\title{A method to access the slot Group from class TrancheDetails}
\usage{
\S4method{Group}{TrancheDetails}(object)
}
\arguments{
\item{object}{an S4 object of the type TrancheDetails}
}
\description{
A method to access the slot Group from class TrancheDetails
}

