% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\name{Group}
\alias{Group}
\title{A generic function to access the slot Group}
\usage{
Group(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A generic function to access the slot Group
}

