% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBSDetails.R
\name{GWac,MBSDetails-method}
\alias{GWac,MBSDetails-method}
\title{Method to extract the slot GWac from the class MBSDetails}
\usage{
\S4method{GWac}{MBSDetails}(object)
}
\arguments{
\item{object}{the name of the object of the type MBSDetails}
}
\description{
Method to extract the slot GWac from the class MBSDetails
}
