% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopyOfGWacTypes.R
\docType{methods}
\name{GWacBasis<-,GWacTypes-method}
\alias{GWacBasis<-,GWacTypes-method}
\title{A method to replace the value of GWacBasis slot in the object GWacTypes}
\usage{
\S4method{GWacBasis}{GWacTypes}(object) <- value
}
\arguments{
\item{object}{an S4 object of the type GWacTypes}

\item{value}{the replacement value of the slot}
}
\description{
A method to replace the value of GWacBasis slot in the object GWacTypes
}

