% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWACTypes.R
\name{GWacBasis<-}
\alias{GWacBasis<-}
\title{A standard generic function to replace the slot GwacBasis}
\usage{
GWacBasis(object) <- value

GWacBasis(object) <- value
}
\arguments{
\item{object}{an S4 class object of the type GWacTypes  #'}

\item{value}{the replacement value of the slot}

\item{object}{an S4 class object of the type GWacTypes  #'}

\item{value}{the replacement value of the slot}
}
\description{
A standard generic function to replace the slot GwacBasis

A standard generic function to replace the slot GwacBasis
}

