% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopyOfGWacTypes.R
\name{GWacBasis}
\alias{GWacBasis}
\title{A standard generic function to access the slot GWacBasis}
\usage{
GWacBasis(object)
}
\arguments{
\item{object}{an S4 class object of the type GWacTypes}
}
\description{
A standard generic function to access the slot GWacBasis
}

