% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWacTypes.R
\name{GWacDecimal<-,GWacTypes-method}
\alias{GWacDecimal<-,GWacTypes-method}
\title{A method to set the vlaue of GWacDecimal slot in the object GWacTypes}
\usage{
\S4method{GWacDecimal}{GWacTypes}(object) <- value
}
\arguments{
\item{object}{an S4 object of the type GWacTypes}

\item{value}{the replacement value of the slot}
}
\description{
A method to set the vlaue of GWacDecimal slot in the object GWacTypes
}
