% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopyOfGWacTypes.R
\name{GWacDecimal<-}
\alias{GWacDecimal<-}
\title{A standard generic function to replace the value of slot GWacDecimal}
\usage{
GWacDecimal(object) <- value
}
\arguments{
\item{object}{an S4 class object of the type GWacTypes  #'}

\item{value}{the replacement value of the slot}
}
\description{
A standard generic function to replace the value of slot GWacDecimal
}

