% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWACTypes.R
\name{GWacDecimal}
\alias{GWacDecimal}
\title{A standard generic function to access the slot GWacDecimal}
\usage{
GWacDecimal(object)
}
\arguments{
\item{object}{an S4 class object of the type GWACTypes}
}
\description{
A standard generic function to access the slot GWacDecimal
}

