% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWACTypes.R
\docType{methods}
\name{GWacDecimalString,GWacTypes-method}
\alias{GWacDecimalString,GWacTypes-method}
\title{A method to extract the value of the slot GWacDecimalString slot in the
object GWacTypes}
\usage{
\S4method{GWacDecimalString}{GWacTypes}(object)
}
\arguments{
\item{object}{an S4 object of the type GwacTypes}
}
\description{
A method to extract the value of the slot GWacDecimalString slot in the
object GWacTypes
}

