% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWacTypes.R
\docType{methods}
\name{GWacDecimalString<-,GWacTypes-method}
\alias{GWacDecimalString<-,GWacTypes-method}
\title{A method to replace the value of GWacDecimalString slot in the 
object GWacTypes}
\usage{
\S4method{GWacDecimalString}{GWacTypes}(object) <- value
}
\arguments{
\item{object}{an S4 object of the type GWacTypes}

\item{value}{the replacement value of the slot}
}
\description{
A method to replace the value of GWacDecimalString slot in the 
object GWacTypes
}

