% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWACTypes.R
\name{GWacDecimalString<-}
\alias{GWacDecimalString<-}
\title{A standard generic function to replace the slot GWacDecimalString}
\usage{
GWacDecimalString(object) <- value

GWacDecimalString(object) <- value
}
\arguments{
\item{object}{an S4 object of the type GWacDecimalBasis  #'}

\item{value}{the replacement value of the slot}

\item{object}{an S4 object of the type GWacDecimalBasis  #'}

\item{value}{the replacement value of the slot}
}
\description{
A standard generic function to replace the slot GWacDecimalString

A standard generic function to replace the slot GWacDecimalString
}

