% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWacTypes.R
\name{GWacDecimalString}
\alias{GWacDecimalString}
\title{A standard generic function to get the slot GWacDecimalString}
\usage{
GWacDecimalString(object)
}
\arguments{
\item{object}{an S4 object of the type GWacTypes}
}
\description{
A standard generic function to get the slot GWacDecimalString
}

