% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWACTypes.R
\docType{class}
\name{GWacTypes-class}
\alias{GWacTypes-class}
\title{An S4 class representing the borrower Gwac borrower (note rate)}
\description{
This class is used to create and pass GWac (note rate) Types reported
to the investor and used in the cash flow analytics.  For example, GWac
is reported to investors as numeric decimal or numeric decimal string but
the basis is used to determine the interest paid to the trust.
}
\section{Slots}{

\describe{
\item{\code{GWacDecimal}}{A numeric value the GWac (borrower note rate)
expressed as numeric decimal}

\item{\code{GWacBasis}}{A numeric value the GWac expressed in basis notation e.g.
0.055}

\item{\code{GWacDecimalString}}{A character value the GWac (borrower note rate)
expressed as numeric decimal notation e.g. "5.50}
}}

