% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWACTypes.R
\name{GWacTypes}
\alias{GWacTypes}
\title{GWacTypes is a constructor function for the GwacTypes class}
\usage{
GWacTypes(GWac = numeric())

GWacTypes(GWac = numeric())
}
\arguments{
\item{GWac}{a numeric value the Gross WAC reported to the investor.  This 
may also be the note rate in the case of a loan}

\item{GWac}{a numeric value the Gross WAC reported to the investor.  This 
may also be the note rate in the case of a loan}
}
\description{
The GWacTypes class converts Gross Wac a decimal numeric value to a 
GWacBasis a numeric value and GWacDecimalString a character string

The GWacTypes class converts Gross Wac a decimal numeric value to a 
GWacBasis a numeric value and GWacDecimalString a character string
}

