% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\docType{methods}
\name{Horizon,Mtg.Scenario-method}
\alias{Horizon,Mtg.Scenario-method}
\title{A method to extract Horizon (horizon period) from an S4 class Mtg.Scenario}
\usage{
\S4method{Horizon}{Mtg.Scenario}(object)
}
\arguments{
\item{object}{the name of an S4 class of type of Mtg.Scenario}
}
\description{
A method to extract Horizon (horizon period) from an S4 class Mtg.Scenario
}

