% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\docType{methods}
\name{Horizon,MtgScenario-method}
\alias{Horizon,MtgScenario-method}
\title{A method to extract Horizon (horizon period) from an S4 class MtgScenario}
\usage{
\S4method{Horizon}{MtgScenario}(object)
}
\arguments{
\item{object}{the name of an S4 class of type of MtgScenario}
}
\description{
A method to extract Horizon (horizon period) from an S4 class MtgScenario
}

