% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScenarioConstructor.R
\docType{methods}
\name{Horizon,Scenario-method}
\alias{Horizon,Scenario-method}
\title{A Method to extract scenario Horizon from an S4 class Scenario}
\usage{
\S4method{Horizon}{Scenario}(object)
}
\arguments{
\item{object}{The name of the object of the S4 class of type Scenario}
}
\description{
A Method to extract scenario Horizon from an S4 class Scenario
}

