% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\name{Horizon}
\alias{Horizon}
\title{A standard generic to access the slot Horizon}
\usage{
Horizon(object)
}
\arguments{
\item{object}{An S4 class object of the type Scenario}
}
\description{
A standard generic to access the slot Horizon
}

