% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectionFunctions.R
\name{HorizonBond}
\alias{HorizonBond}
\title{A connection function to the horizon bond object}
\usage{
HorizonBond()
}
\description{
A read connection to the horizon bond details
No inputs are required
}
\examples{
\dontrun{
HorizonBond()}
}

