% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\name{HorizonCurrBal}
\alias{HorizonCurrBal}
\title{HorizonCurrBal generic}
\usage{
HorizonCurrBal(object)
}
\arguments{
\item{object}{An object of the type MtgScenario}
}
\description{
A generic function for method dispatch
}
\seealso{
Other Scenario Analysis: \code{\link{CouponIncome}},
  \code{\link{HorizonCurveShift}},
  \code{\link{HorizonMos}}, \code{\link{HorizonPrice}},
  \code{\link{HorizonReturn}},
  \code{\link{Name,Scenario-method}},
  \code{\link{PrepaidPrinReceived}},
  \code{\link{PrincipalReceived}},
  \code{\link{ReinvestmentIncome}},
  \code{\link{Scenario-class}},
  \code{\link{ScenarioCurve-class}},
  \code{\link{ScheduledPrinReceived}}
}
