% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScenarioConstructor.R
\docType{methods}
\name{HorizonCurve,ScenarioCurve-method}
\alias{HorizonCurve,ScenarioCurve-method}
\title{Horizon Curve the Scenario Horizon Curve}
\usage{
\S4method{HorizonCurve}{ScenarioCurve}(object)
}
\arguments{
\item{object}{The name of the object of type Horizon Curve}
}
\description{
a method to get the \strong{Horizon Curve} from the HorizonCurve object
}
\seealso{
Other Scenario: \code{\link{HorizonTermStrc,ScenarioCurve-method}},
  \code{\link{StartCurve,ScenarioCurve-method}},
  \code{\link{StartTermStrc,ScenarioCurve-method}}
}
