% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScenarioConstructor.R
\name{HorizonCurve}
\alias{HorizonCurve}
\title{A standard generic function to access the HorizonCurve slot HorizonCurve}
\usage{
HorizonCurve(object)
}
\arguments{
\item{object}{An S4 class object of the type HorizonCurve}
}
\description{
A standard generic function to access the HorizonCurve slot HorizonCurve
}

