% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScenarioConstructor.R
\name{HorizonCurveShift}
\alias{HorizonCurveShift}
\title{a constructor function of the class HorizonCurve}
\usage{
HorizonCurveShift(
  rates.data,
  settlement.date,
  horizon.months = 12,
  scenario = "NC",
  method = "dl"
)
}
\arguments{
\item{rates.data}{A character string referencing a rates object}

\item{settlement.date}{A character string the settlement date 'mm-dd-YYYY'}

\item{horizon.months}{A numeric value the horizon in months}

\item{scenario}{A character string the scenario}

\item{method}{A character string indicating the fitting method ns = Nelson
Siegel, dl = Diebond Lee, sv = Severson, asv = adjusted Severson, 
cs = cubic spline (not yet implemented).  For additional details see the termstrc
documentation.}
}
\description{
a function to construct the HorizonCurve class.  Currently 
supported scenarios
\itemize{
\item{D300 - parallel down 300 basis points}
\item{D275 - parallel down 275 basis points}
\item{D250 - parallel down 250 basis points}
\item{D225 - parallel down 225 basis points}
\item{D200 - parallel down 200 basis points}
\item{D175 - parallel down 175 basis points}
\item{D150 - parallel down 150 basis points}
\item{D125 - parallel down 125 basis points}
\item{D100 - parallel down 100 basis points}
\item{D75 - parallel down 75 basis points}
\item{D50 - parallel down 50 basis points}
\item{D25 - parallel down 25 basis points}
\item{NC - No Change}
\item{U25 - parallel up 25 basis points}
\item{U50 - parallel up 50 basis points}
\item{U75 - parallel up 75 basis points}
\item{U100 - parallel up 100 basis points}
\item{U125 - parallel up 125 basis points}
\item{U150 - parallel up 150 basis points}
\item{U175 - parallel up 175 basis points}
\item{U200 - parallel up 200 basis points}
\item{U225 - parallel up 225 basis points}
\item{U250 - parallel up 250 basis points}
\item{U275 - parallel up 275 basis points}
\item{U300 - parallel up 300 basis points}
}
}
\seealso{
Other Scenario Analysis: 
\code{\link{CouponIncome}()},
\code{\link{HorizonCurrBal}()},
\code{\link{HorizonMos}()},
\code{\link{HorizonPrice}()},
\code{\link{HorizonReturn}()},
\code{\link{Name,Scenario-method}},
\code{\link{PrepaidPrinReceived}()},
\code{\link{PrincipalReceived}()},
\code{\link{ReinvestmentIncome}()},
\code{\link{Scenario-class}},
\code{\link{ScenarioCurve-class}},
\code{\link{ScheduledPrinReceived}()}
}
\concept{Scenario Analysis}
