% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectionFunctions.R
\name{HorizonMBS}
\alias{HorizonMBS}
\title{A connection function to the horzon MBS pass-though object}
\usage{
HorizonMBS()
}
\description{
A read connection function to the horizon MBS details.  
No inputs are required
}
\examples{
\dontrun{
HorizonMBS()}
}

