% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\name{HorizonMos}
\alias{HorizonMos}
\title{A standard generic function to access the slot HorizonMos}
\usage{
HorizonMos(object)
}
\arguments{
\item{object}{An S4 object of type MtgScenario}
}
\description{
A standard generic function to access the slot HorizonMos
}

