% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondScenario.R
\docType{methods}
\name{HorizonPrice,BondReturn-method}
\alias{HorizonPrice,BondReturn-method}
\title{A method to extract HorizonPrice from S4 class BondReturn}
\usage{
\S4method{HorizonPrice}{BondReturn}(object)
}
\arguments{
\item{object}{the name of an S4 class of tyoe BondReturn}
}
\description{
A method to extract HorizonPrice from S4 class BondReturn
}

