% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\docType{methods}
\name{HorizonPrice,MortgageReturn-method}
\alias{HorizonPrice,MortgageReturn-method}
\title{A method to extract HorizonPrice from S4 class MortgageReturn}
\usage{
\S4method{HorizonPrice}{MortgageReturn}(object)
}
\arguments{
\item{object}{the name of an S4 class of tyoe MortgageReturn}
}
\description{
A method to extract HorizonPrice from S4 class MortgageReturn
}

