% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\docType{methods}
\name{HorizonPrice,MtgScenario-method}
\alias{HorizonPrice,MtgScenario-method}
\title{A method to extract HorizonPrice from S4 class MtgScenario}
\usage{
\S4method{HorizonPrice}{MtgScenario}(object)
}
\arguments{
\item{object}{the name of an S4 class of tyoe MtgScenario}
}
\description{
A method to extract HorizonPrice from S4 class MtgScenario
}

