% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\name{HorizonPrice}
\alias{HorizonPrice}
\title{A standard generic function ot access the slot HorizonPrice}
\usage{
HorizonPrice(object)
}
\arguments{
\item{object}{an S4 object of type MtgScenario}
}
\description{
A standard generic function ot access the slot HorizonPrice
}

