% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\docType{methods}
\name{HorizonReturn,MortgageReturn-method}
\alias{HorizonReturn,MortgageReturn-method}
\title{A method to extract HorizonReturn from S4 class MtgScenario}
\usage{
\S4method{HorizonReturn}{MortgageReturn}(object)
}
\arguments{
\item{object}{the name of an S4 class of type MtgScenario}
}
\description{
A method to extract HorizonReturn from S4 class MtgScenario
}

