% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\docType{methods}
\name{HorizonReturn,Mtg.Scenario-method}
\alias{HorizonReturn,Mtg.Scenario-method}
\title{A method to extract HorizonReturn from S4 class Mtg.Scenario}
\usage{
\S4method{HorizonReturn}{Mtg.Scenario}(object)
}
\arguments{
\item{object}{the name of the an S4 class of type Mtg.Scenario}
}
\description{
A method to extract HorizonReturn from S4 class Mtg.Scenario
}

