% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScenarioConstructor.R
\name{HorizonTermStrc,ScenarioCurve-method}
\alias{HorizonTermStrc,ScenarioCurve-method}
\title{HorizonTermStruc the Scenario Horizon Term Structure Assumption}
\usage{
\S4method{HorizonTermStrc}{ScenarioCurve}(object)
}
\arguments{
\item{object}{The name of the object of the type Horizon Curve}
}
\description{
a method to get the \strong{Horizon Term Structure} from the Horizon Curve object
}
\seealso{
Other Scenario: 
\code{\link{HorizonCurve,ScenarioCurve-method}},
\code{\link{StartCurve,ScenarioCurve-method}},
\code{\link{StartTermStrc,ScenarioCurve-method}}
}
\concept{Scenario}
