% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScenarioConstructor.R
\name{HorizonTermStrc}
\alias{HorizonTermStrc}
\title{A standard generic function to access the HorizonCurve slot HorizonTermStruc}
\usage{
HorizonTermStrc(object)
}
\arguments{
\item{object}{An S4 class object of the HorizonCurve}
}
\description{
A standard generic function to access the HorizonCurve slot HorizonTermStruc
}
