% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/HomePriceSim.R
\name{HPISim}
\alias{HPISim}
\title{A function to simulate home prices given a short-term rate}
\usage{
HPISim(shortrate = numeric(), LongTermGrowth = numeric(), T = numeric(),
  step = numeric(), sigma = numeric(), N = numeric())
}
\arguments{
\item{shortrate}{A numeric value the short term forward rate given by the interest rate generation process}

\item{LongTermGrowth}{A numeric value the expected long term home price growth potential}

\item{T}{A numeric value the time in years}

\item{step}{A numeric value the time step interval}

\item{sigma}{A numeric value the volatility of home price}

\item{N}{A numeric value the number of paths}
}
\description{
Simulates home prices as a function of the short-term rate and the long term growth of home prices
}

