% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HullWhite.R
\name{HullWhite}
\alias{HullWhite}
\title{The single factor Hull White Model used to simulate the short rate}
\usage{
HullWhite(short.rate, mean.reversion, theta, time, delta.time, volatility,
  trade.days, num.paths)
}
\arguments{
\item{short.rate}{The short term rate}

\item{mean.reversion}{The rate of mean reversion}

\item{theta}{The paramter to fit to the curve}

\item{time}{The time period in years}

\item{delta.time}{The time step}

\item{volatility}{The market volatility stated in annual terms}

\item{trade.days}{The number of trading days}

\item{num.paths}{Then number of short rate paths}
}
\description{
A function to simulate the short rate using a single factor 
Hull White model.
}
\seealso{
Other Interest Rate Models: \code{\link{Bterm}},
  \code{\link{CreateCashFlowMatrix}},
  \code{\link{CreateMaturitiesMatrix}}
}
