% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BillDetails.R
\docType{methods}
\name{ID,BillDetails-method}
\alias{ID,BillDetails-method}
\title{Method to extract ID from an S4 class of type BillDetail}
\usage{
\S4method{ID}{BillDetails}(object)
}
\arguments{
\item{object}{The name of the S4 object of type BillDetails}
}
\description{
Method to extract ID from an S4 class of type BillDetail
}
