% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondDetails.R
\docType{methods}
\name{ID,BondDetails-method}
\alias{ID,BondDetails-method}
\title{Method to extract ID from an S4 class of type BondDetails}
\usage{
\S4method{ID}{BondDetails}(object)
}
\arguments{
\item{object}{The name of the S4 object BondDetails}
}
\description{
Method to extract ID from an S4 class of type BondDetails
}
