% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModel.R
\docType{methods}
\name{Incentive,PrepaymentModel-method}
\alias{Incentive,PrepaymentModel-method}
\title{A method to extract Incentive from S4 class PrepaymentModel}
\usage{
\S4method{Incentive}{PrepaymentModel}(object)
}
\arguments{
\item{object}{the name of the S4 class}
}
\description{
A method to extract Incentive from S4 class PrepaymentModel
}
