% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelTune.R
\docType{methods}
\name{IncentiveFastEta,PrepaymentModelTune-method}
\alias{IncentiveFastEta,PrepaymentModelTune-method}
\title{A method to extract Incentive.Fast.beta from class PrepaymentModelTune}
\usage{
\S4method{IncentiveFastEta}{PrepaymentModelTune}(object)
}
\arguments{
\item{object}{The name of the object of class PrepaymentModelTune}
}
\description{
A method to extract Incentive.Fast.beta from class PrepaymentModelTune
}

