% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelTune.R
\name{IncentiveFastEta}
\alias{IncentiveFastEta}
\title{A standard generic function to access the slot Incentive.Fast.eta}
\usage{
IncentiveFastEta(object)
}
\arguments{
\item{object}{an S4 object}
}
\description{
A standard generic function to access the slot Incentive.Fast.eta
}
