% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelTune.R
\name{IncentiveFastThetaOne}
\alias{IncentiveFastThetaOne}
\title{A standard generic function to access the slot Incentive.Fast.theta.1}
\usage{
IncentiveFastThetaOne(object)
}
\arguments{
\item{object}{an S4 object}
}
\description{
A standard generic function to access the slot Incentive.Fast.theta.1
}
