% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelTune.R
\docType{methods}
\name{IncentiveFastThetaTwo,PrepaymentModelTune-method}
\alias{IncentiveFastThetaTwo,PrepaymentModelTune-method}
\title{A method to extract Incentive.Fast.theta.2 from class PrepaymentModelTune}
\usage{
\S4method{IncentiveFastThetaTwo}{PrepaymentModelTune}(object)
}
\arguments{
\item{object}{The name of the object of class PrepaymentModelTune}
}
\description{
A method to extract Incentive.Fast.theta.2 from class PrepaymentModelTune
}
