% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelTune.R
\name{IncentiveFastThetaTwo}
\alias{IncentiveFastThetaTwo}
\title{A standard generic function to access the slot Incentive.Fast.theta.2}
\usage{
IncentiveFastThetaTwo(object)
}
\arguments{
\item{object}{an S4 object}
}
\description{
A standard generic function to access the slot Incentive.Fast.theta.2
}
