% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelTune.R
\docType{methods}
\name{IncentiveSlowBeta,PrepaymentModelTune-method}
\alias{IncentiveSlowBeta,PrepaymentModelTune-method}
\title{A method to extract Incentive.Slow.beta from class PrepaymentModelTune}
\usage{
\S4method{IncentiveSlowBeta}{PrepaymentModelTune}(object)
}
\arguments{
\item{object}{The name of the object of class PrepaymentModelTune}
}
\description{
A method to extract Incentive.Slow.beta from class PrepaymentModelTune
}
