% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelTune.R
\docType{methods}
\name{IncentiveSlowEta,PrepaymentModelTune-method}
\alias{IncentiveSlowEta,PrepaymentModelTune-method}
\title{A method to extract Incentive.Slow.eta from class PrepaymentModelTune}
\usage{
\S4method{IncentiveSlowEta}{PrepaymentModelTune}(object)
}
\arguments{
\item{object}{The name of the object of class PrepaymentModelTune}
}
\description{
A method to extract Incentive.Slow.eta from class PrepaymentModelTune
}
