% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelTune.R
\name{IncentiveSlowEta}
\alias{IncentiveSlowEta}
\title{A standard generic function to access the slot Incentive.Slow.beta}
\usage{
IncentiveSlowEta(object)
}
\arguments{
\item{object}{an S4 object}
}
\description{
A standard generic function to access the slot Incentive.Slow.beta
}
