% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelTune.R
\docType{methods}
\name{IncentiveSlowThetaTwo,PrepaymentModelTune-method}
\alias{IncentiveSlowThetaTwo,PrepaymentModelTune-method}
\title{A method to extract Incentive.Slow.theta.2 from class PrepaymentModelTune}
\usage{
\S4method{IncentiveSlowThetaTwo}{PrepaymentModelTune}(object)
}
\arguments{
\item{object}{The name of the object of class PrepaymentModelTune}
}
\description{
A method to extract Incentive.Slow.theta.2 from class PrepaymentModelTune
}
