% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelTune.R
\name{IncentiveSlowThetaTwo}
\alias{IncentiveSlowThetaTwo}
\title{A standard generic function to access the slot Incentive.Slow.theta.2}
\usage{
IncentiveSlowThetaTwo(object)
}
\arguments{
\item{object}{an S4 object}
}
\description{
A standard generic function to access the slot Incentive.Slow.theta.2
}
