% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBSDetails.R
\name{Index,MBSDetails-method}
\alias{Index,MBSDetails-method}
\title{Method to extract the slot Index from the class MBS Details}
\usage{
\S4method{Index}{MBSDetails}(object)
}
\arguments{
\item{object}{the name of the object of type MBSDetails}
}
\description{
Method to extract the slot Index from the class MBS Details
}
