% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{InitialIndexValue,TrancheDetails-method}
\alias{InitialIndexValue,TrancheDetails-method}
\title{A method to access the slot InitialIndexValue from class TrancheDetails}
\usage{
\S4method{InitialIndexValue}{TrancheDetails}(object)
}
\arguments{
\item{object}{an S4 object of the type TrancheDetails}
}
\description{
A method to access the slot InitialIndexValue from class TrancheDetails
}

