% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\name{InitialIndexValue}
\alias{InitialIndexValue}
\title{A generic functon to access the slot InitialIndexValue}
\usage{
InitialIndexValue(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A generic functon to access the slot InitialIndexValue
}

