% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{InitialReserveFund,RAID-method}
\alias{InitialReserveFund,RAID-method}
\title{A method to extract slot InitialReserveFund from class RAID}
\usage{
\S4method{InitialReserveFund}{RAID}(object)
}
\arguments{
\item{object}{an S4 class of the type RAID}
}
\description{
A method to extract slot InitialReserveFund from class RAID
}

