% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\name{InitialReserveFund}
\alias{InitialReserveFund}
\title{A generic function to access the slot InitialReserveFund}
\usage{
InitialReserveFund(object)
}
\arguments{
\item{object}{an S4 object}
}
\description{
A generic function to access the slot InitialReserveFund
}

