% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBSDetails.R
\name{InterestOnlyPeriod}
\alias{InterestOnlyPeriod}
\title{A standard generic function to access the slot InterestOnlyPeriod}
\usage{
InterestOnlyPeriod(object)
}
\arguments{
\item{object}{S4 class object}
}
\description{
A standard generic function to access the slot InterestOnlyPeriod
}
