% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{InterestPmtFrequency,TrancheDetails-method}
\alias{InterestPmtFrequency,TrancheDetails-method}
\title{A method to access the slot InterestPmtFrequency from class TrancheDetails}
\usage{
\S4method{InterestPmtFrequency}{TrancheDetails}(object)
}
\arguments{
\item{object}{an S4 object of the type InterestPmtFrequency}
}
\description{
A method to access the slot InterestPmtFrequency from class TrancheDetails
}

