% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\name{InterestPmtFrequency}
\alias{InterestPmtFrequency}
\title{A generic functon to access the slot InterestPmtFrequency}
\usage{
InterestPmtFrequency(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A generic functon to access the slot InterestPmtFrequency
}

