% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondBasis.R
\name{is.leapyear}
\alias{is.leapyear}
\title{Is leap year}
\usage{
is.leapyear(year)
}
\arguments{
\item{year}{The year to test as leap year}
}
\description{
Tests if year is a leap year and reurns TRUE or FALSE
}
\seealso{
Other BondBasis: \code{\link{ActualFactor}},
  \code{\link{BondBasisConversion}}
}

