% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BillDetails.R
\name{IssueDate,BillDetails-method}
\alias{IssueDate,BillDetails-method}
\title{Method to extract IssueDate from an S4 class of type BillDetails}
\usage{
\S4method{IssueDate}{BillDetails}(object)
}
\arguments{
\item{object}{The name of the S4 class of type BillDetails}
}
\description{
Method to extract IssueDate from an S4 class of type BillDetails
}
