% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondDetails.R
\name{IssueDate,BondDetails-method}
\alias{IssueDate,BondDetails-method}
\title{Method to extract IssueDate from an S4 class of type BondDetails}
\usage{
\S4method{IssueDate}{BondDetails}(object)
}
\arguments{
\item{object}{the name of the S4 class of type BondDetails}
}
\description{
Method to extract IssueDate from an S4 class of type BondDetails
}
