% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBSDetails.R
\docType{methods}
\name{IssueDate,MBSDetails-method}
\alias{IssueDate,MBSDetails-method}
\title{Method to extract the slot IssueDate from the class MBSDetails}
\usage{
\S4method{IssueDate}{MBSDetails}(object)
}
\arguments{
\item{object}{the name of the object of type MBSDetails}
}
\description{
Method to extract the slot IssueDate from the class MBSDetails
}

