% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBSDetails.R
\name{IssueDate}
\alias{IssueDate}
\title{A standard generic function to access the slot IssueDate}
\usage{
IssueDate(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A standard generic function to access the slot IssueDate
}

