% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BillDetails.R
\docType{methods}
\name{IssueDiscountRate,BillDetails-method}
\alias{IssueDiscountRate,BillDetails-method}
\title{Method to extract IssueDiscountRate from an S4 class of type BillDetails}
\usage{
\S4method{IssueDiscountRate}{BillDetails}(object)
}
\arguments{
\item{object}{The name of the S4 class of type BillDetails}
}
\description{
Method to extract IssueDiscountRate from an S4 class of type BillDetails
}
