% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BillDetails.R
\name{IssueDiscountRate}
\alias{IssueDiscountRate}
\title{A standard generic function to get the IssueDiscount Rate}
\usage{
IssueDiscountRate(object)
}
\arguments{
\item{object}{An S4 class object}
}
\description{
A standard generic function to get the IssueDiscount Rate
}
