% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BillDetails.R
\docType{methods}
\name{Issuer,BillDetails-method}
\alias{Issuer,BillDetails-method}
\title{Method to extract BillType from an S4 class of type BillDetails}
\usage{
\S4method{Issuer}{BillDetails}(object)
}
\arguments{
\item{object}{The name of the S4 class of type BIllDetails}
}
\description{
Method to extract BillType from an S4 class of type BillDetails
}

