% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondDetails.R
\name{Issuer,BondDetails-method}
\alias{Issuer,BondDetails-method}
\title{Method to extract Issuer from an S4 class of type BondDetails}
\usage{
\S4method{Issuer}{BondDetails}(object)
}
\arguments{
\item{object}{The name of the S4 object of type BondDetails}
}
\description{
Method to extract Issuer from an S4 class of type BondDetails
}
