% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBSDetails.R
\name{Issuer,MBSDetails-method}
\alias{Issuer,MBSDetails-method}
\title{Method to extreact the slot Issuer from the class MBSDetails}
\usage{
\S4method{Issuer}{MBSDetails}(object)
}
\arguments{
\item{object}{the name of the object of type MBSDetails}
}
\description{
Method to extreact the slot Issuer from the class MBSDetails
}
