% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageOAS.R
\docType{methods}
\name{Issuer,MortgageOAS-method}
\alias{Issuer,MortgageOAS-method}
\title{A method to access the slot Issuer}
\usage{
\S4method{Issuer}{MortgageOAS}(object)
}
\arguments{
\item{object}{MortgageOAS object}
}
\description{
Access the slot Issuer from object MortgageOAS
}
