% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageKeyRate.R
\docType{methods}
\name{Issuer,MortgageTermStructure-method}
\alias{Issuer,MortgageTermStructure-method}
\title{Issuer}
\usage{
\S4method{Issuer}{MortgageTermStructure}(object)
}
\arguments{
\item{object}{MortgageTermStructure}
}
\description{
Get Issuer from object of type MortgageTermStrucutre
}
