% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\name{JumpReferenceSpeed1}
\alias{JumpReferenceSpeed1}
\title{A generic function to access the slot JumpReferenceSpeed1}
\usage{
JumpReferenceSpeed1(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A generic function to access the slot JumpReferenceSpeed1
}

