% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{JumpReferenceSpeed2,RAID-method}
\alias{JumpReferenceSpeed2,RAID-method}
\title{A method to access the slot JumpReferenceSpeed2 from class RAID}
\usage{
\S4method{JumpReferenceSpeed2}{RAID}(object)
}
\arguments{
\item{object}{an S4 object of the type RAID}
}
\description{
A method to access the slot JumpReferenceSpeed2 from class RAID
}

