% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\name{JumpReferenceSpeed2}
\alias{JumpReferenceSpeed2}
\title{A generic function to access the slot JumpReferenceSpeed2}
\usage{
JumpReferenceSpeed2(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A generic function to access the slot JumpReferenceSpeed2
}

