% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\name{JumpReferenceSpeed3}
\alias{JumpReferenceSpeed3}
\title{A generic function to access the slot JumpReferenceSpeed3}
\usage{
JumpReferenceSpeed3(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A generic function to access the slot JumpReferenceSpeed3
}

