% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{JumpReferenceSpeed4,RAID-method}
\alias{JumpReferenceSpeed4,RAID-method}
\title{A method to access the slot JumpReferenceSpeed4 from class RAID}
\usage{
\S4method{JumpReferenceSpeed4}{RAID}(object)
}
\arguments{
\item{object}{an S4 class of type RAID}
}
\description{
A method to access the slot JumpReferenceSpeed4 from class RAID
}

