% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\name{JumpReferenceSpeed4}
\alias{JumpReferenceSpeed4}
\title{A generic function to access the slot JumpReferenceSpeed4}
\usage{
JumpReferenceSpeed4(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A generic function to access the slot JumpReferenceSpeed4
}

