% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{JumpReferenceSpeed5,RAID-method}
\alias{JumpReferenceSpeed5,RAID-method}
\title{A method to access the slot JumpReferenceSpeed5 from class RAID}
\usage{
\S4method{JumpReferenceSpeed5}{RAID}(object)
}
\arguments{
\item{object}{an S4 class of typre RAID}
}
\description{
A method to access the slot JumpReferenceSpeed5 from class RAID
}

