% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\name{JumpReferenceSpeed5}
\alias{JumpReferenceSpeed5}
\title{generic function to access the slot JumpReferenceSpeed5}
\usage{
JumpReferenceSpeed5(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
generic function to access the slot JumpReferenceSpeed5
}

