% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondTermStructure.R
\docType{methods}
\name{KeyRateConvexity,BondTermStructure-method}
\alias{KeyRateConvexity,BondTermStructure-method}
\title{A method to get KeyRateConvexity from the class BondTermStructure}
\usage{
\S4method{KeyRateConvexity}{BondTermStructure}(object)
}
\arguments{
\item{object}{an object of the type BondTermStructure}
}
\description{
A method to get KeyRateConvexity from the class BondTermStructure
}
