% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageKeyRate.R
\name{KeyRateConvexity,MortgageTermStructure-method}
\alias{KeyRateConvexity,MortgageTermStructure-method}
\title{Method to extract the KeyRateConvexity from S4 class}
\usage{
\S4method{KeyRateConvexity}{MortgageTermStructure}(object)
}
\arguments{
\item{object}{The name of the S4 object of type MortgageTermStructure}
}
\description{
Method to extract the KeyRateConvexity from S4 class
}
