% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondTermStructure.R
\docType{methods}
\name{KeyRateDuration,BondTermStructure-method}
\alias{KeyRateDuration,BondTermStructure-method}
\title{A method to get KeyRateDuration from the class BondTermStructure}
\usage{
\S4method{KeyRateDuration}{BondTermStructure}(object)
}
\arguments{
\item{object}{an object of the type BondTermStructure}
}
\description{
A method to get KeyRateDuration from the class BondTermStructure
}
