% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageKeyRate.R
\docType{methods}
\name{KeyRateDuration,MortgageTermStructure-method}
\alias{KeyRateDuration,MortgageTermStructure-method}
\title{Method to extract the KeyRateDuration from S4 class}
\usage{
\S4method{KeyRateDuration}{MortgageTermStructure}(object)
}
\arguments{
\item{object}{The name of the S4 object of type MortgageTermStructure}
}
\description{
Method to extract the KeyRateDuration from S4 class
}
