% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\docType{methods}
\name{KeyRateDuration,Mtg.Scenario-method}
\alias{KeyRateDuration,Mtg.Scenario-method}
\title{A method to extract KeyRateDuration from S4 class Mtg.Scenario}
\usage{
\S4method{KeyRateDuration}{Mtg.Scenario}(object)
}
\arguments{
\item{object}{the name of an S4 class of type Mtg.Scenario}
}
\description{
A method to extract KeyRateDuration from S4 class Mtg.Scenario
}

