% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondTermStructure.R
\docType{methods}
\name{KeyRateTenor,BondTermStructure-method}
\alias{KeyRateTenor,BondTermStructure-method}
\title{A method to get KeyRateTenor from the class BondTermStructure}
\usage{
\S4method{KeyRateTenor}{BondTermStructure}(object)
}
\arguments{
\item{object}{an object of the type BondTermStructure}
}
\description{
A method to get KeyRateTenor from the class BondTermStructure
}

