% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageOAS.R
\docType{methods}
\name{KeyRateTenor,MortgageOAS-method}
\alias{KeyRateTenor,MortgageOAS-method}
\title{A method to extract KeyRateTenor from S4 object MortgageOAS}
\usage{
\S4method{KeyRateTenor}{MortgageOAS}(object)
}
\arguments{
\item{object}{an S4 object of the typre MortgageOAS}
}
\description{
A method to extract KeyRateTenor from S4 object MortgageOAS
}

