% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageKeyRate.R
\docType{methods}
\name{KeyRateTenor,MortgageTermStructure-method}
\alias{KeyRateTenor,MortgageTermStructure-method}
\title{Method to get the KeyRateTenor from S4 class}
\usage{
\S4method{KeyRateTenor}{MortgageTermStructure}(object)
}
\arguments{
\item{object}{The name of the S4 object of type MortgageTermStructure}
}
\description{
Method to get the KeyRateTenor from S4 class
}
