% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaymentDateFunctions.R
\name{LastandNextPmtDate}
\alias{LastandNextPmtDate}
\title{function to determine the last and next payment date of a standard bond}
\usage{
LastandNextPmtDate(
  issue.date,
  dated.date,
  maturity.date,
  settlement.date,
  frequency,
  bond.basis
)
}
\arguments{
\item{issue.date}{A character, the issue date mm-dd-YYYY}

\item{dated.date}{A character, the dated date mm-dd-YYYY}

\item{maturity.date}{A character, the maturity date mm-dd-YYYY}

\item{settlement.date}{A character, the settlement date mm-dd-YYYY}

\item{frequency}{The frequency of interest payments made to the investor
\itemize{
\item Annual = 1
\item Semi Annual = 2
\item Quarterly = 4
\item Monthly = 12}}

\item{bond.basis}{The day count convention used to calculate interest (see BondBasisConversion)
Currently supported bond day count conventions:
\itemize{
\item{'30360' }{Agency MBS}
\item{'30E360' }{European 30360}
\item{'Actual360' }{used in money market}
\item{'Actual365' }{}
\item{'ActualActual' }
}}
}
\description{
The function returns a list the first position is the last payment date and the 
second position is the next payment date
}
\concept{payment dates}
