% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondDetails.R
\name{LastPmtDate,BondDetails-method}
\alias{LastPmtDate,BondDetails-method}
\title{Method to extract LastPmtDate from an S4 class of type BondDetails}
\usage{
\S4method{LastPmtDate}{BondDetails}(object)
}
\arguments{
\item{object}{the name of the S4 class of type BondDetails}
}
\description{
Method to extract LastPmtDate from an S4 class of type BondDetails
}
