% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBSDetails.R
\docType{methods}
\name{LastPmtDate,MBSDetails-method}
\alias{LastPmtDate,MBSDetails-method}
\title{Method to extract the slot LastPmtDate from the class MBSDetails}
\usage{
\S4method{LastPmtDate}{MBSDetails}(object)
}
\arguments{
\item{object}{the name of the object of type MBSDetails}
}
\description{
Method to extract the slot LastPmtDate from the class MBSDetails
}

