% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModel.R
\name{LastPmtDate,PrepaymentModel-method}
\alias{LastPmtDate,PrepaymentModel-method}
\title{A method to extract LastPmtDate from S4 class PrepaymentModel}
\usage{
\S4method{LastPmtDate}{PrepaymentModel}(object)
}
\arguments{
\item{object}{the name of the S4 class}
}
\description{
A method to extract LastPmtDate from S4 class PrepaymentModel
}
