% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondConstructor.R
\docType{methods}
\name{LastPmtDate<-,BondDetails-method}
\alias{LastPmtDate<-,BondDetails-method}
\title{Method to replace the slot LastPmtDate from class BondDetails}
\usage{
\S4method{LastPmtDate}{BondDetails}(object) <- value
}
\arguments{
\item{object}{the name of the S4 object of the type BondDetails}

\item{value}{the value of the replacement}
}
\description{
Method to replace the slot LastPmtDate from class BondDetails
}

