% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBSDetails.R
\name{LastPmtDate<-,MBSDetails-method}
\alias{LastPmtDate<-,MBSDetails-method}
\title{Method to replace slot LastPmtDate from in class MBSDetails}
\usage{
\S4method{LastPmtDate}{MBSDetails}(object) <- value
}
\arguments{
\item{object}{the name of the object of type MBSDetails}

\item{value}{the value of the replacement}
}
\description{
Method to replace slot LastPmtDate from in class MBSDetails
}
