% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PassThroughConstructor.R
\name{LastPmtDate<-}
\alias{LastPmtDate<-}
\title{A standard generic function to replace the slot LastPmtDate}
\usage{
LastPmtDate(object) <- value
}
\arguments{
\item{object}{an S4 class object}

\item{value}{the replacement value of the slot}
}
\description{
A standard generic function to replace the slot LastPmtDate
}

