% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBSDetails.R
\name{LastPmtDate}
\alias{LastPmtDate}
\title{A standard generic function to access the slot LastPmtDate}
\usage{
LastPmtDate(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A standard generic function to access the slot LastPmtDate
}

