% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModel.R
\docType{methods}
\name{LoanAge,PrepaymentModel-method}
\alias{LoanAge,PrepaymentModel-method}
\title{A method to extract LoanAge from S4 class PrepaymentModel}
\usage{
\S4method{LoanAge}{PrepaymentModel}(object)
}
\arguments{
\item{object}{the name of the S4 class}
}
\description{
A method to extract LoanAge from S4 class PrepaymentModel
}

