% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondDetails.R
\name{MakeBondDetails}
\alias{MakeBondDetails}
\title{MakeBondDetails}
\usage{
MakeBondDetails(Cusip, ID, BondType, Sector, Issuer, Underwriter, OfferAmount,
  Coupon, IssueDate, DatedDate, Maturity, LastPmtDate, NextPmtDate, Moody, SP,
  BondLab, Frequency, BondBasis, Callable, Putable, SinkingFund)
}
\arguments{
\item{Cusip}{A character string the bond's cusip number}

\item{ID}{A character string the bond's ID}

\item{BondType}{A character string the type of bond}

\item{Sector}{A character string the bond sector}

\item{Issuer}{A character string the bond issuer}

\item{Underwriter}{A character string the bond underwriter}

\item{OfferAmount}{A numeric value the offer amount of the bond}

\item{Coupon}{A numeric value the coupon (interest rate) 
paid to the investor}

\item{IssueDate}{A character string the issue date of the bond}

\item{DatedDate}{A character string the bond's dated date}

\item{Maturity}{A character string the bond's maturity date of final 
principal payment date}

\item{LastPmtDate}{A character string the most recent payment 
date to the investor}

\item{NextPmtDate}{A character string the next payment date due to 
the investor}

\item{Moody}{A character string the Moody assigned rating}

\item{SP}{A character string the SP assigned rating}

\item{BondLab}{A character string the Bond Lab assigned rating}

\item{Frequency}{A numeric value the payment frequency}

\item{BondBasis}{A character string the interest calculation basis}

\item{Callable}{A character string the bond's call schedule}

\item{Putable}{A caracter string the bond's put schedule}

\item{SinkingFund}{a character strign the bond sinking fund schedule}
}
\description{
Creates standard bond cusip information and
saves to the BondData folder useful for working on a local computer
}
\seealso{
Other BondDetails: \code{\link{BondDetails}}
}

