% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelFunctions.R
\name{MakeModelFunctions}
\alias{MakeModelFunctions}
\title{A constructor function to create the PrepaymentModelFunctions class}
\usage{
MakeModelFunctions()
}
\description{
There are no inputs model functions are coded into the function constructor
}

