% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PrepaymentModelTune.R
\name{MakeModelTune}
\alias{MakeModelTune}
\title{A constructor function to create the prepayment model tuning parameters}
\usage{
MakeModelTune(ModelName = "character", TurnoverRate = "numeric",
  Turnover.alpha = "numeric", Turnover.beta = "numeric",
  Turnover.theta = "numeric", Seasonality.alpha = "numeric",
  Seasonality.theta = "numeric", Incentive.Fast.theta.1 = "numeric",
  Incentive.Fast.theta.2 = "numeric", Incentive.Fast.beta = "numeric",
  Incentive.Fast.eta = "numeric", Incentive.Slow.theta.1 = "numeric",
  Incentive.Slow.theta.2 = "numeric", Incentive.Slow.beta = "numeric",
  Incentive.Slow.eta = "numeric", Burnout.beta.1 = "numeric",
  Burnout.beta.2 = "numeric", BeginCDR = "numeric", PeakCDR = "numeric",
  EndCDR = "numeric", PeakMonth = "numeric", PlateauMonths = "numeric",
  EndMonth = "numeric", MinOrigLTV = "numeric", MaxOrigLTV = "numeric",
  MinOrigMultiplier = "numeric", MaxOrigMultiplier = "numeric",
  UpdatedLTV.beta = "numeric", SATO.beta = "numeric")
}
\arguments{
\item{ModelName}{a character string.  The name of the model.}

\item{TurnoverRate}{a numeric value.  The estimated housing turnover rate.}

\item{Turnover.alpha}{a numeric value. The alpha value of the seasoning curve function.}

\item{Turnover.beta}{a numeric value.  The beta value of the seasoning curve function.}

\item{Turnover.theta}{a numeric value. The theta value of the seasoning curve function.}

\item{Seasonality.alpha}{a numeric value. The alpha value of the seasonality function.}

\item{Seasonality.theta}{a numeric value. The beta value of the seasonality function.}

\item{Incentive.Fast.theta.1}{a numeric value. The first theta value of the fast payer S-curve.}

\item{Incentive.Fast.theta.2}{a numeric value. The second theta value of fast payer S-curve.}

\item{Incentive.Fast.beta}{a numeric value. The beta value of the fast payer S-curve.}

\item{Incentive.Fast.eta}{a numeric value.  The eta value of the fast payer S-curve.}

\item{Incentive.Slow.theta.1}{a numeric value. The first theta value of the slow payer S-curve.}

\item{Incentive.Slow.theta.2}{a numeric value. The second theta value of the slow payer S-curve.}

\item{Incentive.Slow.beta}{a numeric value.  The beta value of the slow payer S-curve.}

\item{Incentive.Slow.eta}{a numeric value. The eta value of the slow payer S-curve.}

\item{Burnout.beta.1}{a numeric value. The burnout first beta value to assign fast and slow payer.}

\item{Burnout.beta.2}{a numeric value. The burnout second beta value to assign fast and slow payer.}

\item{BeginCDR}{a numeric value. The beginning CDR ramp.}

\item{PeakCDR}{a numeric value. The peak of the CDR ramp.}

\item{EndCDR}{a numeric value. The end value of the CDR ramp.}

\item{PeakMonth}{a numeric value. The month at which the CDR ramp reaches its peak.}

\item{PlateauMonths}{a numeric value. The number of months the CDR remains at its peak value.}

\item{EndMonth}{a numeric value. The month at which the CDR ramp peak ends.}

\item{MinOrigLTV}{a numeric value. The minimum original LTV which ties the
end minimum point of the orginal LTV multiplier function}

\item{MaxOrigLTV}{a numeric value. The maximum orginal LTV which ties the
maxumum point of the original LTV multiplier}

\item{MinOrigMultiplier}{a numeric value. The minimum OLTV multiplier applied to the baseline}

\item{MaxOrigMultiplier}{a numeric value. The maximum OLTV multipler applied to the baseline}

\item{UpdatedLTV.beta}{a numeric value,  The beta value of the update loan to value}

\item{SATO.beta}{A numeric value, the beta value of SATO}
}
\description{
Create custom prepayment model tuning parameters that may
be applied to specific collateral cohorts
}

