% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\name{MakeRDME}
\alias{MakeRDME}
\title{A constructor function for the REMIC Disclosure Month End (RDME) file}
\usage{
MakeRDME(DealName = "character", TrancheNumber = numeric(),
  Cusip = "character", PaymentDate = "character", Coupon = numeric(),
  Factor = numeric())
}
\arguments{
\item{DealName}{A charcter string the deal name}

\item{TrancheNumber}{A numeric value the number of the Tranche}

\item{Cusip}{A character string the tranche cusip}

\item{PaymentDate}{A character string the payment date coinciding with the factor data}

\item{Coupon}{A numeric value the tranche's coupon}

\item{Factor}{A numeric value the tranche's factor}
}
\description{
A constructore for the REMIC Month End Discloure.  This file the monthly factor data for each Tranche
}
\examples{
\dontrun{
MakeRDME(DealName = "BondLabPACInverse",
TrancheNumber = 1,
Cusip = "BondLabPAC2",
"PaymentDate = "01-01-2013",
Coupon = 2.25,
Factor = 1)

MakeRDME(DealName = "BondLabPACInverse",
      TrancheNumber = 2,
      Cusip = "BondLabFltr",
      Coupon = 0.55,
      Factor = 1)

MakeRDME(DealName = "BondLabPACInverse",
      TrancheNumber = 3,
      Cusip = "BondLabCMP1",
      PaymentDate = "1-01-2013",
      Coupon = 9.21,
      Factor = 1)

MakeRDME(DealName = "BondLabPACInverse",
      TrancheNumber = 4,
      Cusip = "BondLabPACIO",
      PaymentDate = "1-01-2013",
      Coupon = 1.75,
      Factor = 1)
}
}

