% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\name{MakeSchedule}
\alias{MakeSchedule}
\title{A constructor function to create the PAC Bond Sinking Fund Schedule file}
\usage{
MakeSchedule(bond.id = "character", DealName = "character",
  Group = "character", original.bal = numeric(), trade.date = "character",
  settlement.date = "character", first.pmtdate = "character",
  price = numeric(), begin.cpr = numeric(), end.cpr = numeric(),
  seasoning.period = numeric(), lower.PSA = numeric(),
  upper.PSA = numeric())
}
\arguments{
\item{bond.id}{A character string the cusip or id}

\item{DealName}{A character string the transaction deal name}

\item{Group}{A character string the tranche's collateral group}

\item{original.bal}{A numeric value the collateral group original balance}

\item{trade.date}{A character string the trade date}

\item{settlement.date}{A character string the settlement date}

\item{first.pmtdate}{A character string the bond first payment date}

\item{price}{A numeric value the price of the underlying collateral}

\item{begin.cpr}{A numeric value the beginning value of the PPC assumption}

\item{end.cpr}{A numeric value the ending value of the PPC assumption}

\item{seasoning.period}{A numeric value the length of the PPC ramp}

\item{lower.PSA}{A numeric value the lower PSA band}

\item{upper.PSA}{A numeric value the upper PSA band}
}
\description{
Function to create a PAC bond sinking fund schedule file
}
\examples{
\dontrun{MakeSchedule(bond.id = "BondLabMBS4",DealName = "BondLabPAC01",Group = 1,
original.bal = 200000000,first.pmtdate = "01-25-2013",trade.date = "01-10-2013",
settlement.date = "01-13-2013",price = 105.75,begin.cpr = .2,end.cpr = 6,
seasoning.period = 30,lower.PSA = 75, upper.PSA = 250)}
}

