% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BillDetails.R
\docType{methods}
\name{Maturity,BillDetails-method}
\alias{Maturity,BillDetails-method}
\title{Method to extract MaturityDate from an S4 class of type BillDetails}
\usage{
\S4method{Maturity}{BillDetails}(object)
}
\arguments{
\item{object}{The name of the S4 class of type BillDetails}
}
\description{
Method to extract MaturityDate from an S4 class of type BillDetails
}
