% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondDetails.R
\docType{methods}
\name{Maturity,BondDetails-method}
\alias{Maturity,BondDetails-method}
\title{Method to extract Maturity from an S4 class of type BondDetails}
\usage{
\S4method{Maturity}{BondDetails}(object)
}
\arguments{
\item{object}{the name of the S4 class of type BondDetails}
}
\description{
Method to extract Maturity from an S4 class of type BondDetails
}
