% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBSDetails.R
\docType{methods}
\name{Maturity,MBSDetails-method}
\alias{Maturity,MBSDetails-method}
\title{Method to extract the slot Maturity from the class MBSDetails}
\usage{
\S4method{Maturity}{MBSDetails}(object)
}
\arguments{
\item{object}{the name of the object of type MBSDetails}
}
\description{
Method to extract the slot Maturity from the class MBSDetails
}
