% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelTune.R
\docType{methods}
\name{MaxOrigLTV,PrepaymentModelTune-method}
\alias{MaxOrigLTV,PrepaymentModelTune-method}
\title{A method to extract the MaxOrigLTV from class PrepaymentModelTune}
\usage{
\S4method{MaxOrigLTV}{PrepaymentModelTune}(object)
}
\arguments{
\item{object}{the name of the object of class PrepaymentModelTune}
}
\description{
A method to extract the MaxOrigLTV from class PrepaymentModelTune
}
