% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelTune.R
\name{MaxOrigLTV}
\alias{MaxOrigLTV}
\title{A standard generic functin to access the slot MaxOrigLTV}
\usage{
MaxOrigLTV(object)
}
\arguments{
\item{object}{an S4 object}
}
\description{
A standard generic functin to access the slot MaxOrigLTV
}
