% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelTune.R
\docType{methods}
\name{MaxOrigMultiplier,PrepaymentModelTune-method}
\alias{MaxOrigMultiplier,PrepaymentModelTune-method}
\title{A method to extract the MaxOrigMultiplier from class PrepaymentModelTune}
\usage{
\S4method{MaxOrigMultiplier}{PrepaymentModelTune}(object)
}
\arguments{
\item{object}{the name of the object of class PrepaymentModelTune}
}
\description{
A method to extract the MaxOrigMultiplier from class PrepaymentModelTune
}

