% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ReadConnectionFunctions.R
\name{MBS}
\alias{MBS}
\title{A connection function to the BondData folder to call the MBS cusip objects}
\usage{
MBS(MBS.id = "character")
}
\arguments{
\item{MBS.id}{A character string the MBS id or cusip number}
}
\description{
Opens a connection to the BondData folder to call MBS cusip data
}

