% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MBSDetails-class}
\alias{MBSDetails-class}
\title{An S4 class to represent a mortgage pass-through security}
\description{
An S4 class to represent a mortgage pass-through security
}
\section{Slots}{

\describe{
\item{\code{Cusip}}{A character string the pass-through cusip}

\item{\code{ID}}{A character string identifying the agency and pool number}

\item{\code{BondType}}{A character sting the bond type}

\item{\code{Sector}}{A character string the sector}

\item{\code{Coupon}}{A coupon a numeric value the annual coupon rate}

\item{\code{IssueDate}}{A character string the issue date}

\item{\code{DatedDate}}{A character string the dated date}

\item{\code{Maturity}}{A character string the maturity date}

\item{\code{LastPmtDate}}{A character string the last payment advanced}

\item{\code{NextPmtDate}}{A character string the next payment date}

\item{\code{Term}}{A numeric value the original term of the underlying mortgage}

\item{\code{WALA}}{A numeric value the weighted average loan age}

\item{\code{WAM}}{A numeric value the weighted average maturity}

\item{\code{PaymentDelay}}{A numeric value the payment delay}

\item{\code{Moody}}{A character string the Moody credit rating}

\item{\code{SP}}{A character string the Standard and Poors credit rating}

\item{\code{BondLab}}{A character string the BondLab credit rating}

\item{\code{Frequency}}{A numeric value the frequency of payments (annual = 1, semi-annual = 2, monthly = 12)}

\item{\code{BondBasis}}{A character string the day count used ("actual360",30360", etc.)}

\item{\code{GWac}}{A numeric string the weighted average note rate}

\item{\code{OrigLoanBal}}{A numeric value the original loan balance}

\item{\code{OrigLTV}}{A numeric value the original loan to value ratio}

\item{\code{AmortizationType}}{A character string "fixed" or "arm"}

\item{\code{AmortizationTerm}}{A numeric value the term of the underlying mortgage}

\item{\code{Index}}{A character string the index to which the arm coupon is related}

\item{\code{Margin}}{A numeric value the spread (margin) over the index}

\item{\code{FirstPmtDate}}{A character string the mortgage first payment date}

\item{\code{FinalPmtDate}}{A character string the mortgage final payment date}

\item{\code{Servicing}}{A numeric value the serving strip}

\item{\code{PMI}}{A numeric value the primary mortgage insurance paid}

\item{\code{GFee}}{A numeric value the guarantee fee paid}

\item{\code{InitialInterest}}{A logical value TRUE or FALSE interest only mortgage}

\item{\code{FirstPrincipalPaymentDate}}{A character string the date of the first principal payment}

\item{\code{BalloonPmt}}{A logical value TRUE or FALSE a balloon payment due}

\item{\code{BalloonDate}}{A character string the balloon payment date}

\item{\code{MBSFactor}}{A numeric value the pass-through principal balance outstanding expressed as
a percentage of the original balance}

\item{\code{Model}}{A character string the prepayment model tune file}

\item{\code{Burnout}}{A numeric value the burnout value}

\item{\code{SATO}}{A numeric value the mortgage rate spread at origination}
}}

