% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBSDetails.R
\name{MBSDetails}
\alias{MBSDetails}
\title{MBSDetails}
\usage{
MBSDetails(Cusip = "character", ID = "character", BondType = "character",
  Issuer = "character", Underwriter = "character", Sector = "character",
  Coupon = numeric(), IssueDate = "character", DatedDate = "character",
  Maturity = "character", LastPmtDate = "character",
  NextPmtDate = "character", Term = numeric(), WALA = numeric(),
  WAM = numeric(), PaymentDelay = numeric(), Moody = "character",
  SP = "character", BondLab = "character", Frequency = numeric(),
  BondBasis = "character", GWac = numeric(), OrigLoanBal = numeric(),
  OrigLTV = numeric(), AmortizationType = "character",
  AmortizationTerm = numeric(), Index = "character", Margin = numeric(),
  FirstPmtDate = "character", FinalPmtDate = "character",
  Servicing = numeric(), PMI = numeric(), GFeePremium = numeric(),
  InitialInterest = "logical", InterestOnlyPeriod = numeric(),
  FirstPrinPaymentDate = "character", BalloonPmt = "logical",
  BalloonDate = "character", MBSFactor = numeric(),
  OriginalBal = numeric(), CurrentBal = numeric(), Model = "character",
  Burnout = numeric(), SATO = numeric())
}
\arguments{
\item{Cusip}{A character the Pass Through MBS cusip.}

\item{ID}{A character string the pool number.}

\item{BondType}{A character string the type of Bond MBS, etc.}

\item{Issuer}{A character string the Issuer}

\item{Underwriter}{A character string the Underwriter}

\item{Sector}{A character string description of the Sector Mortgage Sector}

\item{Coupon}{A numeric value the Bond Coupon.}

\item{IssueDate}{A character string the issue date of the security.}

\item{DatedDate}{A character sting The date following the issue when 
interest begins to accure.}

\item{Maturity}{A character sting the final payment date to the investor
in the case MBS the final payment data assuming 0 CPR.}

\item{LastPmtDate}{A character string the date the last scheduled 
payment to the investor.}

\item{NextPmtDate}{A character string the date of the next scheduled 
payment to the investor.}

\item{Term}{A numeric value the original term of the underlying mortgages}

\item{WALA}{A numeric value the weighted average loan age of the 
underlying mortgages}

\item{WAM}{A numeric value the weighted average maturity of the 
underlying mortgages}

\item{PaymentDelay}{A numeric value in the case of MBS the delay of the 
payment from the trust to the investor}

\item{Moody}{A character string Moody's assigned credit rating}

\item{SP}{A character string SP's assigned credit rating}

\item{BondLab}{A character string BondLab's or the user's 
assigned credit rating}

\item{Frequency}{A numeric value string the frequency of payments made 
to the investor}

\item{BondBasis}{A character string the basis on which 
interest is calculated}

\item{GWac}{A numeric value the borrower's note rate}

\item{OrigLoanBal}{A numeric value the original balance of the loan}

\item{OrigLTV}{A numeric value the borrower's original loan to value ratio}

\item{AmortizationType}{A character sting the type of the 
loan 'fixed' or 'arm'.
These values are used by the prepayment model to drive the mortgage 
rate either fixed or adjustable mortgage rate}

\item{AmortizationTerm}{A numeric value the term of the loan in years}

\item{Index}{A character string if the amortization type is 
adjustable the Index to which the note rate references}

\item{Margin}{A numeric value the spread over the index used to determine 
the borrower's note rate}

\item{FirstPmtDate}{A character string the date of the first payment 
of the borrower's note.}

\item{FinalPmtDate}{A character string the date of thee final payment 
of the borrower's note.  In the case of an MBS the final payment made to 
the bondholder}

\item{Servicing}{A numeric value the servicing spread from the Gross WAC 
to the servicer
of the mortgage pool's underlying loans.}

\item{PMI}{A numeric value the primary mortage insurance paid by the 
borrower to the PMI provider}

\item{GFeePremium}{A numeric value the guarantee fee taken from the 
borrower's note rate to guarantee timely payment of principal and interest.  
Applicable in the case of Fannie Mae, Freddie Mac, or Ginnie Mae pools.}

\item{InitialInterest}{A logical indicating the note carries an interest 
only period}

\item{InterestOnlyPeriod}{A numeric value indicating the note's 
interest only period}

\item{FirstPrinPaymentDate}{A character string indicating the first 
principal payment date due of the mortgage.}

\item{BalloonPmt}{A logical indicating the mortgage carries a balloon pmt.}

\item{BalloonDate}{A character string the balloon payment date.}

\item{MBSFactor}{A numeric value the current factor of the MBS.}

\item{OriginalBal}{A numeric value the original balance of the MBS.}

\item{CurrentBal}{A numeric value the current balance of the MBS.}

\item{Model}{A character string the prepayment model to use.}

\item{Burnout}{A numeric model the value of the borrower burnout.}

\item{SATO}{A numeric value the borrrowers Spread AT Origination over the}
}
\description{
MBSDetails creates an MBS pass-through cusip object in the 
local environment
}
\seealso{
Other MBSDetails: \code{\link{MakeMBSDetails}}
}
