% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBSDetails.R
\name{MBSFactor,MBSDetails-method}
\alias{MBSFactor,MBSDetails-method}
\title{Method to extract the slot MBSFactor from the class MBSDetails}
\usage{
\S4method{MBSFactor}{MBSDetails}(object)
}
\arguments{
\item{object}{the name of the object of the type MBSDetails}
}
\description{
Method to extract the slot MBSFactor from the class MBSDetails
}
