% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PassThroughConstructor.R
\name{MBSFactor<-}
\alias{MBSFactor<-}
\title{A standard generic function to replace the slot MBSFactor}
\usage{
MBSFactor(object) <- value
}
\arguments{
\item{object}{an S4 class object}

\item{value}{the replacement value}
}
\description{
A standard generic function to replace the slot MBSFactor
}

