% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModel.R
\docType{methods}
\name{MDR<-,PrepaymentModel-method}
\alias{MDR<-,PrepaymentModel-method}
\title{A method to replace the MDR value in the class PrepaymentModel}
\usage{
\S4method{MDR}{PrepaymentModel}(object) <- value
}
\arguments{
\item{object}{The name of the S4 class object}

\item{value}{the replacement value of the slot}
}
\description{
A method to replace the MDR value in the class PrepaymentModel
}
