% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModel.R
\name{MDR<-}
\alias{MDR<-}
\title{A standard generic function to replace the value of the slot MDR}
\usage{
MDR(object) <- value
}
\arguments{
\item{object}{An S4 class object of the type PrepaymentModel}

\item{value}{the replacement value of the slot}
}
\description{
A standard generic function to replace the value of the slot MDR
}
