% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModel.R
\name{MDR}
\alias{MDR}
\title{A standard generic function to access the slot MDR}
\usage{
MDR(object)
}
\arguments{
\item{object}{An S4 class object of the type PrepaymentAssumption}
}
\description{
A standard generic function to access the slot MDR
}

