% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelTune.R
\docType{methods}
\name{MinOrigLTV,PrepaymentModelTune-method}
\alias{MinOrigLTV,PrepaymentModelTune-method}
\title{A method to extract the MinOrigLTV from class PrepaymentModelTune}
\usage{
\S4method{MinOrigLTV}{PrepaymentModelTune}(object)
}
\arguments{
\item{object}{the name of the object of class PrepaymentModelTune}
}
\description{
A method to extract the MinOrigLTV from class PrepaymentModelTune
}
