% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelTune.R
\docType{methods}
\name{MinOrigMultiplier,PrepaymentModelTune-method}
\alias{MinOrigMultiplier,PrepaymentModelTune-method}
\title{A method to extract the MinOrigMultiplier from class PrepaymentModelTune}
\usage{
\S4method{MinOrigMultiplier}{PrepaymentModelTune}(object)
}
\arguments{
\item{object}{the name of the object of class PrepaymentModelTune}
}
\description{
A method to extract the MinOrigMultiplier from class PrepaymentModelTune
}

