% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelTune.R
\name{MinOrigMultiplier}
\alias{MinOrigMultiplier}
\title{A standard generic function to access the slot MinOrigMultiplier}
\usage{
MinOrigMultiplier(object)
}
\arguments{
\item{object}{an S4 object}
}
\description{
A standard generic function to access the slot MinOrigMultiplier
}
