% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BillCashFlows.R
\docType{methods}
\name{ModDuration,BillCashFlows-method}
\alias{ModDuration,BillCashFlows-method}
\title{Method to get modified duration from S4 class BillCashFlows}
\usage{
\S4method{ModDuration}{BillCashFlows}(object)
}
\arguments{
\item{object}{the name of the S4 object BillCashFlows}
}
\description{
Method to get modified duration from S4 class BillCashFlows
}
