% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondCashFlow.R
\docType{methods}
\name{ModDuration,BondCashFlows-method}
\alias{ModDuration,BondCashFlows-method}
\title{Method to extract ModDuration from S4 class BondCashFlows}
\usage{
\S4method{ModDuration}{BondCashFlows}(object)
}
\arguments{
\item{object}{the name of the S4 object}
}
\description{
Method to extract ModDuration from S4 class BondCashFlows
}

