% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageCashFlow.R
\docType{methods}
\name{ModDuration,MortgageCashFlow-method}
\alias{ModDuration,MortgageCashFlow-method}
\title{Method to extract Modified Duration from S4 class MortgageCashFlow}
\usage{
\S4method{ModDuration}{MortgageCashFlow}(object)
}
\arguments{
\item{object}{the name of the S4 object}
}
\description{
Method to extract Modified Duration from S4 class MortgageCashFlow
}
