% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageCashFlow.R
\name{ModDuration}
\alias{ModDuration}
\title{A standard generic function to access the slot ModDuration}
\usage{
ModDuration(object)
}
\arguments{
\item{object}{an S4 class object of type MortgageCashFlow}
}
\description{
A standard generic function to access the slot ModDuration
}
