% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBSDetails.R
\docType{methods}
\name{Model,MBSDetails-method}
\alias{Model,MBSDetails-method}
\title{Method to extract the slot Model from the class MBSDetails}
\usage{
\S4method{Model}{MBSDetails}(object)
}
\arguments{
\item{object}{the name of the object of the type MBSDetails}
}
\description{
Method to extract the slot Model from the class MBSDetails
}

