% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectionFunctions.R
\name{ModelFunctions}
\alias{ModelFunctions}
\title{A connection function read the Prepayment Model Folder ModelFunctions 
class Opens a connection to the PrepaymentModel folder to read the 
ModelFunctions class}
\usage{
ModelFunctions()
}
\description{
A connection function read the Prepayment Model Folder ModelFunctions 
class Opens a connection to the PrepaymentModel folder to read the 
ModelFunctions class
}
