% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelToCPR.R
\docType{class}
\name{ModelToCPR-class}
\alias{ModelToCPR-class}
\title{An S4 class to hold CPRLife value}
\description{
The CPRLife value is the CPR which equates the MBS yield to maturity to that
of the prepayment model
}
\section{Slots}{

\describe{
\item{\code{CPRLife}}{The CPR which ties the MBS yield to maturity to that of the 
prepayment vector yield to maturity}
}}

