% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ReadConnectionFunctions.R
\name{ModelTune}
\alias{ModelTune}
\title{A connection function to the PrepaymentModel folder to call the prepayment model tuning}
\usage{
ModelTune(bond.id = "character")
}
\arguments{
\item{bond.id}{A character string the bond id or cusip}
}
\description{
A connection function to the mortgage rate function to call prepayment model tuning
}

