% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageOAS.R
\name{MonteCarloMeanReversion}
\alias{MonteCarloMeanReversion}
\title{MonteCarloMeanReversion}
\usage{
MonteCarloMeanReversion(initial.value, mean, sigma, reversion.speed,
  step = 1/12)
}
\arguments{
\item{initial.value}{the initial value}

\item{mean}{the mean of the asset to simulate}

\item{sigma}{the volatility of the asset to simulate}

\item{reversion.speed}{the speed of the mean reversion}

\item{step}{the time step defaults to 1/12}
}
\description{
Monte Carlo Simulation
}
\seealso{
Other mortgage pass through option adjusted spread: \code{\link{CalibrateRates}},
  \code{\link{FitMarket}}, \code{\link{MortgageOAS-class}},
  \code{\link{OASRatesArray}}, \code{\link{SimCurve}},
  \code{\link{SimRates}}
}

