% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageCashFlow.R
\name{MonthlyInterest,MortgageCashFlow-method}
\alias{MonthlyInterest,MortgageCashFlow-method}
\title{Method to extract the MonthlyInterest from class MortgageCashFlow}
\usage{
\S4method{MonthlyInterest}{MortgageCashFlow}(object)
}
\arguments{
\item{object}{the name of the object of type MortgageCashFlow}
}
\description{
Method to extract the MonthlyInterest from class MortgageCashFlow
}
