% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageCashFlow.R
\name{MonthlyInterest}
\alias{MonthlyInterest}
\title{A generic function to access the slot MonthlyInterest}
\usage{
MonthlyInterest(object)
}
\arguments{
\item{object}{an S4 class object of type MortgageCashFlow}
}
\description{
A generic function to access the slot MonthlyInterest
}

