% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageCashFlow.R
\name{MonthlyPmt,MortgageCashFlow-method}
\alias{MonthlyPmt,MortgageCashFlow-method}
\title{Method to extract the MonthlyPmt from class MortgageCashFlow}
\usage{
\S4method{MonthlyPmt}{MortgageCashFlow}(object)
}
\arguments{
\item{object}{the name of the object of type MortgageCashFlow}
}
\description{
Method to extract the MonthlyPmt from class MortgageCashFlow
}
