% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondDetails.R
\docType{methods}
\name{MoodyRating,BondDetails-method}
\alias{MoodyRating,BondDetails-method}
\title{Method to extract Moody rating from S4 class of type BondDetails}
\usage{
\S4method{MoodyRating}{BondDetails}(object)
}
\arguments{
\item{object}{the name of the object of type BondDetails}
}
\description{
Method to extract Moody rating from S4 class of type BondDetails
}
