% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Functions.R
\name{Mortgage.Monthly.Payment}
\alias{Mortgage.Monthly.Payment}
\title{A function to calculate the monthly payment of a mortgage}
\usage{
Mortgage.Monthly.Payment(orig.bal = numeric(), note.rate = numeric(),
  term.mos = numeric())
}
\arguments{
\item{orig.bal}{A numeric value, the original loan amount}

\item{note.rate}{A numeric vlaue, the borrower's note rate}

\item{term.mos}{A numeric value, the original term of the loan in months}
}
\description{
A standard generic function to compute the monthly payment of a mortgage
}
\examples{
Mortgage.Monthly.Payment(orig.bal = 100000, note.rate = .04, term.mos = 360)
}

