% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageOAS.R
\name{Mortgage.OAS}
\alias{Mortgage.OAS}
\title{Mortgage OAS the OAS engine for pass through OAS}
\usage{
Mortgage.OAS(bond.id = "character", trade.date = "character",
  settlement.date = "character", original.bal = numeric(),
  price = numeric(), sigma = numeric(), paths = numeric())
}
\arguments{
\item{bond.id}{A character string the bond id or cusip number}

\item{trade.date}{A character string the trade date}

\item{settlement.date}{A character string the settlment date}

\item{original.bal}{A numeric value the pool original balance}

\item{price}{A numeric value the price of the pass through}

\item{sigma}{A numeric value the volatility assumption (not annualized)}

\item{paths}{A numeric value the number of paths}
}
\description{
Pass through OAS engine
}

