% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageCashFlow.R
\docType{class}
\name{MortgageCashFlow-class}
\alias{MortgageCashFlow-class}
\title{An S4 class MortgageCashFlow containing cashflow data 
for a mortgage pass-through security}
\description{
An S4 class MortgageCashFlow containing cashflow data 
for a mortgage pass-through security
}
\section{Slots}{

\describe{
\item{\code{Price}}{A numeric value the price of the pass-through.}

\item{\code{Accrued}}{A numeric value the accrued interest as of settlement date.}

\item{\code{YieldToMaturity}}{A numeric value the yield to maturity.}

\item{\code{WAL}}{A numeric value the weighted average life of the pass-thorough.}

\item{\code{ModDuration}}{A numeric value the Modified Duration pass-through.}

\item{\code{Convexity}}{A numeric value the Convexity of the pass-through.}

\item{\code{Period}}{A numeric value the period in which the cash-flow is received.}

\item{\code{PmtDate}}{A character string the date in which the cash-flow is received.}

\item{\code{TimePeriod}}{A numeric value the time weight to apply to the principal cash-flow
and discount factors.}

\item{\code{BeginningBal}}{A numeric value the Beginning Balance in the period.}

\item{\code{MonthlyPmt}}{A numeric value the borrower's monthly payment.}

\item{\code{MonthlyInterest}}{A numeric value the borrower's monthly interest.}

\item{\code{PassThroughInterest}}{A numeric value the pass-through interest paid
to the investor in the pool.}

\item{\code{ScheduledPrin}}{A numeric value the scheduled principal due in the period.}

\item{\code{PrepaidPrin}}{A numeric value the prepaid principal in the period.}

\item{\code{DefaultedPrin}}{A numeric value the default principal in the period.}

\item{\code{LossAmount}}{A numeric value the loss amount in the period.}

\item{\code{RecoveredAmount}}{A numeric value the recovered amount in the period.}

\item{\code{EndingBal}}{A numeric value the ending balance in the period.}

\item{\code{ServicingIncome}}{A numeric value the servicing recevied in the period.}

\item{\code{PMIPremium}}{A numeric value the PMI paid in the period.}

\item{\code{GFeePremium}}{A numeric value the GFee paid in the period.}

\item{\code{TotalCashFlow}}{A numeric value the total cashflow paid in the period.}
}}

