% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageCashFlow_Array.R
\name{MortgageCashFlow_Array}
\alias{MortgageCashFlow_Array}
\title{An alternative function to compute pool cash flow metrics}
\usage{
MortgageCashFlow_Array(CashFlowData = "character", price = numeric(),
  original.bal = numeric(), settlement.date = "character",
  bond.id = "character")
}
\arguments{
\item{CashFlowData}{a character string referencing the cash flow file whose extension is .cfm}

\item{price}{the price of the MBS. Price is input as a whole number for example 102 is input as 102 not 1.02}

\item{original.bal}{the original balance of the MBS pool}

\item{settlement.date}{the settlement date of the MBS trade}

\item{bond.id}{a character string the bond.id}
}
\description{
This function by passes the mortgage cash flow engine used in the MortgageCashFlow function
thereby allowing the user to create a custom cashflow array possible use of this function is the loan level
agregation of either pools or REMIC cash flow.
}
\details{
The cashflow array passed to this function must be n rows by 23 columns the file must have headers for each column
as follows: Period, Date, Time, Begin Bal, Monthly Pmt, Scheduled Int, Scheduled Prin, Prepaid Prin, Defaulted Prin,
Loss Amount, Recovered Amount, Ending Bal, Servicing, PMI, GFee, Pass Through Interest, Investor CashFlow, 
Present Value Factor, Present Value, Duration, Convexity Time, CashFlow Convexity, Convexity.  Values in the array are referenced
by name thus the data may be presented in any order however the column headers must match the above exactly.
}
\examples{
\dontrun{MortgageCashFlow_KDS(CashFlowData = "blx_test", price = 100, settlement.date = "01-10-2013", 
original.bal = 102, bond.id = "bondlabMBS4")}
}

