% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageEffectiveMeasures.R
\docType{class}
\name{MortgageEffectiveMeasures-class}
\alias{MortgageEffectiveMeasures-class}
\title{An S4 class MortgageEffectiveMeasures}
\description{
An S4 class MortgageEffectiveMeasures
}
\section{Slots}{

\describe{
\item{\code{EffDuration}}{A numeric value the effective duration}

\item{\code{EffConvexity}}{A numeric value the effective convexity}
}}

