% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageEffectiveMeasures.R
\name{MortgageEffectiveMeasures}
\alias{MortgageEffectiveMeasures}
\title{A function to calculate mortgage effective duration and convexity}
\usage{
MortgageEffectiveMeasures(bond.id = "character",
  settlement.date = "character", rates.data = "character",
  TermStructure = "character", price = "character")
}
\arguments{
\item{bond.id}{A character string referencing  an object of type bond.id}

\item{settlement.date}{A character string the settlement date}

\item{rates.data}{A character string referencing and object rates}

\item{TermStructure}{A character string referencing an object of type
TermStructure}

\item{price}{A character the price of the MBS.  Price may be entered in
decimal notation or 32nds notation}
}
\description{
A function to calculate mortgage effective duration and convexity
}
