% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageOAS.R
\docType{class}
\name{MortgageOAS-class}
\alias{MortgageOAS-class}
\title{An S4 class the Mortgage.OAS constructor function}
\description{
An S4 class the Mortgage.OAS constructor function
}
\section{Slots}{

\describe{
\item{\code{OAS}}{A numeric value the option adjusted spread}

\item{\code{OptionAdjDur}}{A numeric value the option adjusted duration}

\item{\code{OptionAdjCvx}}{A numeric value the option adjusted convexity}

\item{\code{ZeroVolSpread}}{A numeric value the zero volatility spread}

\item{\code{SpreadToCurve}}{A numeric value the spread to the curve}

\item{\code{EffDuration}}{A numeric value the effective duration}

\item{\code{EffConvexity}}{A numeric value the effective convexity}

\item{\code{KeyRateTenor}}{A numeric value the Key Rate Tenor}

\item{\code{KeyRateDuration}}{A numeric value the Key Rate Duration}

\item{\code{KeyRateConvexity}}{A numeric value the Key Rate Convexity}

\item{\code{PriceDist}}{A numeric value the price distribution from the OAS model}

\item{\code{PathSpread}}{A numeric value the spot rate spread along each path}

\item{\code{PathWAL}}{A numeric value the WAL along each path}

\item{\code{PathModDur}}{A numeric value the modified duration along each path}

\item{\code{PathYTM}}{A numeric value the yield to maturity along each path}

\item{\code{OASTermStructure}}{A S4 object the zero volatility Term Structure}

\item{\code{RatePaths}}{A matrix the OAS Rate Paths}

\item{\code{PrepaymentVectors}}{A matrix the prepayment vectors returned along 
each interest rate path}
}}

