% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageOAS.R
\name{MortgageOAS}
\alias{MortgageOAS}
\title{Mortgage PassThrough Option Adjusted Spread}
\usage{
MortgageOAS(bond.id, price, trade.date, settlement.date, OAS.array,
  num.periods = 480)
}
\arguments{
\item{bond.id}{A character the bond.id}

\item{price}{A character the price of the bond given in decimal of 32nds}

\item{trade.date}{A character the trade date mm-dd-YYYY}

\item{settlement.date}{A character the settlement date mm-dd-YYYY}

\item{OAS.array}{An array of simulated rate vectors from the function OASArray}

\item{num.periods}{the number of payment periods defaults to 480 do not change 
this unless you know what you are doing}
}
\description{
A function to compute mortgage passthrough option adjusted spread
}
