% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageRateFunctions.R
\name{MortgageRate}
\alias{MortgageRate}
\title{A constructor function for the class Mortgage Rate}
\usage{
MortgageRate(two, ten, sato)
}
\arguments{
\item{two}{A numeric value the two year rate}

\item{ten}{A numeric value the ten year rate}

\item{sato}{A numeric value the borrower SATO}
}
\description{
A constructor function for the class Mortgage Rate
}

