% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\docType{class}
\name{MortgageScenario-class}
\alias{MortgageScenario-class}
\title{An S4 Class representing the results of mortgage return scenario analysis}
\description{
The SuperClass MtgScenario holds the results of a scenario analysis run
}

