% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\name{MortgageScenario}
\alias{MortgageScenario}
\title{Mortgage Scenario Analysis}
\usage{
MortgageScenario(bond.id, settlement.date, rates.data, price, original.bal,
  scenario, horizon.months, method, prepayment, ...,
  horizon.spot.spread = NULL, horizon.nominal.spread = NULL,
  horizon.OAS = NULL, horizon.price = NULL, begin.cpr, end.cpr,
  seasoning.period, CPR)
}
\arguments{
\item{bond.id}{A character string referencing an object of type MBSDetails}

\item{settlement.date}{A charcter string the settlement date}

\item{rates.data}{A character string the trade date mm-dd-yyyy}

\item{price}{A character string in decimal equivalent (.) or 32nds (-)}

\item{original.bal}{A numeric value the price}

\item{scenario}{A character string the scenario}

\item{horizon.months}{A numeric value the time horizon}

\item{method}{A character string the method used to fit the 
term structure}

\item{prepayment}{A character string the prepayment assumption}

\item{...}{Optional values when PSA or CPR is used or Yield Curve
is used}

\item{horizon.spot.spread}{A numeric value the horizon zero volatility 
spread}

\item{horizon.nominal.spread}{A numeric value the horizon spread}

\item{horizon.OAS}{A numeric value the horizon option adjusted spread
(not currently implemented)}

\item{horizon.price}{A numeric value the horizon price in decimal form}

\item{begin.cpr}{A numeric value the beginning CPR value}

\item{end.cpr}{A numeric value the ending CPR value}

\item{seasoning.period}{A numeric value the length of the 
seasoning ramp}

\item{CPR}{A numeric value the CPR speed}
}
\description{
A function to compute the total return of mortgage pass-throughs MBS
}

