% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\name{MortgageScenarioSet}
\alias{MortgageScenarioSet}
\title{MortgageScenarioSet is the constructor function for the class Mortgage
ScenarioSet}
\usage{
MortgageScenarioSet(ScenarioSet = "list")
}
\arguments{
\item{ScenarioList}{a list the of scenario classes for analysis}
}
\description{
MortgageScenarioSet is the constructor function for the class Mortgage
ScenarioSet
}

