% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageKeyRate.R
\docType{class}
\name{MortgageTermStructure-class}
\alias{MortgageTermStructure-class}
\title{An S4 class MortgageTermStructure}
\description{
An S4 class MortgageTermStructure
}
\section{Slots}{

\describe{
\item{\code{Cusip}}{A character the cusip number}

\item{\code{Issuer}}{A character the Issuer name}

\item{\code{Coupon}}{A numeric value the coupon}

\item{\code{Term}}{A numeric value the amorization term}

\item{\code{ZeroVolSpread}}{A numeric value the spread to the spot rate curve}

\item{\code{EffDuration}}{A numeric value the effective duration}

\item{\code{EffConvexity}}{A numeric value the effective convexity}

\item{\code{KeyRateTenor}}{A numeric value the Key Rate Tenor}

\item{\code{KeyRateDuration}}{A numeric value the Key Rate Duration}

\item{\code{KeyRateConvexity}}{A numeric value the Key Rate Convexity}
}}

