% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MtgScenario.R
\name{Mtg.Scenario}
\alias{Mtg.Scenario}
\title{Mortgage Scenario Analysis}
\usage{
Mtg.Scenario(bond.id = "character", trade.date = "character",
  settlement.date = "character", price = numeric(), proceeds = numeric(),
  spot.spread = numeric(), original.bal = numeric(),
  scenario.set = vector(), rates.data = "character", method = "character",
  PrepaymentAssumption = "character", ..., ModelTune = "character",
  Burnout = numeric(), begin.cpr = numeric(), end.cpr = numeric(),
  seasoning.period = numeric(), CPR = numeric())
}
\arguments{
\item{bond.id}{A character string to object of type MBS details}

\item{price}{A character string the trade date mm-dd-yyyy}

\item{original.bal}{A numeric value the price}

\item{rates.data}{A character string the trade date mm-dd-yyyy}

\item{method}{A character string the method used to fit the term structure}

\item{scenario}{A character string the scenario}
}
\description{
A function to compute the total return of mortgage pass-throughs MBS
}

