% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\docType{class}
\name{Mtg.ScenarioSet-class}
\alias{Mtg.ScenarioSet-class}
\title{An S4 Class A list of Scenario outcomes}
\description{
The class Mtg.ScenarioSet is a list of classes Scenario
}
\section{Slots}{

\describe{
\item{\code{Scenario}}{A list of the class Scenario}
}}

