% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\docType{methods}
\name{Mtg.ScenarioSet,Mtg.ScenarioSet-method}
\alias{Mtg.ScenarioSet,Mtg.ScenarioSet-method}
\title{A method to extract the Scenario list from S4 class Mtg.ScenarioSet}
\usage{
\S4method{Mtg.ScenarioSet}{Mtg.ScenarioSet}(object)
}
\arguments{
\item{object}{the name of the S4 object of the type Scenario}
}
\description{
A method to extract the Scenario list from S4 class Mtg.ScenarioSet
}

