% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\name{Mtg.ScenarioSet}
\alias{Mtg.ScenarioSet}
\title{A standard generic function to extract the slot Scenario}
\usage{
Mtg.ScenarioSet(object)
}
\arguments{
\item{object}{An S4 class object}
}
\description{
A standard generic function to extract the slot Scenario
}

