% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModel.R
\docType{methods}
\name{MtgRateFwd,PrepaymentAssumption-method}
\alias{MtgRateFwd,PrepaymentAssumption-method}
\title{A method to extract MtgRateFwd from S4 class PrepaymentAssumption}
\usage{
\S4method{MtgRateFwd}{PrepaymentAssumption}(object)
}
\arguments{
\item{object}{the name of the S4 class}
}
\description{
A method to extract MtgRateFwd from S4 class PrepaymentAssumption
}

