% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModel.R
\name{MtgRateFwd}
\alias{MtgRateFwd}
\title{A standard generic function to access the slot MtgRateFwd}
\usage{
MtgRateFwd(object)
}
\arguments{
\item{object}{An S4 class object of the type PrepaymentModel}
}
\description{
A standard generic function to access the slot MtgRateFwd
}

