% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\name{Name}
\alias{Name}
\title{A standard generic function to extract the slot}
\usage{
Name(object)
}
\arguments{
\item{object}{An S4 class object}
}
\description{
A standard generic function to extract the slot
}

