% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondConstructor.R
\docType{methods}
\name{NextPmtDate,BondDetails-method}
\alias{NextPmtDate,BondDetails-method}
\title{Method extract the slot NextPmtDate from class BondDetails}
\usage{
\S4method{NextPmtDate}{BondDetails}(object)
}
\arguments{
\item{object}{the name of the S4 object of the type BondDetails}
}
\description{
Method extract the slot NextPmtDate from class BondDetails
}

