% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBSDetails.R
\name{NextPmtDate,MBSDetails-method}
\alias{NextPmtDate,MBSDetails-method}
\title{Method to extract the slot NextPmtDate from the class MBSDetails}
\usage{
\S4method{NextPmtDate}{MBSDetails}(object)
}
\arguments{
\item{object}{the name of the object of the type MBSDetails}
}
\description{
Method to extract the slot NextPmtDate from the class MBSDetails
}
