% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondDetails.R
\docType{methods}
\name{NextPmtDate<-,BondDetails-method}
\alias{NextPmtDate<-,BondDetails-method}
\title{Method to replace the slot NextPmtDate from class BondDetails}
\usage{
\S4method{NextPmtDate}{BondDetails}(object) <- value
}
\arguments{
\item{object}{the name of the S4 object of the type BondDetails}

\item{value}{the value of the replacement}
}
\description{
Method to replace the slot NextPmtDate from class BondDetails
}
