% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBSDetails.R
\docType{methods}
\name{NextPmtDate<-,MBSDetails-method}
\alias{NextPmtDate<-,MBSDetails-method}
\title{Method to extract the slot NextPmtDate from the class MBSDetails}
\usage{
\S4method{NextPmtDate}{MBSDetails}(object) <- value
}
\arguments{
\item{object}{the name of the object of the type MBSDetails}

\item{value}{the value of the slot replacement}
}
\description{
Method to extract the slot NextPmtDate from the class MBSDetails
}
