% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBSDetails.R
\name{NextPmtDate}
\alias{NextPmtDate}
\title{A standard generic function to access the slot NextPmtDate}
\usage{
NextPmtDate(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A standard generic function to access the slot NextPmtDate
}
