% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModel.R
\name{NoteRate}
\alias{NoteRate}
\title{A standard generic function to access the slot NoteRate}
\usage{
NoteRate(object)
}
\arguments{
\item{object}{An S4 class object of the type PrepaymentModel}
}
\description{
A standard generic function to access the slot NoteRate
}

