% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{NumberofCombinedTranches,RAID-method}
\alias{NumberofCombinedTranches,RAID-method}
\title{A method to access the slot NumberofCombinedTranches from class RAID}
\usage{
\S4method{NumberofCombinedTranches}{RAID}(object)
}
\arguments{
\item{object}{an S4 class of type RAID}
}
\description{
A method to access the slot NumberofCombinedTranches from class RAID
}

