% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{NumberofComponents,TrancheDetails-method}
\alias{NumberofComponents,TrancheDetails-method}
\title{A method to access the slot NumberofComponents from class TrancheDetails}
\usage{
\S4method{NumberofComponents}{TrancheDetails}(object)
}
\arguments{
\item{object}{an S4 object of the type TrancheDetails}
}
\description{
A method to access the slot NumberofComponents from class TrancheDetails
}

