% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{NumberofPacSchedules,RAID-method}
\alias{NumberofPacSchedules,RAID-method}
\title{A method to access the slot NumberofPacSchedules from class RAID}
\usage{
\S4method{NumberofPacSchedules}{RAID}(object)
}
\arguments{
\item{object}{an S4 class object of type RAID}
}
\description{
A method to access the slot NumberofPacSchedules from class RAID
}

