% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{NumberofPools,RAID-method}
\alias{NumberofPools,RAID-method}
\title{A method to access the slot NumberofPools from class RAID}
\usage{
\S4method{NumberofPools}{RAID}(object)
}
\arguments{
\item{object}{an S4 class of type RAID}
}
\description{
A method to access the slot NumberofPools from class RAID
}

