% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\name{NumberofPools}
\alias{NumberofPools}
\title{A generic function to access the slot NumberofPools}
\usage{
NumberofPools(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A generic function to access the slot NumberofPools
}

