% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{NumberofTranches,RAID-method}
\alias{NumberofTranches,RAID-method}
\title{A method to access the slot NumberofTranches from class RAID}
\usage{
\S4method{NumberofTranches}{RAID}(object)
}
\arguments{
\item{object}{an S4 class of type RAID}
}
\description{
A method to access the slot NumberofTranches from class RAID
}

