% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageOAS.R
\name{OAC,MortgageOAS-method}
\alias{OAC,MortgageOAS-method}
\title{A Method to access the slot OAC}
\usage{
\S4method{OAC}{MortgageOAS}(object)
}
\arguments{
\item{object}{MortgageOAS}
}
\description{
Access the slot OAC from object MortgageOAS
}
