% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageOAS.R
\name{OAD,MortgageOAS-method}
\alias{OAD,MortgageOAS-method}
\title{A Method to access the slot OAD}
\usage{
\S4method{OAD}{MortgageOAS}(object)
}
\arguments{
\item{object}{MortgageOAS}
}
\description{
Access the slot OAD from object MortgageOAS
}
