% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageOAS.R
\name{OAS,MortgageOAS-method}
\alias{OAS,MortgageOAS-method}
\title{A Method to access the slot OAS}
\usage{
\S4method{OAS}{MortgageOAS}(object)
}
\arguments{
\item{object}{MortgageOAS}
}
\description{
Access the slot OAS from object MortgageOAS
}
