% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageOAS.R
\name{OASRatesArray}
\alias{OASRatesArray}
\title{Option Adjusted Spread Rates Array}
\usage{
OASRatesArray(rates.data, num.paths = 300, num.periods = 480, calibration)
}
\arguments{
\item{rates.data}{A character referencing a rates.data object}

\item{num.paths}{The number of simulated paths defaults to 300}

\item{num.periods}{The number of mortgage payment periods defaults to 480
do not change unless you know what you are doing}

\item{calibration}{A character referencing the rate calibration object}
}
\description{
Simulated rates paths for option adjusted spread analysis
}
\seealso{
Other mortgage pass through option adjusted spread: \code{\link{CalibrateRates}},
  \code{\link{FitMarket}}, \code{\link{MortgageOAS-class}},
  \code{\link{SimCurve}}, \code{\link{SimRates}}
}

