% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BillDetails.R
\docType{methods}
\name{OfferAmount,BillDetails-method}
\alias{OfferAmount,BillDetails-method}
\title{Method to extract OfferAmount from an S4 class of type BillDetails}
\usage{
\S4method{OfferAmount}{BillDetails}(object)
}
\arguments{
\item{object}{The name of the S4 class of type BillDetails}
}
\description{
Method to extract OfferAmount from an S4 class of type BillDetails
}
