% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondDetails.R
\docType{methods}
\name{OfferAmount,BondDetails-method}
\alias{OfferAmount,BondDetails-method}
\title{Method to get OfferAmount from an S4 class of type BondDetails}
\usage{
\S4method{OfferAmount}{BondDetails}(object)
}
\arguments{
\item{object}{the name of the S4 object of the type BondDetails}
}
\description{
Method to get OfferAmount from an S4 class of type BondDetails
}

