% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondDetails.R
\name{OfferAmount}
\alias{OfferAmount}
\title{A standard generic function to get the par amount isssued traded}
\usage{
OfferAmount(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A standard generic function to get the par amount isssued traded
}
