% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageOAS.R
\docType{methods}
\name{OptionAdjCvx,MortgageOAS-method}
\alias{OptionAdjCvx,MortgageOAS-method}
\title{A method to extract the OptionAdjDur from S4 object MortgageOAS}
\usage{
\S4method{OptionAdjCvx}{MortgageOAS}(object)
}
\arguments{
\item{object}{An S4 object of the type MortgageOAS}
}
\description{
A method to extract the OptionAdjDur from S4 object MortgageOAS
}

