% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageOAS.R
\name{OptionAdjDur}
\alias{OptionAdjDur}
\title{A standard generic function to access the slot OptionAdjDur}
\usage{
OptionAdjDur(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A standard generic function to access the slot OptionAdjDur
}

