% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{OrigBalList,Collateral-method}
\alias{OrigBalList,Collateral-method}
\title{A method to access the original bal list from S4 class Collateral}
\usage{
\S4method{OrigBalList}{Collateral}(object)
}
\arguments{
\item{object}{An S4 class object of the type Collaeral}
}
\description{
A method to access the original bal list from S4 class Collateral
}

