% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\name{OrigBalList}
\alias{OrigBalList}
\title{A generic function to access a OrigBal list slot}
\usage{
OrigBalList(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A generic function to access a OrigBal list slot
}

