% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBSDetails.R
\name{OriginalBal,MBSDetails-method}
\alias{OriginalBal,MBSDetails-method}
\title{Method to extract the slot OriginalBal from the class MBSDetails}
\usage{
\S4method{OriginalBal}{MBSDetails}(object)
}
\arguments{
\item{object}{the name of the object of the type MBSDetails}
}
\description{
Method to extract the slot OriginalBal from the class MBSDetails
}
