% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBSDetails.R
\docType{methods}
\name{OrigLoanBal,MBSDetails-method}
\alias{OrigLoanBal,MBSDetails-method}
\title{Method to extract the slot OrigLoanBal from the class MBSDetails}
\usage{
\S4method{OrigLoanBal}{MBSDetails}(object)
}
\arguments{
\item{object}{the name of the object of the type MBSDetails}
}
\description{
Method to extract the slot OrigLoanBal from the class MBSDetails
}
