% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBSDetails.R
\name{OrigLoanBal}
\alias{OrigLoanBal}
\title{A standard generic function to access the slot OrigLoanBal}
\usage{
OrigLoanBal(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A standard generic function to access the slot OrigLoanBal
}

