% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PassThroughConstructor.R
\docType{methods}
\name{OrigLTV,MBSDetails-method}
\alias{OrigLTV,MBSDetails-method}
\title{Method to extract the slot OrigLTV from the class MBSDetails}
\usage{
\S4method{OrigLTV}{MBSDetails}(object)
}
\arguments{
\item{object}{the name of the object of the type MBSDetails}
}
\description{
Method to extract the slot OrigLTV from the class MBSDetails
}

