% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{PacLowBand,TrancheDetails-method}
\alias{PacLowBand,TrancheDetails-method}
\title{A method to access the slot PacLowBand from class TrancheDetails}
\usage{
\S4method{PacLowBand}{TrancheDetails}(object)
}
\arguments{
\item{object}{an S4 object of type TrancheDetails}
}
\description{
A method to access the slot PacLowBand from class TrancheDetails
}

