% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{PacSchedulesIncluded,RAID-method}
\alias{PacSchedulesIncluded,RAID-method}
\title{A method to access the slot PacSchedulesIncluded from class RAID}
\usage{
\S4method{PacSchedulesIncluded}{RAID}(object)
}
\arguments{
\item{object}{an S4 class of type RAID}
}
\description{
A method to access the slot PacSchedulesIncluded from class RAID
}

