% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PassThroughAnalysis.R
\name{PassThroughAnalytics}
\alias{PassThroughAnalytics}
\title{Pass Through Analytics is the Analytic engine for a pass-through security}
\usage{
PassThroughAnalytics(bond.id = "character", original.bal = numeric(),
  price = numeric(), trade.date = "character",
  settlement.date = "character", method = "character",
  scenario.set = vector(), PrepaymentAssumption = "character", ...,
  begin.cpr = numeric(), end.cpr = numeric(),
  seasoning.period = numeric(), CPR = numeric())
}
\arguments{
\item{bond.id}{A character string the bond id or cusip}

\item{original.bal}{A numeric value the original balance of the pass through}

\item{price}{A numeric value the price of the pass through}

\item{trade.date}{A character string the trade date}

\item{settlement.date}{A character string the settlement date}

\item{method}{A character string indicating the fitting method ns = Nelson Siegel, dl = Diebold Lee,
sv = Severson, asv = adjusted Severson, cs = cubic spline (not yet implemented in Bond Lab).
For addiition details see the termstrc documentation.}

\item{scenario.set}{A character vector listing the scenarios to run}

\item{PrepaymentAssumption}{A character string the prepayment assumption used "MODEL", "PPC", or "CPR"}

\item{...}{Optional values when "PPC" or "CPR" is used}

\item{begin.cpr}{A numeric value the beginning CPR assumption}

\item{end.cpr}{A numeric value the ending CPR assumption}

\item{seasoning.period}{A numeric value the length of the seasoning ramp}

\item{CPR}{A numeric value the CPR assumption}
}
\description{
Analytic engine for pass through security returns pass through analytics object
}
\examples{
\dontrun{PassThroughAnalytics(bond.id = "bondlabMBS55", original.bal = 100000,
price = 107.5, trade.date = "01-10-2013", settlement.date = "01-13-2013",
scenario.set = c("DA25", "NC", "UA50", "UA100", "UA150", "UA200"), PrepaymentAssumption = "MODEL")}
}

