% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageCashFlow.R
\name{PassThroughInterest,MortgageCashFlow-method}
\alias{PassThroughInterest,MortgageCashFlow-method}
\title{Method to extract the PassThroughInterest from class MortgageCashFlow}
\usage{
\S4method{PassThroughInterest}{MortgageCashFlow}(object)
}
\arguments{
\item{object}{the name of the object of type MortgageCashFlow}
}
\description{
Method to extract the PassThroughInterest from class MortgageCashFlow
}
