% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageCashFlow.R
\name{PassThroughInterest}
\alias{PassThroughInterest}
\title{A generic function to access the slot PassThroughInterest}
\usage{
PassThroughInterest(object)
}
\arguments{
\item{object}{an S4 class object of the type MortgageCashFlow}
}
\description{
A generic function to access the slot PassThroughInterest
}
