% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PassThroughOAS.R
\docType{class}
\name{PassThroughOAS-class}
\alias{PassThroughOAS-class}
\title{An S4 class containing the classes MBSDetails, MortgageCashFlow, 
and MortgageOAS}
\description{
An S4 class containing the classes MBSDetails, MortgageCashFlow, 
and MortgageOAS
}

