% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageOAS.R
\name{PathModDur}
\alias{PathModDur}
\title{A standard generic function to access the slot PathModDuration}
\usage{
PathModDur(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A standard generic function to access the slot PathModDuration
}

