% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageOAS.R
\docType{methods}
\name{PathWAL,MortgageOAS-method}
\alias{PathWAL,MortgageOAS-method}
\title{A method to extract PathWAL from S4 class MortgageOAS}
\usage{
\S4method{PathWAL}{MortgageOAS}(object)
}
\arguments{
\item{object}{an S4 object of the type MortgageOAS}
}
\description{
A method to extract PathWAL from S4 class MortgageOAS
}

