% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageOAS.R
\name{PathYTM}
\alias{PathYTM}
\title{A standard generic function to access the slot PathYTM}
\usage{
PathYTM(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A standard generic function to access the slot PathYTM
}

