% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBSDetails.R
\name{PaymentDelay,MBSDetails-method}
\alias{PaymentDelay,MBSDetails-method}
\title{Method to extract the slot PaymentDelay from the class MBSDetails}
\usage{
\S4method{PaymentDelay}{MBSDetails}(object)
}
\arguments{
\item{object}{the name of the object of the type MBSDetails}
}
\description{
Method to extract the slot PaymentDelay from the class MBSDetails
}
