% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PassThroughConstructor.R
\name{PaymentDelay}
\alias{PaymentDelay}
\title{A standard generic function to access the slot PaymentDelay}
\usage{
PaymentDelay(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A standard generic function to access the slot PaymentDelay
}

