% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelTune.R
\docType{methods}
\name{PeakCDR,PrepaymentModelTune-method}
\alias{PeakCDR,PrepaymentModelTune-method}
\title{A method to extract PeakCDR from class PrepaymentModelTune}
\usage{
\S4method{PeakCDR}{PrepaymentModelTune}(object)
}
\arguments{
\item{object}{The name of the object of class PrepaymentModelTune}
}
\description{
A method to extract PeakCDR from class PrepaymentModelTune
}

