% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelTune.R
\name{PeakCDR}
\alias{PeakCDR}
\title{A standard generic functin to access the slot PeakCDR}
\usage{
PeakCDR(object)
}
\arguments{
\item{object}{an S4 object}
}
\description{
A standard generic functin to access the slot PeakCDR
}
