% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelTune.R
\name{PeakMonth}
\alias{PeakMonth}
\title{A standard generic function to access the slot PeakMonth}
\usage{
PeakMonth(object)
}
\arguments{
\item{object}{an S4 object}
}
\description{
A standard generic function to access the slot PeakMonth
}
