% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BillCashFlows.R
\docType{methods}
\name{Period,BillCashFlows-method}
\alias{Period,BillCashFlows-method}
\title{Method to get Period from S4 class BillCashFlows}
\usage{
\S4method{Period}{BillCashFlows}(object)
}
\arguments{
\item{object}{the name of the S4 object BillCashFlows}
}
\description{
Method to get Period from S4 class BillCashFlows
}
