% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageCashFlow.R
\docType{methods}
\name{Period,MortgageCashFlow-method}
\alias{Period,MortgageCashFlow-method}
\title{Method to extract Period from S4 class MortgageCashFlow}
\usage{
\S4method{Period}{MortgageCashFlow}(object)
}
\arguments{
\item{object}{the name of the S4 object}
}
\description{
Method to extract Period from S4 class MortgageCashFlow
}

