% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\docType{methods}
\name{Period,Mtg.Scenario-method}
\alias{Period,Mtg.Scenario-method}
\title{A method to extract Period from S4 class Mtg.Scenario}
\usage{
\S4method{Period}{Mtg.Scenario}(object)
}
\arguments{
\item{object}{The name of an S4 class of type Mtg.Scenario}
}
\description{
A method to extract Period from S4 class Mtg.Scenario
}

