% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\docType{methods}
\name{Period,MtgScenario-method}
\alias{Period,MtgScenario-method}
\title{A method to extract Period from S4 class MtgScenario}
\usage{
\S4method{Period}{MtgScenario}(object)
}
\arguments{
\item{object}{The name of an S4 class of type MtgScenario}
}
\description{
A method to extract Period from S4 class MtgScenario
}

