% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModel.R
\docType{methods}
\name{Period,PrepaymentAssumption-method}
\alias{Period,PrepaymentAssumption-method}
\title{A method to extract Period from S4 class PrepaymentAssumption}
\usage{
\S4method{Period}{PrepaymentAssumption}(object)
}
\arguments{
\item{object}{the name of the S4 class}
}
\description{
A method to extract Period from S4 class PrepaymentAssumption
}

