% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModel.R
\docType{methods}
\name{Period,PrepaymentModel-method}
\alias{Period,PrepaymentModel-method}
\title{A method to extract Period from S4 class PrepaymentModel}
\usage{
\S4method{Period}{PrepaymentModel}(object)
}
\arguments{
\item{object}{the name of the S4 class}
}
\description{
A method to extract Period from S4 class PrepaymentModel
}
