% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TermStructure.R
\name{Period<-,TermStructure-method}
\alias{Period<-,TermStructure-method}
\title{Method to replace Period in S4 class TermStructure}
\usage{
\S4method{Period}{TermStructure}(object) <- value
}
\arguments{
\item{object}{the name of the S4 object of type TermStrucuture}

\item{value}{the replacement value of the slot}
}
\description{
Method to replace Period in S4 class TermStructure
}
