% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TermStructure.R
\name{Period<-}
\alias{Period<-}
\title{a generic function to replace the slot Period}
\usage{
Period(object) <- value
}
\arguments{
\item{object}{an S4 class object}

\item{value}{the replacement value of the slot}
}
\description{
a generic function to replace the slot Period
}
