% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TermStructure.R
\docType{methods}
\name{Period,TermStructure-method}
\alias{Period,TermStructure-method}
\title{Method to extract Period from S4 class}
\usage{
\S4method{Period}{TermStructure}(object)
}
\arguments{
\item{object}{the name of the S4 object}
}
\description{
Method to extract Period from S4 class
}

