% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelTune.R
\docType{methods}
\name{PlateauMonths,PrepaymentModelTune-method}
\alias{PlateauMonths,PrepaymentModelTune-method}
\title{A method to extract PlateauMonths from class PrepaymentModelTune}
\usage{
\S4method{PlateauMonths}{PrepaymentModelTune}(object)
}
\arguments{
\item{object}{The name of the object of class PrepaymentModelTune}
}
\description{
A method to extract PlateauMonths from class PrepaymentModelTune
}
