% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelTune.R
\name{PlateauMonths}
\alias{PlateauMonths}
\title{A standard generic function to access the slot PlateauMonths}
\usage{
PlateauMonths(object)
}
\arguments{
\item{object}{an S4 object}
}
\description{
A standard generic function to access the slot PlateauMonths
}

