% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageKeyRate.R
\name{PlotKeyRate}
\alias{PlotKeyRate}
\title{Plot Key Rate Duration}
\usage{
PlotKeyRate(object)
}
\arguments{
\item{object}{MortgageTermStructure object}
}
\description{
Generic function to plot key rate duration
}

