% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageKeyRate.R
\name{plotkrc,MortgageTermStructure-method}
\alias{plotkrc,MortgageTermStructure-method}
\title{Plot Mortgage Key Rate Convexity}
\usage{
\S4method{plotkrc}{MortgageTermStructure}(object)
}
\arguments{
\item{object}{MortgageTermStructure object}
}
\description{
A method to plot mortgage key rate convexity
}
