% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageKeyRate.R
\name{plotkrc}
\alias{plotkrc}
\title{Plot Key Rate Convexity}
\usage{
plotkrc(object)
}
\arguments{
\item{object}{MortgageTermStructure object}
}
\description{
Generic function to plot key rate convexity
}

