% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageKeyRate.R
\docType{methods}
\name{plotkrd,MortgageTermStructure-method}
\alias{plotkrd,MortgageTermStructure-method}
\title{Plot Mortgage Key Rate Duration}
\usage{
\S4method{plotkrd}{MortgageTermStructure}(object)
}
\arguments{
\item{object}{MortgageTermStructure object}
}
\description{
A method to plot mortgage key rate duration
}
