% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageKeyRate.R
\name{plotkrd}
\alias{plotkrd}
\title{Plot Key Rate Duration}
\usage{
plotkrd(object)
}
\arguments{
\item{object}{MortgageTermStructure object}
}
\description{
Generic function to plot key rate duration
}

