% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageOAS.R
\docType{methods}
\name{PlotOAS,MortgageOAS-method}
\alias{PlotOAS,MortgageOAS-method}
\title{Method to plot MortgageOAS spread distribution}
\usage{
\S4method{PlotOAS}{MortgageOAS}(object)
}
\arguments{
\item{object}{MortgageOAS object}
}
\description{
A method to plot MortgageOAS spread distribution
}

