% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBSDetails.R
\name{PMI,MBSDetails-method}
\alias{PMI,MBSDetails-method}
\title{Method to extract the slot PMI from the class MBSDetails}
\usage{
\S4method{PMI}{MBSDetails}(object)
}
\arguments{
\item{object}{the name of the object of the type MBSDetails}
}
\description{
Method to extract the slot PMI from the class MBSDetails
}
