% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMITypes.R
\docType{methods}
\name{PMIBasis<-,PMITypes-method}
\alias{PMIBasis<-,PMITypes-method}
\title{A method to set the value of the slot PMIBasis in the object PMITypes}
\usage{
\S4method{PMIBasis}{PMITypes}(object) <- value
}
\arguments{
\item{object}{an S4 object of the type PMITypes}

\item{value}{the replacement value of the slot}
}
\description{
A method to set the value of the slot PMIBasis in the object PMITypes
}
