% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMITypes.R
\name{PMIBasis}
\alias{PMIBasis}
\title{A standard generic function to acees the slot PMIBasis}
\usage{
PMIBasis(object)
}
\arguments{
\item{object}{an S4 object of the type PMITypes}
}
\description{
A standard generic function to acees the slot PMIBasis
}
