% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMITypes.R
\docType{methods}
\name{PMIDecimal,PMITypes-method}
\alias{PMIDecimal,PMITypes-method}
\title{A method to access the slot PMIDecimal from the object PMITypes}
\usage{
\S4method{PMIDecimal}{PMITypes}(object)
}
\arguments{
\item{object}{an S4 object of the type PMITypes}
}
\description{
A method to access the slot PMIDecimal from the object PMITypes
}

