% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMITypes.R
\name{PMIDecimal}
\alias{PMIDecimal}
\title{A standard generic function to get the slot PMIDecimal}
\usage{
PMIDecimal(object)
}
\arguments{
\item{object}{an S4 object of the type PMITypes}
}
\description{
A standard generic function to get the slot PMIDecimal
}

