% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMITypes.R
\docType{methods}
\name{PMIDecimalString,PMITypes-method}
\alias{PMIDecimalString,PMITypes-method}
\title{A method to get the slot PMIDecimalString from the object PMITypes}
\usage{
\S4method{PMIDecimalString}{PMITypes}(object)
}
\arguments{
\item{object}{an S4 object of the type PMITypes}
}
\description{
A method to get the slot PMIDecimalString from the object PMITypes
}
