% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMITypes.R
\docType{methods}
\name{PMIDecimalString<-,PMITypes-method}
\alias{PMIDecimalString<-,PMITypes-method}
\title{A method to replace the value of the slot PMIDecimalString in the object PMITypes}
\usage{
\S4method{PMIDecimalString}{PMITypes}(object) <- value
}
\arguments{
\item{object}{an S4 object of the type PMITypes}

\item{value}{the replacement value of the slot}
}
\description{
A method to replace the value of the slot PMIDecimalString in the object PMITypes
}

