% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMITypes.R
\name{PMIDecimalString<-}
\alias{PMIDecimalString<-}
\title{A standard generic function to replace the value of the slot PMIDecimalString}
\usage{
PMIDecimalString(object) <- value
}
\arguments{
\item{object}{an S4 object of the type PMITypes}

\item{value}{the replacement value of the slot}
}
\description{
A standard generic function to replace the value of the slot PMIDecimalString
}

