% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMITypes.R
\name{PMIDecimalString}
\alias{PMIDecimalString}
\title{A standard generic function to get the slot PMIDecimalString}
\usage{
PMIDecimalString(object)
}
\arguments{
\item{object}{an S4 object of the type PMIDecimalString}
}
\description{
A standard generic function to get the slot PMIDecimalString
}

