% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageCashFlow.R
\docType{methods}
\name{PMIPremium,MortgageCashFlow-method}
\alias{PMIPremium,MortgageCashFlow-method}
\title{Method to extract the PMIPremium from the class MortgageCashFlow}
\usage{
\S4method{PMIPremium}{MortgageCashFlow}(object)
}
\arguments{
\item{object}{the name of the object of type MortgageCashFlow}
}
\description{
Method to extract the PMIPremium from the class MortgageCashFlow
}
