% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMITypes.R
\name{PMITypes}
\alias{PMITypes}
\title{PMITypes is a constructor function for the PMITypes class}
\usage{
PMITypes(PMI = numeric())
}
\arguments{
\item{PMI}{a numeric value the PMI paid to the insurer by the lender
expressed in decimal numeric}
}
\description{
PMITypes is a constructor function for the PMITypes class
}
