% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BillCashFlows.R
\docType{methods}
\name{PmtDate,BillCashFlows-method}
\alias{PmtDate,BillCashFlows-method}
\title{Method to get PmtDate from S4 class BillCashFlows}
\usage{
\S4method{PmtDate}{BillCashFlows}(object)
}
\arguments{
\item{object}{the name of the S4 object BillCashFlows}
}
\description{
Method to get PmtDate from S4 class BillCashFlows
}
