% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageCashFlow.R
\docType{methods}
\name{PmtDate,MortgageCashFlow-method}
\alias{PmtDate,MortgageCashFlow-method}
\title{Method to extract PmtDate from S4 class MortgageCashFlow}
\usage{
\S4method{PmtDate}{MortgageCashFlow}(object)
}
\arguments{
\item{object}{the name of the S4 object}
}
\description{
Method to extract PmtDate from S4 class MortgageCashFlow
}

