% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModel.R
\name{PmtDate,PrepaymentModel-method}
\alias{PmtDate,PrepaymentModel-method}
\title{A method to extract PmtDate from S4 class PrepaymentModel}
\usage{
\S4method{PmtDate}{PrepaymentModel}(object)
}
\arguments{
\item{object}{the name of the S4 class}
}
\description{
A method to extract PmtDate from S4 class PrepaymentModel
}
