% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Functions.R
\name{PPC.Ramp}
\alias{PPC.Ramp}
\title{A function to calculate the PPC ramp}
\usage{
PPC.Ramp(season.period = numeric(), begin.cpr = numeric(),
  end.cpr = numeric(), period = numeric())
}
\arguments{
\item{season.period}{A numeric value the time to the peak PPC ramp}

\item{begin.cpr}{A numeric value the starting CPR value}

\item{end.cpr}{A numeric value the ending CPR value}

\item{period}{A numeric value the current period (months from start date)}
}
\description{
A standard generic function for the calculation of the PPC ramp
}
\examples{
PPC.Ramp(season.period = 30, begin.cpr = .002, end.cpr = .06, period = 30)
}

