% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModel.R
\docType{methods}
\name{PPCEnd,PrepaymentModel-method}
\alias{PPCEnd,PrepaymentModel-method}
\title{A method to extract PPCEnd from S4 class PrepaymentModel}
\usage{
\S4method{PPCEnd}{PrepaymentModel}(object)
}
\arguments{
\item{object}{the name of the S4 class}
}
\description{
A method to extract PPCEnd from S4 class PrepaymentModel
}

