% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModel.R
\name{PPCEnd}
\alias{PPCEnd}
\title{A standard generic function to access the slot PPCEnd}
\usage{
PPCEnd(object)
}
\arguments{
\item{object}{An S4 class object of the type PrepaymentAssumption
export PPCEnd}
}
\description{
A standard generic function to access the slot PPCEnd
}

