% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModel.R
\docType{methods}
\name{PPCSeasoning,PrepaymentAssumption-method}
\alias{PPCSeasoning,PrepaymentAssumption-method}
\title{A method to extract PPCSeasoning from S4 class PrepaymentAssumption}
\usage{
\S4method{PPCSeasoning}{PrepaymentAssumption}(object)
}
\arguments{
\item{object}{the name of the S4 class}
}
\description{
A method to extract PPCSeasoning from S4 class PrepaymentAssumption
}

