% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModel.R
\name{PPCSeasoning}
\alias{PPCSeasoning}
\title{A standard generic function to access the slot PPCSeasoning}
\usage{
PPCSeasoning(object)
}
\arguments{
\item{object}{An S4 class object of the type PrepaymentAssumption
export PPCSeasoning}
}
\description{
A standard generic function to access the slot PPCSeasoning
}

