% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModel.R
\name{PPCStart,PrepaymentModel-method}
\alias{PPCStart,PrepaymentModel-method}
\title{A method to extract PPCStart from S4 class PrepaymentModel}
\usage{
\S4method{PPCStart}{PrepaymentModel}(object)
}
\arguments{
\item{object}{the name of the S4 class}
}
\description{
A method to extract PPCStart from S4 class PrepaymentModel
}
