% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModel.R
\name{PPCStart}
\alias{PPCStart}
\title{A standard generic function to access the slot PPCstart}
\usage{
PPCStart(object)
}
\arguments{
\item{object}{An S4 class object of the type PrepaymentModel
export PPCStart}
}
\description{
A standard generic function to access the slot PPCstart
}

