% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageCashFlow.R
\docType{methods}
\name{PrepaidPrin,MortgageCashFlow-method}
\alias{PrepaidPrin,MortgageCashFlow-method}
\title{Method to extract Prepaid Principal from class MortgageCashFlow}
\usage{
\S4method{PrepaidPrin}{MortgageCashFlow}(object)
}
\arguments{
\item{object}{the name of the object of type MortgageCashFlow}
}
\description{
Method to extract Prepaid Principal from class MortgageCashFlow
}

