% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageCashFlow.R
\name{PrepaidPrin}
\alias{PrepaidPrin}
\title{A generic function to access to the slot PrepaidPrin}
\usage{
PrepaidPrin(object)
}
\arguments{
\item{object}{an S4 class object of the type MortgageCashFlow}
}
\description{
A generic function to access to the slot PrepaidPrin
}

