% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{PrepaymentAssumption-class}
\alias{PrepaymentAssumption-class}
\title{An S4 class representing the prepayment vector
given by the BondLab prepayment model}
\description{
An S4 class representing the prepayment vector
given by the BondLab prepayment model
}

