% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModel.R
\docType{methods}
\name{PrepaymentAssumption,PrepaymentAssumption-method}
\alias{PrepaymentAssumption,PrepaymentAssumption-method}
\title{A method to extact PrepaymentAssumption from S4 class PrepaymentAssumption}
\usage{
\S4method{PrepaymentAssumption}{PrepaymentAssumption}(object)
}
\arguments{
\item{object}{the name of the S4 class}
}
\description{
A method to extact PrepaymentAssumption from S4 class PrepaymentAssumption
}

