% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModel.R
\docType{methods}
\name{PrepaymentAssumption,PrepaymentModel-method}
\alias{PrepaymentAssumption,PrepaymentModel-method}
\title{A method to extact PrepaymentAssumption from S4 class PrepaymentModel}
\usage{
\S4method{PrepaymentAssumption}{PrepaymentModel}(object)
}
\arguments{
\item{object}{the name of the S4 class}
}
\description{
A method to extact PrepaymentAssumption from S4 class PrepaymentModel
}

