% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModel.R
\name{PrepaymentAssumption}
\alias{PrepaymentAssumption}
\title{A standard generic function to access the slot PrepaymentAssumption}
\usage{
PrepaymentAssumption(object)
}
\arguments{
\item{object}{An S4 class object of the type PrepaymentModel
export PrepaymentAssumption}
}
\description{
A standard generic function to access the slot PrepaymentAssumption
}

