% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModel.R
\name{PrepaymentModel}
\alias{PrepaymentModel}
\title{A contstructor function for the PrepaymentModel object}
\usage{
PrepaymentModel(bond.id, term.structure, ModelTune, Burnout = 0,
  PrepaymentAssumption = "MODEL", ..., begin.cpr = 0.2, end.cpr = 6,
  seasoning.period = 30, cpr = 6, cdr = 1, severity = 0.25)
}
\arguments{
\item{bond.id}{A character string referring to an object 
of the type MBSDetails}

\item{term.structure}{A character string referring to an object 
of the type TermStructure}

\item{ModelTune}{A character string the prepayment model object}

\item{Burnout}{A numeric value the burnout variable}

\item{PrepaymentAssumption}{A character string the prepayment assumption 
used "model", "ppc", or "cpr"}

\item{...}{Optional values when "ppc" or "cpr" is used}

\item{begin.cpr}{A numeric value the beginning CPR assumption}

\item{end.cpr}{A numeric value the ending CPR assumption}

\item{seasoning.period}{A numeric value the length of the seasoning ramp}

\item{cpr}{A numeric value the CPR assumption (annual prepayment rate)}

\item{cdr}{A numeric value the CDR assumption (annual default rate)}

\item{severity}{A numeric value the loss severity given default}
}
\description{
The function is a constructor function for the PrepaymentModel object
}
