% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelFunctions.R
\name{PrepaymentModelFunctions}
\alias{PrepaymentModelFunctions}
\title{A constructor function for the class ModelFunctions}
\usage{
PrepaymentModelFunctions(TurnoverRate = "numeric",
  SeasoningRamp = "function", Curtailment = "function",
  SeasonalFactors = "function", ArcTanIncentive = "function",
  BorrowerBurnout = "function", DefaultRamp = "function",
  DefaultOrigLTVMult = "function", DefaultUpdatedLTVMult = "function",
  DefaultSATOMult = "function")
}
\arguments{
\item{TurnoverRate}{a numeric value the turnover rate}

\item{SeasoningRamp}{a function defining the seasoning ramp.  The input
variables are alpha the aymptote, beta the intercept, theta the point of
maximum curvature.}

\item{Curtailment}{a function defining the curtailment}

\item{SeasonalFactors}{a function defining the seasonal pattern of prepayments
the input variables are alpha the function's maximum value, Month the numeric
value of the month of the year, theta the point at which the function reaches 
its maximum value}

\item{ArcTanIncentive}{a function defining the borrower incentive response}

\item{BorrowerBurnout}{a function defining borrower burnout}

\item{DefaultRamp}{a function defining the default ramp}

\item{DefaultOrigLTVMult}{a function defining the orig. LTV multipliers}

\item{DefaultUpdatedLTVMult}{a function defining the updated LTV default multipliers}

\item{DefaultSATOMult}{a function defining the SATO multipliers on the default curve}
}
\description{
A constructor function for the class ModelFunctions
}
