% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelTune.R
\docType{class}
\name{PrepaymentModelTune-class}
\alias{PrepaymentModelTune-class}
\title{An S4 class representing the prepayment model tuning parameters}
\description{
An S4 class representing the prepayment model tuning parameters
}
\section{Slots}{

\describe{
\item{\code{TurnoverRate}}{A numeric value the estimated housing turnover rate}

\item{\code{Turnover.alpha}}{A numeric value the asypmtote of the seasoning ramp}

\item{\code{Turnover.beta}}{A numeric value the intercept of the seasoning ramp}

\item{\code{Turnover.theta}}{A numeric value the point of maximum curvature of 
the seasoning ramp}

\item{\code{Seasonality.alpha}}{A numeric value setting the maximum value}

\item{\code{Seasonality.theta}}{A numeric value setting the point at which the 
function reaches its maximum value}

\item{\code{Incentive.Fast.theta.1}}{A numeric value used to tune the ArcTanIncentive Function}

\item{\code{Incentive.Fast.theta.2}}{A numeric value used to tune the ArcTanIncentive Function}

\item{\code{Incentive.Fast.beta}}{A numeric value used to tune the ArcTanIncentive Function}

\item{\code{Incentive.Fast.eta}}{A numeric value used to tune the ArcTanIncentive Function}

\item{\code{Incentive.Slow.theta.1}}{A numeric value used to tune the ArcTanIncentive Function}

\item{\code{Incentive.Slow.theta.2}}{A numeric value used to tune the ArcTanIncentive Function}

\item{\code{Incentive.Slow.beta}}{A numeric value used to tune the ArcTanIncentive Function}

\item{\code{Incentive.Slow.eta}}{A numeric value used to tune the ArcTanIncentive Function}

\item{\code{Burnout.beta.1}}{A numeric value used to tune the Burnout Function}

\item{\code{Burnout.beta.2}}{A numeric value used to tune the Burnout Function}

\item{\code{BeginCDR}}{A numeric value beginning CDR of the default ramp}

\item{\code{PeakCDR}}{A numeric value the Peak CDR of the default ramp}

\item{\code{EndCDR}}{A numeric value the Ending CDR of the default ramp}

\item{\code{PeakMonth}}{A numeric value the month at which the default ramp peaks}

\item{\code{PlateauMonths}}{A numeric value the number of months before the CDR declines}

\item{\code{EndMonth}}{A numeric value the ending point of the CDR ramp decline}

\item{\code{MinOrigLTV}}{the minimum LTV for which to apply the default multiplier}

\item{\code{MaxOrigLTV}}{the maximum LTV for which to apply the default multiplier}

\item{\code{MaxOrigMultiplier}}{the maximum orig LTV multiplier}

\item{\code{MinOrigMultiplier}}{the minimum orig LTV multiplier}

\item{\code{UpdatedLTV.beta}}{A numeric value the beta of ULTV function}

\item{\code{SATO.beta}}{A numeric value the beta of the SATO function}
}}

