% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageCashFlow.R
\name{PrepaymentScenario,PriceYieldScenarioSet-method}
\alias{PrepaymentScenario,PriceYieldScenarioSet-method}
\title{A method to extract the PrepaymentScenario from class PriceYieldScenarioSet}
\usage{
\S4method{PrepaymentScenario}{PriceYieldScenarioSet}(object)
}
\arguments{
\item{object}{an S4 object of the type PriceYieldScenarioSet}
}
\description{
A method to extract the PrepaymentScenario from class PriceYieldScenarioSet
}
