% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageCashFlow.R
\name{PrepaymentScenario}
\alias{PrepaymentScenario}
\title{A generic function to access the slot PrepaymentScenario}
\usage{
PrepaymentScenario(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A generic function to access the slot PrepaymentScenario
}

