% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageCashFlow.R
\docType{methods}
\name{PrepaymentType,PriceYieldScenarioSet-method}
\alias{PrepaymentType,PriceYieldScenarioSet-method}
\title{A method to extract the PrepaymentType from class PriceYieldScenarioSet}
\usage{
\S4method{PrepaymentType}{PriceYieldScenarioSet}(object)
}
\arguments{
\item{object}{an S4 object of the typre PriceYieldScenarioSet}
}
\description{
A method to extract the PrepaymentType from class PriceYieldScenarioSet
}

