% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageCashFlow.R
\name{PrepaymentType}
\alias{PrepaymentType}
\title{A generic functon to access the slot PrepaymentType}
\usage{
PrepaymentType(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A generic functon to access the slot PrepaymentType
}

