% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BillCashFlows.R
\name{Price,BillCashFlows-method}
\alias{Price,BillCashFlows-method}
\title{Method to get price from S4 class BillCashFlows}
\usage{
\S4method{Price}{BillCashFlows}(object)
}
\arguments{
\item{object}{the name of the S4 object BillCashFlows}
}
\description{
Method to get price from S4 class BillCashFlows
}
