% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PriceYieldClasses.R
\docType{class}
\name{Price-class}
\alias{Price-class}
\title{An S4 class representating bond price}
\description{
An S4 class representating bond price
}
\section{Slots}{

\describe{
\item{\code{PriceDecimal}}{A numeric value the price using decimal notation}

\item{\code{Price32nds}}{A character the price using 32nds notation}

\item{\code{PriceBasis}}{A numeric value price decimal notation in units of 100}
}}

