% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageCashFlow.R
\docType{methods}
\name{Price,MortgageCashFlow-method}
\alias{Price,MortgageCashFlow-method}
\title{Method to extract Price from S4 class MortgageCashFlow}
\usage{
\S4method{Price}{MortgageCashFlow}(object)
}
\arguments{
\item{object}{the name of the S4 object}
}
\description{
Method to extract Price from S4 class MortgageCashFlow
}

