% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\docType{methods}
\name{Price,MtgScenario-method}
\alias{Price,MtgScenario-method}
\title{A method to extract Price from S4 class MtgScenario}
\usage{
\S4method{Price}{MtgScenario}(object)
}
\arguments{
\item{object}{the name of an S4 class on type MtgScenario}
}
\description{
A method to extract Price from S4 class MtgScenario
}

