% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PriceTypes.R
\docType{methods}
\name{Price32nds,PriceTypes-method}
\alias{Price32nds,PriceTypes-method}
\title{A method to extract Price32nds from slot of class PriceTypes}
\usage{
\S4method{Price32nds}{PriceTypes}(object)
}
\arguments{
\item{object}{an S4 object of the type PriceTypes}
}
\description{
A method to extract Price32nds from slot of class PriceTypes
}
