% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PriceTypes.R
\name{Price32nds<-,PriceTypes-method}
\alias{Price32nds<-,PriceTypes-method}
\title{A method to set Price32nds in slot of class PriceTypes}
\usage{
\S4method{Price32nds}{PriceTypes}(object) <- value
}
\arguments{
\item{object}{an S4 object of the type PriceTypes}

\item{value}{the replacement value of the slot}
}
\description{
A method to set Price32nds in slot of class PriceTypes
}
