% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PriceTypes.R
\name{Price32nds<-}
\alias{Price32nds<-}
\title{A standard generic function to set the slot Price32nds}
\usage{
Price32nds(object) <- value
}
\arguments{
\item{object}{an S4 object}

\item{value}{the replacement value of the slot}
}
\description{
A standard generic function to set the slot Price32nds
}

