% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PriceTypes.R
\name{Price32nds}
\alias{Price32nds}
\title{A standard generic function to get the slot Price32nds}
\usage{
Price32nds(object)
}
\arguments{
\item{object}{an S4 object}
}
\description{
A standard generic function to get the slot Price32nds
}
