% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PriceTypes.R
\name{PriceBasis,PriceTypes-method}
\alias{PriceBasis,PriceTypes-method}
\title{A method to extract PriceBasis from slot of class PriceTypres}
\usage{
\S4method{PriceBasis}{PriceTypes}(object)
}
\arguments{
\item{object}{an S4 object of the type PriceType}
}
\description{
A method to extract PriceBasis from slot of class PriceTypres
}
