% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PriceTypes.R
\docType{methods}
\name{PriceBasis<-,PriceTypes-method}
\alias{PriceBasis<-,PriceTypes-method}
\title{a method to set PriceBasis from slot of class PriceTypes}
\usage{
\S4method{PriceBasis}{PriceTypes}(object) <- value
}
\arguments{
\item{object}{an S4 object of the type PriceTypes}

\item{value}{the replacement value of the slot}
}
\description{
a method to set PriceBasis from slot of class PriceTypes
}

