% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PriceTypes.R
\name{PriceBasis<-}
\alias{PriceBasis<-}
\title{A standard generic to replace the slot PriceBasis}
\usage{
PriceBasis(object) <- value
}
\arguments{
\item{object}{A S4 object of type PriceTypes}

\item{value}{the replacement value of the slot}
}
\description{
A standard generic to replace the slot PriceBasis
}

