% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PriceYieldClasses.R
\name{PriceBasis}
\alias{PriceBasis}
\title{A standard generic to access the slot PriceBasis}
\usage{
PriceBasis(object)
}
\arguments{
\item{object}{an S4 object}
}
\description{
A standard generic to access the slot PriceBasis
}

