% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PriceClasses.R
\docType{methods}
\name{PriceDecimal,PriceTypes-method}
\alias{PriceDecimal,PriceTypes-method}
\title{A method to extract PriceDecimal from slot of class PriceTypes}
\usage{
\S4method{PriceDecimal}{PriceTypes}(object)
}
\arguments{
\item{object}{an S4 object of the type PriceTypes}
}
\description{
A method to extract PriceDecimal from slot of class PriceTypes
}

