% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PriceTypes.R
\docType{methods}
\name{PriceDecimal<-,PriceTypes-method}
\alias{PriceDecimal<-,PriceTypes-method}
\title{A method to replace PriceDecimal in slot of class PriceTypes}
\usage{
\S4method{PriceDecimal}{PriceTypes}(object) <- value
}
\arguments{
\item{object}{an S4 object of the typre PriceTypes}

\item{value}{the replacement value of the slot}
}
\description{
A method to replace PriceDecimal in slot of class PriceTypes
}

