% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PriceTypes.R
\name{PriceDecimal<-}
\alias{PriceDecimal<-}
\title{A standard generic function to replace the slot PriceDecimal}
\usage{
PriceDecimal(object) <- value
}
\arguments{
\item{object}{an S4 object}

\item{value}{the replacement value of the slot}
}
\description{
A standard generic function to replace the slot PriceDecimal
}

