% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PriceYieldClasses.R
\name{PriceDecimal}
\alias{PriceDecimal}
\title{A standard generic function access the slot PriceDecimal}
\usage{
PriceDecimal(object)
}
\arguments{
\item{object}{an S4 object}
}
\description{
A standard generic function access the slot PriceDecimal
}

