% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PriceTypes.R
\name{PriceDecimalString,PriceTypes-method}
\alias{PriceDecimalString,PriceTypes-method}
\title{A method to extract PriceDecimalString from slot of class PriceTypes}
\usage{
\S4method{PriceDecimalString}{PriceTypes}(object)
}
\arguments{
\item{object}{an S4 object of type PriceTypes}
}
\description{
A method to extract PriceDecimalString from slot of class PriceTypes
}
