% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PriceTypes.R
\name{PriceDecimalString}
\alias{PriceDecimalString}
\title{A standard generic to get the slot PriceDecimalString}
\usage{
PriceDecimalString(object)
}
\arguments{
\item{object}{A S4 object of the type PriceTypes}
}
\description{
A standard generic to get the slot PriceDecimalString
}

