% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageOAS.R
\docType{methods}
\name{PriceDist,MortgageOAS-method}
\alias{PriceDist,MortgageOAS-method}
\title{A method to extract PriceDist from S4 class MortgageOAS}
\usage{
\S4method{PriceDist}{MortgageOAS}(object)
}
\arguments{
\item{object}{an S4 object of the type MortgageOAS}
}
\description{
A method to extract PriceDist from S4 class MortgageOAS
}

