% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageOAS.R
\name{PriceDist}
\alias{PriceDist}
\title{A standard generic function to access the slot PriceDist}
\usage{
PriceDist(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A standard generic function to access the slot PriceDist
}

