% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscountToPrice.R
\name{PriceToDiscountYield}
\alias{PriceToDiscountYield}
\title{Price to Discount Yield bill}
\usage{
PriceToDiscountYield(bill.id, price, day.count, settlement.date)
}
\arguments{
\item{bill.id}{A character string the cusip number or bill.id}

\item{price}{A character string the bill price}

\item{day.count}{A numeric value the day count used for pricing}

\item{settlement.date}{A character the settlement date mm-dd-YYYY}
}
\description{
Returns the discount yield of bill given price
}
\seealso{
Other Pricing: \code{\link{BenchMark,CurveSpreads-method}},
  \code{\link{CurveSpreads-class}},
  \code{\link{CurveSpreads}},
  \code{\link{DiscountToPrice}},
  \code{\link{PriceToBondYield}},
  \code{\link{SpreadToBenchmark,CurveSpreads-method}},
  \code{\link{SpreadToCurve,CurveSpreads-method}},
  \code{\link{SpreadToPriceBond}},
  \code{\link{SpreadToPriceMBS}},
  \code{\link{ZeroVolSpread,CurveSpreads-method}}
}
