% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PriceTypes.R
\name{PriceTypes}
\alias{PriceTypes}
\title{PriceTypes is a constructor function for the class PriceTypes}
\usage{
PriceTypes(price)
}
\arguments{
\item{price}{character the price in either 
decimal notation (example "100.125") or 32nds notation (example "100-4")}
}
\description{
PriceTypes is a constructor function for the class PriceTypes
}

