% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageCashFlow.R
\name{PriceYieldScenario,PriceYieldScenarioSet-method}
\alias{PriceYieldScenario,PriceYieldScenarioSet-method}
\title{A method to extract a Price/Yield Scenario Pair}
\usage{
\S4method{PriceYieldScenario}{PriceYieldScenarioSet}(object, scenario = numeric())
}
\arguments{
\item{object}{an S4 object of type PriceYieldScenarioSet}

\item{scenario}{an numeric value the location of the scenario}
}
\description{
A method to extract a Price/Yield Scenario Pair
}
