% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageCashFlow.R
\name{PriceYieldScenario}
\alias{PriceYieldScenario}
\title{generic function to access both slots of PrepaymentScenario}
\usage{
PriceYieldScenario(object, ...)
}
\arguments{
\item{object}{an S4 class object}

\item{...}{optional arguments}
}
\description{
generic function to access both slots of PrepaymentScenario
}

