% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageCashFlow.R
\docType{class}
\name{PriceYieldScenarioSet-class}
\alias{PriceYieldScenarioSet-class}
\title{An S4 class PassThroughScenarios}
\description{
An S4 class PassThroughScenarios
}
\section{Slots}{

\describe{
\item{\code{PrepaymentType}}{A character the prepayment type (i.e. "CPR", "MODEL")}

\item{\code{PrepaymentScenario}}{A character prepayment value (i.e. "25", "D25s")}
}}

