% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageCashFlow.R
\name{PriceYieldScenarioSet}
\alias{PriceYieldScenarioSet}
\title{A function to convert Price Yield data to class PriceYieldScenarioSet}
\usage{
PriceYieldScenarioSet(PrepaymentType = "list", PrepaymentScenario = "list")
}
\arguments{
\item{PrepaymentType}{a list of the prepayment type}

\item{PrepaymentScenario}{a list of the prepayment scenario corresponding
to the prepayment type}
}
\description{
A function to convert Price Yield data to class PriceYieldScenarioSet
}

