% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{PricingSpeed,RAID-method}
\alias{PricingSpeed,RAID-method}
\title{A method to access the slot PricingSpeed from class RAID}
\usage{
\S4method{PricingSpeed}{RAID}(object)
}
\arguments{
\item{object}{an S4 object of the type RAID}
}
\description{
A method to access the slot PricingSpeed from class RAID
}

