% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BillCashFlows.R
\name{PrincipalOutstanding,BillCashFlows-method}
\alias{PrincipalOutstanding,BillCashFlows-method}
\title{Method to get Principal outstanding from S4 class BillCashFlows}
\usage{
\S4method{PrincipalOutstanding}{BillCashFlows}(object)
}
\arguments{
\item{object}{the name of the S$ object BillCashFlows}
}
\description{
Method to get Principal outstanding from S4 class BillCashFlows
}
