% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondCashFlows.R
\name{PrincipalOutstanding,BondCashFlows-method}
\alias{PrincipalOutstanding,BondCashFlows-method}
\title{Method to get PrincipalOutstanding from S4 class BondCashFlows}
\usage{
\S4method{PrincipalOutstanding}{BondCashFlows}(object)
}
\arguments{
\item{object}{the name of the S4 object}
}
\description{
Method to get PrincipalOutstanding from S4 class BondCashFlows
}
