% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrincipalPmtDateCalcs.R
\name{PrincipalPmtDate}
\alias{PrincipalPmtDate}
\title{A function to report the first and last principal payment dates}
\usage{
PrincipalPmtDate(
  date.vector = "character",
  principal.vector = "character",
  type = "character"
)
}
\arguments{
\item{date.vector}{A character string referencing the payment date slot of
class PassThrough.}

\item{principal.vector}{A character string referencing a principal payment slot}

\item{type}{A character string the payment date either "FirstPmt" or "LastPmt"}
}
\description{
A function to report the first and last principal payment dates
}
