% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondScenario.R
\name{PrincipalReceived}
\alias{PrincipalReceived}
\title{PrincipalReceived generic}
\usage{
PrincipalReceived(object)
}
\arguments{
\item{object}{an object of the type BondScenario}
}
\description{
A  generic function for method dispatch
}
\seealso{
Other Scenario Analysis: \code{\link{CouponIncome}},
  \code{\link{HorizonCurrBal}},
  \code{\link{HorizonCurve-class}},
  \code{\link{HorizonCurve}}, \code{\link{HorizonMos}},
  \code{\link{HorizonPrice}}, \code{\link{HorizonReturn}},
  \code{\link{Name,Scenario-method}},
  \code{\link{PrepaidPrinReceived}},
  \code{\link{ReinvestmentIncome}},
  \code{\link{Scenario-class}},
  \code{\link{ScheduledPrinReceived}}
}
