% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\name{PrinPmtFrequency}
\alias{PrinPmtFrequency}
\title{A generic function to access the slot PrinPmtFrequency}
\usage{
PrinPmtFrequency(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A generic function to access the slot PrinPmtFrequency
}

