% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModel.R
\name{ProjectMortgageRate}
\alias{ProjectMortgageRate}
\title{ProjectMortgage Rate}
\usage{
ProjectMortgageRate(bond.id, term.structure)
}
\arguments{
\item{bond.id}{an object of the type MBSDetails}

\item{term.structure}{an object of the type TermStructure}
}
\description{
Projects the forward mortgage rate used by the Bond Lab prepayment
model based on the two and ten year forward rates
}

