% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondConstructor.R
\docType{methods}
\name{Putable,BondDetails-method}
\alias{Putable,BondDetails-method}
\title{Method to extract the slot Putable from S4 class of type BondDetails}
\usage{
\S4method{Putable}{BondDetails}(object)
}
\arguments{
\item{object}{the name of the object of type BondDetials}
}
\description{
Method to extract the slot Putable from S4 class of type BondDetails
}

