% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondDetails.R
\name{Putable}
\alias{Putable}
\title{A standard generic function to access the slot Putable}
\usage{
Putable(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A standard generic function to access the slot Putable
}

