% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageOAS.R
\docType{methods}
\name{RatePaths,MortgageOAS-method}
\alias{RatePaths,MortgageOAS-method}
\title{A method to extact RatePaths from S4 class MortgageOAS}
\usage{
\S4method{RatePaths}{MortgageOAS}(object)
}
\arguments{
\item{object}{an S4 object of the type MortgageOAS}
}
\description{
A method to extact RatePaths from S4 class MortgageOAS
}

