% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageOAS.R
\name{RatePaths}
\alias{RatePaths}
\title{A standard generic function to access the slot RatePaths}
\usage{
RatePaths(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A standard generic function to access the slot RatePaths
}

