% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\name{RDMEData}
\alias{RDMEData}
\title{A function to aggregate monthly updated factor data}
\usage{
RDMEData(NumberofTranches = numeric(), DealName = "character")
}
\arguments{
\item{NumberofTranches}{A numeric value the Number of traches related to the collateral group}

\item{DealName}{A character string the Deal Name}
}
\description{
Aggregator function used by the REMIC constructor to aggregate Monthly Factor Data
}

