% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ConnectionFunctions.R
\name{RDMEFactor}
\alias{RDMEFactor}
\title{A read connection to RDME folder}
\usage{
RDMEFactor(DealName = "character", TrancheNumber = numeric())
}
\arguments{
\item{DealName}{a character string the Deal Name}

\item{TrancheNumber}{a numeric value the tranche number}
}
\description{
The function opens a connection for the aggreation of factor data
which is used by the REMIC constructor
}

