% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ConnectionFunctions.R
\name{ReadRAID}
\alias{ReadRAID}
\title{A connection function to read the RAID information}
\usage{
ReadRAID(RAIDFile = "character")
}
\arguments{
\item{RAIDFile}{a character string the REMIC At Issuance Disclosure}
}
\description{
Opens a connection to the RAID folder and saves RAID information
REMIC at Issuance Disclouse data
}

