% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageCashFlow.R
\docType{methods}
\name{RecoveredAmount,MortgageCashFlow-method}
\alias{RecoveredAmount,MortgageCashFlow-method}
\title{Method to extract the Recovered Amount from class MortgageCashFlow}
\usage{
\S4method{RecoveredAmount}{MortgageCashFlow}(object)
}
\arguments{
\item{object}{the name of the object of type MortgageCashFlow}
}
\description{
Method to extract the Recovered Amount from class MortgageCashFlow
}

