% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageCashFlow.R
\name{RecoveredAmount}
\alias{RecoveredAmount}
\title{A standard generic function to access the slot RecoveredAmount}
\usage{
RecoveredAmount(object)
}
\arguments{
\item{object}{an S4 class object of the type MortgageCashFlow}
}
\description{
A standard generic function to access the slot RecoveredAmount
}

