% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\docType{methods}
\name{ReinvestmentIncome,MortgageReturn-method}
\alias{ReinvestmentIncome,MortgageReturn-method}
\title{A method to extract ReinvestmentIncome from S4 class MortgageReturn}
\usage{
\S4method{ReinvestmentIncome}{MortgageReturn}(object)
}
\arguments{
\item{object}{the name of an S4 class of type MortgageReturn}
}
\description{
A method to extract ReinvestmentIncome from S4 class MortgageReturn
}

