% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Functions.R
\name{Remain.Balance}
\alias{Remain.Balance}
\title{Remain.Balance}
\usage{
Remain.Balance(season.period = numeric(), begin.cpr = numeric(),
  end.cpr = numeric(), period = numeric())
}
\arguments{
\item{period}{A numeric value the numeric of months from the first payment}

\item{balance}{A numeric value the balance of the mortgage}

\item{note.rate}{A numeric value the borrower's note rate}

\item{term.mos}{A numeric value the term of the mortgage}
}
\description{
A standard generic for the calculation of a mortgage remaining balance
}
\examples{
Remain.Balance(balance = 100000, note.rate = .045, term.mos = 360, period = 10)
}

