% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/REMICOAS.R
\name{REMIC.OAS}
\alias{REMIC.OAS}
\title{The REMIC OAS engine. A constructor function of the Mortgage OAS object}
\usage{
REMIC.OAS(bond.id = "character", trade.date = "character",
  settlement.date = "character", tranche.price = numeric(),
  collateral.price = numeric(), sigma = numeric(), paths = numeric())
}
\arguments{
\item{bond.id}{A character string the bond.id or cusip}

\item{trade.date}{A character string the trade date}

\item{settlement.date}{A character string the settlement date}

\item{tranche.price}{A numeric value the tranche price}

\item{collateral.price}{A numeric value the collateral price}

\item{sigma}{A numeric value the volatility (not annualized)}

\item{paths}{A numeric value the number of paths}
}
\description{
REMIC.OAS calculates the option adjusted spread of a REMIC
and creates the MortgageOAS ibject
}
\examples{
REMIC.OAS(bond.id = "BondLabSEQ1", trade.date = "01-10-2013",
settlement.date = "01-13-2013", tranche.price = 100, collateral.price = 100,
sigma = 0.015,paths = 200)
}

