% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/REMICCashFlow.R
\name{REMICCashFlow}
\alias{REMICCashFlow}
\title{The REMIC cash flow engine the constructor function for REMICCashFlow oject}
\usage{
REMICCashFlow(bond.id = "character", trade.date = "character",
  settlement.date = "character", collateral.price = numeric(),
  tranche.price = numeric(), PrepaymentAssumption = "character", ...,
  begin.cpr = numeric(), end.cpr = numeric(),
  seasoning.period = numeric(), CPR = numeric(),
  KeyRateTermStructure = NULL)
}
\arguments{
\item{bond.id}{A character string the cusip of bond id}

\item{trade.date}{A character string the trade date}

\item{settlement.date}{A character string the settlement date}

\item{collateral.price}{A numeric value the price of the collateral}

\item{tranche.price}{A numeric value the price of the tranche}

\item{PrepaymentAssumption}{A character string the prepayment assumption}

\item{...}{Optional values when CPR or PPC input is used}

\item{begin.cpr}{A numeric value the begin CPR of the PPC ramp}

\item{end.cpr}{A numeric value the end CPR of the PPC ramp}

\item{seasoning.period}{A numeric value the length of the PPC ramp seasoning period}

\item{CPR}{A numeric value the CPR assumption used}

\item{KeyRateTermStructure}{default value NULL used for REMIC analytics do not overrided}
}
\description{
The function is the constructor function for REMICCashFlow Object
}
\examples{
\dontrun{
REMICCashFlow(bond.id = "BondLabSMBSIO", trade.date = "01-10-2013",
settlement.date = "01-13-2013", method = "ns", collateral.price = 105.75,
tranche.price = as.numeric(price[i]), PrepaymentAssumption = "CPR", CPR = 6)}
}

