% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ReadConnectionFunctions.R
\name{REMICDeal}
\alias{REMICDeal}
\title{A connection function to the REMICData folder to call REMICs by deal}
\usage{
REMICDeal(remic.deal = "character")
}
\arguments{
\item{remic.deal}{A character string the REMIC deal name}
}
\description{
Opens connection to the REMICData folder calling a REMIC Deal by name
}

