% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ConnectionFunctions.R
\name{REMICGroupConn}
\alias{REMICGroupConn}
\title{A connection function to the groups folder}
\usage{
REMICGroupConn(DealName = "character", Group = numeric())
}
\arguments{
\item{DealName}{A character string the DealName}

\item{Group}{A numeric value the collateral group number.
Note: Group is a counter for this function connection used in the CollateralGroup function}
}
\description{
The connection is a read helper to the aggregator CollateralGroup
}

