% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ConnectionFunctions.R
\name{REMICSchedules}
\alias{REMICSchedules}
\title{A connection function to the REMICSchedules Data calls REMIC structuring element
schedules.  This is the PAC/TAC schedule used by the waterfall}
\usage{
REMICSchedules(REMIC.Tranche = "character")
}
\arguments{
\item{REMIC.Tranche}{the REMIC tranche MBS.id or cusip.  Currently MBS.id is supported}
}
\description{
Opens connection to the Schedules folder by REMIC deal name
this function supports REMIC tranche analysis by provding the REMIC
schedules deal information
}

