% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\name{RemicStructure}
\alias{RemicStructure}
\title{REMIC Constructor}
\usage{
RemicStructure(DealName = "character")
}
\arguments{
\item{DealName}{A character string the deal's names}
}
\description{
Assembles the deal RAID, RDME, Tranches, and Groups into a REMIC structure
}
\examples{
\dontrun{
RemicStructure("BondLabPACInverse")
}
}

