% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ConnectionFunctions.R
\name{REMICWaterFall}
\alias{REMICWaterFall}
\title{A connection function to the REMICWaterFall source script}
\usage{
REMICWaterFall(deal.name = "character")
}
\arguments{
\item{deal.name}{A character string the REMIC deal name}
}
\description{
Opens a connection to the WaterFall folder calling the REMIC waterfall
script by deal name.  This function supports REMIC tranche analysis
}

