% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MtgScenario.R
\name{ReturnAnalysis}
\alias{ReturnAnalysis}
\title{Function to compute the horizon return}
\usage{
ReturnAnalysis(Scenario = "character", settlement.date = "character",
  proceeds = numeric(), spot.spread = numeric(),
  HorizonMonths = numeric(), ReinvestmentRate = numeric())
}
\arguments{
\item{Scenario}{a character string referring to scenario mortgage cashflow}

\item{settlement.date}{a character string the settlment date}

\item{proceeds}{a numeric value the investor initial outlay}

\item{spot.spread}{a numeric value the spread to the spot curve}

\item{HorizonMonths}{a numeric value the investment horizon in months}

\item{ReinvestmentRate}{a numeric value the reinvestment assumption}
}
\description{
A function to compute the horizon return of a PassThrough
}

