% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBSDetails.R
\docType{methods}
\name{SATO,MBSDetails-method}
\alias{SATO,MBSDetails-method}
\title{Method to extract the slot SATO from the class MBSDetails}
\usage{
\S4method{SATO}{MBSDetails}(object)
}
\arguments{
\item{object}{the name of the object of the tyoe MBSDetails}
}
\description{
Method to extract the slot SATO from the class MBSDetails
}

