% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelTune.R
\docType{methods}
\name{SATOBeta,PrepaymentModelTune-method}
\alias{SATOBeta,PrepaymentModelTune-method}
\title{A method to extract the SATO.beta from class PrepaymentModelTune}
\usage{
\S4method{SATOBeta}{PrepaymentModelTune}(object)
}
\arguments{
\item{object}{the name of the object of class PrepaymentModelTune}
}
\description{
A method to extract the SATO.beta from class PrepaymentModelTune
}

