% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelTune.R
\name{SATOBeta}
\alias{SATOBeta}
\title{A standard generic function to access the slot SATO.beta}
\usage{
SATOBeta(object)
}
\arguments{
\item{object}{an S4 object}
}
\description{
A standard generic function to access the slot SATO.beta
}

