% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectionFunctions.R
\name{SaveBond}
\alias{SaveBond}
\title{A connection function to BondData folder saved MBS cusip Detail}
\usage{
SaveBond(filename = "character")
}
\arguments{
\item{filename}{A character string the file name to save the 
default is temp}
}
\description{
Opens a connection to BondData folder to save bond cusip detail
}
