% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ConnectionFunctions.R
\name{SaveCollGroup}
\alias{SaveCollGroup}
\title{A connection functon to the Groups folder}
\usage{
SaveCollGroup(FileName = "character", DealName = "character",
  Group = numeric())
}
\description{
A connection functon to the Groups folder
}

