% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ConnectionFunctions.R
\name{SaveMBS}
\alias{SaveMBS}
\title{A connection function to BondData folder saved MBS cusip Detail}
\usage{
SaveMBS(...)
}
\description{
Opens a connection to BondData folder to save MBS cusip detail
}

