% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectionFunctions.R
\name{SaveModelFunctions}
\alias{SaveModelFunctions}
\title{A connection function to save the Prepayment Model ModelFunctions class}
\usage{
SaveModelFunctions(ModelFile = "character")
}
\arguments{
\item{ModelFile}{A character string the model functions class default 
value is temp}
}
\description{
Opens a connection to the PrepaymentModel folder to save the 
ModelFunctions class
}
