% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectionFunctions.R
\name{SaveMortgageRate}
\alias{SaveMortgageRate}
\title{A connection function to save the PrepaymentModel model tuning parameters}
\usage{
SaveMortgageRate()
}
\arguments{
\item{ModelFile}{A character string the model tuning default value is temp}

\item{ModelName}{A character string the model name used to reference the
prepayment model}
}
\description{
Opens a connection to the PrepaymentModel folder to save the prepayment
model tuning
}

