% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ConnectionFunctions.R
\name{SaveRAID}
\alias{SaveRAID}
\title{A connection function to save the RAID information}
\usage{
SaveRAID(RAIDFile = "character")
}
\description{
Opens a connection to the RAID folder and saves and saves
REMIC at Issuance Disclouse data
}

