% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ConnectionFunctions.R
\name{SaveRDME}
\alias{SaveRDME}
\title{A save connection to the RDME folder}
\usage{
SaveRDME(FileName = "character", DealName = "character",
  TrancheNumber = numeric())
}
\arguments{
\item{FileName}{A character string the FileName default to temp}

\item{DealName}{a character string the DealName}

\item{TrancheNumber}{a numeric value the tranche number}
}
\description{
The function opens a connection to the RDME file to save
REMIC Month End Disclosure data
}

