% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ConnectionFunctions.R
\name{SaveREMIC}
\alias{SaveREMIC}
\title{A save connection to the REMICData folder}
\usage{
SaveREMIC(DealName = "character", file = "character")
}
\arguments{
\item{DealName}{a character string the Deal Name}

\item{file}{a character string the REMIC file name given be constructor}
}
\description{
Opens a connection to REMICData folder saving a REMIC Deal by name
this connection is used by the function REMIC structure
}

