% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ConnectionFunctions.R
\name{SaveSchedules}
\alias{SaveSchedules}
\title{A connection function to the schedules folder}
\usage{
SaveSchedules(DealName = "character", ScheduleFile = "character")
}
\arguments{
\item{DealName}{A character string the deal name}

\item{ScheduleFile}{A character string the name of the schedule file to save}
}
\description{
Opens a connection to the Schedules folder to save PAC or TAC schedules
}

