% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ConnectionFunctions.R
\name{SaveTranche}
\alias{SaveTranche}
\title{A connection function to the Tranche Details information}
\usage{
SaveTranche(Tranche = character())
}
\arguments{
\item{Tranche}{a character string the Tranche}
}
\description{
Opens a connection to the Tranches folder and save Tranche Data
REMIC Tranche information
}

