% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ConnectionFunctions.R
\name{SaveTranches}
\alias{SaveTranches}
\title{A connection function used to assemble the deal tranches as a list}
\usage{
SaveTranches(DealName = "character", TrancheNumber = "character")
}
\arguments{
\item{DealName}{A character string the Deal Name}

\item{TrancheNumber}{A character string the tranche number}
}
\description{
Opens a connection to the Tranches folder and creates of list of deal tranches
used by the REMIC constructor function
}

