% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R, R/ScenarioConstructor.R
\docType{class}
\name{Scenario-class}
\alias{Scenario-class}
\title{An S4 Class representing scenario analysis details}
\description{
The class Scenario is the representation of the 
scenario run by the investor

An S4 class Scenario to describing the interest rate scenario
and interest rate shift function applied the curve
}
\section{Slots}{

\describe{
\item{\code{Name}}{A character the name of the scenario}

\item{\code{Type}}{A character the type of the scenario}

\item{\code{Horizon}}{A character the time horizon over which the 
scenario is run}

\item{\code{ShiftType}}{A character the interest rate shift type (e.g.
twist, parallel, bull flatten, or bull steepen)}

\item{\code{Shiftbps}}{A numeric value the scenario shift in basis points}

\item{\code{Formula}}{A function represnting the shift which applied to 
the term structure}

\item{\code{Name}}{A character the name of the scenario}

\item{\code{Type}}{A character the type of scenario eg("Aggressive", "Gradual",
"Immediate")}

\item{\code{Horizon}}{A character indicating the the time horizon over which the 
scenario evolves}

\item{\code{ShiftType}}{A character indicating the type shift 
eg("Parallel", "Twist)}

\item{\code{Shiftbps}}{A numeric value the interest rate shift in bps}

\item{\code{Formula}}{A function defining the interest rate shift}
}}

