% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScenarioConstructor.R
\docType{class}
\name{ScenarioCurve-class}
\alias{ScenarioCurve-class}
\title{an S4 class Horizon Curve}
\description{
The class HorizonCurve holds both the start and horizon curve
the object returned has both start TermStructure object and the horizon 
term structure object
}
\section{Slots}{

\describe{
\item{\code{HorizonMos}}{the scenario horizon months}

\item{\code{StartCurve}}{the starting curve assumption}

\item{\code{HorizonCurve}}{the horizon curve assumption}

\item{\code{StartTermStrc}}{the starting TermStructure assumption}

\item{\code{HorizonTermStrc}}{the horizon TermStructure assumption}
}}

\seealso{
Other Scenario Analysis: 
\code{\link{CouponIncome}()},
\code{\link{HorizonCurrBal}()},
\code{\link{HorizonCurveShift}()},
\code{\link{HorizonMos}()},
\code{\link{HorizonPrice}()},
\code{\link{HorizonReturn}()},
\code{\link{Name,Scenario-method}},
\code{\link{PrepaidPrinReceived}()},
\code{\link{PrincipalReceived}()},
\code{\link{ReinvestmentIncome}()},
\code{\link{Scenario-class}},
\code{\link{ScheduledPrinReceived}()}
}
\concept{Scenario Analysis}
