% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScenarioConstructor.R
\name{ScenarioFormula}
\alias{ScenarioFormula}
\title{A standard generic to access the slot Formula}
\usage{
ScenarioFormula(object)
}
\arguments{
\item{object}{An S4 class object of the type Scenario}
}
\description{
A standard generic to access the slot Formula
}
