% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScenarioConstructor.R
\name{ScenarioHorizonMos,ScenarioCurve-method}
\alias{ScenarioHorizonMos,ScenarioCurve-method}
\title{HorizonMos method, class ScenarioReturn}
\usage{
\S4method{ScenarioHorizonMos}{ScenarioCurve}(object)
}
\arguments{
\item{object}{the name of an S4 class of type BondReturn}
}
\description{
A method to get \strong{HorizonMos} over which a scenario covers
}
\seealso{
Other Bond Scenario Analysis: 
\code{\link{BondReturn-class}},
\code{\link{BondScenario-class}},
\code{\link{BondScenario}},
\code{\link{CouponIncome,BondReturn-method}},
\code{\link{HorizonCurrBal,BondReturn-method}},
\code{\link{HorizonMos,BondReturn-method}},
\code{\link{HorizonPrice,BondReturn-method}},
\code{\link{HorizonReturn,BondReturn-method}},
\code{\link{PrincipalReceived,BondReturn-method}},
\code{\link{ReinvestmentIncome,BondReturn-method}}
}
\concept{Bond Scenario Analysis}
