% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondScenario.R
\docType{methods}
\name{ScenarioSet,BondScenarioSet-method}
\alias{ScenarioSet,BondScenarioSet-method}
\title{A method to get ScenarioSet from the object BondScenarioSet}
\usage{
\S4method{ScenarioSet}{BondScenarioSet}(object)
}
\arguments{
\item{object}{is the name of the object of the type BondScenarioSet}
}
\description{
A method to get ScenarioSet from BondScenarioSet
\code{ScenarioSet(object)} is a getter method applied to the object
BondScenarioSet the method also applies to the object MortgageScenarioSet
}
\details{
The method "gets" the slot ScenarioSet from the class BondScenarioSet
}
\seealso{
Other Scenario.Analysis: \code{\link{BondReturn-class}},
  \code{\link{BondScenario-class}},
  \code{\link{BondScenarioSet-class}},
  \code{\link{BondScenarioSet}},
  \code{\link{BondScenario}},
  \code{\link{CouponIncome,BondReturn-method}},
  \code{\link{Name,Scenario-method}},
  \code{\link{Scenario-class}}
}

