% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\docType{methods}
\name{ScenarioSet,MortgageScenarioSet-method}
\alias{ScenarioSet,MortgageScenarioSet-method}
\title{A method to extract the Scenario list from S4 class MtgScenarioSet}
\usage{
\S4method{ScenarioSet}{MortgageScenarioSet}(object)
}
\arguments{
\item{object}{a list of the S4 objects of the type Scenario}
}
\description{
A method to extract the Scenario list from S4 class MtgScenarioSet
}

