% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\name{ScenarioSet}
\alias{ScenarioSet}
\title{A standard generic function to extract the slot Scenario}
\usage{
ScenarioSet(object)
}
\arguments{
\item{object}{An S4 class object of the type MtgScenarioSet}
}
\description{
A standard generic function to extract the slot Scenario
}

