% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Functions.R
\name{Sched.Prin}
\alias{Sched.Prin}
\title{Scheduled Principal}
\usage{
Sched.Prin(balance = numeric(), note.rate = numeric(),
  term.mos = numeric(), period = numeric())
}
\arguments{
\item{balance}{A numeric value, the balance of the mortgage}

\item{note.rate}{A numeric value, the borrower's note rate}

\item{term.mos}{A numeric value, the term of the mortgage}

\item{period}{A numeric value, the number of months since the first payment}
}
\description{
Computes the scheduled principal of mortgage loan
}

