% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{class}
\name{Schedule-class}
\alias{Schedule-class}
\title{An S4 class REMIC PAC/TAC schedules and payments}
\description{
Schedule class holds the PAC/TAC schedule payments
}
\section{Slots}{

\describe{
\item{\code{DealName}}{the name of the deal}

\item{\code{Group}}{the group number}

\item{\code{PmtDate}}{the payment date}

\item{\code{Balance}}{the scheduled balance of the PAC/TAC}

\item{\code{ScheduledPmt}}{the scheduled payment to the PAC/TAC bond}
}}

