% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{Schedule,Schedule-method}
\alias{Schedule,Schedule-method}
\title{A method to access the slot Schedule from the class TrancheDetails}
\usage{
\S4method{Schedule}{Schedule}(object)
}
\arguments{
\item{object}{an S4 object of the type TrancheDetails}
}
\description{
A method to access the slot Schedule from the class TrancheDetails
}

