% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageCashFlow.R
\name{ScheduledPrin,MortgageCashFlow-method}
\alias{ScheduledPrin,MortgageCashFlow-method}
\title{Method to extract the ScheduledPrin from class MortgageCashFlow}
\usage{
\S4method{ScheduledPrin}{MortgageCashFlow}(object)
}
\arguments{
\item{object}{the name of the object of type MortgageCashFlow}
}
\description{
Method to extract the ScheduledPrin from class MortgageCashFlow
}
