% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentFunctions.R
\name{SDARamp}
\alias{SDARamp}
\title{Standard default assumption ramp}
\usage{
SDARamp(BeginCDR = numeric(), PeakCDR = numeric(), EndCDR = numeric(),
  PeakMonth = numeric(), PlateauMonths = numeric(), EndMonth = numeric(),
  LoanAge = numeric())
}
\arguments{
\item{BeginCDR}{The beginning condition default rate}

\item{PeakCDR}{The peak conditional default rate}

\item{EndCDR}{The ending conditioin default}

\item{PeakMonth}{The month in which the conditional default rate reaches it
maximum value on the CDR curve.}

\item{PlateauMonths}{The number of months during which the condition default
rate remains at it maximum value.}

\item{EndMonth}{The month at which the conditional default rate reaches its
long term rate.}

\item{LoanAge}{The age of the loan or pool WALA.}
}
\description{
The function is the conditional default ramp for mortgage 
backed securities.  The function is stylized as the standard default assumption
}
\seealso{
Other Prepayment and Default: \code{\link{BorrowerIncentive}},
  \code{\link{BurnoutFunction}}, \code{\link{Seasonals}}
}

