% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelFunctions.R
\docType{methods}
\name{SeasonalFactors,PrepaymentModelFunctions-method}
\alias{SeasonalFactors,PrepaymentModelFunctions-method}
\title{A method to extract the slot SeasonalFactors from the object PrepaymentModelFunctions}
\usage{
\S4method{SeasonalFactors}{PrepaymentModelFunctions}(object)
}
\arguments{
\item{object}{an S4 object of the class PrepaymentModelFunctions}
}
\description{
A method to extract the slot SeasonalFactors from the object PrepaymentModelFunctions
}
