% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelFunctions.R
\name{SeasonalFactors}
\alias{SeasonalFactors}
\title{A standard generic function to access the slot SeasonalFactors from the 
class PrepaymentModelFunctions}
\usage{
SeasonalFactors(object)
}
\arguments{
\item{object}{an S4 object}
}
\description{
A standard generic function to access the slot SeasonalFactors from the 
class PrepaymentModelFunctions
}
