% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelTune.R
\docType{methods}
\name{SeasonalityTheta,PrepaymentModelTune-method}
\alias{SeasonalityTheta,PrepaymentModelTune-method}
\title{A method to extract Seasonality.theta from class PrepaymentModelTune}
\usage{
\S4method{SeasonalityTheta}{PrepaymentModelTune}(object)
}
\arguments{
\item{object}{The name of the object of class PrepaymentModelTune}
}
\description{
A method to extract Seasonality.theta from class PrepaymentModelTune
}
