% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelTune.R
\name{SeasonalityTheta}
\alias{SeasonalityTheta}
\title{A standard generic function to access the slot Seasonality.theta}
\usage{
SeasonalityTheta(object)
}
\arguments{
\item{object}{an S4 object}
}
\description{
A standard generic function to access the slot Seasonality.theta
}
