% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentFunctions.R
\name{Seasonals}
\alias{Seasonals}
\title{Seasonal Factor}
\usage{
Seasonals(alpha = numeric(), month = numeric(), theta = numeric())
}
\arguments{
\item{alpha}{numeric value the function alpha}

\item{month}{numeric value the month of year}

\item{theta}{a numeric value the theta}
}
\description{
The function creates seasonal factors for use in prepayment 
modeling
}

