% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModel.R
\name{Seasoning}
\alias{Seasoning}
\title{A function to define the mortgage seasoning ramp}
\usage{
Seasoning(alpha = numeric(), beta = numeric(), theta = numeric(),
  LoanAge = numeric())
}
\arguments{
\item{alpha}{a numeric value}

\item{beta}{a numeric value}

\item{theta}{a numeric value}

\item{LoanAge}{a numeric value}
}
\description{
A function to define the mortgage seasoning ramp
}

