% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelFunctions.R
\name{SeasoningRamp}
\alias{SeasoningRamp}
\title{A standard generic function to access the slot Seasoning from the class
PrepaymentModelFunctions}
\usage{
SeasoningRamp(object)
}
\arguments{
\item{object}{an S4 object}
}
\description{
A standard generic function to access the slot Seasoning from the class
PrepaymentModelFunctions
}

