% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondConstructor.R
\docType{methods}
\name{Sector,BondDetails-method}
\alias{Sector,BondDetails-method}
\title{A method to extract Sector from an S4 class of type BondDetails}
\usage{
\S4method{Sector}{BondDetails}(object)
}
\arguments{
\item{object}{The name of the S4 object of type BondDetails}
}
\description{
A method to extract Sector from an S4 class of type BondDetails
}

