% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBSDetails.R
\name{Sector}
\alias{Sector}
\title{A standard generic function to access the slot Sector}
\usage{
Sector(object)
}
\arguments{
\item{object}{and S4 class object}
}
\description{
A standard generic function to access the slot Sector
}

