% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBSDetails.R
\name{Servicing,MBSDetails-method}
\alias{Servicing,MBSDetails-method}
\title{Method to extract the slot Servicing from the class MBSDetails}
\usage{
\S4method{Servicing}{MBSDetails}(object)
}
\arguments{
\item{object}{the name of the object of the type MBSDetails}
}
\description{
Method to extract the slot Servicing from the class MBSDetails
}
