% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ServicingFeeTypes.R
\docType{methods}
\name{ServicingFeeBasis<-,ServicingFeeTypes-method}
\alias{ServicingFeeBasis<-,ServicingFeeTypes-method}
\title{A method to set the slot ServicingFeeBasis in object ServicingFeeTypes}
\usage{
\S4method{ServicingFeeBasis}{ServicingFeeTypes}(object) <- value
}
\arguments{
\item{object}{an S4 object of the type ServicingFeeTypes}

\item{value}{the replacement value of the slot}
}
\description{
A method to set the slot ServicingFeeBasis in object ServicingFeeTypes
}

