% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ServicingFeeTypes.R
\name{ServicingFeeBasis}
\alias{ServicingFeeBasis}
\title{A standard generic function to get the value of the slot 
ServicingFeeBasis}
\usage{
ServicingFeeBasis(object)
}
\arguments{
\item{object}{an S4 object of the type ServicingFeeTypes}
}
\description{
A standard generic function to get the value of the slot 
ServicingFeeBasis
}

