% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ServicingFeeTypes.R
\docType{methods}
\name{ServicingFeeDecimal,ServicingFeeTypes-method}
\alias{ServicingFeeDecimal,ServicingFeeTypes-method}
\title{A method to get the slot ServicingFeeDecimal from the object 
ServicingFeeTypes}
\usage{
\S4method{ServicingFeeDecimal}{ServicingFeeTypes}(object)
}
\arguments{
\item{object}{an object of the type ServicingFeeTypes}
}
\description{
A method to get the slot ServicingFeeDecimal from the object 
ServicingFeeTypes
}

