% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ServicingFeeTypes.R
\name{ServicingFeeDecimal<-,ServicingFeeTypes-method}
\alias{ServicingFeeDecimal<-,ServicingFeeTypes-method}
\title{A method to set the slot ServicingFeeBasis in the object 
ServicingFeeTypes}
\usage{
\S4method{ServicingFeeDecimal}{ServicingFeeTypes}(object) <- value
}
\arguments{
\item{object}{an S4 object of the type ServicingFeetypes}

\item{value}{the replacement value of the slot}
}
\description{
A method to set the slot ServicingFeeBasis in the object 
ServicingFeeTypes
}
