% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ServicingFeeTypes.R
\name{ServicingFeeDecimal<-}
\alias{ServicingFeeDecimal<-}
\title{A standard generic function to set the value of the slot 
SerivcingFeeDecimal}
\usage{
ServicingFeeDecimal(object) <- value
}
\arguments{
\item{object}{an S4 object of the type SerivicingFeeTypes}

\item{value}{the replacement value of the slot}
}
\description{
A standard generic function to set the value of the slot 
SerivcingFeeDecimal
}

