% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ServicingFeeTypes.R
\name{ServicingFeeDecimal}
\alias{ServicingFeeDecimal}
\title{A standard generic function to access the slot ServicingFeeDecimal}
\usage{
ServicingFeeDecimal(object)
}
\arguments{
\item{object}{an S4 object of the type ServicingFeeDecimal}
}
\description{
A standard generic function to access the slot ServicingFeeDecimal
}

