% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ServicingFeeTypes.R
\docType{methods}
\name{ServicingFeeDecimalString,ServicingFeeTypes-method}
\alias{ServicingFeeDecimalString,ServicingFeeTypes-method}
\title{A method to get the slot ServicingFeeDecimalString from the object
ServicingFeeTypes}
\usage{
\S4method{ServicingFeeDecimalString}{ServicingFeeTypes}(object)
}
\arguments{
\item{object}{an S4 object of the type ServicingFeeTypes}
}
\description{
A method to get the slot ServicingFeeDecimalString from the object
ServicingFeeTypes
}
