% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ServicingFeeTypes.R
\name{ServicingFeeDecimalString}
\alias{ServicingFeeDecimalString}
\title{A standard generic function to access the value of the slot
ServicingFeeDecimalString}
\usage{
ServicingFeeDecimalString(object)
}
\arguments{
\item{object}{an S4 object of the type ServicingFeeTypes}
}
\description{
A standard generic function to access the value of the slot
ServicingFeeDecimalString
}

