% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ServicingFeeTypes.R
\docType{class}
\name{ServicingFeeTypes-class}
\alias{ServicingFeeTypes-class}
\title{An S4 class representing ServicingFee Types}
\description{
This class is used create and pass SerivicingFee Types reported to the 
investor and used in cashflow analytics.  For example, serivicing fee is 
reported to the investor as numeric decimal or decimal string but basis is
used to determine the servicing fee paid to servicer.
}
\section{Slots}{

\describe{
\item{\code{ServicingFeeDecimal}}{a numeric value servicing fee expressed as numeric
decimal.}

\item{\code{ServicingFeeBasis}}{a numeric value the servicing fee expressed in
basis notation e.g. 0.0025}

\item{\code{ServicingFeeDecimalString}}{a character string the servicing fee
expressed in numeric decimal notation}
}}

