% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ServicingFeeTypes.R
\name{ServicingFeeTypes}
\alias{ServicingFeeTypes}
\title{ServicingFeeTypes is constructor function for the ServicingFeeTypes class}
\usage{
ServicingFeeTypes(ServicingFee = numeric())
}
\arguments{
\item{ServicingFee}{a numeric value the serivicing fee paid to the
servicer of the loan from the trust}
}
\description{
The ServicingFeeTypes class converts ServicingFee, a decimal numeric value,
to ServicingFeeBasis a numeric value and ServicingFeeDecimalString a
character string
}
