% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModel.R
\docType{methods}
\name{Severity,PrepaymentAssumption-method}
\alias{Severity,PrepaymentAssumption-method}
\title{A method to extract Severity from class PrepaymentAssumption}
\usage{
\S4method{Severity}{PrepaymentAssumption}(object)
}
\arguments{
\item{object}{the name of the S4 class}
}
\description{
A method to extract Severity from class PrepaymentAssumption
}

