% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModel.R
\docType{methods}
\name{Severity,PrepaymentModel-method}
\alias{Severity,PrepaymentModel-method}
\title{A method to extract Severity from class PrepaymentModel}
\usage{
\S4method{Severity}{PrepaymentModel}(object)
}
\arguments{
\item{object}{the name of the S4 class}
}
\description{
A method to extract Severity from class PrepaymentModel
}
