% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModel.R
\name{Severity}
\alias{Severity}
\title{A standard generic function to access the slot Severity}
\usage{
Severity(object)
}
\arguments{
\item{object}{An S4 class object of the type PrepaymentAssumption}
}
\description{
A standard generic function to access the slot Severity
}

