% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScenarioConstructor.R
\docType{methods}
\name{Shiftbps,Scenario-method}
\alias{Shiftbps,Scenario-method}
\title{A Method to extract Shiftbps from an S4 class Scenario}
\usage{
\S4method{Shiftbps}{Scenario}(object)
}
\arguments{
\item{object}{The name of the object of the S4 class of type Scenario}
}
\description{
A Method to extract Shiftbps from an S4 class Scenario
}
