% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\name{Shiftbps}
\alias{Shiftbps}
\title{A standard generic to access the slot Shiftbps}
\usage{
Shiftbps(object)
}
\arguments{
\item{object}{An S4 class object of the type Scenario}
}
\description{
A standard generic to access the slot Shiftbps
}

