% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\docType{methods}
\name{ShiftType,MtgScenario-method}
\alias{ShiftType,MtgScenario-method}
\title{A method to extract ShiftType from an S4 class MtgScenario}
\usage{
\S4method{ShiftType}{MtgScenario}(object)
}
\arguments{
\item{object}{the name of an S4 class of type MtgScenario}
}
\description{
A method to extract ShiftType from an S4 class MtgScenario
}

