% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScenarioConstructor.R
\docType{methods}
\name{ShiftType,Scenario-method}
\alias{ShiftType,Scenario-method}
\title{A Method to extract ShiftType from an S4 class Scenario}
\usage{
\S4method{ShiftType}{Scenario}(object)
}
\arguments{
\item{object}{The name of the object of the S4 class of type Scenario}
}
\description{
A Method to extract ShiftType from an S4 class Scenario
}

