% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageOAS.R
\name{SimCurve}
\alias{SimCurve}
\title{SimCurve Yield Curve Simulation}
\usage{
SimCurve(rates.data, num.paths = 300, seed = 42, three.month, two.year,
  ten.year)
}
\arguments{
\item{rates.data}{A character referencing a rates.data object}

\item{num.paths}{The number of simulation paths}

\item{seed}{The value of the random seed defaults to 42}

\item{three.month}{A vector of the mean, standard deviation, and mean 
reversion rate of the three month}

\item{two.year}{A vector of the mean, standard deviation and mean reversion
rate of the two year}

\item{ten.year}{A vector of the mean, standard deviation and mean reversion
rate of the ten year}
}
\description{
Simulation of the 3-mo, 2-year, and 10-year forward.
The function is used by the OASRatesArray function to create the rate array
}
\seealso{
Other mortgage pass through option adjusted spread: \code{\link{CalibrateRates}},
  \code{\link{FitMarket}},
  \code{\link{MonteCarloMeanReversion}},
  \code{\link{MortgageOAS-class}},
  \code{\link{OASRatesArray}}, \code{\link{SimRates}}
}
