% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageOAS.R
\name{SimRates}
\alias{SimRates}
\title{SimRates}
\usage{
SimRates(rates.data, num.paths = 300, seed = 42, rate.to.simulate,
  parameters)
}
\arguments{
\item{rates.data}{A character refrencing a rates.data object}

\item{num.paths}{The number of interest rate paths to simulate defaults to 300}

\item{seed}{The random seed defaults to 42}

\item{rate.to.simulate}{The rate to simulate}

\item{parameters}{A vector parameters mean, standard deviation, and the rate
of mean reversion of the rate simulated}
}
\description{
A function to simulate interest rate paths.  The function is called
by FitMarket and used to fit the forward rate curve
}
\seealso{
Other mortgage pass through option adjusted spread: \code{\link{CalibrateRates}},
  \code{\link{FitMarket}}, \code{\link{MortgageOAS-class}},
  \code{\link{OASRatesArray}}, \code{\link{SimCurve}}
}

