% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondDetails.R
\docType{methods}
\name{SinkingFund,BondDetails-method}
\alias{SinkingFund,BondDetails-method}
\title{Method to extract the slot SinkingFund from S4 class of type BondDetails}
\usage{
\S4method{SinkingFund}{BondDetails}(object)
}
\arguments{
\item{object}{the name of hte object of type BondDetails}
}
\description{
Method to extract the slot SinkingFund from S4 class of type BondDetails
}

