% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondConstructor.R
\name{SinkingFund}
\alias{SinkingFund}
\title{A standard generic function to access the slot SinkinnFund}
\usage{
SinkingFund(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A standard generic function to access the slot SinkinnFund
}

