% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModel.R
\name{SMM,PrepaymentModel-method}
\alias{SMM,PrepaymentModel-method}
\title{A method to extract SMM from S4 class PrepaymentModel}
\usage{
\S4method{SMM}{PrepaymentModel}(object)
}
\arguments{
\item{object}{the name of the S4 class}
}
\description{
A method to extract SMM from S4 class PrepaymentModel
}
