% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModel.R
\docType{methods}
\name{SMM<-,PrepaymentModel-method}
\alias{SMM<-,PrepaymentModel-method}
\title{A method to replace SMM value in the class PrepaymentModel}
\usage{
\S4method{SMM}{PrepaymentModel}(object) <- value
}
\arguments{
\item{object}{the name of the S4 class object}

\item{value}{the replace value of the slot SMM}
}
\description{
A method to replace SMM value in the class PrepaymentModel
}
