% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R, R/MortgageFunctions.R
\name{SMM.To.CPR}
\alias{SMM.To.CPR}
\title{Convert the single monthly mortality rate SMM to CPR}
\usage{
SMM.To.CPR(SMM = numeric())

SMM.To.CPR(SMM = numeric())
}
\arguments{
\item{SMM}{A numeric value the single monthly mortality rate (SMM)}

\item{SMM}{A numeric value the single monthly mortality rate (SMM)}
}
\description{
A standard generic function used to convert the SMM to CPR

A standard generic function used to convert the SMM to CPR
}
\examples{
SMM.To.CPR(SMM = .002)
SMM.To.CPR(SMM = .002)
}

