% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Functions.R
\name{SMMVector.To.CPR}
\alias{SMMVector.To.CPR}
\title{SMMVector.To.CPR}
\usage{
SMMVector.To.CPR(SMM = vector(), period = vector())
}
\arguments{
\item{SMM}{A numeric vector of SMM}

\item{a}{numeric vector of period (Loan Age)}
}
\description{
A standard generic function that converts a time series of SMM
to a time series of CPR.
}

