% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpreadToPriceFunctions.R
\name{SolveSpotSpread}
\alias{SolveSpotSpread}
\title{A function to solve for the spread to the spot rate curve}
\usage{
SolveSpotSpread(
  spot.curve = vector(),
  cash.flow = vector(),
  time.period = vector(),
  proceeds = numeric()
)
}
\arguments{
\item{spot.curve}{A vector representing the spot rate curve}

\item{cash.flow}{A vector representing the cash flow paid to the
investor}

\item{time.period}{A vector representing the time weights used to
discount cash flow}

\item{proceeds}{A numeric value proceeds paid by the investor}
}
\description{
A function to solve for the spread to the spot rate curve
}
