% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StructuringFunctions.R
\name{SplitFloaterInverse}
\alias{SplitFloaterInverse}
\title{Split Bond to Floater Inverse Floater}
\usage{
SplitFloaterInverse(floater.margin, bond.coupon, bond.balance, index, ...,
  inverse.multiplier = NULL, inverse.percentage = NULL)
}
\arguments{
\item{floater.margin}{the floater rate bond's margin to index}

\item{bond.coupon}{the coupon of the parent bond}

\item{bond.balance}{the principal balance of the parent bond}

\item{index}{the floating rate index to which the floating bond is struck}

\item{..., }{optional arguments follow select either inverse.muliplier or inverse.percentage}

\item{inverse.multiplier}{the inverse bond's multipler to the floating index}

\item{inverse.percentage}{the percentage balance of the inverse bond relative to the 
parent bond}
}
\description{
Structuring function to split a bond into a floater and 
inverse floating rate bond.  The optional arguments will determine the principal
balance of the floating rate bond.  Select either inverse.multiplier or inverse.percentage
}

