% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TermStructure.R
\name{SpotRate}
\alias{SpotRate}
\title{A generic function to access the slot SpotRate in class TermStructure}
\usage{
SpotRate(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A generic function to access the slot SpotRate in class TermStructure
}

