% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondTermStructure.R
\docType{methods}
\name{SpotSpread,BondTermStructure-method}
\alias{SpotSpread,BondTermStructure-method}
\title{A method to get SpotSpread from the class BondTermStructure}
\usage{
\S4method{SpotSpread}{BondTermStructure}(object)
}
\arguments{
\item{object}{an object of the type BondTermStructure}
}
\description{
A method to get SpotSpread from the class BondTermStructure
}

