% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageKeyRate.R
\docType{methods}
\name{SpotSpread,MortgageTermStructure-method}
\alias{SpotSpread,MortgageTermStructure-method}
\title{Method to extract SpotSpread from S4 class}
\usage{
\S4method{SpotSpread}{MortgageTermStructure}(object)
}
\arguments{
\item{object}{The name of the S4 object of type MortgageTermStructure}
}
\description{
Method to extract SpotSpread from S4 class
}

