% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BillDetails.R
\docType{methods}
\name{SPRating,BillDetails-method}
\alias{SPRating,BillDetails-method}
\title{Method to extract S&P rating from S4 class of type BillDetails}
\usage{
\S4method{SPRating}{BillDetails}(object)
}
\arguments{
\item{object}{The name of the S4 class of type BillDetails}
}
\description{
Method to extract S&P rating from S4 class of type BillDetails
}

