% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondDetails.R
\name{SPRating,BondDetails-method}
\alias{SPRating,BondDetails-method}
\title{Method to extract SP rating from S4 class of type BondDetails}
\usage{
\S4method{SPRating}{BondDetails}(object)
}
\arguments{
\item{object}{the name of the object of type BondDetails}
}
\description{
Method to extract SP rating from S4 class of type BondDetails
}
