% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpreadTypes.R
\name{SpreadBasis,SpreadTypes-method}
\alias{SpreadBasis,SpreadTypes-method}
\title{A method to get the value of the slot SpreadBasisString from the object
of class SpreadTypes}
\usage{
\S4method{SpreadBasis}{SpreadTypes}(object)
}
\arguments{
\item{object}{an S4 object of the class SpreadTypes}
}
\description{
A method to get the value of the slot SpreadBasisString from the object
of class SpreadTypes
}
