% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpreadTypes.R
\name{SpreadBasis}
\alias{SpreadBasis}
\title{A standard generic function to get the slot SpreadBasis}
\usage{
SpreadBasis(object)
}
\arguments{
\item{object}{an S4 object of the class SpreadTypes}
}
\description{
A standard generic function to get the slot SpreadBasis
}

