% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpreadTypes.R
\docType{methods}
\name{SpreadBasisDecimal<-,SpreadTypes-method}
\alias{SpreadBasisDecimal<-,SpreadTypes-method}
\title{A method to set the value of the slot SpreadBasisDecimal of the 
class SpreadTypes}
\usage{
\S4method{SpreadBasisDecimal}{SpreadTypes}(object) <- value
}
\arguments{
\item{object}{an object of the class SpreadTypes}

\item{value}{the replacement value of the slot}
}
\description{
A method to set the value of the slot SpreadBasisDecimal of the 
class SpreadTypes
}

