% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpreadTypes.R
\docType{methods}
\name{SpreadBasisDecimal,SpreadTypes-method}
\alias{SpreadBasisDecimal,SpreadTypes-method}
\title{A method to get SpreadBasisDecimal from slot of class SpreadTypes}
\usage{
\S4method{SpreadBasisDecimal}{SpreadTypes}(object)
}
\arguments{
\item{object}{an object of the type SpreadTypes}
}
\description{
A method to get SpreadBasisDecimal from slot of class SpreadTypes
}

