% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpreadTypes.R
\docType{methods}
\name{SpreadBasisString<-,SpreadTypes-method}
\alias{SpreadBasisString<-,SpreadTypes-method}
\title{A method to set the value of the slot SpreadBasisString of the
class SpreadTypes}
\usage{
\S4method{SpreadBasisString}{SpreadTypes}(object) <- value
}
\arguments{
\item{object}{and object of the class SpreadTypes}

\item{value}{the replacement value of the slot}
}
\description{
A method to set the value of the slot SpreadBasisString of the
class SpreadTypes
}

