% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpreadTypes.R
\name{SpreadBasisString}
\alias{SpreadBasisString}
\title{A standard generic function to access the slot SpreadBasisString}
\usage{
SpreadBasisString(object)
}
\arguments{
\item{object}{an S4 object of the type SpreadTypes}
}
\description{
A standard generic function to access the slot SpreadBasisString
}

