% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpreadTypes.R
\name{SpreadDecimal,SpreadTypes-method}
\alias{SpreadDecimal,SpreadTypes-method}
\title{A method to get the value of the slot SpreadDecimal from the class SpreadTypes}
\usage{
\S4method{SpreadDecimal}{SpreadTypes}(object)
}
\arguments{
\item{object}{an object of the class SpreadTypes}
}
\description{
A method to get the value of the slot SpreadDecimal from the class SpreadTypes
}
