% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpreadTypes.R
\docType{methods}
\name{SpreadDecimalString,SpreadTypes-method}
\alias{SpreadDecimalString,SpreadTypes-method}
\title{A method to get the value of the slot SpreadDecimalString from the
class SpreadTypes}
\usage{
\S4method{SpreadDecimalString}{SpreadTypes}(object)
}
\arguments{
\item{object}{an object of the class SpreadTypes}
}
\description{
A method to get the value of the slot SpreadDecimalString from the
class SpreadTypes
}

