% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageOAS.R
\docType{methods}
\name{Spreads,MortgageOAS-method}
\alias{Spreads,MortgageOAS-method}
\title{A Method to access the slot Spreads}
\usage{
\S4method{Spreads}{MortgageOAS}(object)
}
\arguments{
\item{object}{MortgageOAS}
}
\description{
Access the slot Spreads from object MortgageOAS
}

