% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageOAS.R
\name{Spreads}
\alias{Spreads}
\title{A standard generic to access the slot Spreads}
\usage{
Spreads(object)
}
\arguments{
\item{object}{MortgageOAS}
}
\description{
Access the slot Spreads from object MortgageOAS
}
