% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurveSpreads.R
\docType{methods}
\name{SpreadToBenchmark,CurveSpreads-method}
\alias{SpreadToBenchmark,CurveSpreads-method}
\title{A method to extract the SpreadToBenchmark from object CurveSpreads}
\usage{
\S4method{SpreadToBenchmark}{CurveSpreads}(object)
}
\arguments{
\item{object}{An object of type CurveSpreads}
}
\description{
A method to extract the SpreadToBenchmark from object CurveSpreads
}

