% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurveSpreads.R
\name{SpreadToBenchmark}
\alias{SpreadToBenchmark}
\title{A standard generic function to access SpreadToBenchmark}
\usage{
SpreadToBenchmark(object)
}
\arguments{
\item{object}{An object of the type CurveSpreads}
}
\description{
A standard generic function to access SpreadToBenchmark
}
