% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurveSpreads.R
\docType{methods}
\name{SpreadToCurve,CurveSpreads-method}
\alias{SpreadToCurve,CurveSpreads-method}
\title{A method to extract the SpreadToCurve from object CurveSpreads}
\usage{
\S4method{SpreadToCurve}{CurveSpreads}(object)
}
\arguments{
\item{object}{An object of type CurveSpreads}
}
\description{
A method to extract the SpreadToCurve from object CurveSpreads
}

