% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurveSpreads.R
\name{SpreadToCurve}
\alias{SpreadToCurve}
\title{A standard generic function to access SpreadToCurve}
\usage{
SpreadToCurve(object)
}
\arguments{
\item{object}{An object of the type SpreadToCurve}
}
\description{
A standard generic function to access SpreadToCurve
}

