% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\name{SpreadToInterCurve}
\alias{SpreadToInterCurve}
\title{A standard generic function to access the slot SpreadToInterCurve}
\usage{
SpreadToInterCurve(object)
}
\arguments{
\item{object}{An S4 object of type Mtg.Scenario}
}
\description{
A standard generic function to access the slot SpreadToInterCurve
}

