% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpreadToPriceFunctions.R
\name{SpreadToPrice}
\alias{SpreadToPrice}
\title{SpreadToPrice Function}
\usage{
SpreadToPrice(bond.id = "character", trade.date = "character",
  settlement.date = "character", PrepaymentAssumption = "character",
  spread = "character", CPR = numeric(), ..., benchmark = NULL)
}
\arguments{
\item{bond.id}{A character string the cusip number or bond.id}

\item{trade.date}{A character string the trade date mm-dd-YYYY}

\item{settlement.date}{A character string the settlement date mm-dd-YYYY}

\item{PrepaymentAssumption}{the assumption must be: "CPR", "PPC", "MODEL"}

\item{spread}{a charcter string the spread to the interpolated curve
entered in basis points}

\item{CPR}{a numeric value the CPR assumption used to price the MBS.  For
example 16 CPR is entered 16.}

\item{...}{optional values follow}

\item{benchmark}{optionally the user can pass pricing benchmark}
}
\description{
Returns the clean price of the MBS given a prepayment speed
and spread to the curve
}

