% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpreadToPriceFunctions.R
\name{SpreadToPriceBond}
\alias{SpreadToPriceBond}
\title{Spread to Price Bond}
\usage{
SpreadToPriceBond(bond.id, trade.date, settlement.date, spread,
  benchmark = NULL)
}
\arguments{
\item{bond.id}{a character or connection referencing an object of type BondDetails}

\item{trade.date}{a character the trade date mm-dd-YYYY}

\item{settlement.date}{a character the settlement.date}

\item{spread}{the spread to the benchmark given in basis points.}

\item{benchmark}{the pricing benchmark.  The default in NULL in which case the
function will determine the nearest pricing benchmark.  The user override values are
1, 2, 3, 4, 5, 7, 10, 30}
}
\description{
returns the clean price of a bond given a spread and benchmark.
market convention is to quote spread to a benchmark in basis points.  Bond Lab
follows the market convection.  The user specified spread to the benchmark in basis points.
}
\seealso{
Other Pricing: \code{\link{BenchMark,CurveSpreads-method}},
  \code{\link{BillPriceToYield}},
  \code{\link{BillYieldToPrice}},
  \code{\link{CurveSpreads-class}},
  \code{\link{CurveSpreads}},
  \code{\link{SpreadToBenchmark,CurveSpreads-method}},
  \code{\link{SpreadToCurve,CurveSpreads-method}},
  \code{\link{SpreadToPriceMBS}},
  \code{\link{ZeroVolSpread,CurveSpreads-method}}
}
