% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpreadTypes.R
\docType{class}
\name{SpreadTypes-class}
\alias{SpreadTypes-class}
\title{A S4 class representing SpreadTypes (Spread the the Curve)}
\description{
The is class is used to calculate SpreadTypes which are in turne 
used as input to calculate price given a yield curve, 
WAL or benchmark, and spread to curve.
}
\section{Slots}{

\describe{
\item{\code{SpreadBasisDecimal}}{a numeric value the spread expressed as numeric 
decimal (bps)}

\item{\code{SpreadBasisString}}{a character the spread expressed as numeric decimal (bps)}

\item{\code{SpreadDecimal}}{a numeric value the spread expressed as numeric decimal}

\item{\code{SpreadDecimalString}}{a character the spread expressed as numeric decimal}
}}

