% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpreadTypes.R
\name{SpreadTypes}
\alias{SpreadTypes}
\title{SpreadTypes is a constructor function for the class SpreadTypes}
\usage{
SpreadTypes(spread)
}
\arguments{
\item{spread}{a character string the spread entered in basis points}
}
\description{
SpreadTypes is a constructor function for the class SpreadTypes
}

