% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScenarioConstructor.R
\docType{methods}
\name{StartCurve,ScenarioCurve-method}
\alias{StartCurve,ScenarioCurve-method}
\title{Start Curve the Scenario Starting Curve}
\usage{
\S4method{StartCurve}{ScenarioCurve}(object)
}
\arguments{
\item{object}{The name of the object of type HorizonCurve}
}
\description{
a method to get the \strong{Start Curve} from the HorizonCurve object
}
\seealso{
Other Scenario: \code{\link{HorizonCurve,ScenarioCurve-method}},
  \code{\link{HorizonTermStrc,ScenarioCurve-method}},
  \code{\link{StartTermStrc,ScenarioCurve-method}}
}
