% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScenarioConstructor.R
\name{StartCurve}
\alias{StartCurve}
\title{A standard generic function to access the HorizonCurve slot StartCurve}
\usage{
StartCurve(object)
}
\arguments{
\item{object}{An S4 class object of the type HorizonCurve}
}
\description{
A standard generic function to access the HorizonCurve slot StartCurve
}
