% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScenarioConstructor.R
\docType{methods}
\name{StartTermStrc,ScenarioCurve-method}
\alias{StartTermStrc,ScenarioCurve-method}
\title{StartTermStruc the Scenario Start Term Structure Assumption}
\usage{
\S4method{StartTermStrc}{ScenarioCurve}(object)
}
\arguments{
\item{object}{The name of the object of type Horizon Curve}
}
\description{
a method to get the \strong{Start Term Structure} from the Horizon Curve object
}
\seealso{
Other Scenario: \code{\link{HorizonCurve,ScenarioCurve-method}},
  \code{\link{HorizonTermStrc,ScenarioCurve-method}},
  \code{\link{StartCurve,ScenarioCurve-method}}
}

