% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScenarioConstructor.R
\name{StartTermStrc}
\alias{StartTermStrc}
\title{A standard generic function to access the HorizonCurve slot StartTermStruc}
\usage{
StartTermStrc(object)
}
\arguments{
\item{object}{An S4 class object of the type HorizonCurve}
}
\description{
A standard generic function to access the HorizonCurve slot StartTermStruc
}

