% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TermStructure.R
\docType{methods}
\name{TenYearForward<-,TermStructure-method}
\alias{TenYearForward<-,TermStructure-method}
\title{Method to replace the TenYearForward in the class TermStructure}
\usage{
\S4method{TenYearForward}{TermStructure}(object) <- value
}
\arguments{
\item{object}{the name of the object of the type TermStructure}

\item{value}{the value of the replacement}
}
\description{
Method to replace the TenYearForward in the class TermStructure
}
