% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TermStructure.R
\docType{methods}
\name{TenYearForward,TermStructure-method}
\alias{TenYearForward,TermStructure-method}
\title{Method to extract the TenYearForward from the class TermStructure}
\usage{
\S4method{TenYearForward}{TermStructure}(object)
}
\arguments{
\item{object}{the name of the object of the type TermStructure}
}
\description{
Method to extract the TenYearForward from the class TermStructure
}

