% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageOAS.R
\name{Term,MortgageOAS-method}
\alias{Term,MortgageOAS-method}
\title{A Method to access the slot Term}
\usage{
\S4method{Term}{MortgageOAS}(object)
}
\arguments{
\item{object}{MortgageOAS}
}
\description{
Access the slot Term from object MortgageOAS
}
