% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageKeyRate.R
\name{Term,MortgageTermStructure-method}
\alias{Term,MortgageTermStructure-method}
\title{Term}
\usage{
\S4method{Term}{MortgageTermStructure}(object)
}
\arguments{
\item{object}{MortgageTermStructure}
}
\description{
Get Term from object of type MortgageTermStructure
}
