% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TermStructure.R
\docType{class}
\name{TermStructure-class}
\alias{TermStructure-class}
\title{An S4 class the term structure data needed to price bonds}
\description{
An S4 class the term structure data needed to price bonds
}
\section{Slots}{

\describe{
\item{\code{tradedate}}{A character string the trade date in the format
mm/dd/YYYY}

\item{\code{period}}{A numeric value the period index of the next cash flow}

\item{\code{date}}{A numeric value the date of the next cash flow}

\item{\code{spotrate}}{A numeric value the one-month spot rate}

\item{\code{forwardrate}}{A numeric value the one-month forward rate}

\item{\code{TwoYearFwd}}{A numeric vlaue the two-year forward rate}

\item{\code{TenYearFwd}}{A numeric value the ten-year forward rate}
}}

