% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BillCashFlows.R
\docType{methods}
\name{TimePeriod,BillCashFlows-method}
\alias{TimePeriod,BillCashFlows-method}
\title{Method to get TimePeriod from S4 class BillCashFlows}
\usage{
\S4method{TimePeriod}{BillCashFlows}(object)
}
\arguments{
\item{object}{the name of the S4 object BillCashFlows}
}
\description{
Method to get TimePeriod from S4 class BillCashFlows
}

