% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondCashFlows.R
\docType{methods}
\name{TimePeriod,BondCashFlows-method}
\alias{TimePeriod,BondCashFlows-method}
\title{Method to extract TimePeriod from S4 class BondCashFlows}
\usage{
\S4method{TimePeriod}{BondCashFlows}(object)
}
\arguments{
\item{object}{the name of the S4 object}
}
\description{
Method to extract TimePeriod from S4 class BondCashFlows
}

