% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageCashFlow.R
\name{TimePeriod,MortgageCashFlow-method}
\alias{TimePeriod,MortgageCashFlow-method}
\title{Method to extract the TimePeriod from class MortgageCashFlow}
\usage{
\S4method{TimePeriod}{MortgageCashFlow}(object)
}
\arguments{
\item{object}{the name of the object of type MortgageCashFlow}
}
\description{
Method to extract the TimePeriod from class MortgageCashFlow
}
