% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\docType{methods}
\name{TimePeriod,MtgScenario-method}
\alias{TimePeriod,MtgScenario-method}
\title{A method to extract TimePeriod from S4 class MtgScenario}
\usage{
\S4method{TimePeriod}{MtgScenario}(object)
}
\arguments{
\item{object}{The name of an S4 class of type MtgScenario}
}
\description{
A method to extract TimePeriod from S4 class MtgScenario
}

