% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TermStructure.R
\docType{methods}
\name{TimePeriod<-,TermStructure-method}
\alias{TimePeriod<-,TermStructure-method}
\title{Method to replace TimePeriod in the class TermStructure}
\usage{
\S4method{TimePeriod}{TermStructure}(object) <- value
}
\arguments{
\item{object}{the name of the object of type TermStructure}

\item{value}{the replacement value of the slot}
}
\description{
Method to replace TimePeriod in the class TermStructure
}
