% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageCashFlow.R
\name{TimePeriod<-}
\alias{TimePeriod<-}
\title{A standard generic function to replace the slot TimePeriod}
\usage{
TimePeriod(object) <- value
}
\arguments{
\item{object}{an S4 class object}

\item{value}{replacement value of the slot}
}
\description{
A standard generic function to replace the slot TimePeriod
}

