% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TermStructure.R
\docType{methods}
\name{TimePeriod,TermStructure-method}
\alias{TimePeriod,TermStructure-method}
\title{Method to get TimePeriod in the class TermStructure}
\usage{
\S4method{TimePeriod}{TermStructure}(object)
}
\arguments{
\item{object}{the name of the object of type TermStructure}
}
\description{
Method to get TimePeriod in the class TermStructure
}

