% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageCashFlow.R
\name{TimePeriod}
\alias{TimePeriod}
\title{A standard generic function to access the slot TimePeriod}
\usage{
TimePeriod(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A standard generic function to access the slot TimePeriod
}

