% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BillCashFlows.R
\docType{methods}
\name{TotalCashFlow,BillCashFlows-method}
\alias{TotalCashFlow,BillCashFlows-method}
\title{Method to get TotalCashFlows from S4 class BillCashFlows}
\usage{
\S4method{TotalCashFlow}{BillCashFlows}(object)
}
\arguments{
\item{object}{the name of the S4 object}
}
\description{
Method to get TotalCashFlows from S4 class BillCashFlows
}

