% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondCashFlow.R
\docType{methods}
\name{TotalCashFlow,BondCashFlows-method}
\alias{TotalCashFlow,BondCashFlows-method}
\title{Method to extract TotalCashFlow from S4 class BondCashFlows}
\usage{
\S4method{TotalCashFlow}{BondCashFlows}(object)
}
\arguments{
\item{object}{the name of the S4 object}
}
\description{
Method to extract TotalCashFlow from S4 class BondCashFlows
}

