% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageCashFlow.R
\name{TotalCashFlow,MortgageCashFlow-method}
\alias{TotalCashFlow,MortgageCashFlow-method}
\title{Method to extract the TotalCashFlow from the class MortgageCashFlow}
\usage{
\S4method{TotalCashFlow}{MortgageCashFlow}(object)
}
\arguments{
\item{object}{the name of the object of type MortgageCashFlow}
}
\description{
Method to extract the TotalCashFlow from the class MortgageCashFlow
}
