% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TermStructure.R
\docType{methods}
\name{TradeDate<-,TermStructure-method}
\alias{TradeDate<-,TermStructure-method}
\title{Method to replace TradeDate in class TermStructure}
\usage{
\S4method{TradeDate}{TermStructure}(object) <- value
}
\arguments{
\item{object}{the name of the object of type TermStructure}

\item{value}{the replacement value of the slot}
}
\description{
Method to replace TradeDate in class TermStructure
}
