% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TermStructure.R
\name{TradeDate<-}
\alias{TradeDate<-}
\title{A generic function to replace the slot TradeDate}
\usage{
TradeDate(object) <- value
}
\arguments{
\item{object}{an S4 class object}

\item{value}{the replacement value}
}
\description{
A generic function to replace the slot TradeDate
}

