% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TermStructure.R
\name{TradeDate,TermStructure-method}
\alias{TradeDate,TermStructure-method}
\title{Method to extract TradeDate from the class TermStructure}
\usage{
\S4method{TradeDate}{TermStructure}(object)
}
\arguments{
\item{object}{the name of the object of type TermStructure}
}
\description{
Method to extract TradeDate from the class TermStructure
}
