% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{TrancheAvgLife,TrancheDetails-method}
\alias{TrancheAvgLife,TrancheDetails-method}
\title{A method to access the slot TrancheAvgLife from class TrancheDetails}
\usage{
\S4method{TrancheAvgLife}{TrancheDetails}(object)
}
\arguments{
\item{object}{an S4 class of type TrancheDetails}
}
\description{
A method to access the slot TrancheAvgLife from class TrancheDetails
}

