% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\name{TrancheAvgLife}
\alias{TrancheAvgLife}
\title{A generic functionto access the slot TrancheAvgLife}
\usage{
TrancheAvgLife(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A generic functionto access the slot TrancheAvgLife
}

