% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{TrancheDatedDate,TrancheDetails-method}
\alias{TrancheDatedDate,TrancheDetails-method}
\title{A method to access the slot TrancheDatedDate from class TrancheDetails}
\usage{
\S4method{TrancheDatedDate}{TrancheDetails}(object)
}
\arguments{
\item{object}{an S4 class of type TrancheDetails}
}
\description{
A method to access the slot TrancheDatedDate from class TrancheDetails
}

