% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\name{TrancheDatedDate}
\alias{TrancheDatedDate}
\title{A generic function to access the TrancheDatedDate}
\usage{
TrancheDatedDate(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A generic function to access the TrancheDatedDate
}

