% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{class}
\name{TrancheDetails-class}
\alias{TrancheDetails-class}
\title{An S4 Class TrancheDetails at Issuance}
\description{
Tranche Details is static data Tranche changes are updated
via REMIC Data Month End (RDME) classes
}
\section{Slots}{

\describe{
\item{\code{DealName}}{A character the deal name}

\item{\code{TrancheNumber}}{A character the tranche number}

\item{\code{NumberofComponents}}{A numeric value the number of components}

\item{\code{ComponentofTranches}}{A character the component of tranches}

\item{\code{TrancheName}}{A character the tranche name}

\item{\code{TranchePrincipal}}{A character the principal}

\item{\code{TranchePrincipalDesc}}{A character the tranche principal description}

\item{\code{TrancheInterestDesc}}{A character the tranche interest description}

\item{\code{TrancheOtherDescription}}{A character additional tranche decriptive 
information}

\item{\code{Cusip}}{A character the tranche cusip identifier}

\item{\code{TrancheOrigBal}}{A numeric value the tranche original balance}

\item{\code{TrancheInterest}}{A character the tranche interest}

\item{\code{TrancheCoupon}}{A numeric value the tranche coupon}

\item{\code{AccrualRate}}{A numeric value the tranche accrual rate}

\item{\code{TreasuryMaturity}}{A numeric value the maturity of the 
benchmark Treasury at issuance}

\item{\code{TreasuryYield}}{A numueric value the yield of the benchmark Treasury}

\item{\code{TreasurySpread}}{A numeric value the tranche spread over the 
benchmark Treasury at issuance}

\item{\code{TrancheYield}}{A numeric value the tranche yield at issuance}

\item{\code{TranchePrice}}{A numeric value the tranche price at issuance}

\item{\code{TrancheProceedsWithInterest}}{A numeric value the tranche proceeds at
issuance principal and interest}

\item{\code{TrancheAvgLife}}{A numeric vlaue the tranche average life at issuance}

\item{\code{TrancheDuration}}{A numeric value the tranche duration at issuance}

\item{\code{TrancheDatedDate}}{A character the dated date of the tranche}

\item{\code{TrancheFirstPmtDate}}{A character the first payment date of the tranche}

\item{\code{TrancheLastPmtDate}}{A character the last payment date of the tranche}

\item{\code{TrancheNextPmtDate}}{A character the next payment date of the tranche}

\item{\code{TrancheFinalPmtDate}}{A character the final payment date of the tranche}

\item{\code{Delay}}{A numeric value the delay days of payment to the investor}

\item{\code{InterestPmtFrequency}}{A numeric value the frequency of 
interest payments to the investor}

\item{\code{PrinPmtFrequency}}{A numeric value the frequency of principal
payments to the investor}

\item{\code{PacLowBand}}{A numeric value lower PAC band}

\item{\code{PacHighBand}}{A numeric value upper PAC band}

\item{\code{FloaterIndex}}{A character the floating rate index used to determine
the floating rate bond coupon}

\item{\code{InitialIndexValue}}{A numeric value the value of the floater index at
issuance}

\item{\code{FloaterMargin}}{A numeric value the floater margin}

\item{\code{FloaterMultiplier}}{A numeric value the floater multiplier}

\item{\code{FloaterCap}}{A numeric vlaue the floating rate bond cap}

\item{\code{FloaterFloor}}{A numeric value the floating rate bond floor}

\item{\code{FloaterInitialCoupon}}{A numeric value the floater initial coupon}

\item{\code{FloaterResetFrequency}}{A numeric value the frequency at which the 
floating coupon changes}

\item{\code{FloaterFirstResetDate}}{a numeric vlaue the first reset date of the 
floating rate bond}

\item{\code{FloaterFormula}}{A function used to calculate the floating rate bond
coupon rate}

\item{\code{Group}}{A numeric value the collateral group to which the 
tranche belongs}

\item{\code{TrancheType}}{A character the tranche type}

\item{\code{Schedule}}{A logical indicating the bond payments are tied to 
a schedule}

\item{\code{Fixed}}{A logical indicating the bond coupon rate is fixed}
}}

