% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{TrancheDuration,TrancheDetails-method}
\alias{TrancheDuration,TrancheDetails-method}
\title{A method to access the slot TrancheDuration from class TrancheDetails}
\usage{
\S4method{TrancheDuration}{TrancheDetails}(object)
}
\arguments{
\item{object}{an S4 class of type TrancheDetails}
}
\description{
A method to access the slot TrancheDuration from class TrancheDetails
}

