% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\name{TrancheDuration}
\alias{TrancheDuration}
\title{A generic function to access the slot TrancheDuration}
\usage{
TrancheDuration(object)
}
\arguments{
\item{object}{an S4 class oject}
}
\description{
A generic function to access the slot TrancheDuration
}

