% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{TrancheFinalPmtDate,TrancheDetails-method}
\alias{TrancheFinalPmtDate,TrancheDetails-method}
\title{A method to access the slot TrancheFinalPmtDate from class TrancheDetails}
\usage{
\S4method{TrancheFinalPmtDate}{TrancheDetails}(object)
}
\arguments{
\item{object}{an S4 class of type TrancheDetails}
}
\description{
A method to access the slot TrancheFinalPmtDate from class TrancheDetails
}

