% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{TrancheFirstPmtDate,TrancheDetails-method}
\alias{TrancheFirstPmtDate,TrancheDetails-method}
\title{A method to access the slot TrancheFirstPmtDate from class TrancheDetails}
\usage{
\S4method{TrancheFirstPmtDate}{TrancheDetails}(object)
}
\arguments{
\item{object}{an S4 class of type TrancheDetails}
}
\description{
A method to access the slot TrancheFirstPmtDate from class TrancheDetails
}

