% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\name{TrancheInterest}
\alias{TrancheInterest}
\title{A generic function access the slot TrancheInterest}
\usage{
TrancheInterest(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A generic function access the slot TrancheInterest
}

