% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{TrancheInterestDesc,TrancheDetails-method}
\alias{TrancheInterestDesc,TrancheDetails-method}
\title{A method to access the slot TrancheInterestDesc from class TrancheDetails}
\usage{
\S4method{TrancheInterestDesc}{TrancheDetails}(object)
}
\arguments{
\item{object}{an S4 object of the type TrancheInterestDesc}
}
\description{
A method to access the slot TrancheInterestDesc from class TrancheDetails
}

