% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{TrancheLastPmtDate,TrancheDetails-method}
\alias{TrancheLastPmtDate,TrancheDetails-method}
\title{A method to access the slot TrancheLastPmtDate from class TrancheDetails}
\usage{
\S4method{TrancheLastPmtDate}{TrancheDetails}(object)
}
\arguments{
\item{object}{an S4 clss of type TrancheDetails}
}
\description{
A method to access the slot TrancheLastPmtDate from class TrancheDetails
}

