% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{TrancheName,TrancheDetails-method}
\alias{TrancheName,TrancheDetails-method}
\title{A method to access the slot TrancheName from class TrancheDetails}
\usage{
\S4method{TrancheName}{TrancheDetails}(object)
}
\arguments{
\item{object}{an S4 object of the type TrancheDetails}
}
\description{
A method to access the slot TrancheName from class TrancheDetails
}

