% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{TrancheNextPmtDate,TrancheDetails-method}
\alias{TrancheNextPmtDate,TrancheDetails-method}
\title{A method to access the slot TrancheNextPmtDate from class TrancheDetails}
\usage{
\S4method{TrancheNextPmtDate}{TrancheDetails}(object)
}
\arguments{
\item{object}{an S4 class of type TrancheDetails}
}
\description{
A method to access the slot TrancheNextPmtDate from class TrancheDetails
}

