% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{TrancheNumber,TrancheDetails-method}
\alias{TrancheNumber,TrancheDetails-method}
\title{A method to access the slot TrancheNumber from class TrancheDetails}
\usage{
\S4method{TrancheNumber}{TrancheDetails}(object)
}
\arguments{
\item{object}{an S4 class of type TrancheDetails}
}
\description{
A method to access the slot TrancheNumber from class TrancheDetails
}

