% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{TrancheOrigBal,TrancheDetails-method}
\alias{TrancheOrigBal,TrancheDetails-method}
\title{A method to access the slot TrancheOrigBal from class TrancheDetails}
\usage{
\S4method{TrancheOrigBal}{TrancheDetails}(object)
}
\arguments{
\item{object}{an S4 class of type TrancheDetails}
}
\description{
A method to access the slot TrancheOrigBal from class TrancheDetails
}

