% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{TranchePrice,TrancheDetails-method}
\alias{TranchePrice,TrancheDetails-method}
\title{A method to access the slot TranchePrice from class TrancheDetails}
\usage{
\S4method{TranchePrice}{TrancheDetails}(object)
}
\arguments{
\item{object}{an S4 class of type TrancheDetails}
}
\description{
A method to access the slot TranchePrice from class TrancheDetails
}

