% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{TranchePrincipal,TrancheDetails-method}
\alias{TranchePrincipal,TrancheDetails-method}
\title{A method to access the slot TranchePrincipal from class TrancheDetails}
\usage{
\S4method{TranchePrincipal}{TrancheDetails}(object)
}
\arguments{
\item{object}{an S4 object of the type TrancheDetails}
}
\description{
A method to access the slot TranchePrincipal from class TrancheDetails
}

