% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{TranchePrincipalDesc,TrancheDetails-method}
\alias{TranchePrincipalDesc,TrancheDetails-method}
\title{A method to access the slot TranchePrincipalDesc from class TrancheDetails}
\usage{
\S4method{TranchePrincipalDesc}{TrancheDetails}(object)
}
\arguments{
\item{object}{an S4 object of the type TranchePrincipalDesc}
}
\description{
A method to access the slot TranchePrincipalDesc from class TrancheDetails
}

