% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\name{TrancheProceedsWithInterest}
\alias{TrancheProceedsWithInterest}
\title{A generic function to access the slot TrancheProceedsWithInterest}
\usage{
TrancheProceedsWithInterest(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A generic function to access the slot TrancheProceedsWithInterest
}

