% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{class}
\name{Tranches-class}
\alias{Tranches-class}
\title{A S4 class Tranches for the REMIC}
\description{
Tranches is a list of the tranche details of each tranche collateralized
}
\section{Slots}{

\describe{
\item{\code{Tranches}}{a list of TrancheDetail classes}
}}

