% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{Tranches,Tranches-method}
\alias{Tranches,Tranches-method}
\title{A method to access the Tranches from class Tranches}
\usage{
\S4method{Tranches}{Tranches}(object)
}
\arguments{
\item{object}{an S4 class object of type Tranches}
}
\description{
A method to access the Tranches from class Tranches
}

