% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\name{Tranches}
\alias{Tranches}
\title{Aggregator Function for REMIC constructor}
\usage{
Tranches(NumberofTranches = numeric(), DealName = "character")
}
\arguments{
\item{NumberofTranches}{A numeric value the number of tranches in the deal}

\item{DealName}{A character string the Deal Name}
}
\description{
Aggregates Tranche data for REMIC constructor
}

