% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\name{TranchesList}
\alias{TranchesList}
\title{Aggregator Function for REMIC constructor}
\usage{
TranchesList(NumberofTranches = numeric(), DealName = "character")
}
\arguments{
\item{NumberofTranches}{A numeric value the number of tranches in the deal}

\item{DealName}{A character string the Deal Name}
}
\description{
Aggregates Tranche data for REMIC constructor
}

