% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{TrancheType,TrancheDetails-method}
\alias{TrancheType,TrancheDetails-method}
\title{A method to access the slot TrancheType from class TrancheDetails}
\usage{
\S4method{TrancheType}{TrancheDetails}(object)
}
\arguments{
\item{object}{an S4 object of the typre TrancheDetails}
}
\description{
A method to access the slot TrancheType from class TrancheDetails
}

