% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{TrancheYield,TrancheDetails-method}
\alias{TrancheYield,TrancheDetails-method}
\title{A method to access the slot TrancheYield from class TrancheDetails}
\usage{
\S4method{TrancheYield}{TrancheDetails}(object)
}
\arguments{
\item{object}{an S4 class of type TrancheDetails}
}
\description{
A method to access the slot TrancheYield from class TrancheDetails
}

