% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{TreasuryMaturity,TrancheDetails-method}
\alias{TreasuryMaturity,TrancheDetails-method}
\title{A method to access the slot TreasuryMaturity from class TrancheDetails}
\usage{
\S4method{TreasuryMaturity}{TrancheDetails}(object)
}
\arguments{
\item{object}{an S4 class of type TrancheDetails}
}
\description{
A method to access the slot TreasuryMaturity from class TrancheDetails
}

