% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{TreasurySpread,TrancheDetails-method}
\alias{TreasurySpread,TrancheDetails-method}
\title{A method to access the slot TreasurySpread from class TrancheDetails}
\usage{
\S4method{TreasurySpread}{TrancheDetails}(object)
}
\arguments{
\item{object}{an S4 class of type TrancheDetails}
}
\description{
A method to access the slot TreasurySpread from class TrancheDetails
}

