% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{TreasuryYield,TrancheDetails-method}
\alias{TreasuryYield,TrancheDetails-method}
\title{A method to access the slot TreasuryYield from class TrancheDetails}
\usage{
\S4method{TreasuryYield}{TrancheDetails}(object)
}
\arguments{
\item{object}{an S4 class of type TrancheDetails}
}
\description{
A method to access the slot TreasuryYield from class TrancheDetails
}

