% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{Trustee,RAID-method}
\alias{Trustee,RAID-method}
\title{A method to access the slot Trustee from class RAID}
\usage{
\S4method{Trustee}{RAID}(object)
}
\arguments{
\item{object}{an S4 object of the type RAID}
}
\description{
A method to access the slot Trustee from class RAID
}

