% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentFunctions.R
\name{Turnover}
\alias{Turnover}
\title{Housing Turnover}
\usage{
Turnover(alpha = 1, beta = 0.879, theta = 0.192, loan.age)
}
\arguments{
\item{alpha}{The seasoning curve aysmptote (default value is 1.0)}

\item{beta}{The seasoning curve intercept default value is 0.879}

\item{theta}{The point of maximum steepness of the seasoning curve}

\item{loan.age}{The age (WALA) of the mortgage loan or pool of mortgage loans}
}
\description{
The function defines the housing turnover ramp (seasoning)
}
\seealso{
Other Prepayment and Default: \code{\link{BorrowerIncentive}},
  \code{\link{BurnoutFunction}}, \code{\link{SDARamp}},
  \code{\link{Seasonals}}
}

