% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelTune.R
\docType{methods}
\name{TurnoverAlpha,PrepaymentModelTune-method}
\alias{TurnoverAlpha,PrepaymentModelTune-method}
\title{A method to extract Turnover.alpha from class PrepaymentModelTune}
\usage{
\S4method{TurnoverAlpha}{PrepaymentModelTune}(object)
}
\arguments{
\item{object}{The name of the object of class PrepaymentModelTune}
}
\description{
A method to extract Turnover.alpha from class PrepaymentModelTune
}
