% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelTune.R
\name{TurnoverAlpha}
\alias{TurnoverAlpha}
\title{A standard generic function to access the slot Turnover.alpha}
\usage{
TurnoverAlpha(object)
}
\arguments{
\item{object}{an S4 object}
}
\description{
A standard generic function to access the slot Turnover.alpha
}

