% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelTune.R
\docType{methods}
\name{TurnoverBeta,PrepaymentModelTune-method}
\alias{TurnoverBeta,PrepaymentModelTune-method}
\title{A method to extract Turnover.beta from class PrepaymentModelTune}
\usage{
\S4method{TurnoverBeta}{PrepaymentModelTune}(object)
}
\arguments{
\item{object}{The name of the object of class PrepaymentModelTune}
}
\description{
A method to extract Turnover.beta from class PrepaymentModelTune
}
