% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelFunctions.R
\docType{methods}
\name{TurnoverRate,PrepaymentModelFunctions-method}
\alias{TurnoverRate,PrepaymentModelFunctions-method}
\title{A method to extract the slot TurnoverRate from the object PrepaymentModelFunctions}
\usage{
\S4method{TurnoverRate}{PrepaymentModelFunctions}(object)
}
\arguments{
\item{object}{an S4 object of the class PrepaymentModelFunctions}
}
\description{
A method to extract the slot TurnoverRate from the object PrepaymentModelFunctions
}

