% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelTune.R
\docType{methods}
\name{TurnoverRate,PrepaymentModelTune-method}
\alias{TurnoverRate,PrepaymentModelTune-method}
\title{A method to extract TurnoverRate from class PrepaymentModelTune}
\usage{
\S4method{TurnoverRate}{PrepaymentModelTune}(object)
}
\arguments{
\item{object}{The name of the object of class PrepaymentModelTune}
}
\description{
A method to extract TurnoverRate from class PrepaymentModelTune
}

