% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelTune.R
\name{TurnoverRate}
\alias{TurnoverRate}
\title{A standard generic function to access the slot TuroverRate}
\usage{
TurnoverRate(object)
}
\arguments{
\item{object}{an S4 object}
}
\description{
A standard generic function to access the slot TuroverRate
}

