% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelTune.R
\docType{methods}
\name{TurnoverTheta,PrepaymentModelTune-method}
\alias{TurnoverTheta,PrepaymentModelTune-method}
\title{A method to extract Turnover.theta from class PrepaymentModelTune}
\usage{
\S4method{TurnoverTheta}{PrepaymentModelTune}(object)
}
\arguments{
\item{object}{The name of the object of class PrepaymentModelTune}
}
\description{
A method to extract Turnover.theta from class PrepaymentModelTune
}

