% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelTune.R
\name{TurnoverTheta}
\alias{TurnoverTheta}
\title{A standard generic function to access the slot Turnover.theta}
\usage{
TurnoverTheta(object)
}
\arguments{
\item{object}{an S4 object}
}
\description{
A standard generic function to access the slot Turnover.theta
}

