% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TermStructure.R
\name{TwoYearForward<-,TermStructure-method}
\alias{TwoYearForward<-,TermStructure-method}
\title{Method to replace the TwoYearForward in the class TermStructure}
\usage{
\S4method{TwoYearForward}{TermStructure}(object) <- value
}
\arguments{
\item{object}{the name of the object of the Type TermStructure}

\item{value}{the replacement value}
}
\description{
Method to replace the TwoYearForward in the class TermStructure
}
