% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TermStructure.R
\name{TwoYearForward<-}
\alias{TwoYearForward<-}
\title{A generic function to replace the slot TwoYearForward in class 
TermStructure}
\usage{
TwoYearForward(object) <- value
}
\arguments{
\item{object}{an S4 class object}

\item{value}{the value of the replacement}
}
\description{
A generic function to replace the slot TwoYearForward in class 
TermStructure
}

