% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TermStructure.R
\name{TwoYearForward,TermStructure-method}
\alias{TwoYearForward,TermStructure-method}
\title{Method to extract the TwoYearForward from the class TermStructure}
\usage{
\S4method{TwoYearForward}{TermStructure}(object)
}
\arguments{
\item{object}{the name of the object of the type TermStructure}
}
\description{
Method to extract the TwoYearForward from the class TermStructure
}
