% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\docType{methods}
\name{Type,MtgScenario-method}
\alias{Type,MtgScenario-method}
\title{A method to extract Type (scenario type) from S4 class MtgScenario}
\usage{
\S4method{Type}{MtgScenario}(object)
}
\arguments{
\item{object}{the name of an S4 class of type MtgScenario}
}
\description{
A method to extract Type (scenario type) from S4 class MtgScenario
}

