% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\docType{methods}
\name{Type,Scenario-method}
\alias{Type,Scenario-method}
\title{A Method to extract scenario type from an S4 class Scenario}
\usage{
\S4method{Type}{Scenario}(object)
}
\arguments{
\item{object}{The name of the object of the S4 class of type Scenario}
}
\description{
A Method to extract scenario type from an S4 class Scenario
}

