% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\name{Type}
\alias{Type}
\title{A standard generic to access the slot Type}
\usage{
Type(object)
}
\arguments{
\item{object}{An S4 class object of the type Scenario}
}
\description{
A standard generic to access the slot Type
}

