% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ConnectionFunctions.R
\name{ULTV}
\alias{ULTV}
\title{Function to calculate the updated loan to value}
\usage{
ULTV()
}
\description{
A read connection function to the updated loan to value function.  No inputs are
required the function propogates the updated loan to value based on amortization
and updated home price.
}
\examples{
\dontrun{
ULTV()}
}

