% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondDetails.R
\docType{methods}
\name{Underwriter,BondDetails-method}
\alias{Underwriter,BondDetails-method}
\title{Method to extract Underwriter from an S$ class of type BondDetails}
\usage{
\S4method{Underwriter}{BondDetails}(object)
}
\arguments{
\item{object}{The name of the S4 object of type BondDetails}
}
\description{
Method to extract Underwriter from an S$ class of type BondDetails
}

