% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{Underwriter,RAID-method}
\alias{Underwriter,RAID-method}
\title{A method to access the slot Underwriter from class RAID}
\usage{
\S4method{Underwriter}{RAID}(object)
}
\arguments{
\item{object}{an S4 object of the type RAID}
}
\description{
A method to access the slot Underwriter from class RAID
}

