% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBSDetails.R
\name{Underwriter}
\alias{Underwriter}
\title{A generic function to access the slot Underwriter}
\usage{
Underwriter(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A generic function to access the slot Underwriter
}

