% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelTune.R
\docType{methods}
\name{UpdatedLTVBeta,PrepaymentModelTune-method}
\alias{UpdatedLTVBeta,PrepaymentModelTune-method}
\title{A method to extract the UpdatedLTV.beta from class PrepaymentModelTune}
\usage{
\S4method{UpdatedLTVBeta}{PrepaymentModelTune}(object)
}
\arguments{
\item{object}{the name of the object of class PrepaymentModelTune}
}
\description{
A method to extract the UpdatedLTV.beta from class PrepaymentModelTune
}

