% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageCashFlow.R
\name{ValidPriceYieldScenarioSet}
\alias{ValidPriceYieldScenarioSet}
\title{A function to test the validity of PriceYieldScenarioSet object}
\usage{
ValidPriceYieldScenarioSet(object)
}
\arguments{
\item{object}{an S4 object of type PriceYieldScenarioSet}
}
\description{
Validity test of the PriceYieldScenarioSet object test if the length
of the PrepaymentType and the PrepaymentScenario are equal
}

