% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BillCashFlows.R
\docType{methods}
\name{WAL,BillCashFlows-method}
\alias{WAL,BillCashFlows-method}
\title{Method to get WAL from S4 class BillCashFlows}
\usage{
\S4method{WAL}{BillCashFlows}(object)
}
\arguments{
\item{object}{the name of the S4 object BillCashFlows}
}
\description{
Method to get WAL from S4 class BillCashFlows
}
