% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageCashFlow.R
\docType{methods}
\name{WAL,MortgageCashFlow-method}
\alias{WAL,MortgageCashFlow-method}
\title{Method to extract WAL from S4 class}
\usage{
\S4method{WAL}{MortgageCashFlow}(object)
}
\arguments{
\item{object}{the name of the S4 object}
}
\description{
Method to extract WAL from S4 class
}

