% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageCashFlow.R
\name{WAL}
\alias{WAL}
\title{A standard generic function to access the slot WAL}
\usage{
WAL(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A standard generic function to access the slot WAL
}

