% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBSDetails.R
\docType{methods}
\name{WALA<-,MBSDetails-method}
\alias{WALA<-,MBSDetails-method}
\title{Method to replace the slot WALA in the class MBSDetails}
\usage{
\S4method{WALA}{MBSDetails}(object) <- value
}
\arguments{
\item{object}{the name of the object of the type MBSDetails}

\item{value}{the replacement value of the slot}
}
\description{
Method to replace the slot WALA in the class MBSDetails
}

