% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBSDetails.R
\name{WALA<-}
\alias{WALA<-}
\title{A standard generic function to replace the slot WALA}
\usage{
WALA(object) <- value
}
\arguments{
\item{object}{an S4 class object}

\item{value}{the replacement value of the slot}
}
\description{
A standard generic function to replace the slot WALA
}

