% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PassThroughConstructor.R
\name{WALA}
\alias{WALA}
\title{A standard generic function to access the slot WALA}
\usage{
WALA(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A standard generic function to access the slot WALA
}

