% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBSDetails.R
\name{WAM,MBSDetails-method}
\alias{WAM,MBSDetails-method}
\title{Method to extract the slot WAM from the class MBSDetails}
\usage{
\S4method{WAM}{MBSDetails}(object)
}
\arguments{
\item{object}{the name of the object of the type MBSDetails}
}
\description{
Method to extract the slot WAM from the class MBSDetails
}
