% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBSDetails.R
\name{WAM<-,MBSDetails-method}
\alias{WAM<-,MBSDetails-method}
\title{Method to replace the slot WAM in the class MBSDetails}
\usage{
\S4method{WAM}{MBSDetails}(object) <- value
}
\arguments{
\item{object}{the name of the object of the type MBSDetails}

\item{value}{the replacement value of the slot}
}
\description{
Method to replace the slot WAM in the class MBSDetails
}
