% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PassThroughConstructor.R
\name{WAM<-}
\alias{WAM<-}
\title{A standard generic function to replace the slot WAM}
\usage{
WAM(object) <- value
}
\arguments{
\item{object}{an S4 class object}

\item{value}{the replacement value of the slot}
}
\description{
A standard generic function to replace the slot WAM
}

