% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBSDetails.R
\name{WAM}
\alias{WAM}
\title{A standard generic function to access the slot WAM}
\usage{
WAM(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A standard generic function to access the slot WAM
}

