% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YieldTypes.R
\docType{methods}
\name{YieldBasis<-,YieldTypes-method}
\alias{YieldBasis<-,YieldTypes-method}
\title{A method to replace the value in the slot YieldBasis}
\usage{
\S4method{YieldBasis}{YieldTypes}(object) <- value
}
\arguments{
\item{object}{an S4 class object of the type YieldTypes}

\item{value}{an S4 calss object of the type Yieldtypes}
}
\description{
A method to replace the value in the slot YieldBasis
}

