% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YieldTypes.R
\name{YieldBasis<-}
\alias{YieldBasis<-}
\title{A standard generic function set the value in the slot YieldBasis}
\usage{
YieldBasis(object) <- value
}
\arguments{
\item{object}{an S4 object of the type YieldBasis}

\item{value}{the replacement value of the slot}
}
\description{
A standard generic function set the value in the slot YieldBasis
}

