% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YieldTypes.R
\docType{methods}
\name{YieldBasis,YieldTypes-method}
\alias{YieldBasis,YieldTypes-method}
\title{A method to extract YieldBasis from slot of YieldTypes}
\usage{
\S4method{YieldBasis}{YieldTypes}(object)
}
\arguments{
\item{object}{an S4 class object of the type YieldTypes}
}
\description{
A method to extract YieldBasis from slot of YieldTypes
}

