% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YieldTypes.R
\name{YieldBasis}
\alias{YieldBasis}
\title{A standard generic function to get the slot YieldBasis}
\usage{
YieldBasis(object)
}
\arguments{
\item{object}{an S4 object of the type YieldBasis}
}
\description{
A standard generic function to get the slot YieldBasis
}

