% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YieldTypes.R
\docType{methods}
\name{YieldDecimal<-,YieldTypes-method}
\alias{YieldDecimal<-,YieldTypes-method}
\title{A method to replace the vlaue in the slot YieldDecimal}
\usage{
\S4method{YieldDecimal}{YieldTypes}(object) <- value
}
\arguments{
\item{object}{an S4 class of the type YieldTypes}

\item{value}{the replacement value of the slot}
}
\description{
A method to replace the vlaue in the slot YieldDecimal
}

