% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YieldTypes.R
\name{YieldDecimal<-}
\alias{YieldDecimal<-}
\title{A standard generic function to set the value in the slot YieldDecimal}
\usage{
YieldDecimal(object) <- value
}
\arguments{
\item{object}{an S4 object of the type YieldTypes}

\item{value}{the replacement value of the slot}
}
\description{
A standard generic function to set the value in the slot YieldDecimal
}

