% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YieldTypes.R
\docType{methods}
\name{YieldDecimalString<-,YieldTypes-method}
\alias{YieldDecimalString<-,YieldTypes-method}
\title{A method to replace the slot YieldDecimalstring from the class YieldTypes}
\usage{
\S4method{YieldDecimalString}{YieldTypes}(object) <- value
}
\arguments{
\item{object}{an S4 class object of the type YieldTypes}

\item{value}{the replacement value of the slot}
}
\description{
A method to replace the slot YieldDecimalstring from the class YieldTypes
}

