% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YieldTypes.R
\name{YieldDecimalString,YieldTypes-method}
\alias{YieldDecimalString,YieldTypes-method}
\title{A method to extract the slot YieldDecimalString from the class YieldTypes}
\usage{
\S4method{YieldDecimalString}{YieldTypes}(object)
}
\arguments{
\item{object}{an S4 class object of the type YieldTypes}
}
\description{
A method to extract the slot YieldDecimalString from the class YieldTypes
}
