% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YieldTypes.R
\name{YieldDecimalString}
\alias{YieldDecimalString}
\title{A standard generic function to get the slot YieldDecimalString}
\usage{
YieldDecimalString(object)
}
\arguments{
\item{object}{an S4 object of the type YieldTypes}
}
\description{
A standard generic function to get the slot YieldDecimalString
}

