% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BillCashFlows.R
\docType{methods}
\name{YieldToMaturity,BillCashFlows-method}
\alias{YieldToMaturity,BillCashFlows-method}
\title{Method to get YieldToMaturity from S4 class}
\usage{
\S4method{YieldToMaturity}{BillCashFlows}(object)
}
\arguments{
\item{object}{the name of the S4 object BillCashFlows}
}
\description{
Method to get YieldToMaturity from S4 class
}

