% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageCashFlow.R
\docType{methods}
\name{YieldToMaturity,MortgageCashFlow-method}
\alias{YieldToMaturity,MortgageCashFlow-method}
\title{Method to extract YieldToMaturity from S4 class MortgageCashFlow}
\usage{
\S4method{YieldToMaturity}{MortgageCashFlow}(object)
}
\arguments{
\item{object}{the name of the S4 object}
}
\description{
Method to extract YieldToMaturity from S4 class MortgageCashFlow
}

