% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageCashFlow.R
\name{YieldToMaturity}
\alias{YieldToMaturity}
\title{A standard generic function to access the slot YieldToMaturity}
\usage{
YieldToMaturity(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A standard generic function to access the slot YieldToMaturity
}
