% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YieldTypes.R
\docType{class}
\name{YieldTypes-class}
\alias{YieldTypes-class}
\title{A S4 class representing yield to maturity}
\description{
This class is used to create and pass YieldTypes reported by Bond Lab 
analytics and reported to investors.  It does not calculate yield to maturity
rather it holds the value in decimal number, basis, and decimal string form.
}
\section{Slots}{

\describe{
\item{\code{YieldDecimal}}{a numeric value the yield to maturity in decimal number form}

\item{\code{YieldBasis}}{a numeric value the yield to maturity in basis form}

\item{\code{YieldDecimalString}}{a character string the yield to maturity}
}}

