% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YieldTypes.R
\name{YieldTypes}
\alias{YieldTypes}
\title{YieldTypes is the constructor function for the class YieldTypes}
\usage{
YieldTypes(yield = numeric())
}
\arguments{
\item{yield}{a numeric value the yield expresssed in basis format
(expample 5.0 is 0.5)}
}
\description{
YieldTypes is the constructor function for the class YieldTypes
}
