% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageRateFunctions.R
\name{yr15}
\alias{yr15}
\title{A standard generic function to access the slot yr15 from class MortgageRate}
\usage{
yr15(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A standard generic function to access the slot yr15 from class MortgageRate
}

