% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HullWhite.R
\name{ZeroBondPrice}
\alias{ZeroBondPrice}
\title{Zero Bond price}
\usage{
ZeroBondPrice(r, t)
}
\arguments{
\item{r}{the interest rate used to discount}

\item{t}{time time period for discounting}
}
\description{
A function ot compute bond price per Hull White
}
