% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondTermStructure.R
\docType{methods}
\name{ZeroVolSpread,BondTermStructure-method}
\alias{ZeroVolSpread,BondTermStructure-method}
\title{Zero Volatility Spread}
\usage{
\S4method{ZeroVolSpread}{BondTermStructure}(object)
}
\arguments{
\item{object}{object of type BondTermStructure}
}
\description{
Zero Volatility Spread from object BondTermStructure
}

