% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurveSpreads.R
\docType{methods}
\name{ZeroVolSpread,CurveSpreads-method}
\alias{ZeroVolSpread,CurveSpreads-method}
\title{A method to extract the ZeroVol spread from object CurveSpreads}
\usage{
\S4method{ZeroVolSpread}{CurveSpreads}(object)
}
\arguments{
\item{object}{An object of type CurveSpreads}
}
\description{
A method to extract the ZeroVol spread from object CurveSpreads
}

