% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageOAS.R
\docType{methods}
\name{ZeroVolSpread,MortgageOAS-method}
\alias{ZeroVolSpread,MortgageOAS-method}
\title{A Method to access the slot ZeroVolSpread}
\usage{
\S4method{ZeroVolSpread}{MortgageOAS}(object)
}
\arguments{
\item{object}{MortgageOAS}
}
\description{
Access the slot ZeroVolSpread from object MortgageOAS
}

