% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageKeyRate.R
\docType{methods}
\name{ZeroVolSpread,MortgageTermStructure-method}
\alias{ZeroVolSpread,MortgageTermStructure-method}
\title{Zero Volatility Spread}
\usage{
\S4method{ZeroVolSpread}{MortgageTermStructure}(object)
}
\arguments{
\item{object}{MortgageTermStructure object}
}
\description{
Get ZeroVolSpread from object of type MortgageTermStructure
}

