% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurveSpreads.R
\name{ZeroVolSpread}
\alias{ZeroVolSpread}
\title{A standard generic function to access ZeroVol Spread}
\usage{
ZeroVolSpread(object)
}
\arguments{
\item{object}{An object of the type SpreadToCurve}
}
\description{
A standard generic function to access ZeroVol Spread
}

