% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondBasis.R
\name{ActualFactor}
\alias{ActualFactor}
\title{Actual/Actual day count}
\usage{
ActualFactor(settlement.date, nextpmt.date)
}
\arguments{
\item{settlement.date}{The settlement date}

\item{nextpmt.date}{The next scheduled interest payment date}
}
\description{
Calculates actual days using leap year IDSA method
}
\seealso{
Other BondBasis: \code{\link{ActFactor}},
  \code{\link{BondBasisConversion}},
  \code{\link{is.leapyear}}
}
\concept{BondBasis}
