% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondBasis.R
\name{ActualFactor}
\alias{ActualFactor}
\title{Actual day count}
\usage{
ActualFactor(settlement.date, nextpmt.date)
}
\arguments{
\item{settlement.date}{The settlement date}

\item{nextpmt.date}{The next scheduled interest payment date}
}
\description{
Calculates actual days using leap year
}
\seealso{
Other BondBasis: \code{\link{BondBasisConversion}},
  \code{\link{is.leapyear}}
}

