% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrincipalPmtDateCalcs.R
\name{AddMonths}
\alias{AddMonths}
\title{Principal payment dates}
\usage{
AddMonths(Date = "character", Months = numeric())
}
\arguments{
\item{Date}{A character string the date to which the number of months are added}

\item{Months}{A numeric value the number of months to add to Date}
}
\description{
In the case of mortgage and asset backed securities the 
first and last principal payment date are of interest to the investors. 
This funnction computes the first principal payment and last principal 
payment dates.
}
\examples{
AddMonths("03-01-2015", 360)
}

