% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICStructuringFunctions.R
\name{BeginBal}
\alias{BeginBal}
\title{Function to assign the beginning principal balance to REMIC tranche}
\usage{
BeginBal(node, period = numeric())
}
\arguments{
\item{node}{The entry point of the deal's tree structure.}

\item{period}{A numeric value the period in which payment is recieved stated as an integer.}
}
\description{
This function takes the ending balance of the previous period and assigns
the value to the beginning balance in the current period. Structuring functions
run on a tree structure created by the package data.tree.  See data.tree for more details on the package.
}

