% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurveSpreads.R
\docType{methods}
\name{BenchMark,CurveSpreads-method}
\alias{BenchMark,CurveSpreads-method}
\title{BenchMark (Pricing)}
\usage{
\S4method{BenchMark}{CurveSpreads}(object)
}
\arguments{
\item{object}{An object of the type CurveSpreads}
}
\description{
Get the pricing benchmark.  The pricing benchmark is the nearest
marutity along the pricing curve to the weighted average life of the bond.
}
\seealso{
Other Pricing: \code{\link{BillPriceToYield}},
  \code{\link{BillYieldToPrice}},
  \code{\link{CurveSpreads-class}},
  \code{\link{CurveSpreads}},
  \code{\link{SpreadToBenchmark,CurveSpreads-method}},
  \code{\link{SpreadToCurve,CurveSpreads-method}},
  \code{\link{SpreadToPriceBond}},
  \code{\link{SpreadToPriceMBS}},
  \code{\link{ZVSpreadToPriceBond}},
  \code{\link{ZVSpreadToPriceMBS}},
  \code{\link{ZeroVolSpread,CurveSpreads-method}}
}
\concept{Pricing}
