% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectionFunctions.R
\name{Bond}
\alias{Bond}
\title{query sql lite UST Bond cusip database}
\usage{
Bond(cusip)
}
\arguments{
\item{cusip}{the cusip identifier of the pass through must be in quotes}
}
\description{
a function to query the sql lite data base of mortgage
cusips.  The function illustrates how one would connect to a cusip database
and construct the mortgage pass through object for analysis
}

