% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BondAnalytics.R
\name{BondAnalytics}
\alias{BondAnalytics}
\title{Bond Analytics}
\usage{
BondAnalytics(bond.id = "character", principal = numeric(),
  price = numeric(), trade.date = "character",
  settlement.date = "character", method = "character")
}
\arguments{
\item{bond.id}{A character string the cusip number or id of the bond}

\item{principal}{A numeric value the principal amount of the bond}

\item{price}{A numeric value the price of the bond}

\item{trade.date}{A character string the trade date}

\item{settlement.date}{A character string the settlement date}

\item{method}{A character string the method used to fit the term structure passed to the package termstruc}
}
\description{
Analytic engine for standard bond (non-mortgage) analysis
}

