% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BondBasis.R
\name{BondBasisConversion}
\alias{BondBasisConversion}
\title{BondBasisConversion}
\usage{
BondBasisConversion(issue.date, start.date, end.date, settlement.date,
  lastpmt.date, nextpmt.date)
}
\description{
BondBasisConversion
}

