% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondBasis.R
\name{BondBasisConversion}
\alias{BondBasisConversion}
\title{Bond Basis (day count) conventions}
\usage{
BondBasisConversion(issue.date, start.date, end.date, settlement.date,
  lastpmt.date, nextpmt.date, type)
}
\arguments{
\item{issue.date}{A date object, the issue date of the security}

\item{start.date}{A date object, the start date for interest payment}

\item{end.date}{A date object, the final payment date}

\item{settlement.date}{A date object, the settlement date}

\item{lastpmt.date}{A date object, the last payment date}

\item{nextpmt.date}{A date object, the next payment date to the investor}

\item{type}{a character vector the interest basis day count type}
}
\description{
Applies the correct bond basis (day count) convention to interest
calculations. Currently supported bond day count conventions:
\itemize{
\item{'30360' }{Agency MBS}
\item{'30E360' }{European 30360}
\item{'Actual360' }{used in money market}
\item{'Actual365' }{}
\item{'ActualActual' }
}
}
\seealso{
Other BondBasis: \code{\link{ActFactor}},
  \code{\link{ActualFactor}}, \code{\link{is.leapyear}}
}
\concept{BondBasis}
