% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BondCashFlow.R
\name{BondCashFlows}
\alias{BondCashFlows}
\title{Bond cash flow engine for standard non-callable}
\usage{
BondCashFlows(bond.id = "character", principal = numeric(),
  settlement.date = "character", price = numeric())
}
\arguments{
\item{bond.id}{A character string the cusip number or id of the bond}

\item{principal}{A numeric value the principal or face amount of the bond}

\item{settlement.date}{A character string the settlement date}

\item{price}{A numeric value the price of the bond}
}
\description{
Cashflow engine for standard non-callable bond
}
\examples{
\dontrun{BondCashFlows(bond.id = "bondlab10", principal = 1000,
settlement.date = "1-13-2013", price = 100)}
}

