% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondScenario.R
\docType{class}
\name{BondReturn-class}
\alias{BondReturn-class}
\title{An S4 class the results of bond total return scenario analysis}
\description{
The class BondReturn holds the results of bond total return analysis
}
\section{Slots}{

\describe{
\item{\code{CouponIncome}}{A numeric value the coupon income received over the 
investment horizon}

\item{\code{PrincipalReceived}}{A numeric value the scheduled principal 
received over the investment horizon}

\item{\code{ReinvestmentIncome}}{A numeric value the reivestment income received 
over the investment horizon}

\item{\code{HorizonCurrBal}}{A numeric value the current balance at the end of the
investment horizon}

\item{\code{HorizonPrice}}{A numeric the price at the end of the horizon}

\item{\code{HorizonReturn}}{A numeric value the horizon total return}

\item{\code{HorizonMos}}{A numeric value the number of months to 
the scenario horizon date}
}}

