% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondScenario.R
\docType{class}
\name{BondScenario-class}
\alias{BondScenario-class}
\title{An S4 class the bond return scenario analysis}
\description{
The SuperClass BondScenario holds the results of a scenario analysis run
BondScenario contains the following classes: TermStructure, BondCashFlow,
BondTermStructure, BondReturn, CurveSpreads, and Scenario.  BondScenario
inherts the getters and setters of the above classes.
}

