% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondScenario.R
\docType{class}
\name{BondScenario-class}
\alias{BondScenario-class}
\title{An S4 class the bond return scenario analysis}
\description{
The class BondScenario holds the results of a scenario analysis run
BondScenario contains the following super classes: TermStructure, 
BondCashFlow, BondTermStructure, BondReturn, CurveSpreads, and Scenario.
BondScenario inherts the getters and setters of the above classes.
}
\seealso{
Other Scenario.Analysis: \code{\link{BondReturn-class}},
  \code{\link{BondScenario}},
  \code{\link{CouponIncome,BondReturn-method}},
  \code{\link{Name,Scenario-method}},
  \code{\link{Scenario-class}}
}

