% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondScenario.R
\name{BondScenario}
\alias{BondScenario}
\title{BondScenario function}
\usage{
BondScenario(bond.id = "character", settlement.date = "character",
  scenario.curves = "character", price = "character",
  principal = numeric(), ..., horizon.spot.spread = NULL,
  horizon.nominal.spread = NULL, horizon.OAS = NULL, horizon.price = NULL)
}
\arguments{
\item{bond.id}{A character string referencing an object of the type BondDetails}

\item{settlement.date}{A character string the settlement data "mm-dd-YYYY".}

\item{scenario.curves}{A character string an object yield curve}

\item{price}{A character string in decimal equivalent (.) or 32nds (-)}

\item{principal}{A numeric value the par amount.}

\item{...}{Optional values to select term structure and horizon price method}

\item{horizon.spot.spread}{A numeric value the horizon zero volatility 
spread}

\item{horizon.nominal.spread}{A numeric value the horizon nominal spread
or spread to  the curve}

\item{horizon.OAS}{A numeric value the horizon option adjusted spread
(not currently implemented)}

\item{horizon.price}{A numeric value the horizon price}
}
\description{
A function to compute the \strong{total return} of a Bond.  The function 
first calculates the bond cashflows as of settlment date, rolls the bond 
forward per the horizon months, recomputes the cash flows, maturity and 
average life and prices the bond per the user's input.  The cash flow 
received, reinvestment income, and horizon price are used to compute the
investor's total return.  The function returns the class BondScenario
}
\seealso{
Other Bond Scenario Analysis: \code{\link{BondReturn-class}},
  \code{\link{BondScenario-class}},
  \code{\link{CouponIncome,BondReturn-method}},
  \code{\link{HorizonCurrBal,BondReturn-method}},
  \code{\link{HorizonMos,BondReturn-method}},
  \code{\link{HorizonPrice,BondReturn-method}},
  \code{\link{HorizonReturn,BondReturn-method}},
  \code{\link{PrincipalReceived,BondReturn-method}},
  \code{\link{ReinvestmentIncome,BondReturn-method}},
  \code{\link{ScenarioHorizonMos,ScenarioCurve-method}}
}
