% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondScenario.R
\docType{class}
\name{BondScenarioSet-class}
\alias{BondScenarioSet-class}
\title{An S4 class a list of bond scenarios for analysis}
\description{
The class BondScenarioSet is a list of classes of the type scenario
}
\section{Slots}{

\describe{
\item{\code{Scenario}}{A list of the classes of the type scenario}
}}

