% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondScenario.R
\name{BondScenarioSet}
\alias{BondScenarioSet}
\title{BondScenarioSet - constructor function}
\usage{
BondScenarioSet(ScenarioSet = "list")
}
\arguments{
\item{ScenarioSet}{a list of the scenario classes used for analysis}
}
\description{
A constructor function for the class BondScenarioSet.
}
\seealso{
Other Scenario.Analysis: \code{\link{BondReturn-class}},
  \code{\link{BondScenario-class}},
  \code{\link{BondScenarioSet-class}},
  \code{\link{BondScenario}},
  \code{\link{CouponIncome,BondReturn-method}},
  \code{\link{Name,Scenario-method}},
  \code{\link{Scenario-class}},
  \code{\link{ScenarioSet,BondScenarioSet-method}}
}

