% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondTermStructure.R
\name{BondTermStructure}
\alias{BondTermStructure}
\title{A function to calculate a bond key rate duration}
\usage{
BondTermStructure(bond.id, Rate.Delta = 0.5, TermStructure, principal, price,
  cashflow)
}
\arguments{
\item{bond.id}{A character string referencing an object of 
type bond details.}

\item{Rate.Delta}{A numeric value the rate delta used to calcualte the 
bond KRDs.  Default value is 0.50.  Do not change this unless you know what
you are doing.}

\item{TermStructure}{A character string referencing an object of typre
TermStructure.}

\item{principal}{A numeric value the principal or face amount of the bond.}

\item{price}{A character the price of the bond}

\item{cashflow}{A character string referencing an object of the 
BondCashFlow.}
}
\description{
This is a generic function used to contruct the object BondTermStructure
}
