% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterestRateModelCIR.R
\name{CalibrateCIR}
\alias{CalibrateCIR}
\title{A function to calibrate the Cox, Ingersoll, Ross single factor model}
\usage{
CalibrateCIR(trade.date = character, sigma = numeric())
}
\arguments{
\item{trade.date}{A numeric value the trade date}

\item{sigma}{the volatility in absolute percent the annualized value is sigma times sqrt(240)}
}
\description{
The function calibrates the CIR Model to daily swap curve data
}

