% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondCashFlows.R
\name{Convexity,BondCashFlows-method}
\alias{Convexity,BondCashFlows-method}
\title{Method to get Convexity from S4 class BondCashFlows}
\usage{
\S4method{Convexity}{BondCashFlows}(object)
}
\arguments{
\item{object}{the name of the S4 object}
}
\description{
Method to get Convexity from S4 class BondCashFlows
}
