% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageCashFlow.R
\docType{methods}
\name{Convexity,MortgageCashFlow-method}
\alias{Convexity,MortgageCashFlow-method}
\title{Method to extract Convexity from S4 class}
\usage{
\S4method{Convexity}{MortgageCashFlow}(object)
}
\arguments{
\item{object}{the name of the S4 object}
}
\description{
Method to extract Convexity from S4 class
}

