% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondScenario.R
\docType{methods}
\name{CouponIncome,BondReturn-method}
\alias{CouponIncome,BondReturn-method}
\title{BondReturn CouponIncome}
\usage{
\S4method{CouponIncome}{BondReturn}(object)
}
\arguments{
\item{object}{the name of an S4 class of type BondReturn}
}
\description{
A method to get the \strong{CouponIncome} paid over the scenario horzon
to the investor.  Coupon income reported as the sum of the coupon income 
received by the investor.
}
\seealso{
Other Scenario.Analysis: \code{\link{BondReturn-class}},
  \code{\link{BondScenario-class}},
  \code{\link{BondScenarioSet-class}},
  \code{\link{BondScenarioSet}},
  \code{\link{BondScenario}},
  \code{\link{Name,Scenario-method}},
  \code{\link{Scenario-class}},
  \code{\link{ScenarioSet,BondScenarioSet-method}}
}

