% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondScenario.R
\docType{methods}
\name{CouponIncome,BondReturn-method}
\alias{CouponIncome,BondReturn-method}
\title{CouponIncome method, class BondReturn}
\usage{
\S4method{CouponIncome}{BondReturn}(object)
}
\arguments{
\item{object}{An S4 class of type BondReturn}
}
\description{
A method to get the \strong{CouponIncome} paid over the scenario horizon
to the investor.  Coupon income is reported as the sum of the coupon income 
received by the investor.
}
\seealso{
Other Bond Scenario Analysis: \code{\link{BondReturn-class}},
  \code{\link{BondScenario-class}},
  \code{\link{BondScenario}},
  \code{\link{HorizonCurrBal,BondReturn-method}},
  \code{\link{HorizonMos,BondReturn-method}},
  \code{\link{HorizonPrice,BondReturn-method}},
  \code{\link{HorizonReturn,BondReturn-method}},
  \code{\link{PrincipalReceived,BondReturn-method}},
  \code{\link{ReinvestmentIncome,BondReturn-method}},
  \code{\link{ScenarioHorizonMos,ScenarioCurve-method}}
}

