% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Functions.R
\name{CPR.To.SMM}
\alias{CPR.To.SMM}
\title{A function to convert the CPR to a single monthly mortality rate SMM}
\usage{
CPR.To.SMM(CPR = numeric())
}
\arguments{
\item{CPR}{A numeric value}
}
\description{
Standard Generic Function used to convert CPR to SMM
}
\examples{
CPR.To.SMM(CPR = .06)
}

