% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurveSpreads.R
\docType{class}
\name{CurveSpreads-class}
\alias{CurveSpreads-class}
\title{CurveSpreads}
\description{
A class representing the curve spreads used to determine relative
value in the fixed income markets.
}
\section{Slots}{

\describe{
\item{\code{BenchMark}}{The maturity of the nearest maturity on the coupon pricing 
curve.}

\item{\code{SpreadToBenchmark}}{The yield (nominal) spread of the bond over the nearest
maturity point along the coupon pricing curve.}

\item{\code{SpreadToCurve}}{The yield spread over the interpolated point on the 
coupon curve matching the bond's weighted average life.}

\item{\code{ZeroVolSpread}}{The spread over the spot rate curve.  By market convention
this is referred to as the ZeroVolSpread, an output of the OAS model, outside an
OAS framework a more accurate reference to the measure is spread to the spot curve.}
}}

\seealso{
Other Pricing: \code{\link{BenchMark,CurveSpreads-method}},
  \code{\link{BillPriceToYield}},
  \code{\link{BillYieldToPrice}},
  \code{\link{CurveSpreads}},
  \code{\link{SpreadToBenchmark,CurveSpreads-method}},
  \code{\link{SpreadToCurve,CurveSpreads-method}},
  \code{\link{SpreadToPriceBond}},
  \code{\link{SpreadToPriceMBS}},
  \code{\link{ZeroVolSpread,CurveSpreads-method}}
}
