% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurveSpreads.R
\name{CurveSpreads}
\alias{CurveSpreads}
\title{A functon to compute the curve spread metrics}
\usage{
CurveSpreads(rates.data = "character", CashFlow = "character",
  TermStructure = "character", proceeds = numeric())
}
\arguments{
\item{rates.data}{a character string the trade date mm-dd-YYYY}

\item{CashFlow}{a character string the object of type MBSCashFlow}

\item{TermStructure}{a character string the object of type TermStructure}

\item{proceeds}{a numeric value the investor trade proceeeds 
MortgageCashFlow}
}
\description{
A functon to compute the curve spread metrics
}

