% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurveSpreads.R
\name{CurveSpreads}
\alias{CurveSpreads}
\title{A functon to compute the curve spread metrics}
\usage{
CurveSpreads(rates.data = "character", MortgageCashFlow = "character")
}
\arguments{
\item{rates.data}{A character string the trade date mm-dd-YYYY}

\item{MortgageCashFlow}{A character string the for object of type
MortgageCashFlow}
}
\description{
A functon to compute the curve spread metrics
}

