% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurveSpreads.R
\name{CurveSpreads}
\alias{CurveSpreads}
\title{Compute Curve Spreads}
\usage{
CurveSpreads(rates.data, CashFlow, TermStructure, proceeds)
}
\arguments{
\item{rates.data}{a character string referencing a rates.data object.}

\item{CashFlow}{a character string referencing an object of type MBSCashFlow or BondCashFlow}

\item{TermStructure}{a character string referencing an object of type TermStructure}

\item{proceeds}{a numeric value the investor trade proceeeds.}
}
\description{
Given proceeds, coupon curve, term structure, and cashflow compute
the curve spreads used to determine relative value in the fixed income markets.
the function returns the class CurveSpreads which contains the Benchmark, 
SpreadToBenchMark, SpreadToCurve, and ZeroVolSpread.
}
\seealso{
Other Pricing: \code{\link{CurveSpreads-class}},
  \code{\link{SpreadToPrice}}
}

