% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondConstructor.R
\docType{methods}
\name{Cusip,BondDetails-method}
\alias{Cusip,BondDetails-method}
\title{A method to extract Cusip from an S4 class of type BondDetails}
\usage{
\S4method{Cusip}{BondDetails}(object)

\S4method{Cusip}{BondDetails}(object)
}
\arguments{
\item{object}{The name of the S4 object of type BondDetails}

\item{object}{the name of the S4 object of type BondDetails}
}
\description{
A method to extract Cusip from an S4 class of type BondDetails

A method to extract Coupon from an S4 class of type BondDetails
}

