% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Functions.R
\name{Effective.Duration}
\alias{Effective.Duration}
\title{Effective.Duration}
\usage{
Effective.Duration(Rate.Delta = numeric(), cashflow = vector(),
  discount.rates = vector(), discount.rates.up = vector(),
  discount.rates.dwn = vector(), t.period = vector(),
  proceeds = numeric())
}
\arguments{
\item{Rate.Delta}{A numeric value the interest rate shift in basis points}

\item{cashflow}{A numeric vector of cash flow}

\item{discount.rates}{A numeric vector of discount rates}

\item{discount.rates.up}{A numeric vector of the up discount rates}

\item{t.period}{A numeric vector of the time period used for discounting}

\item{discount.rate.dwn}{A numeric vector of the down discount rates}
}
\description{
Calculates effective duarion based on a discount vector (zero coupon)
cash flow vector, and rate delta
}

