% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBSDetails.R
\name{FirstPrinPaymentDate,MBSDetails-method}
\alias{FirstPrinPaymentDate,MBSDetails-method}
\title{Method to extract the slot FirstPrinPaymentDate from the class MBSDetails}
\usage{
\S4method{FirstPrinPaymentDate}{MBSDetails}(object)
}
\arguments{
\item{object}{the name of the object of the type MBSDetails}
}
\description{
Method to extract the slot FirstPrinPaymentDate from the class MBSDetails
}
