% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TermStructure.R
\name{Forward.Rate}
\alias{Forward.Rate}
\title{A function to calculate forward rates}
\usage{
Forward.Rate(term.structure, forward.tenor, type = "C")
}
\arguments{
\item{term.structure}{an S4 object of the type TermStructure}

\item{forward.tenor}{the forward tenor stated in months.  For example the
two-year forward rate is specified as 24.}

\item{type}{the calculation type "P" periodic or "C" continuous}
}
\description{
Calculate forward rate given a vector of spot rates
}
