% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Functions.R
\name{Forward.Rate}
\alias{Forward.Rate}
\title{A function to calculate forward rates}
\usage{
Forward.Rate(SpotRate.Curve = vector(), FwdRate.Tenor)
}
\arguments{
\item{SpotRate.Curve}{A vector of monthly spot rates}

\item{FwdRate.Tenor}{A numeric value the tenor of the forward rate in months}
}
\description{
Calculate forward rate given a vector of spot rates
}

