% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TermStructure.R
\name{ForwardDate<-}
\alias{ForwardDate<-}
\title{A generic function to replace the slot Date in class TermStructure}
\usage{
ForwardDate(object) <- value
}
\arguments{
\item{object}{an S4 class object}

\item{value}{the replacement value of the slot}
}
\description{
A generic function to replace the slot Date in class TermStructure
}
