% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/HorizonBondData.R
\name{ForwardPassThrough}
\alias{ForwardPassThrough}
\title{A Function to "Roll Forward" the mortgage pass through cusip object (MBS Details)}
\usage{
ForwardPassThrough(bond.id = "character", original.bal = numeric(),
  projected.cashflow = "character", horizon.months = numeric())
}
\arguments{
\item{bond.id}{A character string referring to an object type of MBSDetails}

\item{original.bal}{A numeric value the original balance}

\item{projected.cashflow}{A character string referring to an object type of MortgageCashFlows}

\item{horizon.months}{A numeric value the investment horizon in months}
}
\description{
A standard generic function to roll forward MBS cusip detail to the scenario horizon
for pricing using spot spread, nominal spread, or option adjusted spread
}

