% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TermStructure.R
\docType{methods}
\name{ForwardRate,TermStructure-method}
\alias{ForwardRate,TermStructure-method}
\title{Method to extract ForwardRate from the class TermStructure}
\usage{
\S4method{ForwardRate}{TermStructure}(object)
}
\arguments{
\item{object}{the name of the object of type TermStructure}
}
\description{
Method to extract ForwardRate from the class TermStructure
}

