% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWacTypes.R
\name{GWacBasis,GWacTypes-method}
\alias{GWacBasis,GWacTypes-method}
\title{A method to get the value of GWacBasis slot in the object GWacTypes}
\usage{
\S4method{GWacBasis}{GWacTypes}(object)
}
\arguments{
\item{object}{an S4 object of the tyoe GWacTypes}
}
\description{
A method to get the value of GWacBasis slot in the object GWacTypes
}
