% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWacTypes.R
\docType{methods}
\name{GWacDecimal,GWacTypes-method}
\alias{GWacDecimal,GWacTypes-method}
\title{A method to get GWacDecimal from the object GWacTypes}
\usage{
\S4method{GWacDecimal}{GWacTypes}(object)
}
\arguments{
\item{object}{an S4 object of the type GWacDecimal}
}
\description{
A method to get GWacDecimal from the object GWacTypes
}

