% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondScenario.R
\docType{methods}
\name{HorizonCurrBal,BondReturn-method}
\alias{HorizonCurrBal,BondReturn-method}
\title{Bond Return HorizonCurrBal}
\usage{
\S4method{HorizonCurrBal}{BondReturn}(object)
}
\arguments{
\item{object}{An S4 class of type BondReturn}
}
\seealso{
Other Scenario.Analysis: \code{\link{BondReturn-class}},
  \code{\link{BondScenario-class}},
  \code{\link{BondScenario}},
  \code{\link{CouponIncome,BondReturn-method}},
  \code{\link{HorizonMos,BondReturn-method}},
  \code{\link{HorizonPrice,BondReturn-method}},
  \code{\link{HorizonReturn,BondReturn-method}},
  \code{\link{Name,Scenario-method}},
  \code{\link{PrincipalReceived,BondReturn-method}},
  \code{\link{PrincipalReceived}},
  \code{\link{ReinvestmentIncome,BondReturn-method}},
  \code{\link{Scenario-class}}
}

