% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\name{HorizonCurrBal}
\alias{HorizonCurrBal}
\title{A standard generic function to access the slot HorizonCurrBal}
\usage{
HorizonCurrBal(object)
}
\arguments{
\item{object}{an S4 object of type MtgScenario}
}
\description{
A standard generic function to access the slot HorizonCurrBal
}

