% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondScenario.R
\name{HorizonPrice,BondReturn-method}
\alias{HorizonPrice,BondReturn-method}
\title{HorizonPrice method, class BondReturn}
\usage{
\S4method{HorizonPrice}{BondReturn}(object)
}
\arguments{
\item{object}{An S4 class of type BondReturn}
}
\description{
A method to get \strong{HorizonPrice} at the end of the 
scenario horizon.
}
\seealso{
Other Bond Scenario Analysis: 
\code{\link{BondReturn-class}},
\code{\link{BondScenario-class}},
\code{\link{BondScenario}},
\code{\link{CouponIncome,BondReturn-method}},
\code{\link{HorizonCurrBal,BondReturn-method}},
\code{\link{HorizonMos,BondReturn-method}},
\code{\link{HorizonReturn,BondReturn-method}},
\code{\link{PrincipalReceived,BondReturn-method}},
\code{\link{ReinvestmentIncome,BondReturn-method}},
\code{\link{ScenarioHorizonMos,ScenarioCurve-method}}
}
\concept{Bond Scenario Analysis}
