% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\name{HorizonPrice}
\alias{HorizonPrice}
\title{HorizonPrice generic}
\usage{
HorizonPrice(object)
}
\arguments{
\item{object}{an S4 object of the type MtgScenario}
}
\description{
A generic function for method dispatch
}
\seealso{
Other Scenario Analysis: 
\code{\link{CouponIncome}()},
\code{\link{HorizonCurrBal}()},
\code{\link{HorizonCurveShift}()},
\code{\link{HorizonMos}()},
\code{\link{HorizonReturn}()},
\code{\link{Name,Scenario-method}},
\code{\link{PrepaidPrinReceived}()},
\code{\link{PrincipalReceived}()},
\code{\link{ReinvestmentIncome}()},
\code{\link{Scenario-class}},
\code{\link{ScenarioCurve-class}},
\code{\link{ScheduledPrinReceived}()}
}
\concept{Scenario Analysis}
