% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondScenario.R
\name{HorizonReturn,BondReturn-method}
\alias{HorizonReturn,BondReturn-method}
\title{HorizonReturn method, class BondReturn}
\usage{
\S4method{HorizonReturn}{BondReturn}(object)
}
\arguments{
\item{object}{the name of an S4 class of type BondReturn}
}
\description{
A method to get \strong{HorizonReturn}
}
\seealso{
Other Bond Scenario Analysis: 
\code{\link{BondReturn-class}},
\code{\link{BondScenario-class}},
\code{\link{BondScenario}},
\code{\link{CouponIncome,BondReturn-method}},
\code{\link{HorizonCurrBal,BondReturn-method}},
\code{\link{HorizonMos,BondReturn-method}},
\code{\link{HorizonPrice,BondReturn-method}},
\code{\link{PrincipalReceived,BondReturn-method}},
\code{\link{ReinvestmentIncome,BondReturn-method}},
\code{\link{ScenarioHorizonMos,ScenarioCurve-method}}
}
\concept{Bond Scenario Analysis}
