% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\name{HorizonReturn}
\alias{HorizonReturn}
\title{HorizonPrice generic}
\usage{
HorizonReturn(object)
}
\arguments{
\item{object}{An object of the type MtgScenario}
}
\description{
A generic function for method dispatch
}
\seealso{
Other Scenario Analysis: \code{\link{CouponIncome}},
  \code{\link{HorizonCurrBal}}, \code{\link{HorizonMos}},
  \code{\link{HorizonPrice}},
  \code{\link{Name,Scenario-method}},
  \code{\link{PrepaidPrinReceived}},
  \code{\link{PrincipalReceived}},
  \code{\link{ReinvestmentIncome}},
  \code{\link{Scenario-class}},
  \code{\link{ScheduledPrinReceived}}
}

