% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICStructuringFunctions.R
\name{Interest}
\alias{Interest}
\title{Function to calculate interest in the current period based on Beginning 
Balance.}
\usage{
Interest(node, frequency, period = numeric())
}
\arguments{
\item{node}{The entry point of the deal's tree structure}

\item{frequency}{the number of payments received over the calendar year}

\item{period}{A numeric value the period in which payment is received 
stated as integer}
}
\description{
The function takes beginning balance and frequency and period to 
calculate interest due the investor in the period.  The coupon is passed 
via the node argument from the deal's tree structure.  Structuring 
functions run on a tree structure created by the 
package data.tree.  See data.tree for more details on the package.
}

