% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondTermStructure.R
\docType{methods}
\name{Issuer,BondTermStructure-method}
\alias{Issuer,BondTermStructure-method}
\title{Issuer}
\usage{
\S4method{Issuer}{BondTermStructure}(object)
}
\arguments{
\item{object}{object of type BondTermStructure}
}
\description{
Get Issuer from object BondTermStructure
}

