% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BondConstructor.R
\name{MakeBondDetails}
\alias{MakeBondDetails}
\title{A constuctor function to create standard bond cusip information}
\usage{
MakeBondDetails(Cusip = "character", ID = "character",
  BondType = "character", Sector = "character", Coupon = numeric(),
  IssueDate = "character", DatedDate = "character",
  Maturity = "character", LastPmtDate = "character",
  NextPmtDate = "character", Moody = "character", SP = "character",
  BondLab = "character", Frequency = numeric(), BondBasis = "character",
  Callable = "character", Putable = "character",
  SinkingFund = "character")
}
\arguments{
\item{Cusip}{A character string the bond's cusip number}

\item{ID}{A character string the bond's ID}

\item{BondType}{A character string the type of bond}

\item{Sector}{A character string the bond sector}

\item{Coupon}{A character string the coupon (interest rate) paid to the investor}

\item{IssueDate}{A character string the issue date of the bond}

\item{DatedDate}{A character string the bond's dated date}

\item{Maturity}{A character string the bond's maturity date of final principal payment date}

\item{LastPmtDate}{A character string the most recent payment date to the investor}

\item{NextPmtDate}{A character string the next payment date due to the investor}

\item{Moody}{A character string the Moody assigned rating}

\item{SP}{A character string the SP assigned rating}

\item{BondLab}{A character string the Bond Lab assigned rating}

\item{Frequency}{A numeric value the payment frequency}

\item{BondBasis}{A character string the interest calculation basis}

\item{Callable}{A character string the bond's call schedule}

\item{Putable}{A caracter string the bond's put schedule}

\item{SinkingFund}{a character strign the bond sinking fund schedule}
}
\description{
This is a standard generic function used to construct a tradtional bond
}

