% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PassThroughConstructor.R
\name{MakeMBSDetails}
\alias{MakeMBSDetails}
\title{A constructor function to create a mortgage pass through security}
\usage{
MakeMBSDetails(Cusip = "character", ID = "character",
  BondType = "character", Sector = "character", Coupon = numeric(),
  IssueDate = "character", DatedDate = "character",
  Maturity = "character", LastPmtDate = "character",
  NextPmtDate = "character", PaymentDelay = numeric(),
  Moody = "character", SP = "character", BondLab = "character",
  Frequency = numeric(), BondBasis = "character", GWac = numeric(),
  OrigLoanBal = numeric(), OrigLTV = numeric(),
  AmortizationType = "character", AmortizationTerm = numeric(),
  Index = "character", Margin = numeric(), FirstPmtDate = "character",
  FinalPmtDate = "character", Servicing = numeric(), PMI = numeric(),
  Gfee = numeric(), InitialInterest = "character",
  InterestOnlyPeriod = numeric(), FirstPrinPaymentDate = "character",
  BalloonPmt = "character", BalloonDate = "character",
  MBSFactor = numeric(), Model = "character", Burnout = numeric(),
  SATO = numeric())
}
\arguments{
\item{Cusip}{A character the Pass Through MBS cusip.}

\item{ID}{A character string the pool number.}

\item{BondType}{A character string the type of Bond MBS, etc.}

\item{Sector}{A charcter string description of the Sector Mortgage, Utility, Government.}

\item{Coupon}{A numeric value the Bond Coupon.}

\item{IssueDate}{A character string the issue date of the security.}

\item{DatedDate}{A character sting The date following the issue when interest begins to accure.}

\item{Maturity}{A character sting the final payment date to the investor
in the case MBS the final payment data assuming 0 CPR.}

\item{LastPmtDate}{A character string the date the last payment scheduled payment to the investor.}

\item{NextPmtDate}{A character string the date of the next scheduled payment to the investor.}

\item{PaymentDelay}{A numeric value in the case of MBS the delay of the payment
from the trust to the investor}

\item{Moody}{A character string Moody's assigned credit rating}

\item{SP}{A character string SP's assigned credit rating}

\item{BondLab}{A character string BondLab's or the user's assigned credit rating}

\item{Frequency}{A numeric value string the frequency of payments made to the investor}

\item{BondBasis}{A character string the basis on which interest is calculated}

\item{GWac}{A numeric value the borrower's note rate}

\item{OrigLoanBal}{A numeric value the original balance of the loan}

\item{OrigLTV}{A numeric value the borrower's original loan to value ratio}

\item{AmortizationType}{A character sting the type of the loan fixed or floating}

\item{AmortizationTerm}{A numeric value the term of the loan in years}

\item{Index}{A character string if the amortization type is adjustable the Index
to which the note rate references}

\item{Margin}{A numeric value the spread over the index used to determine
the borrower's note rate}

\item{FirstPmtDate}{A character string the date of the first payment
of the borrower's note.  In the case of an the first payment made to the bondholder}

\item{FinalPmtDate}{A character string the date of thee final payment
of the borrower's note.  In the case of an MBS the final payment made to the bondholder}

\item{Servicing}{A numeric value the servicing spread from the Gross WAC to the servicer
of hte mortgage pool's underlying loans.}

\item{PMI}{A numeric value the primary mortage insurance paid by the borrower to the
PMI provider}

\item{Gfee}{A numeric value the guarantee fee taken from the borrower's note rate to
guarantee timely payment of principal and interest.  Applicable in the case of Fannie
Mae, Freddie Mac, or Ginnie Mae pools.}

\item{InitialInterest}{A character string indicating the note carries an interest only period}

\item{InterestOnlyPeriod}{A character string indicating the note's interest only period}

\item{FirstPrinPaymentDate}{A character string indicating the first principal payment date
due on an interest only mortgage.}

\item{BalloonPmt}{A character string indicating the mortgage carries a balloon pmt.}

\item{BalloonDate}{A character string the balloon payment date.}

\item{MBSFactor}{A numeric value the current factor of the MBS.}

\item{Model}{A character string the prepayment model to use.}

\item{Burnout}{A numeric model the value of the borrower burnout.}

\item{SATO}{A numeric value the borrrowers Spread AT Origination over the prime lending rate.}
}
\description{
This is a standard generic function used to construct a MBS pass through security
}

