% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\name{MakeRAID}
\alias{MakeRAID}
\title{A constructor function for the REMIC At Issuance Disclosure (RAID) file}
\usage{
MakeRAID(DealName = "character", Issuer = "character",
  DealPriceDate = "character", DealSettlementDate = "character",
  Underwriter = "character", NumberofTranches = numeric(),
  NumberPacSchedules = numeric(), NumberofGroups = numeric(),
  DealSize = numeric(), CollateralAmount = numeric())
}
\arguments{
\item{DealName}{A character string the deal name}

\item{Issuer}{A character string the Isser Name}

\item{DealPriceDate}{A character string the Deal Pricing Date}

\item{DealSettlementDate}{A character string the Deal Settlement Date}

\item{Underwriter}{A character string the Deal Underwriter}

\item{NumberofTranches}{A numeric string the Number of Tranches}

\item{NumberPacSchedules}{A numeric value the number of PAC schedules}

\item{NumberofGroups}{A numeric value the number of groups}

\item{DealSize}{A numeric value the original balance of all tranches}

\item{CollateralAmount}{A numeric value the current face amount of the collateral}
}
\description{
The RAID function creates the REMIC At Issuance Disclosure file
}
\examples{
\dontrun{
   MakeRAID(DealName = "BondLabPACInverse_test",
   Issuer = "Bondlab",
   DealPriceDate = "12-01-2012",
   DealSettlementDate = "01-01-2013",
   Underwriter = "Bondlab",
   NumberofTranches = 4,
   NumberPacSchedules = 1,
   NumberofGroups = 1,
   DealSize = 200000000,
   CollateralAmount = 200000000)}
}

