% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\name{MakeTranche}
\alias{MakeTranche}
\title{Makes Tranches for REMIC structure}
\usage{
MakeTranche(DealName = "character", TrancheNumber = "character",
  TrancheName = "character", TranchePrincipal = "character",
  TrancheInterest = "character", TranchePrincipalDesc = "character",
  TrancheInterestDesc = "character", Cusip = "character",
  TrancheOrigBal = numeric(), TrancheDatedDate = "character",
  TrancheFirstPmtDate = "character", TrancheLastPmtDate = "character",
  TrancheNextPmtDate = "character", TrancheCoupon = numeric(),
  Delay = numeric(), PrinPmtFrequency = numeric(),
  InterestPmtFrequency = numeric(), FloaterIndex = "character",
  FloaterMargin = numeric(), FloaterCap = numeric(),
  FloaterFloor = numeric(), FloaterFormula = "function",
  PacLowBand = numeric(), PacHighBand = numeric(), Group = numeric(),
  Schedule = "logical", Fixed = "logical")
}
\description{
Make Tranche data for REMIC constructor
}

