% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\name{MakeTranche}
\alias{MakeTranche}
\title{A constructor function for REMIC tranche detail}
\usage{
MakeTranche(DealName = "character", TrancheNumber = "character",
  TrancheName = "character", TranchePrincipal = "character",
  TrancheInterest = "character", TranchePrincipalDesc = "character",
  TrancheInterestDesc = "character", Cusip = "character",
  TrancheOrigBal = numeric(), TrancheDatedDate = "character",
  TrancheFirstPmtDate = "character", TrancheLastPmtDate = "character",
  TrancheNextPmtDate = "character", TrancheCoupon = numeric(),
  Delay = numeric(), PrinPmtFrequency = numeric(),
  InterestPmtFrequency = numeric(), FloaterIndex = "character",
  FloaterMargin = numeric(), FloaterCap = numeric(),
  FloaterFloor = numeric(), FloaterFormula = "function",
  PacLowBand = numeric(), PacHighBand = numeric(), Group = numeric(),
  Schedule = "logical", Fixed = "logical")
}
\arguments{
\item{DealName}{A character string the deal name}

\item{TrancheNumber}{A character string the Tranche Number}

\item{TrancheName}{A character string the the Tranche Name}

\item{TranchePrincipal}{A character string the principal type (pass-through, notional)}

\item{TrancheInterest}{A character string the interest type (fixed or floating)}

\item{TranchePrincipalDesc}{A character string the REMIC principal type (sequential, IO, PAC, etc.)}

\item{TrancheInterestDesc}{A character string the REMIC interest type (Fixed, Floating, Variable)}

\item{Cusip}{A character string the tranche cusip}

\item{TrancheOrigBal}{A character numeric value the original balance}

\item{TrancheDatedDate}{A character value the tranche dated date}

\item{TrancheFirstPmtDate}{A character value the tranche first payment date}

\item{TrancheLastPmtDate}{A character value the tranche last payment date}

\item{TrancheNextPmtDate}{A character value the tranche next payment date}

\item{TrancheCoupon}{A numeric value the tranche coupon}

\item{Delay}{A numeric value the delay days}

\item{PrinPmtFrequency}{A numeric value the principal payment frequency}

\item{InterestPmtFrequency}{A numeric value the interest payment frequency}

\item{FloaterIndex}{A character value the floater index name}

\item{FloaterMargin}{A numeric value the floater margin}

\item{FloaterCap}{A numeric value the floater cap}

\item{FloaterFloor}{A numeric value the floater floor}

\item{FloaterFormula}{A function the floater coupon formula}

\item{PacLowBand}{A numeric value the PAC Lower Band}

\item{PacHighBand}{A numeric value the PAC Upper Band}

\item{Group}{A numeric value the collateral group number}

\item{Schedule}{A logical indicating the PAC/TAC schedule}

\item{Fixed}{A logical indicating Fixed (TRUE) or Floating (FALSE) coupon}
}
\description{
MakeTranche a constructor function used to create a REMIC tranche
}
\examples{
\dontrun{
  MakeTranche(DealName = "BondLabPACInverse_test",
 TrancheNumber = "1",
 TrancheName = "A",
 TranchePrincipal = "Pass_Through",
 TrancheInterest = "Fix",
 TranchePrincipalDesc = "PAC",
 TrancheInterestDesc = "Fix",
 Cusip = "BondLabPAC2",
 TrancheOrigBal = 148769215,
 TrancheDatedDate  = "01-01-2013",
 TrancheFirstPmtDate = "01-15-2013",
 TrancheLastPmtDate = "12-15-2042",
 TrancheNextPmtDate = "01-15-2013",
 TrancheCoupon = 2.25,
 Delay = 15,
 PrinPmtFrequency = 12,
 InterestPmtFrequency = 12,
 FloaterIndex = "999",
 FloaterMargin = 0,
 FloaterCap = 0,
 FloaterFloor = 0,
 FloaterFormula = function(){},
 PacLowBand = 75,
 PacHighBand = 250,
 Group = 1,
 Schedule = TRUE,
 Fixed = TRUE)}
}

