% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ConnectionFunctions.R
\name{MBS}
\alias{MBS}
\title{A connection function to the BondData calling MBS cusips}
\usage{
MBS(MBS.id = "character")
}
\arguments{
\item{MBS.id}{A character string the MBS.id or cusip number current MBS.id is supported}
}
\description{
Opens a connection to the BondData folder to call MBS cusip data
}

