% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MortgageCashFlow.R
\name{MortgageCashFlow}
\alias{MortgageCashFlow}
\title{A function to compute the cash flow of a pool of securitized mortgages}
\usage{
MortgageCashFlow(bond.id = "character", original.bal = numeric(),
  settlement.date = "character", price = numeric(),
  PrepaymentAssumption = "character")
}
\arguments{
\item{bond.id}{A character string referencing the object MBSDetails}

\item{original.bal}{The original balance of the MBS pool}

\item{settlement.date}{The settlment date of the MBS trade}

\item{price}{The price traded.  Price is input as a whole number.
For example $102 is input as 102.00 not 1.02.}

\item{PrepaymentAssumption}{A character string referencing the class object
PrepaymentAssumption}
}
\description{
This is a generic function used to construct the class object MortgageCashFlow
For this function to work properly the classes MBSDetails and PrepaymentAssumption
must be present and loaded into the local environment
}
\examples{
\dontrun{
  MortgageCashFlow(bond.id = "bondlabMBS4", original.bal = 1000000000,
  settlement.date = "01/13/2013", price = 104.00, PrepaymentAssumption = "Prepayment")}
}

