% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageCashFlowArray.R
\docType{class}
\name{MortgageCashFlowArray-class}
\alias{MortgageCashFlowArray-class}
\title{An S4 Class a mortgage cashflow array}
\description{
This function is used in MortgageCashFlow_Array to facilitate
passing a cashflow array from an external cash flow engine
}
\section{Slots}{

\describe{
\item{\code{CashFlowArray}}{an array of mortgage cashflows}
}}

