% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageOAS.R
\docType{class}
\name{MortgageOAS-class}
\alias{MortgageOAS-class}
\title{mortgage oas class}
\description{
mortgage pass through oas class
}
\section{Slots}{

\describe{
\item{\code{Cusip}}{the pass through cusip}

\item{\code{Issuer}}{the pass through issuer}

\item{\code{Coupon}}{the pass through investor coupon}

\item{\code{Term}}{the pass though amortization term}

\item{\code{OAS}}{the mortgage option adjusted spread}

\item{\code{ZeroVolSpread}}{the zero volatility spread}

\item{\code{Spreads}}{the path spread along each simulation}

\item{\code{OAD}}{option adjusted duration}

\item{\code{OAC}}{option adjusted convexity}
}}

\seealso{
Other mortgage pass through option adjusted spread: \code{\link{CalibrateRates}},
  \code{\link{FitMarket}},
  \code{\link{MonteCarloMeanReversion}},
  \code{\link{OASRatesArray}}, \code{\link{SimCurve}},
  \code{\link{SimRates}}
}
