% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\docType{class}
\name{MortgageReturn-class}
\alias{MortgageReturn-class}
\title{An S4 class representing mortgage scenario return analysis}
\description{
The class MortgageReturn holds the results of mortgage return analysis
}
\section{Slots}{

\describe{
\item{\code{CouponIncome}}{A numeric value the Coupon Income received over the 
investment horizon}

\item{\code{ScheduledPrinReceived}}{A numeric value the scheduled principal 
received over the investment horizon}

\item{\code{PrepaidPrinReceived}}{A numeric value the prepaid principal received
over the investment horizon}

\item{\code{ReinvestmentIncome}}{A numeric value the reivestment income received 
over the investment horizon}

\item{\code{HorizonCurrBal}}{A numeric value the current balance at the end of the
investment horizon}

\item{\code{HorizonPrice}}{A numeric the price at the end of the horizon}

\item{\code{HorizonReturn}}{A numeric value the horizon total return}

\item{\code{HorizonMos}}{A numeric value the number of months to 
the scenario horizon date}
}}

