% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\docType{class}
\name{MortgageReturn-class}
\alias{MortgageReturn-class}
\title{MBS Total Return Metrics}
\description{
The class MortgageReturn holds the results of mortgage return analysis.
MortgageReturn is the sum of coupon income, principal received, reinvestment
income and price appreciation or depreciation.  In addition, most investors
also track horizon current balance.
}
\section{Slots}{

\describe{
\item{\code{CouponIncome}}{A numeric value the coupon income received over the 
investment horizon}

\item{\code{ScheduledPrinReceived}}{A numeric value the scheduled principal 
received over the investment horizon}

\item{\code{PrepaidPrinReceived}}{A numeric value the prepaid principal received
over the investment horizon}

\item{\code{ReinvestmentIncome}}{A numeric value the reivestment income received 
over the investment horizon}

\item{\code{HorizonCurrBal}}{A numeric value the current balance at the end of the
investment horizon}

\item{\code{HorizonPrice}}{A numeric the price at the end of the horizon}

\item{\code{HorizonReturn}}{A numeric value the horizon total return}

\item{\code{HorizonMos}}{A numeric value the number of months to 
the scenario horizon date}
}}
\seealso{
Other Mortgage Scenario Analysis: \code{\link{CouponIncome,MortgageReturn-method}},
  \code{\link{ScheduledPrinReceived,MortgageReturn-method}}
}

