% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\docType{class}
\name{MortgageScenario-class}
\alias{MortgageScenario-class}
\title{An S4 Class the results of mortgage return scenario analysis}
\description{
The SuperClass MortgageScenario holds the results of a scenario analysis run
MortgageScenario contains the following classes: TermStructure, 
PrepaymentAssumption, MortgageCashFlow, MortgageTermStructure, MortgageReturn,
ModelToCPR. CurveSpreads, and Scenario.  MortgageScenario inherits the 
getters of each of the above classes.
}

