% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\docType{class}
\name{MortgageScenario-class}
\alias{MortgageScenario-class}
\title{MortgageScenario class}
\description{
The call MortgageScenario contains the following super classes:
TermStructure, PrepaymentAssumption, MortgageCashFlow, MortgageTermStructure
MortgageReturn, ModelToCPR. CurveSpreads, and Scenario.  MortgageScenario
inherits the  getters and setter of each of the above super classes.
}
\seealso{
Other Mortgage Scenario Analysis: \code{\link{CouponIncome,MortgageReturn-method}},
  \code{\link{HorizonCurrBal,MortgageReturn-method}},
  \code{\link{HorizonMos,MortgageReturn-method}},
  \code{\link{HorizonPrice,MortgageReturn-method}},
  \code{\link{HorizonReturn,MortgageReturn-method}},
  \code{\link{MortgageReturn-class}},
  \code{\link{PrepaidPrinReceived,MortgageReturn-method}},
  \code{\link{ReinvestmentIncome,MortgageReturn-method}},
  \code{\link{ScheduledPrinReceived,MortgageReturn-method}}
}
\concept{Mortgage Scenario Analysis}
