% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Mtg.ScenarioSet-class}
\alias{Mtg.ScenarioSet-class}
\title{An S4 class representing a call to scenarios.}
\description{
An S4 class representing a call to scenarios.
}
\section{Slots}{

\describe{
\item{\code{Scenario}}{is a list which is used to call
the interest rate scenarios used for total
return and horizon analysis}
}}

