% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MortgageKeyRate.R
\name{MtgTermStructure}
\alias{MtgTermStructure}
\title{A function to calculate mortgage key rate duration}
\usage{
MtgTermStructure(bond.id = "character", original.bal = numeric(),
  Rate.Delta = numeric(), TermStructure = "character",
  settlement.date = "character", principal = numeric(), price = numeric(),
  cashflow = "character")
}
\arguments{
\item{bond.id}{A character string refrencing an object of type MBS details}

\item{original.bal}{A numeric value the original balance}

\item{Rate.Delta}{A numeric value the rate delta used to calculate KRDs}

\item{TermStructure}{A character string referencing an object of the type TermStructure}

\item{settlement.date}{a character string the settlement date}

\item{principal}{A numeric value the principal balance.  The principal balance is the
original balance multiplied by the MBS factor}

\item{price}{A numeric value the price paid}

\item{cashflow}{A character string referencing an object of the type MortgageCashFlow}
}
\description{
This is a generic function used to construct the MortgageTermStructure object
}

