% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScenarioConstructor.R
\docType{methods}
\name{Name,Scenario-method}
\alias{Name,Scenario-method}
\title{Scenario Name}
\usage{
\S4method{Name}{Scenario}(object)
}
\arguments{
\item{object}{The name of the object of the S4 class of type Scenario}
}
\description{
A method to get \strong{Name} from the object scenario.  This
can be useful for those investors saving the results of multiple scenario
analysis.
}
\seealso{
Other Scenario Analysis: \code{\link{CouponIncome}},
  \code{\link{HorizonCurrBal}},
  \code{\link{HorizonCurveShift}},
  \code{\link{HorizonMos}}, \code{\link{HorizonPrice}},
  \code{\link{HorizonReturn}},
  \code{\link{PrepaidPrinReceived}},
  \code{\link{PrincipalReceived}},
  \code{\link{ReinvestmentIncome}},
  \code{\link{Scenario-class}},
  \code{\link{ScenarioCurve-class}},
  \code{\link{ScheduledPrinReceived}}
}
