% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\docType{methods}
\name{Name,Scenario-method}
\alias{Name,Scenario-method}
\title{A method to extract Scenario description from S4 class Scenario}
\usage{
\S4method{Name}{Scenario}(object)
}
\arguments{
\item{object}{The name of the S4 object of class Scnenario}
}
\description{
A method to extract Scenario description from S4 class Scenario
}

