% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondCashFlows.R
\docType{methods}
\name{Period,BondCashFlows-method}
\alias{Period,BondCashFlows-method}
\title{Method to extract Period from S4 class BondCashFlows}
\usage{
\S4method{Period}{BondCashFlows}(object)
}
\arguments{
\item{object}{the name of the S4 object}
}
\description{
Method to extract Period from S4 class BondCashFlows
}

