% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TermStructure.R
\name{Period,TermStructure-method}
\alias{Period,TermStructure-method}
\title{Method to extract Period from S4 class TermStructure}
\usage{
\S4method{Period}{TermStructure}(object)
}
\arguments{
\item{object}{the name of the S4 object TermStructure}
}
\description{
Method to extract Period from S4 class TermStructure
}
