% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageCashFlow.R
\name{PrepaidPrin}
\alias{PrepaidPrin}
\title{A standard generic function to access the slot PrepaidPrin}
\usage{
PrepaidPrin(object)
}
\arguments{
\item{object}{an object whose method signature is KDS_PassThrough}
}
\description{
A standard generic function to access the slot PrepaidPrin
}

