% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\docType{methods}
\name{PrepaidPrinReceived,MortgageReturn-method}
\alias{PrepaidPrinReceived,MortgageReturn-method}
\title{PrepaidPrinReceived method, class MortgageReturn}
\usage{
\S4method{PrepaidPrinReceived}{MortgageReturn}(object)
}
\arguments{
\item{object}{An S4 class of type MortgageReturn}
}
\description{
A method to get \strong{PrepaidPrinReceived} paid to the 
investor over the scenario horizon.  PrepaidPrinReceived is reported as
the sum of the prepaid principal received over the horizon.
}

