% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\docType{methods}
\name{PrepaidPrinReceived,MortgageReturn-method}
\alias{PrepaidPrinReceived,MortgageReturn-method}
\title{PrepaidPrinReceived method, class MortgageReturn}
\usage{
\S4method{PrepaidPrinReceived}{MortgageReturn}(object)
}
\arguments{
\item{object}{An S4 class of type MortgageReturn}
}
\description{
A method to get \strong{PrepaidPrinReceived} paid to the 
investor over the scenario horizon.  PrepaidPrinReceived is reported as
the sum of the prepaid principal received over the horizon.
}
\seealso{
Other Mortgage Scenario Analysis: \code{\link{CouponIncome,MortgageReturn-method}},
  \code{\link{HorizonCurrBal,MortgageReturn-method}},
  \code{\link{HorizonMos,MortgageReturn-method}},
  \code{\link{HorizonPrice,MortgageReturn-method}},
  \code{\link{HorizonReturn,MortgageReturn-method}},
  \code{\link{MortgageReturn-class}},
  \code{\link{MortgageScenario-class}},
  \code{\link{ReinvestmentIncome,MortgageReturn-method}},
  \code{\link{ScheduledPrinReceived,MortgageReturn-method}}
}

