% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\name{PrepaidPrinReceived}
\alias{PrepaidPrinReceived}
\title{A standard generic function to access the slot PrepaidPrinReceived}
\usage{
PrepaidPrinReceived(object)
}
\arguments{
\item{object}{an S4 object of type MtgScenario}
}
\description{
A standard generic function to access the slot PrepaidPrinReceived
}

