% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PrepaymentModel.R
\name{PrepaymentAssumption}
\alias{PrepaymentAssumption}
\title{A contstructor function for the PrepaymentAssumption object}
\usage{
PrepaymentAssumption(bond.id = "character", TermStructure = "character",
  MortgageRate = "character", ModelTune = "character",
  Burnout = numeric(), PrepaymentAssumption = "character", ...,
  begin.cpr = numeric(), end.cpr = numeric(),
  seasoning.period = numeric(), CPR = numeric(), HomePrice = NULL,
  Severity = numeric())
}
\arguments{
\item{bond.id}{A character string the bond id or cusip}

\item{TermStructure}{A character string the method used to fit the term structure}

\item{MortgageRate}{A character string the input value of mortgagerate.rds.  Prepayment Assumption does
open Mtg.Rate connection directly rather takes the argument as an object.}

\item{ModelTune}{A character string the prepayment model tune object}

\item{Burnout}{A numeric value the burnout variable}

\item{PrepaymentAssumption}{A character string the prepayment assumption used "MODEL", "PPC", or "CPR"}

\item{...}{Optional values when "PPC" or "CPR" is used}

\item{begin.cpr}{A numeric value the beginning CPR assumption}

\item{end.cpr}{A numeric value the ending CPR assumption}

\item{seasoning.period}{A numeric value the length of the seasoning ramp}

\item{CPR}{A numeric value the CPR assumption}

\item{HomePrice}{NULL do not override value}

\item{Severity}{A numeric value the loss severity given default}
}
\description{
The function is a constructor function for the PrepaymentAssumption object
}

