% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModel.R
\name{PrepaymentModel}
\alias{PrepaymentModel}
\title{A contstructor function for the PrepaymentModel object}
\usage{
PrepaymentModel(bond.id, term.structure, prepayment.assumption = "MODEL",
  ..., begin.cpr = 0.002, end.cpr = 0.06, seasoning.period = 30,
  cpr = 0.06, cdr = 0.01, severity = 0.25)
}
\arguments{
\item{bond.id}{A character string referring to an object 
of the type MBSDetails}

\item{term.structure}{A character string referring to an object 
of the type TermStructure}

\item{prepayment.assumption}{A character string the prepayment assumption 
used "MODEL", "PPC", or "CPR"}

\item{...}{Optional values when "PPC" or "CPR" is used}

\item{begin.cpr}{A numeric value the beginning CPR assumption}

\item{end.cpr}{A numeric value the ending CPR assumption}

\item{seasoning.period}{A numeric value the length of the seasoning ramp}

\item{cpr}{A numeric value the CPR assumption (annual prepayment rate)}

\item{cdr}{A numeric value the CDR assumption (annual default rate)}

\item{severity}{A numeric value the loss severity given default}
}
\description{
The function is a constructor function for the PrepaymentModel object
}
