% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelFunctions.R
\docType{class}
\name{PrepaymentModelFunctions-class}
\alias{PrepaymentModelFunctions-class}
\title{An S4 class PrepaymentModelFunctions containing functions
for building an agency prepayment model}
\description{
An S4 class PrepaymentModelFunctions containing functions
for building an agency prepayment model
}
\section{Slots}{

\describe{
\item{\code{TurnoverRate}}{an numeric value the Turnover Rate}

\item{\code{SeasoningRamp}}{a function defining the seasoning ramp}
}}

