% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepaymentModelFunctions.R
\name{PrepaymentModelFunctions}
\alias{PrepaymentModelFunctions}
\title{A constructor function for the class ModelFunctions}
\usage{
PrepaymentModelFunctions(TurnoverRate = "numeric",
  SeasoningRamp = "function", Curtailment = "function",
  ArcTanIncentive = "function", BorrowerBurnout = "function",
  DefaultRamp = "function", DefaultOrigLTVMult = "function",
  DefaultUpdatedLTVMult = "function", DefaultSATOMult = "function")
}
\arguments{
\item{TurnoverRate}{a function defining the turnover rate}

\item{SeasoningRamp}{a function defining the seasoning ramp}

\item{Curtailment}{a function defining the curtailment}

\item{ArcTanIncentive}{a function defining the borrower incentive response}

\item{BorrowerBurnout}{a function defining borrower burnout}

\item{DefaultRamp}{a function defining the default ramp}

\item{DefaultOrigLTVMult}{a function defining the orig. LTV multipliers}

\item{DefaultUpdatedLTVMult}{a function defining the updated LTV default multipliers}

\item{DefaultSATOMult}{a function defining the SATO multipliers on the default curve}
}
\description{
A constructor function for the class ModelFunctions
}

