% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondCashFlows.R
\docType{methods}
\name{Price,BondCashFlows-method}
\alias{Price,BondCashFlows-method}
\title{Method to get Price from S4 class}
\usage{
\S4method{Price}{BondCashFlows}(object)
}
\arguments{
\item{object}{the name of the S4 object BondCashFlows}
}
\description{
Method to get Price from S4 class
}
