% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PriceTypes.R
\docType{methods}
\name{PriceDecimalString<-,PriceTypes-method}
\alias{PriceDecimalString<-,PriceTypes-method}
\title{A method to set PriceDecimalString from slot of class PriceTypes}
\usage{
\S4method{PriceDecimalString}{PriceTypes}(object) <- value
}
\arguments{
\item{object}{an S4 object of type PriceTypes}

\item{value}{the replacement value of the slot}
}
\description{
A method to set PriceDecimalString from slot of class PriceTypes
}
