% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PriceTypes.R
\name{PriceDecimalString<-}
\alias{PriceDecimalString<-}
\title{A standard generic to set the slot PriceDecimalString}
\usage{
PriceDecimalString(object) <- value
}
\arguments{
\item{object}{A S4 object of the type PriceTypes}

\item{value}{The replacement value of the slot}
}
\description{
A standard generic to set the slot PriceDecimalString
}

