% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PriceYieldClasses.R
\name{PriceTypes}
\alias{PriceTypes}
\title{A constructor function for the class PriceTypes}
\usage{
PriceTypes(Price = "character")
}
\arguments{
\item{Price}{character the price in either 
decimal notation (example "100.125") or 32nds notation (example "100-4")}
}
\description{
A constructor function for the class PriceTypes
}

