% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PriceYieldClasses.R
\name{PriceTypes}
\alias{PriceTypes}
\title{A constructor function for the class PriceTypes}
\usage{
PriceTypes(Price = numeric())
}
\arguments{
\item{Price}{the price in decimal notation (example 100.125)}
}
\description{
A constructor function for the class PriceTypes
}

