% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondScenario.R
\docType{methods}
\name{PrincipalReceived,BondReturn-method}
\alias{PrincipalReceived,BondReturn-method}
\title{Bond Return PrincipalReceived}
\usage{
\S4method{PrincipalReceived}{BondReturn}(object)
}
\arguments{
\item{object}{An S4 class of type BondReturn}
}
\description{
A method to get \strong{ScheduledPrinReceived} over the 
scenario horizon to the investor.  SceheduledPrincipalReceived is reported 
as the sum of scheduled principal received by the investor.
}
\seealso{
Other Scenario Analysis: \code{\link{BondReturn-class}},
  \code{\link{BondScenario-class}},
  \code{\link{BondScenario}},
  \code{\link{CouponIncome,BondReturn-method}},
  \code{\link{CouponIncome}},
  \code{\link{HorizonCurrBal,BondReturn-method}},
  \code{\link{HorizonMos,BondReturn-method}},
  \code{\link{HorizonPrice,BondReturn-method}},
  \code{\link{HorizonReturn,BondReturn-method}},
  \code{\link{MortgageReturn-class}},
  \code{\link{Name,Scenario-method}},
  \code{\link{PrepaidPrinReceived}},
  \code{\link{PrincipalReceived}},
  \code{\link{ReinvestmentIncome,BondReturn-method}},
  \code{\link{ReinvestmentIncome}},
  \code{\link{Scenario-class}},
  \code{\link{ScheduledPrinReceived}}
}

