% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondScenario.R
\name{PrincipalReceived}
\alias{PrincipalReceived}
\title{PrincipalReceived generic}
\usage{
PrincipalReceived(object)
}
\arguments{
\item{object}{an object of the type BondScenario}
}
\description{
A  generic function for method dispatch
}
\seealso{
Other Scenario.Analysis: \code{\link{BondReturn-class}},
  \code{\link{BondScenario-class}},
  \code{\link{BondScenario}},
  \code{\link{CouponIncome,BondReturn-method}},
  \code{\link{HorizonCurrBal,BondReturn-method}},
  \code{\link{HorizonMos,BondReturn-method}},
  \code{\link{HorizonPrice,BondReturn-method}},
  \code{\link{HorizonReturn,BondReturn-method}},
  \code{\link{Name,Scenario-method}},
  \code{\link{PrincipalReceived,BondReturn-method}},
  \code{\link{ReinvestmentIncome,BondReturn-method}},
  \code{\link{Scenario-class}}
}

