% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ConnectionFunctions.R
\name{Rates}
\alias{Rates}
\title{A connection function to the RatesData folder to call swap curve data}
\usage{
Rates(trade.date = "character")
}
\arguments{
\item{trade.date}{A character string the trade date}
}
\description{
A read connection function to the rates data base folder to call swap rate data
}

