% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\name{ReinvestmentIncome,MortgageReturn-method}
\alias{ReinvestmentIncome,MortgageReturn-method}
\title{ReinvestmentIncome method, class MortgageReturn}
\usage{
\S4method{ReinvestmentIncome}{MortgageReturn}(object)
}
\arguments{
\item{object}{An S4 class of the type MortgageReturn}
}
\description{
A method to get \strong{ReinvestmentIncome} paid to the 
investor over the scenario horizon.  ReinvesmentIncome is reported as
the sum of the reinvestment income received over the horizon.
}
\seealso{
Other Mortgage Scenario Analysis: 
\code{\link{CouponIncome,MortgageReturn-method}},
\code{\link{HorizonCurrBal,MortgageReturn-method}},
\code{\link{HorizonMos,MortgageReturn-method}},
\code{\link{HorizonPrice,MortgageReturn-method}},
\code{\link{HorizonReturn,MortgageReturn-method}},
\code{\link{MortgageReturn-class}},
\code{\link{MortgageScenario-class}},
\code{\link{PrepaidPrinReceived,MortgageReturn-method}},
\code{\link{ScheduledPrinReceived,MortgageReturn-method}}
}
\concept{Mortgage Scenario Analysis}
