% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\name{ReinvestmentIncome}
\alias{ReinvestmentIncome}
\title{ReinvestmentIncome generic}
\usage{
ReinvestmentIncome(object)
}
\arguments{
\item{object}{an object of the type MtgScenario}
}
\description{
A generic function for method dispatch
}
\seealso{
Other Scenario Analysis: \code{\link{CouponIncome}},
  \code{\link{HorizonCurrBal}},
  \code{\link{HorizonCurveShift}},
  \code{\link{HorizonMos}}, \code{\link{HorizonPrice}},
  \code{\link{HorizonReturn}},
  \code{\link{Name,Scenario-method}},
  \code{\link{PrepaidPrinReceived}},
  \code{\link{PrincipalReceived}},
  \code{\link{Scenario-class}},
  \code{\link{ScenarioCurve-class}},
  \code{\link{ScheduledPrinReceived}}
}
