% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Functions.R
\name{Remain.Balance}
\alias{Remain.Balance}
\title{A function to compute the remaining balance of a mortgage}
\usage{
Remain.Balance(balance = numeric(), note.rate = numeric(),
  term.mos = numeric(), period = numeric())
}
\arguments{
\item{balance}{A numeric value the balance of the mortgage}

\item{note.rate}{A numeric value the borrower's note rate}

\item{term.mos}{A numeric value the term of the mortgage}

\item{period}{A numeric value the numeric of months from the first payment}
}
\description{
A standard generic for the calculation of a mortgage remaining balance
}
\examples{
Remain.Balance(balance = 100000, note.rate = .045, term.mos = 360, period = 10)
}

