% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/REMICAnalytics.R
\name{REMICAnalytics}
\alias{REMICAnalytics}
\title{The REMIC Analytics engine the constructor function for REMIC Analytics}
\usage{
REMICAnalytics(bond.id = "character", trade.date = "character",
  settlement.date = "character", method = "character",
  collateral.price = numeric(), tranche.price = numeric(),
  PrepaymentAssumption = "character", ..., begin.cpr = numeric(),
  end.cpr = numeric(), seasoning.period = numeric(), CPR = numeric())
}
\arguments{
\item{bond.id}{A character string the bond.id or cusip number}

\item{trade.date}{A character string the trade date}

\item{settlement.date}{A character string the settlement date}

\item{method}{A character string the method used to fit the term strucutre}

\item{collateral.price}{A numeric value the price of the underlying collateral}

\item{tranche.price}{A numeric value the price of the tranche}

\item{PrepaymentAssumption}{A character string defining the prepayment assumption used}

\item{...}{Optional values when CPR and PPC are input}

\item{begin.cpr}{optional, a numeric value used when applying a PPC assumption}

\item{end.cpr}{optional, a numeric value used when applying a PPC assumption}

\item{seasoning.period}{optional, a numeric value used when applying a PPC assumption}

\item{CPR}{optional, a numeric value used when applying a CPR assumption}
}
\description{
Cash flow analysis and risk metrics of REMIC Tranches including spot spread analysis
effective duration, effective convexity, Key rate duration analysis
}
\examples{
\dontrun{REMICAnalytics(bond.id = "BondLabSMBSIO", trade.date = "01-10-2013",
settlement.date = "01-13-2013", method = "ns", collateral.price = 105.75,
tranche.price = 25, PrepaymentAssumption = "CPR", CPR = 6)}
}

