% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ConnectionFunctions.R
\name{REMICDeal}
\alias{REMICDeal}
\title{A read connection function to the REMICData folder}
\usage{
REMICDeal(remic.deal = "character")
}
\arguments{
\item{remic.deal}{A character string the REMIC deal name}
}
\description{
Opens a connection to the REMICData folder calling a REMIC Deal by name
this function supports REMIC tranche analysis by providing the REMIC
deal information as the umbrella structure over the tranche
}

