% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\name{RemicStructure}
\alias{RemicStructure}
\title{The constructor function to build the REMIC deal from each of its elements}
\usage{
RemicStructure(DealName = "character")
}
\arguments{
\item{DealName}{A character string the deal's names}
}
\description{
The function Assembles the deal RAID, RDME, Tranches, and Groups files into a REMIC structure
}
\examples{
\dontrun{
RemicStructure("BondLabPACInverse")
}
}

