% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ConnectionFunctions.R
\name{SaveCollGroup}
\alias{SaveCollGroup}
\title{A connection functon to the Groups folder}
\usage{
SaveCollGroup(FileName = "character", DealName = "character",
  Group = numeric())
}
\arguments{
\item{FileName}{A character string the FileName default to temp}

\item{DealName}{A character string the DealName}

\item{Group}{A numeric value the Group number
Note: Group is a counter for this function connection used in the CollateralGroup function.}
}
\description{
Opens a connection to the Groups folder to save collateral group
}

