% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ConnectionFunctions.R
\name{SaveModelTune}
\alias{SaveModelTune}
\title{A connection function to sace the PrepaymentModel model tuning parameters}
\usage{
SaveModelTune(ModelFile = "character", ModelName = "character")
}
\arguments{
\item{ModelFile}{A character string the model tuning default value is temp}

\item{ModelName}{A character string the model name used to reference the prepayment model}
}
\description{
Opens a connection to the PrepaymentModel folder to save the prepayment model tuning
}

