% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectionFunctions.R
\name{SaveMortgageRate}
\alias{SaveMortgageRate}
\title{A connection function to save the prepayment model folder to 
save teh mortgage rate function class}
\usage{
SaveMortgageRate(ModelFile, ModelName)
}
\arguments{
\item{ModelFile}{A character string the model tuning default value is temp}

\item{ModelName}{A character string the model name used to reference the
prepayment model}
}
\description{
A connection function to save the prepayment model folder to 
save teh mortgage rate function class
}

