% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectionFunctions.R
\name{SaveMortgageRate}
\alias{SaveMortgageRate}
\title{A connection function to save the PrepaymentModel model tuning parameters}
\usage{
SaveMortgageRate()
}
\description{
Opens a connection to the PrepaymentModel folder to save the MortgageRate
object
}
