% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectionFunctions.R
\name{SaveScenario}
\alias{SaveScenario}
\title{A save connection function to the interest rate scenario data folder}
\usage{
SaveScenario(Scenario = "character", ScenarioFile = "character")
}
\arguments{
\item{Scenario}{A character string the name of the Scenario}

\item{ScenarioFile}{the scenario file to save}
}
\description{
Opens a connection to the Scenarios folder to save user defined scenarios by name
}

