% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\docType{class}
\name{Scenario-class}
\alias{Scenario-class}
\title{An S4 Class representing scenario analysis details}
\description{
The class Scenario is the representation of the 
scenario run by the investor
}
\section{Slots}{

\describe{
\item{\code{Name}}{A character the name of the scenario}

\item{\code{Type}}{A character the type of the scenario}

\item{\code{Horizon}}{A character the time horizon over which the 
scenario is run}

\item{\code{ShiftType}}{A character the interest rate shift type (e.g.
twist, parallel, bull flatten, or bull steepen)}

\item{\code{Shiftbps}}{A numeric value the scenario shift in basis points}

\item{\code{Formula}}{A function represnting the shift which applied to 
the term structure}
}}

