% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScenarioConstructor.R
\docType{class}
\name{Scenario-class}
\alias{Scenario-class}
\title{An S4 class Scenario}
\description{
The interest rate shift function applied the yield curve.  The shift may
be applied to the spot rate curve slot of the object TermStructure or the
rates data object.  The BondLab convention is to identify shifts applied to
the sopt rate curve with a trailing for example D50s is down 50bps applied 
to the spot rate curve.  D50 is down 50bps applied to the rates data object.
In all cases the curve shift is not allowed to result in a negative rate. 
Thus, all shifts are subject to a floor of 0.01.  It is recommend that those
users wishing to create custom yield shift scenarios avoid 0 or negative 
rates when creating custom shift scenarios.
}
\section{Slots}{

\describe{
\item{\code{Name}}{A character the name of the scenario}

\item{\code{Type}}{A character the type of scenario eg("Immediate", "Gradual")}

\item{\code{ShiftType}}{A character indicating the type shift 
eg("Parallel", "Twist")}

\item{\code{Shiftbps}}{A numeric value the interest rate shift in bps}

\item{\code{Formula}}{A function defining the interest rate shift}
}}

\seealso{
Other Scenario Analysis: \code{\link{CouponIncome}},
  \code{\link{HorizonCurrBal}}, \code{\link{HorizonMos}},
  \code{\link{HorizonPrice}}, \code{\link{HorizonReturn}},
  \code{\link{Name,Scenario-method}},
  \code{\link{PrepaidPrinReceived}},
  \code{\link{PrincipalReceived}},
  \code{\link{ReinvestmentIncome}},
  \code{\link{ScheduledPrinReceived}}
}
