% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScenarioConstructor.R
\name{Scenario}
\alias{Scenario}
\title{A constructor function of the class Scenario}
\usage{
Scenario(
  Name = "character",
  Type = "character",
  ShiftType = "character",
  Shiftbps = "character",
  Formula = "function"
)
}
\arguments{
\item{Name}{A character string indicating the unique name which 
identifies the scenario}

\item{Type}{A character string indicating the type of scenario 
"aggressive", "moderate"}

\item{ShiftType}{A character string indicating the type interest rate 
shift "Parallel", "Twist", "etc"}

\item{Shiftbps}{A numeric vector indicating the shift applied to each 
point on the curve}

\item{Formula}{A function to apply the shift to each pont on the curve}
}
\description{
This is a standard generic function used to construct the class Scenario.
The Scenario class is applied to the either a Rates object or a 
TermStructure object
}
\examples{
\dontrun{
MakeScenario(Name = "Flat50",
Type = "Immediate",
ShiftType = "Twist",
Shiftbps = -75,
Formula = function(rates.data, Shiftbps){
as.character(as.numeric(rates.data[1,2:length(rates.data)]) + 
Shiftbps/yield.basis)})}  
}
