% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ConnectionFunctions.R
\name{ScenarioCall}
\alias{ScenarioCall}
\title{A connection function to the interest rate scenario data folder}
\usage{
ScenarioCall(Scenario = "character")
}
\arguments{
\item{Scenario}{a character string the scenario used in the analysis}
}
\description{
Opens connection to the Scenarios folder calling user defined scenarios by name
}

