% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondScenario.R
\docType{methods}
\name{ScenarioSet,BondScenarioSet-method}
\alias{ScenarioSet,BondScenarioSet-method}
\title{A Method to extract the slot Scenario from the object BondScenarioSet}
\usage{
\S4method{ScenarioSet}{BondScenarioSet}(object)
}
\arguments{
\item{object}{The name of an S4 object of the type BondScenarioSet}
}
\description{
A Method to extract the slot Scenario from the object BondScenarioSet
}

