% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Functions.R
\name{Sched.Prin}
\alias{Sched.Prin}
\title{A function to calculate the scheduled principal of mortgage}
\usage{
Sched.Prin(balance = numeric(), note.rate = numeric(),
  term.mos = numeric(), period = numeric())
}
\arguments{
\item{balance}{A numeric value, the balance of the mortgage}

\item{note.rate}{A numeric value, the borrower's note rate}

\item{term.mos}{A numeric value, the term of the mortgage}

\item{period}{A numeric value the number of months since the first payment}
}
\description{
A standard generic for the calculation of scheduled principal of mortgage loan
}
\examples{
Sched.Prin(balance = 100000, note.rate = .045, term.mos = 360, period = 2)
}

