% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageScenario.R
\docType{methods}
\name{ScheduledPrinReceived,MortgageReturn-method}
\alias{ScheduledPrinReceived,MortgageReturn-method}
\title{ScheduledPrinReceived method, class MortgageReturn}
\usage{
\S4method{ScheduledPrinReceived}{MortgageReturn}(object)
}
\arguments{
\item{object}{An S4 class of type MortgageReturn}
}
\description{
A method to get the \strong{ScheduledPrinRecieved} paid to the
investor over the scenario horizon.  ScheduledPrinReceived is reported as 
the sum of the sceduled principal received over the horizon.
}
\seealso{
Other Mortgage Scenario Analysis: \code{\link{CouponIncome,MortgageReturn-method}},
  \code{\link{HorizonCurrBal,MortgageReturn-method}},
  \code{\link{HorizonMos,MortgageReturn-method}},
  \code{\link{HorizonPrice,MortgageReturn-method}},
  \code{\link{HorizonReturn,MortgageReturn-method}},
  \code{\link{MortgageReturn-class}},
  \code{\link{MortgageScenario-class}},
  \code{\link{PrepaidPrinReceived,MortgageReturn-method}},
  \code{\link{ReinvestmentIncome,MortgageReturn-method}}
}
\concept{Mortgage Scenario Analysis}
