% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ServicingFeeTypes.R
\name{ServicingFeeBasis,ServicingFeeTypes-method}
\alias{ServicingFeeBasis,ServicingFeeTypes-method}
\title{A method to get the slot ServicingFeeBasis from the object 
ServicingFeeTypes}
\usage{
\S4method{ServicingFeeBasis}{ServicingFeeTypes}(object)
}
\arguments{
\item{object}{an S4 object of the type ServicingFeeTypes}
}
\description{
A method to get the slot ServicingFeeBasis from the object 
ServicingFeeTypes
}
