% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondDetails.R
\name{SinkingFund}
\alias{SinkingFund}
\title{A standard generic function to get the slot SinkingFund}
\usage{
SinkingFund(object)
}
\arguments{
\item{object}{an S4 class object}
}
\description{
A standard generic function to get the slot SinkingFund
}
