% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Functions.R
\name{SMMVector.To.CPR}
\alias{SMMVector.To.CPR}
\title{A function to convert  a time series of SMM to CPR}
\usage{
SMMVector.To.CPR(SMM = vector(), num.period = vector())
}
\arguments{
\item{SMM}{A numeric vector of SMM}

\item{num.period}{a numeric vector of period (Loan Age)}
}
\description{
A standard generic function for the conversion a SMM time series to CPR
to a time series of CPR.
}

