% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TermStructure.R
\name{SpotRate<-}
\alias{SpotRate<-}
\title{A generic function to replace the SpotRate in class TermStructure}
\usage{
SpotRate(object) <- value
}
\arguments{
\item{object}{an S4 class object}

\item{value}{the replacement value}
}
\description{
A generic function to replace the SpotRate in class TermStructure
}

