% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpreadTypes.R
\name{SpreadBasisString<-}
\alias{SpreadBasisString<-}
\title{A standard generic functioin to replace the value of the slot SpreadBasisString}
\usage{
SpreadBasisString(object) <- value
}
\arguments{
\item{object}{an S4 object of the type SpreadTypes}

\item{value}{the replacement value of the slot}
}
\description{
A standard generic functioin to replace the value of the slot SpreadBasisString
}

