% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpreadTypes.R
\name{SpreadDecimal}
\alias{SpreadDecimal}
\title{A standard generic function to get the slot SpreadDecimal}
\usage{
SpreadDecimal(object)
}
\arguments{
\item{object}{an S4 object of the class SpreadTypes}
}
\description{
A standard generic function to get the slot SpreadDecimal
}

