% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurveSpreads.R
\docType{methods}
\name{SpreadToBenchmark,CurveSpreads-method}
\alias{SpreadToBenchmark,CurveSpreads-method}
\title{Spread To Benchmark}
\usage{
\S4method{SpreadToBenchmark}{CurveSpreads}(object)
}
\arguments{
\item{object}{An object of type CurveSpreads}
}
\description{
The bond yield (nominal) spread relative to that of the nearest
pricing benchmark along the curve.
}
\seealso{
Other Pricing: \code{\link{BenchMark,CurveSpreads-method}},
  \code{\link{CurveSpreads-class}},
  \code{\link{CurveSpreads}},
  \code{\link{DiscountToPrice}},
  \code{\link{SpreadToCurve,CurveSpreads-method}},
  \code{\link{SpreadToPriceBond}},
  \code{\link{SpreadToPriceMBS}},
  \code{\link{ZeroVolSpread,CurveSpreads-method}}
}
