% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurveSpreads.R
\docType{methods}
\name{SpreadToCurve,CurveSpreads-method}
\alias{SpreadToCurve,CurveSpreads-method}
\title{Spread To curve}
\usage{
\S4method{SpreadToCurve}{CurveSpreads}(object)
}
\arguments{
\item{object}{An object of type CurveSpreads}
}
\description{
The bond yield spread (nominal) relative to the interpolated
point along the couopn curve matching the weighted average life.
}
\seealso{
Other Pricing: \code{\link{BenchMark,CurveSpreads-method}},
  \code{\link{CurveSpreads-class}},
  \code{\link{CurveSpreads}},
  \code{\link{DiscountToPrice}},
  \code{\link{PriceToBondYield}},
  \code{\link{PriceToDiscountYield}},
  \code{\link{SpreadToBenchmark,CurveSpreads-method}},
  \code{\link{SpreadToPriceBond}},
  \code{\link{SpreadToPriceMBS}},
  \code{\link{ZeroVolSpread,CurveSpreads-method}}
}
