% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpreadToPriceFunctions.R
\name{SpreadToPrice}
\alias{SpreadToPrice}
\title{A function to solve for the price of a MBS pass-through given
both pricing speed and spread to the interpolated curve.  The function
returns the clean price of the bond}
\usage{
SpreadToPrice(bond.id = "character", trade.date = "character",
  settlement.date = "character", PrepaymentAssumption = "character",
  spread = numeric(), CPR = numeric(), ..., benchmark = NULL)
}
\arguments{
\item{bond.id}{A character string the cusip number or bond.id}

\item{trade.date}{A character string the trade date mm-dd-YYYY}

\item{settlement.date}{A character string the settlement date mm-dd-YYYY}

\item{PrepaymentAssumption}{the assumption must be: "CPR", "PPC", "MODEL"}

\item{spread}{a numeric value the spread to the interpolated curve
entered in percent}

\item{CPR}{a numeric value the CPR assumption used to price the MBS
e.g. 100 basis points = 1}

\item{...}{optional values follow}

\item{benchmark}{optionally the user can pass pricing benchmark}
}
\description{
A function to solve for the price of a MBS pass-through given
both pricing speed and spread to the interpolated curve.  The function
returns the clean price of the bond
}

