% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondTermStructure.R
\name{Term,BondTermStructure-method}
\alias{Term,BondTermStructure-method}
\title{Maturity}
\usage{
\S4method{Term}{BondTermStructure}(object)
}
\arguments{
\item{object}{object of type BondTermStructure}
}
\description{
Get Term from object BondTermStructure
}
