% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TermStructure.R
\docType{class}
\name{TermStructure-class}
\alias{TermStructure-class}
\title{An S4 class the term structure data needed to price bonds}
\description{
An S4 class the term structure data needed to price bonds
}
\section{Slots}{

\describe{
\item{\code{TradeDate}}{A character string the trade date in the format
mm/dd/YYYY}

\item{\code{Period}}{A numeric value the period index of the next cash flow}

\item{\code{Date}}{A numeric value the date of the next cash flow}

\item{\code{TimePeriod}}{a numeric value the time period between payments made
to the investor}

\item{\code{SpotRate}}{A numeric value the one-month spot rate}

\item{\code{ForwardRate}}{A numeric value the one-month forward rate}

\item{\code{DiscRate}}{A numeric value the discount rate curve}

\item{\code{TwoYearFwd}}{A numeric vlaue the two-year forward rate}

\item{\code{TenYearFwd}}{A numeric value the ten-year forward rate}
}}

