% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TermStructure.R
\name{TermStructure}
\alias{TermStructure}
\title{The TermStructure constructor function it is a wrapper function around the package termstrc}
\usage{
TermStructure(rates.data = "character", method = "character")
}
\arguments{
\item{rates.data}{A character string representing the data for which the user
would like to call the swap curve}

\item{method}{A character string indicating the fitting method ns = Nelson Siegel, dl = Diebold Lee,
sv = Severson, asv = adjusted Severson, cs = cubic spline (not yet implemented in Bond Lab).
For addiition details see the termstrc documentation.}
}
\description{
This is a wrapper function around the R package termstrc.  The function passes swap rate data
cash flows the to termstrc and creates the TermStructure object used by Bondlab.
The function call rates data processes the yield curve and derives cashflow
for the daily close swap curve. A Rates object must be called in the local
environment for this function to work.
}
\examples{
\dontrun{
TermStructure(rates.data = "01-10-2013", method = "ns")}
}

