% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMICConstructor.R
\docType{methods}
\name{TrancheProceedsWithInterest,TrancheDetails-method}
\alias{TrancheProceedsWithInterest,TrancheDetails-method}
\title{A method to access TrancheProceedsWithInterest from class TranchePrice}
\usage{
\S4method{TrancheProceedsWithInterest}{TrancheDetails}(object)
}
\arguments{
\item{object}{an S4 class of type TrancheDetails}
}
\description{
A method to access TrancheProceedsWithInterest from class TranchePrice
}

