% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YieldTypes.R
\docType{methods}
\name{YieldBasis<-,YieldTypes-method}
\alias{YieldBasis<-,YieldTypes-method}
\title{A method to set the value of the slot YieldBasis}
\usage{
\S4method{YieldBasis}{YieldTypes}(object) <- value
}
\arguments{
\item{object}{an S4 object class of the type YieldTypes}

\item{value}{the replacement value of the slot}
}
\description{
A method to set the value of the slot YieldBasis
}

