% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YieldTypes.R
\name{YieldBasis,YieldTypes-method}
\alias{YieldBasis,YieldTypes-method}
\title{A method to extract the slot YieldBasis from the class YieldTypes}
\usage{
\S4method{YieldBasis}{YieldTypes}(object)
}
\arguments{
\item{object}{an S4 object class of the type YieldTypes}
}
\description{
A method to extract the slot YieldBasis from the class YieldTypes
}
