% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YieldTypes.R
\docType{methods}
\name{YieldDecimal,YieldTypes-method}
\alias{YieldDecimal,YieldTypes-method}
\title{A method to extract the slot YieldDecimal from the class YieldTypes}
\usage{
\S4method{YieldDecimal}{YieldTypes}(object)
}
\arguments{
\item{object}{an S4 class object of the type YieldTypes}
}
\description{
A method to extract the slot YieldDecimal from the class YieldTypes
}
