% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondCashFlow.R
\docType{methods}
\name{YieldToMaturity,BondCashFlows-method}
\alias{YieldToMaturity,BondCashFlows-method}
\title{Method to extract YieldToMaturity from S4 class BondCashFlows}
\usage{
\S4method{YieldToMaturity}{BondCashFlows}(object)
}
\arguments{
\item{object}{the name of the S4 object}
}
\description{
Method to extract YieldToMaturity from S4 class BondCashFlows
}

