% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YieldTypes.R
\name{YieldTypes}
\alias{YieldTypes}
\title{YieldTypes is a constructor function for the class YieldTypes}
\usage{
YieldTypes(yield = numeric())
}
\arguments{
\item{yield}{a numeric value the bond yield to maturity in expressed in
basis terms expample(5.0 is 0.050).}
}
\description{
YieldTypes is a constructor function for the class YieldTypes
}

