% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortgageOAS.R
\docType{methods}
\name{ZeroVolSpread,MortgageOAS-method}
\alias{ZeroVolSpread,MortgageOAS-method}
\title{A method to extract ZeroVolSpread from S4 object MortgageOAS}
\usage{
\S4method{ZeroVolSpread}{MortgageOAS}(object)
}
\arguments{
\item{object}{an S4 object of the type MortgageOAS}
}
\description{
A method to extract ZeroVolSpread from S4 object MortgageOAS
}

